% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{add_header}
\alias{add_header}
\title{Add a Header to a \code{teal} App}
\usage{
add_header(messages, header = "")
}
\arguments{
\item{header}{(\code{shiny.tag} or \code{character(1)}) The header content to set. Defaults to an empty paragraph tag.}

\item{app}{(\code{environment}) The \code{teal} app environment.}
}
\description{
Adds a header to the \code{teal} app.
}
\examples{
app <- init(
  data = teal_data(IRIS = iris),
  modules = modules(example_module())
) |>
  add_header(
    tags$div(
      h3("Custom header")
    )
  )

shinyApp(app$ui, app$server)
}
