% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teal_modifiers.R
\name{add_landing_popup}
\alias{add_landing_popup}
\title{Add a Landing Popup to \code{teal} Application}
\usage{
add_landing_popup(
  x,
  title = NULL,
  content = NULL,
  buttons = modalButton("Accept")
)
}
\arguments{
\item{x}{(\code{list}) The \code{teal} ui and server object created using \code{init}.}

\item{title}{(\code{character(1)}) Text to be displayed as popup title.}

\item{content}{(\code{character(1)}, \code{shiny.tag} or \code{shiny.tag.list}) with the content of the popup.
Passed to \code{...} of \code{shiny::modalDialog}.}

\item{buttons}{(\code{shiny.tag} or \code{shiny.tag.list}) Typically a \code{modalButton} or \code{actionButton}.}
}
\description{
Adds a landing popup to the \code{teal} app. This popup will be shown when the app starts.
The dialog blocks access to the application and must be closed with a button before the application can be viewed.
}
\examples{
app <- init(
  data = teal_data(IRIS = iris, MTCARS = mtcars),
  modules = modules(example_module())
) |>
  add_landing_popup(
    title = "Welcome",
    content = "This is a landing popup.",
    buttons = modalButton("Accept")
  )

if (interactive()) {
  shinyApp(app$ui, app$server)
}
}
