% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choices_selected.R
\name{add_no_selected_choices}
\alias{add_no_selected_choices}
\title{Add empty choice to choices selected}
\usage{
add_no_selected_choices(x, multiple = FALSE)
}
\arguments{
\item{x}{(\code{choices_selected}) output}

\item{multiple}{(\code{logical}) whether multiple selections are allowed or not}
}
\description{
\ifelse{html}{\out{<a href='https://www.tidyverse.org/lifecycle/#maturing'><img src='figures/lifecycle-maturing.svg' alt='Maturing lifecycle'></a>}}{\strong{Maturing}}
}
