% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choices_icons.R
\name{add_subtext}
\alias{add_subtext}
\title{Add a subtext to a given list of choices}
\usage{
add_subtext(content, subtext)
}
\arguments{
\item{content}{(atomic vector of \code{character}) choices as displayed in UI, not their internal values}

\item{subtext}{(atomic vector of \code{character}) subtexts to display for each item in content}
}
\value{
(atomic vector of \code{character}) HTML contents with content and subtext combined
}
\description{
content and data-subtext don't work in conjunction, subtext creates
\code{<small class="text-muted">Study Identifier</small>}
The output of this function can be added in choiceOpts as \code{list(content = ...)}
}
\examples{
add_subtext(c("SEX", "BRMRKR1"), c("Sex", "Biomarker 1"))

add_subtext(c("SEX", "BRMRKR1"), c("", "Biomarker 1"))
}
