% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{add_title}
\alias{add_title}
\title{Add a Title to a \code{teal} App}
\usage{
add_title(app, title = build_app_title())
}
\arguments{
\item{app}{(\code{environment}) The \code{teal} app environment.}

\item{title}{(\code{shiny.tag} or \code{character(1)}) The title to set. Defaults to the result of \code{\link[=build_app_title]{build_app_title()}}.}
}
\value{
The modified \code{teal} app environment, invisibly.
}
\description{
Adds a browser window title or app title to the \code{teal} app.
}
\examples{
app <- list2env(list(title = NULL))
app <- add_title(app, "My App Title")
}
