% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choices_icons.R
\name{add_variable_type_icons}
\alias{add_variable_type_icons}
\title{Add variable type icons next to variable labels}
\usage{
add_variable_type_icons(columns, data, column_labels = columns)
}
\arguments{
\item{columns}{(atomic vector of \code{character}) columns to take from data}

\item{data}{(\code{data.frame}) data to determine variable types from}

\item{column_labels}{(atomic vector of \code{character}) if not specified, icons are prefixed with column name}
}
\value{
(atomic vector of \code{character}) vector of strings that can be rendered as HTML
}
\description{
If icons are not displayed, this is a CSS problem. Make sure to include the CSS somewhere,
e.g. by calling icon("cog") somewhere and include it in the html
}
\examples{
add_variable_type_icons(
  c("x", "z"),
  data.frame(x = 1:3, y = factor(c("a", "b", "a")), z = c("h1", "h2", "h3"),
    stringsAsFactors = FALSE
  )
)
add_variable_type_icons(
  c("x", "z"),
  data.frame(x = 1:3, y = factor(c("a", "b", "a")), z = c("h1", "h2", "h3"),
    stringsAsFactors = FALSE
  ),
  column_labels = c("x var", "z var")
)

add_variable_type_icons(character(0), data.frame(x = 1:3))
}
