% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/after.R
\name{after}
\alias{after}
\title{Modify \code{teal_module}}
\usage{
after(
  x,
  ui = function(id, elem) elem,
  server = function(input, output, session, data) data,
  ...
)
}
\arguments{
\item{x}{(\code{teal_data})}

\item{ui}{(\verb{function(id, elem, ...)}) function to receive output (\code{shiny.tag}) from \code{x$ui}}

\item{server}{(\verb{function(input, output, session, data, ...)}) function to receive output data from \code{x$server}.}

\item{...}{additional argument passed to \code{ui} and \code{server} by matching their formals names.}
}
\description{
Modify \code{ui} or \code{server} of \code{teal_module} to create a new UI elements or modify \code{teal_data}.
}
\examples{
new_mod <- example_module() |>
  after(
    ui = function(id, elem) {
      tags$div(
        tags$header("This is a module's header"),
        elem
      )
    },
    server = function(data) {
      teal.reporter::report(data) <- c(
        teal.reporter::report(data),
        "# added content",
        "This content has been added through `after`"
      )
      data
    }
  )

app <- init(
  data = teal_data(iris = iris),
  modules = modules(new_mod)
)
if (interactive()) {
  shinyApp(app$ui, app$server)
}
}
