% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_choices.R
\name{all_choices}
\alias{all_choices}
\title{An S3 structure representing the selection of all
possible choices in a \code{filter_spec} or \code{choices_selected} object.}
\usage{
all_choices()
}
\value{
\code{all_choices} object
}
\description{
An S3 structure representing the selection of all
possible choices in a \code{filter_spec} or \code{choices_selected} object.
}
\examples{
# Both structures are semantically identical
filter_spec(
  vars = c("selected_variable"),
  choices = c("value1", "value2"),
  selected = c("value1", "value2")
)

filter_spec(
  vars = c("selected_variable"),
  choices = c("value1", "value2"),
  selected = all_choices()
)

}
