% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{all_true}
\alias{all_true}
\title{Split by sepearator}
\usage{
all_true(lst, fcn)
}
\arguments{
\item{lst}{(\code{list}) List of values}

\item{fcn}{(\code{function}) function to apply
that returns TRUE or FALSE}
}
\description{
Split by sepearator
}
