% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_state_manager.R
\name{app_state_grab}
\alias{app_state_grab}
\title{Grab selection state (value) of all input items in the app.}
\usage{
app_state_grab()
}
\value{
Object of class \code{teal_grab}, describing the state of all inputs in the app (except buttons).
}
\description{
Find all inputs in the app ans store their state.
Buttons are omitted as their values do not (usually) matter
and they are virtually impossible to restore.
}
\seealso{
\code{\link{app_state_store}}, \code{\link{app_state_restore}}, \code{\link{state_manager_module}}
}
\keyword{internal}
