% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_state_manager.R
\name{app_state_restore}
\alias{app_state_restore}
\title{Restore state (value) of all input items in the app according to a grab or file.}
\usage{
app_state_restore(grab, file)
}
\arguments{
\item{grab}{optional \code{teal_grab}}

\item{file}{optional \code{path} to a .json file}
}
\value{
Returns \code{NULL} invisibly.
}
\description{
Restore state (value) of all input items in the app according to a grab or file.
}
\seealso{
\code{\link{app_state_grab}}, \code{\link{app_state_store}}, \code{\link{state_manager_module}}
}
\keyword{internal}
