% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_state_manager.R
\name{app_state_store}
\alias{app_state_store}
\title{Save input grab to JSON file.}
\usage{
app_state_store(grab, file)
}
\arguments{
\item{grab}{\code{teal_grab}}

\item{file}{\code{path} to save the input states to; must be a \code{.json} file; will be overwritten}
}
\value{
Returns \code{NULL} invisibly.
}
\description{
Save input grab to JSON file.
}
\seealso{
\code{\link{app_state_grab}}, \code{\link{app_state_restore}}, \code{\link{state_manager_module}}
}
\keyword{internal}
