% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{as.global}
\alias{as.global}
\title{Copy a variable to the global environment}
\usage{
as.global(x)
}
\arguments{
\item{x}{variable to write to the GlobalEnv. Not non-standard evaluation is
applied for getting the variable name and variable content}
}
\description{
This function is uselful for debugging shiny apps
}
