% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_state_manager.R
\name{as.teal_grab}
\alias{as.teal_grab}
\title{Convert named list to \code{teal_grab}.}
\usage{
as.teal_grab(x)
}
\arguments{
\item{x}{\verb{named list}}
}
\value{
Object of class \code{teal_grab}, which is a list of lists,
each of which has two elements, one named "id" and the other "value".
}
\description{
Convert named list to \code{teal_grab}.
}
\keyword{internal}
