% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_cdisc.R
\name{as_cdisc}
\alias{as_cdisc}
\alias{as_cdisc.Dataset}
\alias{as_cdisc.DatasetConnector}
\title{Convert a \code{Dataset(Connector)} object to a \code{CDISCDataset(Connector)} object}
\usage{
as_cdisc(
  x,
  parent = if (identical(get_dataname(x), "ADSL")) character(0) else "ADSL"
)

\method{as_cdisc}{Dataset}(
  x,
  parent = if (identical(get_dataname(x), "ADSL")) character(0) else "ADSL"
)

\method{as_cdisc}{DatasetConnector}(
  x,
  parent = if (identical(get_dataname(x), "ADSL")) character(0) else "ADSL"
)
}
\arguments{
\item{x}{an object of \code{Dataset} or \code{DatasetConnector} class}

\item{parent}{optional, (\code{character}) \cr
parent dataset name}
}
\value{
(\code{CDISCDataset} or \code{CDISCDatasetConnector}) object
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\details{
Convert a \code{Dataset(Connector)} object to a \code{CDISCDataset(Connector)} object
}
\note{
If passed a \code{cdisc}-flavored object it returns the unmodified object.
}
\examples{
# Dataset --------

library(scda)
as_cdisc(
  dataset(
    "ADSL",
    synthetic_cdisc_data("latest")$adsl,
    keys = get_cdisc_keys("ADSL"),
    code = "ADSL <- synthetic_cdisc_data(\"latest\")$adsl"
  )
)
as_cdisc(
  dataset(
    "ADAE",
    synthetic_cdisc_data("latest")$adae,
    keys = get_cdisc_keys("ADAE"),
    code = "ADAE <- synthetic_cdisc_data(\"latest\")$adae"
  ),
  parent = "ADSL"
)
# DatasetConnector --------

library(scda)
pull_fun_adsl <- callable_function(
  function() {synthetic_cdisc_data("latest")$adsl}
)
as_cdisc(
  dataset_connector(
    "ADSL",
    pull_fun_adsl,
    keys = get_cdisc_keys("ADSL")
  )
)

pull_fun_adae <- callable_function(
  function() {synthetic_cdisc_data("latest")$adae}
)
as_cdisc(
  dataset_connector(
    "ADAE",
    pull_fun_adae,
    keys = get_cdisc_keys("ADAE")
  ),
  parent = "ADSL"
)

}
