% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_relational.R
\name{as_cdisc_relational}
\alias{as_cdisc_relational}
\title{Convert a \verb{Named<...>} object to a \verb{Relational<...>} object}
\usage{
as_cdisc_relational(x, dataname, label = character(0))
}
\arguments{
\item{x}{an object of (\code{NamedDataset} or \code{NamedDatasetConnector}) class}

\item{dataname}{(\code{character})\cr
A given name for the dataset it may not contain spaces}

\item{label}{(\code{character})\cr
Label to describe the dataset}
}
\value{
(\code{RelationalDataset} or \code{RelationalDatasetConnector}) object
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#defunct}{\figure{lifecycle-defunct.svg}{options: alt='[Defunct]'}}}{\strong{[Defunct]}}
}
\details{
Convert a \verb{Named<...>} to a \verb{Relational<...>}
}
