% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_relational.R
\name{as_relational}
\alias{as_relational}
\alias{as_cdisc_relational}
\title{Convert a \verb{Named<...>} object to a \verb{Relational<...>} object}
\usage{
as_relational(x, keys)

as_cdisc_relational(x, dataname, label = character(0))
}
\arguments{
\item{x}{an object of (\code{NamedDataset} or \code{NamedDatasetConnector}) class}

\item{keys}{optional, (\code{character})\cr
vector with primary keys}

\item{dataname}{(\code{character})\cr
A given name for the dataset it may not contain spaces}

\item{label}{(\code{character})\cr
Label to describe the dataset}
}
\value{
(\code{RelationalDataset} or \code{RelationalDatasetConnector}) object
}
\description{
\ifelse{html}{\figure{lifecycle-soft-deprecated.svg}{options: alt='Soft-deprecated lifecycle'}}{\strong{Soft-deprecated}}

\ifelse{html}{\figure{lifecycle-defunct.svg}{options: alt='Defunct lifecycle'}}{\strong{Defunct}}
}
\details{
Convert a \verb{Named<...>} to a \verb{Relational<...>}

\code{as_cdisc_relational} will derive the keys by the \code{dataname} and therefore
does not need the \code{keys} argument to be specified
}
