% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_layout.R
\name{badge_dropdown}
\alias{badge_dropdown}
\title{Dropdown badge}
\usage{
badge_dropdown(id, label, content)
}
\arguments{
\item{id}{(\code{character(1)}) shiny module's id}

\item{label}{(\code{shiny.tag}) Label displayed on a badge.}

\item{...}{(\code{shiny.tag}) Content of a dropdown.}
}
\description{
Dropdown button in a form of a badge with \code{bg-primary} as default style
Clicking badge shows a dropdown containing any \code{HTML} element. Folded dropdown
doesn't trigger display output which means that items rendered using \verb{render*}
will be recomputed only when dropdown is show.
}
