% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blueprint.R
\name{blueprint-actors}
\alias{blueprint-actors}
\title{Actors}
\description{
\if{html}{\out{<div class="sourceCode {r, child = "man/blueprint/blueprint-actors.Rmd"}">}}\preformatted{}\if{html}{\out{</div>}}
}
\keyword{internal}
