% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blueprint.R
\name{blueprint-filter_panel}
\alias{blueprint-filter_panel}
\title{Filter Panel}
\description{
\if{html}{\out{<div class="sourceCode {r, child = "man/blueprint/blueprint-filter_panel.Rmd"}">}}\preformatted{}\if{html}{\out{</div>}}
}
\keyword{internal}
