% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blueprint.R
\name{blueprint-intro}
\alias{blueprint-intro}
\title{Introduction}
\description{
\code{teal} is an interactive data exploration framework based on Shiny, designed with emphasis on analyzing CDISC clinical trial data.

A Shiny Application created with \code{teal} offers users the ability to:
\itemize{
\item Import data from external sources;
\item Dynamically filter data for analysis;
\item Generate reproducible code for future analysis
\item Create downloadable reports of analysis results \emph{(for analysis modules that support reporting)}.
}

Moreover, the \code{teal} framework provides application developers with a wide range of customizable standard analysis modules to integrate into their applications, along with a logging framework that helps facilitate debugging.
Additionally, advanced users of the framework can develop new analysis modules and easily integrate them into any \code{teal} application.

The \code{teal} framework's functionality draws heavily from the following packages:\tabular{rl}{
   R package \tab Description \cr
   \href{https://github.com/insightsengineering/teal}{\code{teal}} \tab \code{shiny}-based interactive exploration framework for analyzing data \cr
   \href{https://github.com/insightsengineering/teal.data}{\code{teal.data}} \tab creating and loading the data needed for \code{teal} applications \cr
   \href{https://github.com/insightsengineering/teal.widgets}{\code{teal.widgets}} \tab \code{shiny} UI components used within \code{teal} \cr
   \href{https://github.com/insightsengineering/teal.slice}{\code{teal.slice}} \tab provides a filtering panel to allow filtering of data \cr
   \href{https://github.com/insightsengineering/teal.code}{\code{teal.code}} \tab handles reproducibility of outputs \cr
   \href{https://github.com/insightsengineering/teal.transform}{\code{teal.transform}} \tab standardizes extracting and merging data \cr
   \href{https://github.com/insightsengineering/teal.logger}{\code{teal.logger}} \tab standardizes logging within \code{teal} framework \cr
   \href{https://github.com/insightsengineering/teal.reporter}{\code{teal.reporter}} \tab allows \code{teal} applications to generate reports \cr
}


Although these packages are mentioned in the material, we strongly recommend visiting their vignettes to learn more about them.

Learn on how to make your first \code{teal} application \href{https://insightsengineering.github.io/teal/main/articles/teal.html}{here}!
}
\keyword{internal}
