% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blueprint.R
\name{blueprint-module_encapsulation}
\alias{blueprint-module_encapsulation}
\title{Module Encapsulation}
\description{
\subsection{Introduction}{

\code{teal} leverages the \href{https://rstudio.github.io/shiny/reference/moduleServer.html}{Shiny module} concept to enable seamless communication between the framework and \code{teal} modules, while also promoting encapsulation.

\code{teal} modules heavily depends on this core Shiny feature and is a powerful extension that allows for a seamless
}

\subsection{Benefits}{

The shiny module technique provides an efficient and structured approach when developing large shiny apps.
By incorporating the shiny module concept into \code{teal}, developers can break down complex functionalities into smaller, more manageable modules, encapsulating specific functionality within each module.

This encapsulation ensures that the variables, functions, and logic within a module are self-contained and independent from other parts of the application.
As a result, developers can focus on building self-contained components within each module, making it easier to understand, test, and enhance different aspects of the application independently.

By embracing encapsulation through shiny modules, \code{teal} \strong{empowers developers to} create modular and maintainable shiny apps, enhancing code organization and facilitating collaboration.
}
}
\keyword{internal}
