% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blueprint.R
\name{blueprint-products_map}
\alias{blueprint-products_map}
\title{Products Map}
\description{
\if{html}{\out{<div class="sourceCode {r, child = "man/blueprint/blueprint-products_map.Rmd"}">}}\preformatted{}\if{html}{\out{</div>}}
}
\keyword{internal}
