% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules_debugging.R
\name{bookmark_module}
\alias{bookmark_module}
\title{Bookmarking module}
\usage{
bookmark_module(label = "Bookmark Module")
}
\arguments{
\item{label}{\code{character} label of module}
}
\description{
Presents a button that bookmarks the current state.
Note that Shiny does not bookmark uploaded files except when
\code{enableBookmarking} is set to \code{server}.
The module also prints a nice error message if bookmarking is not enabled.
}
\keyword{internal}
