% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{build_app_title}
\alias{build_app_title}
\title{Build app title with favicon}
\usage{
build_app_title(
  title = "teal app",
  favicon =
    "https://raw.githubusercontent.com/insightsengineering/hex-stickers/main/PNG/nest.png"
)
}
\arguments{
\item{title}{(\code{character}) The browser title for the \code{teal} app.}

\item{favicon}{(\code{character}) The path for the icon for the title.
The image/icon path can be remote or the static path accessible by \code{shiny}, like the \verb{www/}}
}
\value{
A \code{shiny.tag} containing the element that adds the title and logo to the \code{shiny} app.
}
\description{
A helper function to create the browser title along with a logo.
}
