% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_nested_tabs.R
\name{calculate_hashes}
\alias{calculate_hashes}
\title{Get the hash of a dataset}
\usage{
calculate_hashes(datanames, datasets)
}
\arguments{
\item{datanames}{(\code{character}) names of datasets}

\item{datasets}{(\code{FilteredData}) object holding the data}
}
\value{
A list of hashes per dataset.
}
\description{
Get the hash of a dataset
}
\keyword{internal}
