% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_utils.R
\name{call_check_parse_varname}
\alias{call_check_parse_varname}
\title{Checks \code{varname} argument and convert to call}
\usage{
call_check_parse_varname(varname)
}
\arguments{
\item{varname}{(\code{name}, \code{call} or \code{character(1)})\cr
name of the variable}
}
\description{
Checks \code{varname} type and parse if it's a \code{character}
}
\keyword{internal}
