% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_utils.R
\name{call_condition_choice}
\alias{call_condition_choice}
\title{Choices condition call}
\usage{
call_condition_choice(varname, choices)
}
\arguments{
\item{varname}{(\code{name} or \code{character(1)})\cr
name of the variable}

\item{choices}{(\code{vector})\cr
\code{varname} values to match the \code{==} (single value) or
\code{\%in\%} (vector) condition.}
}
\value{
\code{call}
}
\description{
Compose choices condition call from inputs.
}
\examples{
teal:::call_condition_choice(as.name("SEX"), choices = "F")
teal:::call_condition_choice("SEX", choices = c("F", "M"))
teal:::call_condition_choice("SEX", choices = c(1, 2))
}
