% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_utils.R
\name{call_condition_range_date}
\alias{call_condition_range_date}
\title{\code{Date} range condition call}
\usage{
call_condition_range_date(varname, range)
}
\arguments{
\item{varname}{(\code{name} or \code{character(1)})\cr
name of the variable}

\item{range}{(\code{Date})\cr
range of the variable}
}
\value{
a \code{call}
}
\description{
Compose \code{Date} range condition call from inputs
}
\examples{
call_condition_range_date(
  as.name("date"),
  range = c(Sys.Date(), Sys.Date() + 1)
)
}
\keyword{internal}
