% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_utils.R
\name{call_extract_array}
\alias{call_extract_array}
\title{Get call to subset and select array}
\usage{
call_extract_array(dataname = ".", row = NULL, column = NULL, aisle = NULL)
}
\arguments{
\item{dataname}{(\code{character(1)} or \code{name})\cr}

\item{row}{(\code{character(1)} or \code{call})\cr
optional, name of the \code{row} or condition}

\item{column}{(\code{character(1)} or \code{call})\cr
optional, name of the \code{column} or condition}

\item{aisle}{(\code{character(1)} or \code{call})\cr
optional, name of the \code{row} or condition}
}
\value{
\code{[} call with all conditions included

specific \code{\link[base]{Extract}} \code{call}
}
\description{
Get call to subset and select array
}
\examples{
teal:::call_extract_array(
  dataname = "my_array",
  row = teal:::call_condition_choice("my_array$SEX", "M"),
  column = call("c", "SEX", "AGE"),
  aisle = "RNAseq_rnaaccess"
)
teal:::call_extract_array(
  "mae_object",
  column = teal:::call_condition_choice("SEX", "M")
)
}
