% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CallableCode.R
\name{callable_code}
\alias{callable_code}
\title{Create \code{CallableCode} object}
\usage{
callable_code(code)
}
\arguments{
\item{code}{(\code{character})\cr
a string containing R code to reproduce the desired object.}
}
\value{
\code{CallableCode} object
}
\description{
Create \link{CallableCode} object to execute specific code and get reproducible call.
}
\examples{
cf <- callable_code(code = "mtcars")
cf$run()
cf$get_call()
}
