% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CallableCode.R
\name{CallableCode}
\alias{CallableCode}
\title{A \code{CallableCode} class of objects

Object that stores code to reproduce an object. It includes methods to
get or run the code and return the object.}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
}
\section{Super class}{
\code{\link[teal:Callable]{teal::Callable}} -> \code{CallableCode}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{CallableCode$new()}}
\item \href{#method-get_call}{\code{CallableCode$get_call()}}
\item \href{#method-clone}{\code{CallableCode$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="teal" data-topic="Callable" data-id="assign_to_env">}\href{../../teal/html/Callable.html#method-assign_to_env}{\code{teal::Callable$assign_to_env()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="teal" data-topic="Callable" data-id="get_error_message">}\href{../../teal/html/Callable.html#method-get_error_message}{\code{teal::Callable$get_error_message()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="teal" data-topic="Callable" data-id="is_failed">}\href{../../teal/html/Callable.html#method-is_failed}{\code{teal::Callable$is_failed()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="teal" data-topic="Callable" data-id="run">}\href{../../teal/html/Callable.html#method-run}{\code{teal::Callable$run()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new \code{CallableCode} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CallableCode$new(code, env = new.env(parent = parent.env(globalenv())))}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{code}}{(\code{character})\cr
a string containing R code to reproduce the desired object.}

\item{\code{env}}{(\code{environment})\cr
environment where function will be evaluated}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
new \code{CallableCode} object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_call"></a>}}
\if{latex}{\out{\hypertarget{method-get_call}{}}}
\subsection{Method \code{get_call()}}{
Get sequence of calls from the code supplied to produce the object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CallableCode$get_call(deparse = TRUE, args = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deparse}}{(\code{logical} value)\cr
whether to return a deparsed version of call}

\item{\code{args}}{(\code{NULL})\cr
available to be consistent with \code{CallableFunction} but are not used to
retrieve the call.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{list} of \code{calls} or \code{character} depending on \code{deparse} argument
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CallableCode$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
