% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{card_template}
\alias{card_template}
\title{Template function to generate reporter card.}
\usage{
card_template(title, label, description = NULL, with_filter, filter_panel_api)
}
\arguments{
\item{title}{(\code{character(1)}) title of the card (unless overwritten by label)}

\item{label}{(\code{character(1)}) label provided by the user when adding the card}

\item{description}{(\code{character(1)}) optional additional description}

\item{with_filter}{(\code{logical(1)}) flag indicating to add filter state}

\item{filter_panel_api}{(\code{FilterPanelAPI}) object with API that allows the generation
of the filter state in the report}
}
\value{
(\code{TealReportCard}) populated with a title, description and filter state
}
\description{
This function generates a report card with a title,
an optional description, and the option to append the filter state list.
}
