% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cast_to_list}
\alias{cast_to_list}
\title{Cast data in input to list.}
\usage{
cast_to_list(data, names_data)
}
\arguments{
\item{data}{\code{RelationalData} or \code{list} or \code{data.frame}.}

\item{names_data}{\code{character} of given names of data.}
}
\value{
list of \code{RelationalData} or \code{data.frame}.
}
\description{
Cast data in input to list.
}
\details{
wraps the data input as a list when necessary.
}
