% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CDISCTealDataConnector.R
\name{cdisc_data_connector}
\alias{cdisc_data_connector}
\title{The constructor of \code{CDISCTealDataConnector} objects.}
\usage{
cdisc_data_connector(connection, connectors)
}
\arguments{
\item{connection}{(\code{TealDataConnection})\cr
connection to data source}

\item{connectors}{(\code{list} of \code{TealDatasetConnector} elements)\cr
list with dataset connectors}
}
\value{
\code{CDISCTealDataConnector} object
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
}
\examples{
\dontrun{
adsl_cf <- CallableFunction$new(
  function() as.data.frame(as.list(setNames(nm = teal.data::get_cdisc_keys("ADSL"))))
)
adae_cf <- CallableFunction$new(
  function() as.data.frame(as.list(setNames(nm = teal.data::get_cdisc_keys("ADAE"))))
)
adsl <- teal.data:::CDISCTealDatasetConnector$new(
  "ADSL", adsl_cf,
  keys = teal.data::get_cdisc_keys("ADSL"), parent = character(0)
)
adae <- teal.data:::CDISCTealDatasetConnector$new(
  "ADAE", adae_cf,
  keys = teal.data::get_cdisc_keys("ADAE"), parent = "ADSL"
)
data <- cdisc_data_connector(
  connection = data_connection(open_fun = CallableFunction$new(function() "open function")),
  connectors = list(adsl, adae)
)
}
}
