% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdisc_data.R
\name{cdisc_dataset_file}
\alias{cdisc_dataset_file}
\title{Load \code{RelationalDataset} object from a file}
\usage{
cdisc_dataset_file(x)
}
\arguments{
\item{x}{A (\code{connection}) or a (\code{character}) string giving the pathname
of the file or URL to read from. "" indicates the connection \code{stdin}.}
}
\value{
\code{RelationalDataset} object if file returns a \code{RelationalDataset}
  object.
}
\description{
Load \code{RelationalDataset} object from a file
}
\examples{
file_example <- tempfile(fileext = ".R")
writeLines(
  text = c(
    "library(random.cdisc.data)

     cdisc_dataset(dataname = \"ADSL\", # RelationalDataset
                   data = radsl(cached = TRUE),
                   code = \"library(random.cdisc.data)\nADSL <- radsl(cached = TRUE)\")"
  ),
  con = file_example
)

cdisc_dataset_file(file_example)

}
