% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CDISCData.R
\name{CDISCData}
\alias{CDISCData}
\title{Manage multiple \code{CDISCDataConnector}, \code{CDISCDatasetConnector} and \code{CDISCDataset} objects.}
\description{
Class manages \code{CDISCDataConnector}, \code{CDISCDatasetConnector} and
\code{CDISCDataset} objects and aggregate them in one collection.
Class also decides whether to launch app before initialize teal application.
}
\section{Super classes}{
\code{\link[teal:DataAbstract]{teal::DataAbstract}} -> \code{\link[teal:RelationalData]{teal::RelationalData}} -> \code{CDISCData}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{CDISCData$new()}}
\item \href{#method-get_parent}{\code{CDISCData$get_parent()}}
\item \href{#method-check_metadata}{\code{CDISCData$check_metadata()}}
\item \href{#method-clone}{\code{CDISCData$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="teal" data-topic="DataAbstract" data-id="check">}\href{../../teal/html/DataAbstract.html#method-check}{\code{teal::DataAbstract$check()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="teal" data-topic="DataAbstract" data-id="get_check_result">}\href{../../teal/html/DataAbstract.html#method-get_check_result}{\code{teal::DataAbstract$get_check_result()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="teal" data-topic="DataAbstract" data-id="get_code">}\href{../../teal/html/DataAbstract.html#method-get_code}{\code{teal::DataAbstract$get_code()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="teal" data-topic="DataAbstract" data-id="get_code_class">}\href{../../teal/html/DataAbstract.html#method-get_code_class}{\code{teal::DataAbstract$get_code_class()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="teal" data-topic="DataAbstract" data-id="get_dataset">}\href{../../teal/html/DataAbstract.html#method-get_dataset}{\code{teal::DataAbstract$get_dataset()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="teal" data-topic="DataAbstract" data-id="get_datasets">}\href{../../teal/html/DataAbstract.html#method-get_datasets}{\code{teal::DataAbstract$get_datasets()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="teal" data-topic="DataAbstract" data-id="is_pulled">}\href{../../teal/html/DataAbstract.html#method-is_pulled}{\code{teal::DataAbstract$is_pulled()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="teal" data-topic="DataAbstract" data-id="mutate">}\href{../../teal/html/DataAbstract.html#method-mutate}{\code{teal::DataAbstract$mutate()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="teal" data-topic="DataAbstract" data-id="mutate_dataset">}\href{../../teal/html/DataAbstract.html#method-mutate_dataset}{\code{teal::DataAbstract$mutate_dataset()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="teal" data-topic="DataAbstract" data-id="set_check">}\href{../../teal/html/DataAbstract.html#method-set_check}{\code{teal::DataAbstract$set_check()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="teal" data-topic="DataAbstract" data-id="set_pull_code">}\href{../../teal/html/DataAbstract.html#method-set_pull_code}{\code{teal::DataAbstract$set_pull_code()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="teal" data-topic="RelationalData" data-id="get_connectors">}\href{../../teal/html/RelationalData.html#method-get_connectors}{\code{teal::RelationalData$get_connectors()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="teal" data-topic="RelationalData" data-id="get_datanames">}\href{../../teal/html/RelationalData.html#method-get_datanames}{\code{teal::RelationalData$get_datanames()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="teal" data-topic="RelationalData" data-id="get_items">}\href{../../teal/html/RelationalData.html#method-get_items}{\code{teal::RelationalData$get_items()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="teal" data-topic="RelationalData" data-id="get_join_keys">}\href{../../teal/html/RelationalData.html#method-get_join_keys}{\code{teal::RelationalData$get_join_keys()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="teal" data-topic="RelationalData" data-id="get_server">}\href{../../teal/html/RelationalData.html#method-get_server}{\code{teal::RelationalData$get_server()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="teal" data-topic="RelationalData" data-id="get_ui">}\href{../../teal/html/RelationalData.html#method-get_ui}{\code{teal::RelationalData$get_ui()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="teal" data-topic="RelationalData" data-id="launch">}\href{../../teal/html/RelationalData.html#method-launch}{\code{teal::RelationalData$launch()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="teal" data-topic="RelationalData" data-id="mutate_join_keys">}\href{../../teal/html/RelationalData.html#method-mutate_join_keys}{\code{teal::RelationalData$mutate_join_keys()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new object of \code{CDISCData} class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CDISCData$new(..., join_keys)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{(\code{RelationalDataConnector}, \code{Dataset} or
\code{DatasetConnector}) elements to include where \code{ADSL} data is mandatory.}

\item{\code{join_keys}}{(\code{JoinKeys}) or a single (\code{JoinKeySet})\cr
(optional) object with datasets column names used for joining.
If empty then it would be automatically derived basing on intersection of datasets primary keys}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_parent"></a>}}
\if{latex}{\out{\hypertarget{method-get_parent}{}}}
\subsection{Method \code{get_parent()}}{
Get all datasets parent names
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CDISCData$get_parent()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
(named \code{list}) with dataset name and its corresponding parent dataset name
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-check_metadata"></a>}}
\if{latex}{\out{\hypertarget{method-check_metadata}{}}}
\subsection{Method \code{check_metadata()}}{
Check correctness of stored joining keys and presence of keys to parent
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CDISCData$check_metadata()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
raise and error or invisible \code{TRUE}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CDISCData$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
