% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CDISCDataConnector.R
\name{CDISCDataConnector}
\alias{CDISCDataConnector}
\title{Manage multiple and \code{DatasetConnector} of the same type.}
\description{
\ifelse{html}{\out{<a href='https://www.tidyverse.org/lifecycle/#experimental'><img src='figures/lifecycle-experimental.svg' alt='Experimental lifecycle'></a>}}{\strong{Experimental}}

Class manages \code{DatasetConnector} to specify additional dynamic arguments and to
open/close connection.
}
\section{Super classes}{
\code{\link[teal:DataAbstract]{teal::DataAbstract}} -> \code{\link[teal:RelationalDataConnector]{teal::RelationalDataConnector}} -> \code{CDISCDataConnector}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{CDISCDataConnector$new()}}
\item \href{#method-get_parent}{\code{CDISCDataConnector$get_parent()}}
\item \href{#method-clone}{\code{CDISCDataConnector$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="teal" data-topic="DataAbstract" data-id="check">}\href{../../teal/html/DataAbstract.html#method-check}{\code{teal::DataAbstract$check()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="teal" data-topic="DataAbstract" data-id="check_reproducibility">}\href{../../teal/html/DataAbstract.html#method-check_reproducibility}{\code{teal::DataAbstract$check_reproducibility()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="teal" data-topic="DataAbstract" data-id="execute_mutate">}\href{../../teal/html/DataAbstract.html#method-execute_mutate}{\code{teal::DataAbstract$execute_mutate()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="teal" data-topic="DataAbstract" data-id="get_check_result">}\href{../../teal/html/DataAbstract.html#method-get_check_result}{\code{teal::DataAbstract$get_check_result()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="teal" data-topic="DataAbstract" data-id="get_code">}\href{../../teal/html/DataAbstract.html#method-get_code}{\code{teal::DataAbstract$get_code()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="teal" data-topic="DataAbstract" data-id="get_datanames">}\href{../../teal/html/DataAbstract.html#method-get_datanames}{\code{teal::DataAbstract$get_datanames()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="teal" data-topic="DataAbstract" data-id="get_dataset">}\href{../../teal/html/DataAbstract.html#method-get_dataset}{\code{teal::DataAbstract$get_dataset()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="teal" data-topic="DataAbstract" data-id="get_datasets">}\href{../../teal/html/DataAbstract.html#method-get_datasets}{\code{teal::DataAbstract$get_datasets()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="teal" data-topic="DataAbstract" data-id="get_items">}\href{../../teal/html/DataAbstract.html#method-get_items}{\code{teal::DataAbstract$get_items()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="teal" data-topic="DataAbstract" data-id="is_pulled">}\href{../../teal/html/DataAbstract.html#method-is_pulled}{\code{teal::DataAbstract$is_pulled()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="teal" data-topic="DataAbstract" data-id="mutate_dataset">}\href{../../teal/html/DataAbstract.html#method-mutate_dataset}{\code{teal::DataAbstract$mutate_dataset()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="teal" data-topic="DataAbstract" data-id="set_check">}\href{../../teal/html/DataAbstract.html#method-set_check}{\code{teal::DataAbstract$set_check()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="teal" data-topic="DataAbstract" data-id="set_pull_code">}\href{../../teal/html/DataAbstract.html#method-set_pull_code}{\code{teal::DataAbstract$set_pull_code()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="teal" data-topic="RelationalDataConnector" data-id="get_code_class">}\href{../../teal/html/RelationalDataConnector.html#method-get_code_class}{\code{teal::RelationalDataConnector$get_code_class()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="teal" data-topic="RelationalDataConnector" data-id="get_connection">}\href{../../teal/html/RelationalDataConnector.html#method-get_connection}{\code{teal::RelationalDataConnector$get_connection()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="teal" data-topic="RelationalDataConnector" data-id="get_server">}\href{../../teal/html/RelationalDataConnector.html#method-get_server}{\code{teal::RelationalDataConnector$get_server()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="teal" data-topic="RelationalDataConnector" data-id="get_ui">}\href{../../teal/html/RelationalDataConnector.html#method-get_ui}{\code{teal::RelationalDataConnector$get_ui()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="teal" data-topic="RelationalDataConnector" data-id="is_failed">}\href{../../teal/html/RelationalDataConnector.html#method-is_failed}{\code{teal::RelationalDataConnector$is_failed()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="teal" data-topic="RelationalDataConnector" data-id="launch">}\href{../../teal/html/RelationalDataConnector.html#method-launch}{\code{teal::RelationalDataConnector$launch()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="teal" data-topic="RelationalDataConnector" data-id="mutate">}\href{../../teal/html/RelationalDataConnector.html#method-mutate}{\code{teal::RelationalDataConnector$mutate()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="teal" data-topic="RelationalDataConnector" data-id="pull">}\href{../../teal/html/RelationalDataConnector.html#method-pull}{\code{teal::RelationalDataConnector$pull()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="teal" data-topic="RelationalDataConnector" data-id="set_pull_args">}\href{../../teal/html/RelationalDataConnector.html#method-set_pull_args}{\code{teal::RelationalDataConnector$set_pull_args()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="teal" data-topic="RelationalDataConnector" data-id="set_server">}\href{../../teal/html/RelationalDataConnector.html#method-set_server}{\code{teal::RelationalDataConnector$set_server()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="teal" data-topic="RelationalDataConnector" data-id="set_ui">}\href{../../teal/html/RelationalDataConnector.html#method-set_ui}{\code{teal::RelationalDataConnector$set_ui()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new \code{CDISCDataConnector} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CDISCDataConnector$new(connection, connectors)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{connection}}{(\code{DataConnection})\cr
connection to data source}

\item{\code{connectors}}{(\code{list} of \code{DatasetConnector} elements)\cr
list with dataset connectors}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_parent"></a>}}
\if{latex}{\out{\hypertarget{method-get_parent}{}}}
\subsection{Method \code{get_parent()}}{
Get all datasets parent names
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CDISCDataConnector$get_parent()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
(named \code{list}) with dataset name and its corresponding parent dataset name
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CDISCDataConnector$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
