% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CDISCFilteredData.R
\docType{class}
\name{CDISCFilteredData}
\alias{CDISCFilteredData}
\title{Class to encapsulate relational filtered datasets with its parents.}
\description{
Class to encapsulate relational filtered datasets with its parents.

Class to encapsulate relational filtered datasets with its parents.
}
\details{
The \code{CDISCFilteredData} class implements logic to filter a relational
dataset by inheriting from \code{FilteredData}.
A dataset can have up to one parent dataset. Rows are identified by the foreign
key and only those rows that appear in the parent dataset are kept in the filtered
dataset.

The relational model is attached to the data. When \code{set_data} is called with the
data, its \code{keys} attribute has the following elements:
\itemize{
\item \code{primary}: columns based on which unique rows are counted
\item \code{parent}: parent dataset based on which the dataset is filtered (via \code{inner_join})
\item \code{foreign}: keys to join on when filtering with parent dataset
}

The teal UI works with objects of class \code{FilteredData} which may mix CDISC and other
data (e.g. \code{iris}).

It is not too difficult to make this class generalize to more than one parent dataset,
provided each dataset then contains information about its parents together with the
shared keys with each of them.
}
\examples{
library(random.cdisc.data)

ADSL <- radsl(cached = TRUE)
attr(ADSL, "keys") <- get_cdisc_keys("ADSL")
datasets <- teal:::CDISCFilteredData$new()

# to avoid using isolate(), you can provide a default isolate context by calling
# options(shiny.suppressMissingContextError = TRUE) #nolint
# don't forget to deactivate this option at the end
# options(shiny.suppressMissingContextError = FALSE) #nolint

isolate({
  datasets$datanames()
  datasets

  datasets$set_data("ADSL", ADSL)
  datasets$set_join_keys(join_keys())

  datasets$datanames()
  datasets

  datasets$datanames()
  datasets$get_data_info("ADSL", filtered = FALSE)
  # filters dataset to obtain information
  datasets$get_data_info("ADSL", filtered = TRUE)
  datasets$get_filter_info("ADSL")
  df <- datasets$get_data("ADSL", filtered = FALSE)
  # df

  datasets$get_filter_type("ADSL", "SEX")
  datasets$set_filter_state("ADSL", state = list(
    AGE = list(range = c(33, 44), keep_na = FALSE),
    SEX = list(choices = c("M", "F"), keep_na = FALSE)
  ))
  datasets
  datasets$get_filter_type("ADSL", "SEX")
  datasets$get_filter_info("ADSL")[["SEX"]]$type

  # will fail because of invalid range
  # datasets$set_filter_state("ADSL", list(
  #   AGE = list(range = c(3, 7), keep_na = FALSE)
  # ))
  datasets$set_filter_state("ADSL", list(
    AGE = list(range = c(33, 44), keep_na = FALSE)
  ))
  # wrapper functions
  teal:::set_single_filter_state(datasets, "ADSL", "AGE", c(35, 42))
  teal:::set_single_filter_state(
    datasets, "ADSL", "AGE",
    state = list(range = c(33, 44), keep_na = FALSE)
  )
  datasets$set_filter_state(
    "ADSL",
     state = list(
       SEX = list(choices = c("M", "F"), keep_na = FALSE)
    )
  )
  datasets$get_filter_type("ADSL", "SEX")

  datasets$get_filter_state("ADSL")

  datasets$print_filter_info("ADSL")

  # add ADAE to show dependency
  ADAE <- radae(cached = TRUE)
  ADAE_indep <- ADAE
  attr(ADAE, "keys") <- get_cdisc_keys("ADAE")
  attr(ADAE_indep, "keys") <- get_cdisc_keys("ADAE")
  attr(ADAE, "parent") <- "ADSL"

  datasets$set_data("ADAE", ADAE)
  datasets$set_data("ADAEIndep", ADAE_indep)

  datasets$set_join_keys(join_keys(join_key("ADSL", "ADAE", c("STUDYID", "USUBJID"))))

  datasets$get_data_info("ADSL", filtered = FALSE)
  datasets$get_data_info("ADSL", filtered = TRUE)
  datasets$get_data_info("ADAE", filtered = FALSE)
  datasets$get_data_info("ADAE", filtered = TRUE)
  datasets$get_data_info("ADAEIndep", filtered = FALSE)
  datasets$get_data_info("ADAEIndep", filtered = TRUE)

  # see generated code
  datasets$get_filter_expr("ADSL")
  datasets$get_filter_expr("ADAEIndep")
})
}
\seealso{
\code{FilteredData} class
}
\section{Super class}{
\code{\link[teal:FilteredData]{teal::FilteredData}} -> \code{CDISCFilteredData}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{CDISCFilteredData$new()}}
\item \href{#method-datanames}{\code{CDISCFilteredData$datanames()}}
\item \href{#method-get_filterable_datanames}{\code{CDISCFilteredData$get_filterable_datanames()}}
\item \href{#method-get_filterable_varnames}{\code{CDISCFilteredData$get_filterable_varnames()}}
\item \href{#method-get_filter_expr}{\code{CDISCFilteredData$get_filter_expr()}}
\item \href{#method-get_data_info}{\code{CDISCFilteredData$get_data_info()}}
\item \href{#method-get_parentname}{\code{CDISCFilteredData$get_parentname()}}
\item \href{#method-clone}{\code{CDISCFilteredData$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="teal" data-topic="FilteredData" data-id="get_bookmark_state">}\href{../../teal/html/FilteredData.html#method-get_bookmark_state}{\code{teal::FilteredData$get_bookmark_state()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="teal" data-topic="FilteredData" data-id="get_code">}\href{../../teal/html/FilteredData.html#method-get_code}{\code{teal::FilteredData$get_code()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="teal" data-topic="FilteredData" data-id="get_data">}\href{../../teal/html/FilteredData.html#method-get_data}{\code{teal::FilteredData$get_data()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="teal" data-topic="FilteredData" data-id="get_data_attr">}\href{../../teal/html/FilteredData.html#method-get_data_attr}{\code{teal::FilteredData$get_data_attr()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="teal" data-topic="FilteredData" data-id="get_datalabel">}\href{../../teal/html/FilteredData.html#method-get_datalabel}{\code{teal::FilteredData$get_datalabel()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="teal" data-topic="FilteredData" data-id="get_default_filter_state">}\href{../../teal/html/FilteredData.html#method-get_default_filter_state}{\code{teal::FilteredData$get_default_filter_state()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="teal" data-topic="FilteredData" data-id="get_filter_info">}\href{../../teal/html/FilteredData.html#method-get_filter_info}{\code{teal::FilteredData$get_filter_info()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="teal" data-topic="FilteredData" data-id="get_filter_overview_tbl">}\href{../../teal/html/FilteredData.html#method-get_filter_overview_tbl}{\code{teal::FilteredData$get_filter_overview_tbl()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="teal" data-topic="FilteredData" data-id="get_filter_state">}\href{../../teal/html/FilteredData.html#method-get_filter_state}{\code{teal::FilteredData$get_filter_state()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="teal" data-topic="FilteredData" data-id="get_filter_type">}\href{../../teal/html/FilteredData.html#method-get_filter_type}{\code{teal::FilteredData$get_filter_type()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="teal" data-topic="FilteredData" data-id="get_join_keys">}\href{../../teal/html/FilteredData.html#method-get_join_keys}{\code{teal::FilteredData$get_join_keys()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="teal" data-topic="FilteredData" data-id="get_keys">}\href{../../teal/html/FilteredData.html#method-get_keys}{\code{teal::FilteredData$get_keys()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="teal" data-topic="FilteredData" data-id="get_primary_keys">}\href{../../teal/html/FilteredData.html#method-get_primary_keys}{\code{teal::FilteredData$get_primary_keys()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="teal" data-topic="FilteredData" data-id="get_variable_labels">}\href{../../teal/html/FilteredData.html#method-get_variable_labels}{\code{teal::FilteredData$get_variable_labels()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="teal" data-topic="FilteredData" data-id="get_varlabels">}\href{../../teal/html/FilteredData.html#method-get_varlabels}{\code{teal::FilteredData$get_varlabels()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="teal" data-topic="FilteredData" data-id="get_varnames">}\href{../../teal/html/FilteredData.html#method-get_varnames}{\code{teal::FilteredData$get_varnames()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="teal" data-topic="FilteredData" data-id="print">}\href{../../teal/html/FilteredData.html#method-print}{\code{teal::FilteredData$print()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="teal" data-topic="FilteredData" data-id="print_filter_info">}\href{../../teal/html/FilteredData.html#method-print_filter_info}{\code{teal::FilteredData$print_filter_info()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="teal" data-topic="FilteredData" data-id="restore_state_from_bookmark">}\href{../../teal/html/FilteredData.html#method-restore_state_from_bookmark}{\code{teal::FilteredData$restore_state_from_bookmark()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="teal" data-topic="FilteredData" data-id="set_code">}\href{../../teal/html/FilteredData.html#method-set_code}{\code{teal::FilteredData$set_code()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="teal" data-topic="FilteredData" data-id="set_data">}\href{../../teal/html/FilteredData.html#method-set_data}{\code{teal::FilteredData$set_data()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="teal" data-topic="FilteredData" data-id="set_data_attr">}\href{../../teal/html/FilteredData.html#method-set_data_attr}{\code{teal::FilteredData$set_data_attr()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="teal" data-topic="FilteredData" data-id="set_data_attrs">}\href{../../teal/html/FilteredData.html#method-set_data_attrs}{\code{teal::FilteredData$set_data_attrs()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="teal" data-topic="FilteredData" data-id="set_filter_state">}\href{../../teal/html/FilteredData.html#method-set_filter_state}{\code{teal::FilteredData$set_filter_state()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="teal" data-topic="FilteredData" data-id="set_join_keys">}\href{../../teal/html/FilteredData.html#method-set_join_keys}{\code{teal::FilteredData$set_join_keys()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initialize a \code{CDISCFilteredData} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CDISCFilteredData$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-datanames"></a>}}
\if{latex}{\out{\hypertarget{method-datanames}{}}}
\subsection{Method \code{datanames()}}{
Get datanames

The datanames are returned in the order in which they must be
evaluated (in case of dependencies).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CDISCFilteredData$datanames()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
(\code{character} vector) of datanames
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_filterable_datanames"></a>}}
\if{latex}{\out{\hypertarget{method-get_filterable_datanames}{}}}
\subsection{Method \code{get_filterable_datanames()}}{
Get dataset names of a given dataname for the filtering.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CDISCFilteredData$get_filterable_datanames(dataname)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dataname}}{(\code{character} vector) names of the dataset}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
(\code{character} vector) of dataset names
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_filterable_varnames"></a>}}
\if{latex}{\out{\hypertarget{method-get_filterable_varnames}{}}}
\subsection{Method \code{get_filterable_varnames()}}{
Get variable names of a given dataname for the filtering.

This method incorporates parent-child relations, i.e.
parent variable names are removed from child variable names.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CDISCFilteredData$get_filterable_varnames(dataname)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dataname}}{(\code{character}) name of the dataset}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
(\code{character} vector) of variable names
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_filter_expr"></a>}}
\if{latex}{\out{\hypertarget{method-get_filter_expr}{}}}
\subsection{Method \code{get_filter_expr()}}{
Returns the filter expression to filter a single dataset including the \code{inner_join}
with its parent dataset. It assumes that the filtered datasets it depends
on are available.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CDISCFilteredData$get_filter_expr(dataname)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dataname}}{(\code{character}) name of the dataset}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
(\code{expression}) to filter dataset
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_data_info"></a>}}
\if{latex}{\out{\hypertarget{method-get_data_info}{}}}
\subsection{Method \code{get_data_info()}}{
Get info about dataname, i.e. number of rows, subjects.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CDISCFilteredData$get_data_info(dataname, filtered)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dataname}}{(\code{character}) name of the dataset}

\item{\code{filtered}}{(\code{logical}) whether to obtain this info for the
filtered dataset}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a named vector
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_parentname"></a>}}
\if{latex}{\out{\hypertarget{method-get_parentname}{}}}
\subsection{Method \code{get_parentname()}}{
Get parent dataset name
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CDISCFilteredData$get_parentname(dataname)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dataname}}{(\code{character}) name of the dataset}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
(\code{character}) name of parent dataset
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CDISCFilteredData$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
