% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataConnection.R
\name{cdse_connection}
\alias{cdse_connection}
\title{Open connection to \code{CDSE}}
\usage{
cdse_connection(env = "prod")
}
\arguments{
\item{env}{optional, \code{CDSE} environment name.}
}
\value{
(\code{DataConnection}) type of object
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#defunct}{\figure{lifecycle-defunct.svg}{options: alt='[Defunct]'}}}{\strong{[Defunct]}}
}
