% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RelationalDataConnector_constructors.R
\name{cdse_data}
\alias{cdse_data}
\title{\code{RelationalDataConnector} connector for \code{CDSE}}
\usage{
cdse_data(..., connection = cdse_connection())
}
\arguments{
\item{...}{(\code{DatasetConnector} objects)\cr
dataset connectors created using \code{\link{cdse_dataset_connector}}}

\item{connection}{(\code{DataConnection}) object returned from \code{cdse_connection}.}
}
\value{
An object of class \code{RelationalDataConnector}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#defunct}{\figure{lifecycle-defunct.svg}{options: alt='[Defunct]'}}}{\strong{[Defunct]}}
Build data connector for \code{CDSE} datasets
}
