% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_ellipsis}
\alias{check_ellipsis}
\title{Ensure the ellipsis, ..., in method arguments are empty}
\usage{
check_ellipsis(..., stop = FALSE, allowed_args = character(0))
}
\arguments{
\item{...}{it should literally just be ...}

\item{stop}{TRUE to raise an error; FALSE will output warning message}

\item{allowed_args}{character vector naming arguments that are allowed in the \code{...}.
to allow for unnamed arguments, let "" be one of the elements in this character vector.}
}
\value{
\code{NULL} if ... is empty
}
\description{
Ellipsis, ..., are needed as part of method arguments to allow for its arguments to be different from its generic's
arguments and for this to pass check(). Hence, ..., should always be empty. This function will check for this
condition.
}
\examples{
method.class <- function(a, b, c, ...) {
  check_ellipsis(...)
}
method.class <- function(a, b, c, ...) {
  check_ellipsis(..., allowed_args = c("y", "z"))
}
}
\keyword{internal}
