% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_filter_datanames}
\alias{check_filter_datanames}
\title{Check \code{datanames} in filters}
\usage{
check_filter_datanames(filters, datanames)
}
\arguments{
\item{filters}{(\code{teal_slices}) object}

\item{datanames}{(\code{character}) names of datasets available in the \code{data} object}
}
\value{
A \code{character(1)} containing error message or TRUE if validation passes.
}
\description{
This function checks whether \code{datanames} in filters correspond to those in \code{data},
returning character vector with error messages or \code{TRUE} if all checks pass.
}
\keyword{internal}
