% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdisc_data.R
\name{check_foreign_keys}
\alias{check_foreign_keys}
\title{Utility function to check if foreign keys are existing in parent dataset}
\usage{
check_foreign_keys(datasets_keys)
}
\arguments{
\item{datasets_keys}{list of keys}
}
\description{
Utility function to check if foreign keys are existing in parent dataset
}
