% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_in_range}
\alias{check_in_range}
\title{Check that a given range is valid}
\usage{
check_in_range(subinterval, range, pre_msg = "")
}
\arguments{
\item{subinterval}{(\code{numeric} or \code{date})\cr
vector of length 2 to be  compared against the full range.}

\item{range}{(\code{numeric} or \code{date})\cr
vector of length 2 containing the full range to validate against.}

\item{pre_msg}{\code{character} message to print before error for
additional context.}
}
\value{
\code{NULL} if \code{subinterval} is a valid range or error with message
otherwise.
}
\description{
Check that a given range is valid
}
\examples{
\dontrun{
check_in_range(c(3, 1), c(1, 3))
check_in_range(c(0, 3), c(1, 3))
check_in_range(
  c(as.Date("2020-01-01"), as.Date("2020-01-20")),
  c(as.Date("2020-01-01"), as.Date("2020-01-02"))
)
}
}
\keyword{internal}
