% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_in_subset}
\alias{check_in_subset}
\title{Check that one set is a subset of another}
\usage{
check_in_subset(subset, choices, pre_msg = "")
}
\arguments{
\item{subset}{\code{collection-like} should be a subset of \code{choices}}

\item{choices}{\code{collection-like} superset}

\item{pre_msg}{\code{character} message to print before}
}
\description{
Raises an error message if not and says which elements are not in
the allowed \code{choices}.
}
\examples{
check_in_subset(c("a", "b"), c("a", "b", "c"))
\dontrun{
check_in_subset(c("a", "b"), c("b", "c"), pre_msg = "Error: ")
# truncated because too long
check_in_subset("a", LETTERS, pre_msg = "Error: ")
}
}
