% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_pkg_quietly}
\alias{check_pkg_quietly}
\title{Check if package can be loaded}
\usage{
check_pkg_quietly(pckg, msg)
}
\arguments{
\item{pckg}{\code{character} package name.}

\item{msg}{\code{character} error message to display if package is not available.}
}
\value{
Error or invisible NULL.
}
\description{
Check if package can be loaded
}
\keyword{internal}
