% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_required_datanames}
\alias{check_required_datanames}
\alias{check_reserved_datanames}
\alias{check_required_datanames_html}
\title{Check \code{datanames} in modules}
\usage{
check_required_datanames(
  datanames_required,
  datanames_available,
  show_modules_info = FALSE
)

check_reserved_datanames(datanames)

check_required_datanames_html(
  datanames_required = list(),
  datanames_available,
  show_modules_info = FALSE
)
}
\arguments{
\item{datanames_required}{(\verb{named list}) list of datanames required by modules. List element per module, named
after module's label.}

\item{datanames_available}{(\code{character}) names of datasets available in the \code{data} object}
}
\value{
\code{TRUE} if validation passes, otherwise \code{character(1)} or \code{shiny.tag.list}
}
\description{
These functions check if specified \code{datanames} in modules match those in the data object,
returning error messages or \code{TRUE} for successful validation. Two functions return error message
in different forms:
\itemize{
\item \code{check_required_datanames} returns \code{character(1)} for basic assertion usage
\item \code{check_required_datanames_html} returns \code{shiny.tag.list} to display it in the app.
}
}
\keyword{internal}
