% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_setequal}
\alias{check_setequal}
\title{Check that two sets are equal and informative error message otherwise}
\usage{
check_setequal(x, y, pre_msg = "")
}
\arguments{
\item{x}{object to be compared to other}

\item{y}{object to be compared to other}

\item{pre_msg}{\code{character} to be displayed before error message}
}
\description{
Check that two sets are equal and informative error message otherwise
}
\examples{
check_setequal(1:3, 1:3)
check_setequal(c(1, 1, 3), c(1, 3, 3))
try(check_setequal(c(1, 2, 3), c(1, 3, 3), pre_msg = "Not equal: "))
try(check_setequal(c(1, 2, 3), c("a", "b"), pre_msg = "Not equal: "))
}
