% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RelationalDatasetConnector_constructors.R
\name{code_cdisc_dataset_connector}
\alias{code_cdisc_dataset_connector}
\title{Code \code{CDISC} Dataset Connector}
\usage{
code_cdisc_dataset_connector(
  dataname,
  code,
  mutate_code = character(0),
  label = character(0),
  ...
)
}
\arguments{
\item{dataname}{(\code{character})\cr
A given name for the dataset it may not contain spaces}

\item{code}{(\code{character})\cr
String containing the code to produce the object.
The code must end in a call to the object.}

\item{mutate_code}{(\code{character})\cr
String containing the code used to mutate the object
after it is produced.}

\item{label}{(\code{character})\cr
Label to describe the dataset.}

\item{...}{(\code{optional})\cr
Additional arguments applied to pull function.
In case when this object code depends on the \code{raw_data} from the other
\code{NamedDataset}, \code{NamedDatasetConnector} object(s) or other constant value,
this/these object(s) should be included. Please note that \code{vars}
are included to this object as local \code{vars} and they cannot be modified
within another dataset.}
}
\description{
\ifelse{html}{\out{<a href='https://www.tidyverse.org/lifecycle/#experimental'><img src='figures/lifecycle-experimental.svg' alt='Experimental lifecycle'></a>}}{\strong{Experimental}}
Create a \code{RelationalDatasetConnector} from a string of code with keys
assigned automatically by \code{dataname}.
}
