% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_relational.R
\name{code_from_script}
\alias{code_from_script}
\title{Get code from script}
\usage{
code_from_script(code, script, dataname = NULL)
}
\arguments{
\item{code}{(\code{character})\cr
A character string defining the code needed to produce the data set in \code{x}}

\item{script}{(\code{character})\cr
file that contains R Code that can be read using \link{read_script}.
Preferred before \code{code} argument}

\item{dataname}{(\code{character})\cr
A given name for the dataset it may not contain spaces}
}
\value{
code (\code{character})
}
\description{
Get code from script. Switches between \code{code} and \code{script arguments}
to return non-empty one to pass it further to constructors.
}
