% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_code.R
\name{code_remove_library}
\alias{code_remove_library}
\title{Remove library call from code}
\usage{
code_remove_library(lines)
}
\arguments{
\item{lines}{(\code{character}) of code as seperate element in vector}
}
\description{
Removes \code{library()} or \code{require()} calls from code
}
