% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_app_id}
\alias{create_app_id}
\title{Creates App ID as a hash of the input data and modules.}
\usage{
create_app_id(data, modules)
}
\arguments{
\item{data}{data object}

\item{modules}{modules object}
}
\value{
character(1) with hash
}
\description{
Creates App ID as a hash of the input data and modules.
}
\keyword{internal}
