% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_lockfile_future}
\alias{create_lockfile_future}
\title{Create \code{pak} \code{.lock} file}
\usage{
create_lockfile_future(lockfile = "session.lock")
}
\arguments{
\item{lockfile}{lock file name passed to \code{pak::lockfile_create()}}
}
\value{
(\code{character(n)}) content of \code{lockfile}.
}
\description{
Function detects used packages in the session, extends options('repos') with custom repositories,
and then executes \code{pak::lockfile_create()} that creates a \code{.lock} file for future session reproducibility.
}
\keyword{internal}
