% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_renv_lockfile}
\alias{create_renv_lockfile}
\title{Create \code{renv} \code{.lock} file}
\usage{
create_renv_lockfile()
}
\value{
(\code{character(1)}) the path to the \code{lockfile} created in a \code{tempdir()}.
}
\description{
Function allows to create \code{renv::snapshot()} during the \code{teal} initialization.
}
\details{
User is able to provide his/her own predefined \code{.lockfile} by setting the path to the \code{.lockfile} through
\code{options("teal.renv.lockfile")}. Then this function is not used. If this option is empty, by default \code{teal} will
create an \code{implicit} type of the \code{.lockfile}, that uses \code{renv::dependencies()} to detect all R packages in current
project working directory. You can always include a \code{DESCRIPTION} file in your working directory and enable
\code{.lockfile} creation based on this file. To do this, set \code{options(teal.renv.type = "explicit")}.
Naming of \code{type} is the same as in \code{renv::snapshot()}.
}
\note{
This function computes the \code{.lockfile} as a \code{future::future} promise, while
running the evaluation of the process on a separate worker. \code{future::plan()} and \code{shiny::ExtendedTask()} are used to
setup parallel asynchronous computations.
}
\keyword{internal}
