% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RelationalDatasetConnector_constructors.R
\name{csv_cdisc_dataset_connector}
\alias{csv_cdisc_dataset_connector}
\title{\code{csv} \code{CDISC} Dataset Connector}
\usage{
csv_cdisc_dataset_connector(
  dataname,
  file,
  code = character(0),
  script = character(0),
  label = character(0),
  ...
)
}
\arguments{
\item{dataname}{(\code{character})\cr
A given name for the dataset it may not contain spaces}

\item{file}{(\code{character})\cr
path to (\verb{.csv)} (or general delimited) file that contains \code{data.frame} object}

\item{code}{(\code{character})\cr
A character string defining code to modify \code{raw_data} from this dataset. To modify
current dataset code should contain at least one assignment to object defined in \code{dataname}
argument. For example if \code{dataname = ADSL} example code should contain
\verb{ADSL <- <some R code>}. Can't be used simultaneously with \code{script}}

\item{script}{(\code{character})\cr
Alternatively to \code{code} - location of the file containing modification code.
Can't be used simultaneously with \code{script}.}

\item{label}{(\code{character})\cr
Label to describe the dataset.}

\item{...}{(\code{optional})\cr
additional arguments applied to pull function (\code{readr::read_delim}) by default
\code{delim = ","}.}
}
\description{
\ifelse{html}{\out{<a href='https://www.tidyverse.org/lifecycle/#experimental'><img src='figures/lifecycle-experimental.svg' alt='Experimental lifecycle'></a>}}{\strong{Experimental}}
Create a \code{RelationalDatasetConnector} from \code{csv} (or general delimited) file
with keys assigned automatically by \code{dataname}.
}
