% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{custom_modal_dialog}
\alias{custom_modal_dialog}
\title{Custom modalDialog}
\usage{
custom_modal_dialog(id, modal_args, ...)
}
\arguments{
\item{id}{(\code{character(1)}) \code{shiny} module id for the modalDialog.\cr
Note that this id will be used to show/hide using \code{shinyjs::show} and \code{shinyjs::hide}}

\item{modal_args}{(\code{list}) list of arguments for the \code{shiny::modalDialog} function to customize the modal}

\item{...}{(\code{shiny.tag}) shiny UI elements that will be displayed in the modal UI}
}
\value{
(\code{shiny.tag}) returns \code{HTML} for Shiny module UI with modifications that allow for custom show/hide
}
\description{
Alternative to modalDialog. To create custom modalDialogs that can be shown/hidden
using JS using the modal \code{id} without disturbing other modalDialogs
}
\keyword{internal}
