% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilterStates.R
\name{data_choices_labeled}
\alias{data_choices_labeled}
\title{Returns a \code{choices_labeled} object}
\usage{
data_choices_labeled(
  data,
  choices,
  varlabels = character(0),
  keys = character(0)
)
}
\value{
\code{character(0)} if choices are empty; a \code{choices_labeled} object otherwise
}
\description{
Returns a \code{choices_labeled} object
}
