% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataConnection.R
\name{data_connection}
\alias{data_connection}
\title{Public facing object constructor for \code{DataConnection} class.}
\usage{
data_connection(
  open_fun = NULL,
  close_fun = NULL,
  ping_fun = NULL,
  if_conn_obj = FALSE
)
}
\arguments{
\item{open_fun}{(\code{CallableFunction}) function to open connection}

\item{close_fun}{(\code{CallableFunction}) function to close connection}

\item{ping_fun}{(\code{CallableFunction}) function to ping connection}

\item{if_conn_obj}{optional, (\code{logical}) whether to store \code{conn} object returned from opening}
}
\value{
\code{DataConnection} object
}
\description{
Public facing object constructor for \code{DataConnection} class.
}
