% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataConnector.R
\name{data_connector}
\alias{data_connector}
\title{Function to create an object of class DataConnector}
\usage{
data_connector(ui, server)
}
\arguments{
\item{ui}{(\code{function}) A shiny module UI function with the argument \code{ui}}

\item{server}{(\code{function}) A shiny module SERVER function with the
arguments \code{input}, \code{output}, \code{session} and one additional
argument to handover data to the final teal app}
}
\value{
An object of class \code{DataConnector}
}
\description{
Function to create an object of class DataConnector
}
