% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_extract.R
\name{data_extract}
\alias{data_extract}
\title{Data Extract input for teal modules}
\usage{
data_extract(dataname = NULL, keys_filtering = NULL, columns = NULL)
}
\arguments{
\item{dataname}{(\code{character}) Name of a teal data set}

\item{keys_filtering}{(\code{list}) Define how to filter the
key columns of the data set. This is the outcome of \link{keys_filter_from_sep}}

\item{columns}{(\code{choices_selected}) Define which columns of the data set shall
be selected next to the key variables. This shall be the outcome of \link{choices_selected}}
}
\description{
The Data Extract input can be used to filter and select columns from a data
set to be used in teal. From this function's output a \code{data_extract_input} can
be constructed. This input can be read by a \link{data_extractor} module.
}
