% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_extract_spec.R
\name{data_extract_spec}
\alias{data_extract_spec}
\title{Constructor for \link{DataExtractSpec}}
\usage{
data_extract_spec(dataname, columns, filter = NULL)
}
\arguments{
\item{dataname}{(\code{character}) The name of the \code{\link[teal]{teal}} dataset to
be extracted. This dataset has to be handed over to the \code{data} argument of the
\code{teal::}\code{\link[teal]{init}} function.}

\item{columns}{(\code{column_choices_spec-S3 class}) Columns to be selected from the input dataset
mentioned in \code{dataname.} The setup can be created using \code{\link{columns_spec}} function.}

\item{filter}{(\code{filter_choices_spec}-S3-class) Setup of the filtering of
key columns inside the dataset. This setup can be created using the \code{\link{filter_spec}}
function.}
}
\value{
\link{DataExtractSpec} class object
}
\description{
For the description please look into \link{DataExtractSpec}
}
