% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_extract_shiny_module.R
\name{data_extractor}
\alias{data_extractor}
\title{teal Data Column Extraction module server}
\usage{
data_extractor(input, output, session, datasets, constant_values)
}
\arguments{
\item{input}{\code{shiny input}}

\item{output}{\code{shiny output}}

\item{session}{\code{shiny session}}

\item{datasets}{(\code{teal::FilteredData}) Data build up by teal}

\item{constant_values}{(\code{data_extract}) A list of data filter and select
information constructed by \link{data_extract}}
}
\value{
A \code{data.frame} where the keys are filtered and the columns selected
 due to the \link{data_extract_input} that is called with this module.
}
\description{
teal Data Column Extraction module server
}
