% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdisc_data.R
\name{data_label<-}
\alias{data_label<-}
\title{Set dataset label attribute}
\usage{
data_label(x) <- value
}
\arguments{
\item{x}{\code{data.frame} for which attribute is set}

\item{value}{(\code{character}) label}
}
\value{
modified \code{x} object
}
\description{
Set dataset label attribute
}
\examples{
library(random.cdisc.data)
x <- radsl(cached = TRUE)
data_label(x) <- "My custom label"
data_label(x)
}
