% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdisc_data.R
\name{data_label}
\alias{data_label}
\title{Get dataset label attribute}
\usage{
data_label(data)
}
\arguments{
\item{data}{\code{data.frame} from which attribute is extracted}
}
\value{
(\code{character}) label or \code{NULL} if it's missing
}
\description{
\ifelse{html}{\out{<a href='https://www.tidyverse.org/lifecycle/#maturing'><img src='figures/lifecycle-maturing.svg' alt='Maturing lifecycle'></a>}}{\strong{Maturing}}
}
\examples{
library(random.cdisc.data)
data_label(radsl(cached = TRUE))
}
