% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_module.R
\name{data_module}
\alias{data_module}
\title{Data module}
\usage{
data_module(ui, server)
}
\arguments{
\item{ui}{(\verb{function(id)})\cr
\code{shiny} \code{ui} module with \code{id} argument}

\item{server}{(\verb{function(id)})\cr
\code{shiny} server function with \code{id} as argument. Module should return reactive \code{teal_data}.}
}
\description{
Data input for \code{teal::init} in form of a module
}
\examples{
data <- data_module(
  ui = function(id) {
    ns <- NS(id)
    actionButton(ns("submit"), label = "Load data")
  },
  server = function(id) {
    moduleServer(id, function(input, output, session) {
      eventReactive(input$submit, {
        data <- within(
          teal.data::teal_data(),
          {
            dataset1 <- iris
            dataset2 <- mtcars
          }
        )
        teal.data::datanames(data) <- c("iris", "mtcars")

        data
      })
    })
  }
)
}
