% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_data_table.R
\name{data_table_item}
\alias{data_table_item}
\title{Create A Data Table Viewer}
\usage{
data_table_item(label = "data table", variables_selected = NULL)
}
\arguments{
\item{label}{label shown in the navigation for the item}

\item{variables_selected}{a named list that says which variables should be
initially  shown for which dataset}
}
\description{
A data table viewer shows the data using a paginated table.
}
