% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_extract_spec.R
\docType{data}
\name{DataExtractSpec}
\alias{DataExtractSpec}
\alias{data_extract_class}
\title{Data Extract input for teal modules}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
data_extract_class
}
\arguments{
\item{dataname}{(\code{character}) Name of a teal data set}

\item{keys_filtering}{(\code{list}) Define how to filter the
key columns of the data set. This is the outcome of \link{keys_filtering_spec}}

\item{columns}{(\code{choices_selected}) Define which columns of the data set shall
be selected next to the key variables. This shall be the outcome of \link{choices_selected}}
}
\description{
The Data Extract input can be used to filter and select columns from a data
set to be used in teal. From this function's output a \code{data_extract_input} can
be constructed. This input can be read by a \link{data_extract_module} module.
}
\section{Fields}{

\describe{
\item{\code{dataname}}{(\code{character}) Data set to be extracted and selected}

\item{\code{keys_filtering}}{(\code{KeysFilteringSpec}) Setup of the dataset filtering}

\item{\code{columns}}{(\code{choices_selected}) Columns to be selected from the input dataset}
}}

\keyword{datasets}
