% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAETealDataset.R
\name{dataset.MultiAssayExperiment}
\alias{dataset.MultiAssayExperiment}
\title{S3 method to construct an \code{MAETealDataset} object from \code{MultiAssayExperiment}}
\usage{
\method{dataset}{MultiAssayExperiment}(
  dataname,
  x,
  keys = character(0),
  label = data_label(x),
  code = character(0),
  vars = list()
)
}
\arguments{
\item{dataname}{(\code{character})\cr
A given name for the dataset it may not contain spaces}

\item{x}{(\code{MultiAssayExperiment})}

\item{keys}{optional, (\code{character})\cr
vector with primary keys}

\item{label}{(\code{character})\cr
Label to describe the dataset}

\item{code}{(\code{character})\cr
A character string defining the code needed to produce the data set in \code{x}}

\item{vars}{(named \code{list})) \cr
In case when this object code depends on other \code{TealDataset} object(s) or
other constant value, this/these object(s) should be included as named
element(s) of the list. For example if this object code needs \code{ADSL}
object we should specify \verb{vars = list(ADSL = <adsl object>)}.
It's recommended to include \code{TealDataset} or \code{TealDatasetConnector} objects to
the \code{vars} list to preserve reproducibility. Please note that \code{vars}
are included to this object as local \code{vars} and they cannot be modified
within another dataset.}
}
\description{
S3 method to construct an \code{MAETealDataset} object from \code{MultiAssayExperiment}
}
\examples{
# Simple example
mae_d <- dataset("MAE", MultiAssayExperiment::miniACC, keys = c("STUDYID", "USUBJID"))
mae_d$get_dataname()
mae_d$get_dataset_label()
mae_d$get_code()
mae_d$get_raw_data()
}
