% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dataset.R
\name{dataset}
\alias{dataset}
\alias{raw_dataset}
\alias{named_dataset}
\alias{relational_dataset}
\title{Constructor for \code{\link{Dataset}} class}
\usage{
dataset(
  dataname,
  x,
  keys = character(0),
  label = data_label(x),
  code = character(0),
  vars = list()
)

raw_dataset(x)

named_dataset(
  dataname,
  x,
  code = character(0),
  label = character(0),
  vars = list()
)

relational_dataset(
  dataname,
  x,
  keys = character(0),
  code = character(0),
  label = character(0),
  vars = list()
)
}
\arguments{
\item{dataname}{(\code{character})\cr
A given name for the dataset it may not contain spaces}

\item{x}{(\code{data.frame})\cr}

\item{keys}{optional, (\code{character})\cr
vector with primary keys}

\item{label}{(\code{character})\cr
Label to describe the dataset}

\item{code}{(\code{character})\cr
A character string defining the code needed to produce the data set in \code{x}}

\item{vars}{(named \code{list})) \cr
In case when this object code depends on the \code{raw_data} from the other
\code{Dataset} object(s) or other constant value,
this/these object(s) should be included as named element of the list.}
}
\value{
\code{\link{Dataset}} object
}
\description{
\ifelse{html}{\out{<a href='https://www.tidyverse.org/lifecycle/#experimental'><img src='figures/lifecycle-experimental.svg' alt='Experimental lifecycle'></a>}}{\strong{Experimental}}

\ifelse{html}{\figure{lifecycle-defunct.svg}{options: alt='Defunct lifecycle'}}{\strong{Defunct}}

\ifelse{html}{\figure{lifecycle-soft-deprecated.svg}{options: alt='Soft-deprecated lifecycle'}}{\strong{Soft-deprecated}}

\ifelse{html}{\figure{lifecycle-soft-deprecated.svg}{options: alt='Soft-deprecated lifecycle'}}{\strong{Soft-deprecated}}
}
\examples{
# Simple example
dataset("iris", iris)

# Example with more arguments
library(random.cdisc.data)
ADSL <- radsl(cached = TRUE)
ADSL_dataset <- dataset(dataname = "ADSL", x = ADSL)

ADSL_dataset$get_dataname()

ADSL_dataset <- dataset(
  dataname = "ADSL",
  x = ADSL,
  label = "AdAM subject-level dataset",
  code = "ADSL <- radsl(cached = TRUE)"
)

ADSL_dataset$get_dataset_label()
ADSL_dataset$get_code()
}
