% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAETealDataset.R, R/TealDataset.R
\name{MAETealDataset}
\alias{MAETealDataset}
\alias{TealDataset}
\title{R6 Class representing a \code{MultiAssayExperiment} object with its attributes}
\description{
Any \code{MultiAssayExperiment} object can be stored inside this \code{MAETealDataset}.
Some attributes like colnames, dimension or column names for a specific type will
be automatically derived.

Any \code{data.frame} object can be stored inside this object.
Some attributes like colnames, dimension or column names for a specific type will
be automatically derived.
}
\examples{

## ------------------------------------------------
## Method `TealDataset$reassign_datasets_vars`
## ------------------------------------------------

test_dataset <- teal:::TealDataset$new(
  dataname = "iris",
  x = iris,
  vars = list(dep = teal:::TealDataset$new("iris2", iris))
)
test_dataset$reassign_datasets_vars(
  list(iris2 = teal:::TealDataset$new("iris2", head(iris)))
)

}
\section{Super class}{
\code{\link[teal:TealDataset]{teal::TealDataset}} -> \code{MAETealDataset}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{MAETealDataset$new()}}
\item \href{#method-check}{\code{MAETealDataset$check()}}
\item \href{#method-check_keys}{\code{MAETealDataset$check_keys()}}
\item \href{#method-print}{\code{MAETealDataset$print()}}
\item \href{#method-clone}{\code{MAETealDataset$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="teal" data-topic="TealDataset" data-id="get_attrs">}\href{../../teal/html/TealDataset.html#method-get_attrs}{\code{teal::TealDataset$get_attrs()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="teal" data-topic="TealDataset" data-id="get_character_colnames">}\href{../../teal/html/TealDataset.html#method-get_character_colnames}{\code{teal::TealDataset$get_character_colnames()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="teal" data-topic="TealDataset" data-id="get_code">}\href{../../teal/html/TealDataset.html#method-get_code}{\code{teal::TealDataset$get_code()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="teal" data-topic="TealDataset" data-id="get_code_class">}\href{../../teal/html/TealDataset.html#method-get_code_class}{\code{teal::TealDataset$get_code_class()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="teal" data-topic="TealDataset" data-id="get_colnames">}\href{../../teal/html/TealDataset.html#method-get_colnames}{\code{teal::TealDataset$get_colnames()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="teal" data-topic="TealDataset" data-id="get_column_labels">}\href{../../teal/html/TealDataset.html#method-get_column_labels}{\code{teal::TealDataset$get_column_labels()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="teal" data-topic="TealDataset" data-id="get_dataname">}\href{../../teal/html/TealDataset.html#method-get_dataname}{\code{teal::TealDataset$get_dataname()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="teal" data-topic="TealDataset" data-id="get_datanames">}\href{../../teal/html/TealDataset.html#method-get_datanames}{\code{teal::TealDataset$get_datanames()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="teal" data-topic="TealDataset" data-id="get_dataset">}\href{../../teal/html/TealDataset.html#method-get_dataset}{\code{teal::TealDataset$get_dataset()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="teal" data-topic="TealDataset" data-id="get_dataset_label">}\href{../../teal/html/TealDataset.html#method-get_dataset_label}{\code{teal::TealDataset$get_dataset_label()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="teal" data-topic="TealDataset" data-id="get_factor_colnames">}\href{../../teal/html/TealDataset.html#method-get_factor_colnames}{\code{teal::TealDataset$get_factor_colnames()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="teal" data-topic="TealDataset" data-id="get_hash">}\href{../../teal/html/TealDataset.html#method-get_hash}{\code{teal::TealDataset$get_hash()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="teal" data-topic="TealDataset" data-id="get_join_keys">}\href{../../teal/html/TealDataset.html#method-get_join_keys}{\code{teal::TealDataset$get_join_keys()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="teal" data-topic="TealDataset" data-id="get_keys">}\href{../../teal/html/TealDataset.html#method-get_keys}{\code{teal::TealDataset$get_keys()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="teal" data-topic="TealDataset" data-id="get_mutate_code_class">}\href{../../teal/html/TealDataset.html#method-get_mutate_code_class}{\code{teal::TealDataset$get_mutate_code_class()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="teal" data-topic="TealDataset" data-id="get_mutate_vars">}\href{../../teal/html/TealDataset.html#method-get_mutate_vars}{\code{teal::TealDataset$get_mutate_vars()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="teal" data-topic="TealDataset" data-id="get_numeric_colnames">}\href{../../teal/html/TealDataset.html#method-get_numeric_colnames}{\code{teal::TealDataset$get_numeric_colnames()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="teal" data-topic="TealDataset" data-id="get_raw_data">}\href{../../teal/html/TealDataset.html#method-get_raw_data}{\code{teal::TealDataset$get_raw_data()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="teal" data-topic="TealDataset" data-id="get_row_labels">}\href{../../teal/html/TealDataset.html#method-get_row_labels}{\code{teal::TealDataset$get_row_labels()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="teal" data-topic="TealDataset" data-id="get_rownames">}\href{../../teal/html/TealDataset.html#method-get_rownames}{\code{teal::TealDataset$get_rownames()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="teal" data-topic="TealDataset" data-id="get_var_r6">}\href{../../teal/html/TealDataset.html#method-get_var_r6}{\code{teal::TealDataset$get_var_r6()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="teal" data-topic="TealDataset" data-id="get_vars">}\href{../../teal/html/TealDataset.html#method-get_vars}{\code{teal::TealDataset$get_vars()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="teal" data-topic="TealDataset" data-id="is_mutate_delayed">}\href{../../teal/html/TealDataset.html#method-is_mutate_delayed}{\code{teal::TealDataset$is_mutate_delayed()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="teal" data-topic="TealDataset" data-id="is_pulled">}\href{../../teal/html/TealDataset.html#method-is_pulled}{\code{teal::TealDataset$is_pulled()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="teal" data-topic="TealDataset" data-id="merge_join_keys">}\href{../../teal/html/TealDataset.html#method-merge_join_keys}{\code{teal::TealDataset$merge_join_keys()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="teal" data-topic="TealDataset" data-id="mutate">}\href{../../teal/html/TealDataset.html#method-mutate}{\code{teal::TealDataset$mutate()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="teal" data-topic="TealDataset" data-id="mutate_join_keys">}\href{../../teal/html/TealDataset.html#method-mutate_join_keys}{\code{teal::TealDataset$mutate_join_keys()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="teal" data-topic="TealDataset" data-id="reassign_datasets_vars">}\href{../../teal/html/TealDataset.html#method-reassign_datasets_vars}{\code{teal::TealDataset$reassign_datasets_vars()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="teal" data-topic="TealDataset" data-id="recreate">}\href{../../teal/html/TealDataset.html#method-recreate}{\code{teal::TealDataset$recreate()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="teal" data-topic="TealDataset" data-id="set_code">}\href{../../teal/html/TealDataset.html#method-set_code}{\code{teal::TealDataset$set_code()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="teal" data-topic="TealDataset" data-id="set_dataset_label">}\href{../../teal/html/TealDataset.html#method-set_dataset_label}{\code{teal::TealDataset$set_dataset_label()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="teal" data-topic="TealDataset" data-id="set_join_keys">}\href{../../teal/html/TealDataset.html#method-set_join_keys}{\code{teal::TealDataset$set_join_keys()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="teal" data-topic="TealDataset" data-id="set_keys">}\href{../../teal/html/TealDataset.html#method-set_keys}{\code{teal::TealDataset$set_keys()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="teal" data-topic="TealDataset" data-id="set_vars">}\href{../../teal/html/TealDataset.html#method-set_vars}{\code{teal::TealDataset$set_vars()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new object of \code{MAETealDataset} class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MAETealDataset$new(
  dataname,
  x,
  keys = character(0),
  code = character(0),
  label = character(0),
  vars = list()
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dataname}}{(\code{character})\cr
A given name for the dataset it may not contain spaces}

\item{\code{x}}{(\code{MultiAssayExperiment})\cr}

\item{\code{keys}}{optional, (\code{character})\cr
vector with primary keys}

\item{\code{code}}{(\code{character} or \code{CodeClass})\cr
A character string defining the code needed to produce the data set in \code{x}.
\code{initialize()} and \code{recreate()} accept code as \code{CodeClass} also
which is needed to preserve the code uniqueness and correct order.}

\item{\code{label}}{(\code{character})\cr
Label to describe the dataset}

\item{\code{vars}}{(named \code{list})) \cr
In case when this object code depends on other \code{TealDataset} object(s) or
other constant value, this/these object(s) should be included as named
element(s) of the list. For example if this object code needs \code{ADSL}
object we should specify \verb{vars = list(ADSL = <adsl object>)}.
It's recommended to include \code{TealDataset} or \code{TealDatasetConnector} objects to
the \code{vars} list to preserve reproducibility. Please note that \code{vars}
are included to this object as local \code{vars} and they cannot be modified
within another dataset.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-check"></a>}}
\if{latex}{\out{\hypertarget{method-check}{}}}
\subsection{Method \code{check()}}{
Check to determine if the raw data is reproducible from the \code{get_code()} code.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MAETealDataset$check()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{TRUE} if the dataset generated from evaluating the
\code{get_code()} code is identical to the raw data, else \code{FALSE}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-check_keys"></a>}}
\if{latex}{\out{\hypertarget{method-check_keys}{}}}
\subsection{Method \code{check_keys()}}{
Check if keys has been specified correctly for dataset. Set of \code{keys}
should distinguish unique rows or be \code{character(0)}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MAETealDataset$check_keys(keys = private$.keys)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{keys}}{optional, (\code{character})\cr
vector with primary keys}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{TRUE} if dataset has been already pulled, else \code{FALSE}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
Prints this \code{MAETealDataset}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MAETealDataset$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{additional arguments to the printing method}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
invisibly self
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MAETealDataset$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{ncol}}{Number of columns}

\item{\code{nrow}}{Number of rows}

\item{\code{dim}}{Dimension \code{c(x, y)}}

\item{\code{colnames}}{The column names of the data}

\item{\code{rownames}}{The rownames of the data}

\item{\code{raw_data}}{The data.frame behind this R6 class}

\item{\code{data}}{The data.frame behind this R6 class}

\item{\code{var_names}}{The column names of the data}

\item{\code{row_labels}}{Row labels (can have spaces)}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{TealDataset$new()}}
\item \href{#method-recreate}{\code{TealDataset$recreate()}}
\item \href{#method-print}{\code{TealDataset$print()}}
\item \href{#method-get_dataset}{\code{TealDataset$get_dataset()}}
\item \href{#method-get_attrs}{\code{TealDataset$get_attrs()}}
\item \href{#method-get_raw_data}{\code{TealDataset$get_raw_data()}}
\item \href{#method-get_numeric_colnames}{\code{TealDataset$get_numeric_colnames()}}
\item \href{#method-get_character_colnames}{\code{TealDataset$get_character_colnames()}}
\item \href{#method-get_factor_colnames}{\code{TealDataset$get_factor_colnames()}}
\item \href{#method-get_colnames}{\code{TealDataset$get_colnames()}}
\item \href{#method-get_column_labels}{\code{TealDataset$get_column_labels()}}
\item \href{#method-get_rownames}{\code{TealDataset$get_rownames()}}
\item \href{#method-get_row_labels}{\code{TealDataset$get_row_labels()}}
\item \href{#method-get_dataname}{\code{TealDataset$get_dataname()}}
\item \href{#method-get_datanames}{\code{TealDataset$get_datanames()}}
\item \href{#method-get_dataset_label}{\code{TealDataset$get_dataset_label()}}
\item \href{#method-get_keys}{\code{TealDataset$get_keys()}}
\item \href{#method-get_join_keys}{\code{TealDataset$get_join_keys()}}
\item \href{#method-get_hash}{\code{TealDataset$get_hash()}}
\item \href{#method-get_var_r6}{\code{TealDataset$get_var_r6()}}
\item \href{#method-reassign_datasets_vars}{\code{TealDataset$reassign_datasets_vars()}}
\item \href{#method-set_dataset_label}{\code{TealDataset$set_dataset_label()}}
\item \href{#method-set_keys}{\code{TealDataset$set_keys()}}
\item \href{#method-set_join_keys}{\code{TealDataset$set_join_keys()}}
\item \href{#method-merge_join_keys}{\code{TealDataset$merge_join_keys()}}
\item \href{#method-mutate_join_keys}{\code{TealDataset$mutate_join_keys()}}
\item \href{#method-set_vars}{\code{TealDataset$set_vars()}}
\item \href{#method-set_code}{\code{TealDataset$set_code()}}
\item \href{#method-get_code}{\code{TealDataset$get_code()}}
\item \href{#method-get_code_class}{\code{TealDataset$get_code_class()}}
\item \href{#method-get_mutate_code_class}{\code{TealDataset$get_mutate_code_class()}}
\item \href{#method-get_vars}{\code{TealDataset$get_vars()}}
\item \href{#method-get_mutate_vars}{\code{TealDataset$get_mutate_vars()}}
\item \href{#method-is_mutate_delayed}{\code{TealDataset$is_mutate_delayed()}}
\item \href{#method-mutate}{\code{TealDataset$mutate()}}
\item \href{#method-check}{\code{TealDataset$check()}}
\item \href{#method-check_keys}{\code{TealDataset$check_keys()}}
\item \href{#method-is_pulled}{\code{TealDataset$is_pulled()}}
\item \href{#method-clone}{\code{TealDataset$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new object of \code{TealDataset} class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TealDataset$new(
  dataname,
  x,
  keys = character(0),
  code = character(0),
  label = character(0),
  vars = list()
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dataname}}{(\code{character})\cr
A given name for the dataset it may not contain spaces}

\item{\code{x}}{(\code{data.frame})\cr}

\item{\code{keys}}{optional, (\code{character})\cr
Vector with primary keys}

\item{\code{code}}{(\code{character})\cr
A character string defining the code needed to produce the data set in \code{x}.
\code{initialize()} and \code{recreate()} accept code as \code{CodeClass} also
which is needed to preserve the code uniqueness and correct order.}

\item{\code{label}}{(\code{character})\cr
Label to describe the dataset}

\item{\code{vars}}{(named \code{list})) \cr
In case when this object code depends on other \code{TealDataset} object(s) or
other constant value, this/these object(s) should be included as named
element(s) of the list. For example if this object code needs \code{ADSL}
object we should specify \verb{vars = list(ADSL = <adsl object>)}.
It's recommended to include \code{TealDataset} or \code{TealDatasetConnector} objects to
the \code{vars} list to preserve reproducibility. Please note that \code{vars}
are included to this object as local \code{vars} and they cannot be modified
within another dataset.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-recreate"></a>}}
\if{latex}{\out{\hypertarget{method-recreate}{}}}
\subsection{Method \code{recreate()}}{
Recreate this \code{TealDataset} with its current attributes.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TealDataset$recreate(
  dataname = self$get_dataname(),
  x = self$get_raw_data(),
  keys = self$get_keys(),
  code = private$code,
  label = self$get_dataset_label(),
  vars = list()
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dataname}}{(\code{character})\cr
A given name for the dataset it may not contain spaces}

\item{\code{x}}{(\code{data.frame})\cr}

\item{\code{keys}}{optional, (\code{character})\cr
Vector with primary keys}

\item{\code{code}}{(\code{character})\cr
A character string defining the code needed to produce the data set in \code{x}.
\code{initialize()} and \code{recreate()} accept code as \code{CodeClass} also
which is needed to preserve the code uniqueness and correct order.}

\item{\code{label}}{(\code{character})\cr
Label to describe the dataset}

\item{\code{vars}}{(named \code{list})) \cr
In case when this object code depends on other \code{TealDataset} object(s) or
other constant value, this/these object(s) should be included as named
element(s) of the list. For example if this object code needs \code{ADSL}
object we should specify \verb{vars = list(ADSL = <adsl object>)}.
It's recommended to include \code{TealDataset} or \code{TealDatasetConnector} objects to
the \code{vars} list to preserve reproducibility. Please note that \code{vars}
are included to this object as local \code{vars} and they cannot be modified
within another dataset.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a new object of the \code{TealDataset} class
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
Prints this \code{TealDataset}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TealDataset$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{additional arguments to the printing method}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
invisibly self
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_dataset"></a>}}
\if{latex}{\out{\hypertarget{method-get_dataset}{}}}
\subsection{Method \code{get_dataset()}}{
Performs any delayed mutate calls before returning self.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TealDataset$get_dataset()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
dataset (\code{TealDataset})
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_attrs"></a>}}
\if{latex}{\out{\hypertarget{method-get_attrs}{}}}
\subsection{Method \code{get_attrs()}}{
Get all dataset attributes
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TealDataset$get_attrs()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
(named \code{list}) with dataset attributes
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_raw_data"></a>}}
\if{latex}{\out{\hypertarget{method-get_raw_data}{}}}
\subsection{Method \code{get_raw_data()}}{
Derive the raw data frame inside this object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TealDataset$get_raw_data()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{data.frame}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_numeric_colnames"></a>}}
\if{latex}{\out{\hypertarget{method-get_numeric_colnames}{}}}
\subsection{Method \code{get_numeric_colnames()}}{
Derive the names of all \code{numeric} columns
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TealDataset$get_numeric_colnames()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{character} vector.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_character_colnames"></a>}}
\if{latex}{\out{\hypertarget{method-get_character_colnames}{}}}
\subsection{Method \code{get_character_colnames()}}{
Derive the names of all \code{character} columns
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TealDataset$get_character_colnames()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{character} vector.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_factor_colnames"></a>}}
\if{latex}{\out{\hypertarget{method-get_factor_colnames}{}}}
\subsection{Method \code{get_factor_colnames()}}{
Derive the names of all \code{factor} columns
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TealDataset$get_factor_colnames()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{character} vector.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_colnames"></a>}}
\if{latex}{\out{\hypertarget{method-get_colnames}{}}}
\subsection{Method \code{get_colnames()}}{
Derive the column names
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TealDataset$get_colnames()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{character} vector.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_column_labels"></a>}}
\if{latex}{\out{\hypertarget{method-get_column_labels}{}}}
\subsection{Method \code{get_column_labels()}}{
Derive the column labels
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TealDataset$get_column_labels()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{character} vector.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_rownames"></a>}}
\if{latex}{\out{\hypertarget{method-get_rownames}{}}}
\subsection{Method \code{get_rownames()}}{
Derive the row names
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TealDataset$get_rownames()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{character} vector.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_row_labels"></a>}}
\if{latex}{\out{\hypertarget{method-get_row_labels}{}}}
\subsection{Method \code{get_row_labels()}}{
Derive the row labels
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TealDataset$get_row_labels()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{character} vector.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_dataname"></a>}}
\if{latex}{\out{\hypertarget{method-get_dataname}{}}}
\subsection{Method \code{get_dataname()}}{
Derive the \code{name} which was formerly called \code{dataname}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TealDataset$get_dataname()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{character} name of the dataset
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_datanames"></a>}}
\if{latex}{\out{\hypertarget{method-get_datanames}{}}}
\subsection{Method \code{get_datanames()}}{
Derive the dataname
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TealDataset$get_datanames()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{character} name of the dataset
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_dataset_label"></a>}}
\if{latex}{\out{\hypertarget{method-get_dataset_label}{}}}
\subsection{Method \code{get_dataset_label()}}{
Derive the \code{label} which was former called \code{datalabel}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TealDataset$get_dataset_label()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{character} label of the dataset
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_keys"></a>}}
\if{latex}{\out{\hypertarget{method-get_keys}{}}}
\subsection{Method \code{get_keys()}}{
Get primary keys of dataset
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TealDataset$get_keys()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
(\code{character} vector) with dataset primary keys
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_join_keys"></a>}}
\if{latex}{\out{\hypertarget{method-get_join_keys}{}}}
\subsection{Method \code{get_join_keys()}}{
Get \code{JoinKeys} object with keys used for joining.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TealDataset$get_join_keys()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
(\code{JoinKeys})
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_hash"></a>}}
\if{latex}{\out{\hypertarget{method-get_hash}{}}}
\subsection{Method \code{get_hash()}}{
Returns the string representation of the raw data hashed with the MD5 hash algorithm.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TealDataset$get_hash()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{character} the hash of the raw data
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_var_r6"></a>}}
\if{latex}{\out{\hypertarget{method-get_var_r6}{}}}
\subsection{Method \code{get_var_r6()}}{
Get the list of dependencies that are \code{TealDataset} or \code{TealDatasetConnector} objects
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TealDataset$get_var_r6()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{list}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-reassign_datasets_vars"></a>}}
\if{latex}{\out{\hypertarget{method-reassign_datasets_vars}{}}}
\subsection{Method \code{reassign_datasets_vars()}}{
Overwrites \code{TealDataset} or \code{TealDatasetConnector} dependencies of this \code{TealDataset} with
those found in \code{datasets}. Reassignment
refers only to the provided \code{datasets}, other \code{vars} remains the same.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TealDataset$reassign_datasets_vars(datasets)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{datasets}}{(\verb{named list} of \code{TealDataset(s)} or \code{TealDatasetConnector(s)})\cr
objects with valid pointers.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Reassign \code{vars} in this object to keep references up to date after deep clone.
Update is done based on the objects passed in \code{datasets} argument.
Overwrites dependencies with names matching the names of the objects passed
in \code{datasets}.
}

\subsection{Returns}{
NULL invisible
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{test_dataset <- teal:::TealDataset$new(
  dataname = "iris",
  x = iris,
  vars = list(dep = teal:::TealDataset$new("iris2", iris))
)
test_dataset$reassign_datasets_vars(
  list(iris2 = teal:::TealDataset$new("iris2", head(iris)))
)

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_dataset_label"></a>}}
\if{latex}{\out{\hypertarget{method-set_dataset_label}{}}}
\subsection{Method \code{set_dataset_label()}}{
Set the label for the dataset
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TealDataset$set_dataset_label(label)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{label}}{(\code{character})\cr
Label to describe the dataset}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
(\code{self}) invisibly for chaining
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_keys"></a>}}
\if{latex}{\out{\hypertarget{method-set_keys}{}}}
\subsection{Method \code{set_keys()}}{
Set new keys
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TealDataset$set_keys(keys)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{keys}}{optional, (\code{character})\cr
Vector with primary keys}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
(\code{self}) invisibly for chaining.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_join_keys"></a>}}
\if{latex}{\out{\hypertarget{method-set_join_keys}{}}}
\subsection{Method \code{set_join_keys()}}{
set join_keys for a given dataset and object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TealDataset$set_join_keys(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{\code{list} of \code{JoinKeySet} objects (which are created using the \code{join_key} function)
or single \code{JoinKeySet} objects}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
(\code{self}) invisibly for chaining
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-merge_join_keys"></a>}}
\if{latex}{\out{\hypertarget{method-merge_join_keys}{}}}
\subsection{Method \code{merge_join_keys()}}{
merge input join key with join key inside of object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TealDataset$merge_join_keys(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{\code{list} of \code{JoinKeys} objects or single \code{JoinKeys} object}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
(\code{self}) invisibly for chaining
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-mutate_join_keys"></a>}}
\if{latex}{\out{\hypertarget{method-mutate_join_keys}{}}}
\subsection{Method \code{mutate_join_keys()}}{
mutate the join_keys for a given dataset and object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TealDataset$mutate_join_keys(dataset, val)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dataset}}{(\code{character}) dataset for which join_keys are to be set against self}

\item{\code{val}}{(named \code{character}) column names used to join}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
(\code{self}) invisibly for chaining
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_vars"></a>}}
\if{latex}{\out{\hypertarget{method-set_vars}{}}}
\subsection{Method \code{set_vars()}}{
Adds variables which code depends on
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TealDataset$set_vars(vars)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{vars}}{(\verb{named list}) contains any R object which code depends on}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
(\code{self}) invisibly for chaining
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_code"></a>}}
\if{latex}{\out{\hypertarget{method-set_code}{}}}
\subsection{Method \code{set_code()}}{
Sets reproducible code
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TealDataset$set_code(code)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{code}}{(\code{character})\cr
A character string defining the code needed to produce the data set in \code{x}.
\code{initialize()} and \code{recreate()} accept code as \code{CodeClass} also
which is needed to preserve the code uniqueness and correct order.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
(\code{self}) invisibly for chaining
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_code"></a>}}
\if{latex}{\out{\hypertarget{method-get_code}{}}}
\subsection{Method \code{get_code()}}{
Get code to get data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TealDataset$get_code(deparse = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deparse}}{(\code{logical}) whether return deparsed form of a call}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
optionally deparsed \code{call} object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_code_class"></a>}}
\if{latex}{\out{\hypertarget{method-get_code_class}{}}}
\subsection{Method \code{get_code_class()}}{
Get internal \code{CodeClass} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TealDataset$get_code_class(nodeps = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{nodeps}}{(\code{logical(1)}) whether \code{CodeClass} should not contain the code
of the dependent \code{vars}
the \code{mutate}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{CodeClass}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_mutate_code_class"></a>}}
\if{latex}{\out{\hypertarget{method-get_mutate_code_class}{}}}
\subsection{Method \code{get_mutate_code_class()}}{
Get internal \code{CodeClass} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TealDataset$get_mutate_code_class()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{CodeClass}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_vars"></a>}}
\if{latex}{\out{\hypertarget{method-get_vars}{}}}
\subsection{Method \code{get_vars()}}{
Get internal \code{vars} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TealDataset$get_vars()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{list}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_mutate_vars"></a>}}
\if{latex}{\out{\hypertarget{method-get_mutate_vars}{}}}
\subsection{Method \code{get_mutate_vars()}}{
Get internal \code{mutate_vars} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TealDataset$get_mutate_vars()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{list}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-is_mutate_delayed"></a>}}
\if{latex}{\out{\hypertarget{method-is_mutate_delayed}{}}}
\subsection{Method \code{is_mutate_delayed()}}{
Whether mutate code has delayed evaluation.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TealDataset$is_mutate_delayed()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{logical}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-mutate"></a>}}
\if{latex}{\out{\hypertarget{method-mutate}{}}}
\subsection{Method \code{mutate()}}{
Mutate dataset by code
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TealDataset$mutate(code, vars = list(), force_delay = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{code}}{(\code{CodeClass}) or (\code{character}) R expressions to be executed}

\item{\code{vars}}{a named list of R objects that \code{code} depends on to execute}

\item{\code{force_delay}}{(\code{logical}) used by the containing \code{TealDatasetConnector} object

Either code or script must be provided, but not both.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
(\code{self}) invisibly for chaining
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-check"></a>}}
\if{latex}{\out{\hypertarget{method-check}{}}}
\subsection{Method \code{check()}}{
Check to determine if the raw data is reproducible from the \code{get_code()} code.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TealDataset$check()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{TRUE} if the dataset generated from evaluating the
\code{get_code()} code is identical to the raw data, else \code{FALSE}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-check_keys"></a>}}
\if{latex}{\out{\hypertarget{method-check_keys}{}}}
\subsection{Method \code{check_keys()}}{
Check if keys has been specified correctly for dataset. Set of \code{keys}
should distinguish unique rows or be \code{character(0)}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TealDataset$check_keys(keys = private$.keys)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{keys}}{optional, (\code{character})\cr
Vector with primary keys}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{TRUE} if dataset has been already pulled, else \code{FALSE}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-is_pulled"></a>}}
\if{latex}{\out{\hypertarget{method-is_pulled}{}}}
\subsection{Method \code{is_pulled()}}{
Check if dataset has already been pulled.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TealDataset$is_pulled()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{TRUE} if dataset has been already pulled, else \code{FALSE}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TealDataset$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
