% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DatasetConnector_constructors.R
\name{dataset_connector}
\alias{dataset_connector}
\alias{raw_dataset_connector}
\alias{named_dataset_connector}
\alias{relational_dataset_connector}
\title{Create a new \code{DatasetConnector} object}
\usage{
dataset_connector(
  dataname,
  pull_callable,
  keys = character(0),
  label = character(0),
  code = character(0),
  script = character(0),
  vars = list()
)

raw_dataset_connector(pull_callable)

named_dataset_connector(
  dataname,
  pull_callable,
  label = character(0),
  code = character(0),
  script = character(0),
  vars = list()
)

relational_dataset_connector(
  dataname,
  pull_callable,
  keys,
  label = character(0),
  code = character(0),
  script = character(0),
  vars = list()
)
}
\arguments{
\item{dataname}{(\code{character})\cr
A given name for the dataset it may not contain spaces}

\item{pull_callable}{(\code{CallableFunction})\cr
function with necessary arguments set to fetch data from connection.}

\item{keys}{optional, (\code{character})\cr
vector of dataset primary keys column names}

\item{label}{(\code{character})\cr
Label to describe the dataset.}

\item{code}{(\code{character})\cr
A character string defining code to modify \code{raw_data} from this dataset. To modify
current dataset code should contain at least one assignment to object defined in \code{dataname}
argument. For example if \code{dataname = ADSL} example code should contain
\verb{ADSL <- <some R code>}. Can't be used simultaneously with \code{script}}

\item{script}{(\code{character})\cr
Alternatively to \code{code} - location of the file containing modification code.
Can't be used simultaneously with \code{script}.}

\item{vars}{(named \code{list})) \cr
In case when this object code depends on other \code{Dataset} object(s) or
other constant value, this/these object(s) should be included as named
element(s) of the list. For example if this object code needs \code{ADSL}
object we should specify \verb{vars = list(ADSL = <adsl object>)}.
It's recommended to include \code{Dataset} or \code{DatasetConnector} objects to
the \code{vars} list to preserve reproducibility. Please note that \code{vars}
are included to this object as local \code{vars} and they cannot be modified
within another dataset.}
}
\value{
new \code{DatasetConnector} object
}
\description{
Create \code{DatasetConnector} from \code{\link{callable_function}}.

\ifelse{html}{\figure{lifecycle-defunct.svg}{options: alt='Defunct lifecycle'}}{\strong{Defunct}}

\ifelse{html}{\figure{lifecycle-soft-deprecated.svg}{options: alt='Soft-deprecated lifecycle'}}{\strong{Soft-deprecated}}

\ifelse{html}{\figure{lifecycle-soft-deprecated.svg}{options: alt='Soft-deprecated lifecycle'}}{\strong{Soft-deprecated}}
}
\details{
\ifelse{html}{\out{<a href='https://www.tidyverse.org/lifecycle/#experimental'><img src='figures/lifecycle-experimental.svg' alt='Experimental lifecycle'></a>}}{\strong{Experimental}}
}
