% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdisc_data.R
\name{dataset_file}
\alias{dataset_file}
\title{Load \code{RelationalDataset} object from a file}
\usage{
dataset_file(path, code = get_code(path))
}
\arguments{
\item{path}{(\code{character}) string giving the pathname of the file to read from.}

\item{code}{(\code{character}) reproducible code to re-create object}
}
\value{
\code{RelationalDataset} object
}
\description{
\ifelse{html}{\out{<a href='https://www.tidyverse.org/lifecycle/#experimental'><img src='figures/lifecycle-experimental.svg' alt='Experimental lifecycle'></a>}}{\strong{Experimental}}
Please note that the script has to end with a call creating desired object. The error will be raised otherwise.
}
\examples{
# simple example
file_example <- tempfile(fileext = ".R")
writeLines(
  text = c(
    "library(teal)
     library(random.cdisc.data)

     dataset(dataname = \"ADSL\",
             data = radsl(cached = TRUE),
             keys = get_cdisc_keys(\"ADSL\"),
             code = \"library(random.cdisc.data)\nADSL <- radsl(cached = TRUE)\")"
  ),
  con = file_example
)
x <- dataset_file(file_example, code = character(0))
get_code(x)

# custom code
file_example <- tempfile(fileext = ".R")
writeLines(
  text = c(
    "library(teal)

     # code>
     library(random.cdisc.data)
     adsl <- radsl(cached = TRUE)
     adsl$a1 <- 1
     adsl$a2 <- 2

     # <code
     dataset(dataname = \"ADSL\", data = adsl, keys = get_cdisc_keys(\"ADSL\"))"
  ),
  con = file_example
)
x <- dataset_file(file_example)
get_code(x)
}
