% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddl.R
\name{ddl}
\alias{ddl}
\title{DDL object}
\usage{
ddl(
  expr,
  code,
  ui = submit_button_ui,
  input_mask = list(),
  server = submit_button_server,
  join_keys = teal.data::join_keys(),
  datanames = names(join_keys$get())
)
}
\arguments{
\item{expr}{(\code{expression})\cr
Syntatically valid R code to be executed in the shiny session.
shouldn't be specified when \code{code} is specified.}

\item{code}{(\code{character}, \code{language})\cr
Object containing code to be evaluated to load data. Shouldn't be specified when \code{expr}
is specified.}

\item{ui}{(\code{shiny.tag})\cr
\code{shiny} ui module containing inputs which \code{id} correspond to the
args in the \code{code}.}

\item{input_mask}{(\code{list} named)\cr
arguments to be substituted in the \code{code}. These
argument are going to replace arguments set through
\code{ui} and \code{server}. Example use case is when app user
is asked to input a password and we'd like to skip this
input in the reproducible code. Typically users password
is substituted with \code{askpass::askpass()} call, so the
returned code is still executable but secure.}

\item{server}{(\code{function})\cr
\code{shiny} server module returning data. This server should execute
\code{code} and return a reactive data containing necessary data. To handle
evaluation and code masking process it is recommended to use \code{ddl_run}.
Package provides universal \code{username_password_server} which
runs \code{ddl_run} function, which returns \code{teal_data} object.
Details in the the example}

\item{join_keys}{(\code{JoinKeys}) or a single (\code{JoinKeySet})\cr
(optional) object with dataset column relationships used for joining.
If empty then no joins between pairs of objects}

\item{datanames}{(\code{character})\cr
Names of the objects to be created from the code evaluation.
If not specified (\code{character(0)}), all objects will be used to \code{teal_data} function
(via \code{env_list} in \code{postprocess_fun}).}
}
\description{
Object to execute custom DDL code in the shiny session
}
