% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-data-utils.R
\name{ddl_login_password}
\alias{ddl_login_password}
\title{Convenience wrapper for ddl_login_password}
\usage{
ddl_login_password(data, code, input_mask)
}
\description{
Convenience wrapper for ddl_login_password
}
