% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddl-modules.R
\name{ddl_module}
\alias{ddl_module}
\alias{ddl_server}
\alias{ddl_ui}
\title{Wrapper for \code{ui} and \code{server} in \code{ddl} object}
\usage{
ddl_server(id, x)

ddl_ui(id, x)
}
\arguments{
\item{id}{(\code{character}) \code{shiny} module id.}

\item{x}{(\code{ddl}) object.}
}
\value{
\code{shiny} module
}
\description{
Convenience wrapper for \code{ui} and \code{server} functions in \code{ddl} object.
On the \code{server} side, function calls \code{shiny} module and adjusts the arguments
to the formals of the \code{server} function.
}
\keyword{internal}
