% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddl-run.R
\name{ddl_run}
\alias{ddl_run}
\title{Run code and mask inputs}
\usage{
ddl_run(
  input = list(),
  code,
  input_mask = list(),
  join_keys = teal.data::join_keys(),
  datanames = names(join_keys$get())
)
}
\arguments{
\item{input}{(\code{list}) containing inputs to be used in the \code{code}}

\item{code}{(\code{language}) code to be executed}

\item{input_mask}{(\code{list}) containing inputs to be masked in the \code{code}}

\item{join_keys}{(\code{join_keys}) object}

\item{datanames}{(\code{character}) names of the objects to be created from the code evaluation}
}
\value{
\code{teal_data} object
}
\description{
Function runs the \code{code}, masks the \code{code} and creates \code{teal_data} object.
}
