% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddl.R
\name{ddl_run2}
\alias{ddl_run2}
\title{Run code and mask inputs}
\usage{
ddl_run2(x, input = list())
}
\arguments{
\item{x}{(\code{ddl}) object}

\item{input}{(\code{list}) input to be used in the \code{code}}
}
\value{
\code{teal_data} object
}
\description{
Function runs the \code{code}, masks the \code{code} and creates \code{teal_data} object.
}
