% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules_debugging.R
\name{debug_browser_module}
\alias{debug_browser_module}
\title{Module that calls \code{browser()} on button click}
\usage{
debug_browser_module(label = "Browser Debug Module")
}
\arguments{
\item{label}{\code{character} label of module}
}
\description{
The module presents a button that will call \code{browser()}.
This is useful as breakpoints or global variables can be set in this way.
When you are developing another module and find a function not working
as expected, without restarting the app, you can call \code{debug(your_fcn)} and
then resume execution. On the next invocation of the function, it will
debug it.
}
