% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkmate.R
\name{decorate_err_msg}
\alias{decorate_err_msg}
\title{Capture error and decorate error message.}
\usage{
decorate_err_msg(x, prefix = character(0), suffix = character(0))
}
\arguments{
\item{x}{object to evaluate}

\item{prefix}{(\code{character(1)}) prefix to error message}

\item{suffix}{(\code{character(1)}) suffix to error message}
}
\value{
\code{x} if no error, otherwise throws error with decorated message
}
\description{
Capture error and decorate error message.
}
\keyword{internal}
