% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decorate.R
\name{decorate_teal_data}
\alias{decorate_teal_data}
\title{Modify object in teal_data environment}
\usage{
decorate_teal_data(ui = function(id) NULL, server, expr)
}
\description{
Modify object in teal_data environment
}
