% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{deep_clone_r6}
\alias{deep_clone_r6}
\title{Helper function to deep copy \code{R6} object}
\usage{
deep_clone_r6(name, value)
}
\arguments{
\item{name}{(\code{character}) argument passed by \code{deep_clone} function.}

\item{value}{(any \code{R} object) argument passed by \code{deep_clone} function.}
}
\description{
When cloning an R6 object the private function
\code{deep_clone} is automatically used. To ensure a complete
clone the private function should call this function
}
\keyword{internal}
