% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teal_slices.R
\name{deep_copy_filter}
\alias{deep_copy_filter}
\title{Deep copy \code{teal_slices}}
\usage{
deep_copy_filter(filter)
}
\arguments{
\item{filter}{(\code{teal_slices})}
}
\value{
\code{teal_slices}
}
\description{
it's important to create a new copy of \code{teal_slices} when
starting a new \code{shiny} session. Otherwise, object will be shared
by multiple users as it is created in global environment before
\code{shiny} session starts.
}
\keyword{internal}
