% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-transform_module.R
\name{delayed_data}
\alias{delayed_data}
\title{\code{delayed_data} for \code{teal_data}}
\usage{
delayed_data(ui, server, ...)
}
\arguments{
\item{ui}{(\verb{function(id)}) function to create UI}

\item{server}{(\verb{function(id)}) \code{shiny} server which returns \code{teal_data} object wrapped in
\code{reactive}. \code{server} should have \code{id} argument and exactly the same formals as specified in \code{...}.}

\item{...}{(\code{any}) arguments passed to \code{server} function.}
}
\description{
Function creates object of class \code{delayed_data} which allows
\code{teal} app developer to transform freely \code{teal_data} object passed to \code{data} argument in
\code{\link[=init]{init()}}. This helps in case when app developer wants to use \code{teal} app
where \code{data} can be influenced by app user. For example, app developer can create
\code{teal} app which allows user to connect to database and then use data from this database.
}
