% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DelayedRelationalData.R
\name{DelayedRelationalData}
\alias{DelayedRelationalData}
\title{Manage multiple \code{RelationalDataConnector} and \code{RelationalDatasetConnector}}
\description{
Class manages  \code{RelationalDataConnector} and \code{RelationalDatasetConnector} objects
and aggregate them in one application. Class also decides whether to launch app
before initialize teal application.
}
\examples{
library(random.cdisc.data)
x <- rcd_cdisc_data( # RelationalDataConnector
  rcd_cdisc_dataset_connector("ADSL", radsl, cached = TRUE),
  rcd_cdisc_dataset_connector("ADLB", radlb, cached = TRUE)
)
x2 <- rcd_cdisc_data( # RelationalDataConnector
  rcd_cdisc_dataset_connector("ADRS", radrs, cached = TRUE)
)

tc <- teal:::DelayedRelationalData$new(x, x2)
\dontrun{
tc$launch()
tc$get_cdisc_data()
}
}
\section{Super class}{
\code{\link[teal:RelationalData]{teal::RelationalData}} -> \code{DelayedRelationalData}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{DelayedRelationalData$new()}}
\item \href{#method-get_datanames}{\code{DelayedRelationalData$get_datanames()}}
\item \href{#method-get_code}{\code{DelayedRelationalData$get_code()}}
\item \href{#method-get_ui}{\code{DelayedRelationalData$get_ui()}}
\item \href{#method-get_dataset_connectors}{\code{DelayedRelationalData$get_dataset_connectors()}}
\item \href{#method-get_data_connectors}{\code{DelayedRelationalData$get_data_connectors()}}
\item \href{#method-get_server}{\code{DelayedRelationalData$get_server()}}
\item \href{#method-append}{\code{DelayedRelationalData$append()}}
\item \href{#method-launch}{\code{DelayedRelationalData$launch()}}
\item \href{#method-clone}{\code{DelayedRelationalData$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="teal" data-topic="RelationalData" data-id="get_cdisc_data">}\href{../../teal/html/RelationalData.html#method-get_cdisc_data}{\code{teal::RelationalData$get_cdisc_data()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="teal" data-topic="RelationalData" data-id="get_datasets">}\href{../../teal/html/RelationalData.html#method-get_datasets}{\code{teal::RelationalData$get_datasets()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DelayedRelationalData$new(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{(\code{RelationalDataConnector}, \code{RelationalDataset},
\code{RelationalDatasetConnector}) object}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_datanames"></a>}}
\subsection{Method \code{get_datanames()}}{
Derive the names of all datasets
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DelayedRelationalData$get_datanames()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{character} vector with names
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_code"></a>}}
\subsection{Method \code{get_code()}}{
Derive the code for all datasets
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DelayedRelationalData$get_code()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{list} of \code{character} containing code
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_ui"></a>}}
\subsection{Method \code{get_ui()}}{
Get a shiny-module UI to render the necessary app to
derive \code{RelationalDataConnector} object's data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DelayedRelationalData$get_ui(id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{(\code{character}) item ID for the shiny module
Get dataset connectors.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_dataset_connectors"></a>}}
\subsection{Method \code{get_dataset_connectors()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DelayedRelationalData$get_dataset_connectors()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{list} with all \code{RelationalDatasetConnector} objects.
Get data connectors.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_data_connectors"></a>}}
\subsection{Method \code{get_data_connectors()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DelayedRelationalData$get_data_connectors()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{list} with all \code{RelationalDataConnector} objects.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_server"></a>}}
\subsection{Method \code{get_server()}}{
Get a shiny-module server to render the necessary app to
derive \code{RelationalDataConnector} object's data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DelayedRelationalData$get_server()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-append"></a>}}
\subsection{Method \code{append()}}{
Append datasets to this class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DelayedRelationalData$append(datasets)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{datasets}}{(\code{RelationalDataset}) use a already rendered dataset
and add it}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-launch"></a>}}
\subsection{Method \code{launch()}}{
Launch an app that allows to run the user-interfaces of all
\code{DataConnector} and \code{DatasetConnector} modules

This piece is mainly used for debugging.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DelayedRelationalData$launch()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DelayedRelationalData$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
