% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teal_reporter.R
\name{disable_report}
\alias{disable_report}
\title{Disable \code{teal_module}(s) report}
\usage{
disable_report(x)
}
\arguments{
\item{x}{(\code{teal_module}, \code{teal_modules})}
}
\description{
Disables "Add to report" functionality of \code{teal_module}(s).
}
