% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_code.R
\name{disable_show_r_code}
\alias{disable_show_r_code}
\title{Disable the "Show R Code" global button in the UI}
\usage{
disable_show_r_code(x)
}
\arguments{
\item{x}{(\code{teal_module}) a \code{teal_module} object.}
}
\value{
modified data object that indicates that it should not show the "Show R Code"
button in the UI.
}
\description{
Convenience function that disables the user's ability to see the code of the module.
}
\examples{
app <- init(
  data = within(teal_data(), iris <- iris),
  modules = modules(
    example_module(label = "example teal module") |> disable_show_r_code()
  )
)
if (interactive()) {
  shinyApp(app$ui, app$server)
}
}
