% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_snapshot_manager.R
\name{disassemble_slices}
\alias{disassemble_slices}
\title{Convert teal_slices and to list of lists (drop classes), while maintaining attributes.
Adds special class so that the reverse action can have assertion on argument type.}
\usage{
disassemble_slices(tss)
}
\arguments{
\item{tss}{(\code{teal_slices})}
}
\value{
Object of class \code{teal_slices_snapshot}, which is a list of the same length as \code{tss},
where each \code{teal_slice} has been converted to a list.
}
\description{
Convert teal_slices and to list of lists (drop classes), while maintaining attributes.
Adds special class so that the reverse action can have assertion on argument type.
}
\keyword{internal}
