% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_init_data.R
\name{.add_signature_to_data}
\alias{.add_signature_to_data}
\title{Adds signature protection to the \code{datanames} in the data}
\usage{
.add_signature_to_data(data)
}
\arguments{
\item{data}{(\code{teal_data})}
}
\value{
\code{teal_data} with additional code that has signature of the \code{datanames}
}
\description{
Adds signature protection to the \code{datanames} in the data
}
\keyword{internal}
