% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teal_slices.R
\name{.copy_from_namespace}
\alias{.copy_from_namespace}
\title{Function to copy functions from other namespaces}
\usage{
.copy_from_namespace(x, ns)
}
\arguments{
\item{x}{an object name (given as a character string).}

\item{ns}{a namespace, or character string giving the namespace.}
}
\description{
It differs from \code{getFromNamespace} as it returns a copy of the function
so it is a part of \code{teal} namespace. Useful when we require function from other namespace
where this function calls other functions from \code{teal} namespace (see \code{as.teal_slices}, \code{c.teal_slices}).
}
\keyword{internal}
