% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teal_slices.R
\name{.copy_to_teal}
\alias{.copy_to_teal}
\title{Copy functions to \code{teal} namespace}
\usage{
.copy_to_teal(fun)
}
\description{
Useful when we require function from other namespace where this function
calls other functions from \code{teal} namespace (see \code{as.teal_slices}, \code{c.teal_slices}).
}
\keyword{internal}
