% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_nested_tabs.R
\name{.datasets_to_data}
\alias{.datasets_to_data}
\title{Convert \code{FilteredData} to reactive list of datasets of the \code{teal_data} type.}
\usage{
.datasets_to_data(module, datasets, check = FALSE)
}
\arguments{
\item{module}{(\code{teal_module}) module where needed filters are taken from}

\item{datasets}{(\code{FilteredData}) object where needed data are taken from}

\item{check}{(\code{logical}) reproducibility check - whether to perform a check that the data creation
code included in the object definitions actually produces those objects.}
}
\value{
A \code{teal_data} object.
}
\description{
Converts \code{FilteredData} object to \code{teal_data} object containing datasets needed for a specific module.
Please note that if module needs dataset which has a parent, then parent will be also returned.
A hash per \code{dataset} is calculated internally and returned in the code.
}
\keyword{internal}
