% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_teal_data.R
\name{.fallback_on_failure}
\alias{.fallback_on_failure}
\title{Fallback on failure}
\usage{
.fallback_on_failure(reactive_input, label)
}
\arguments{
\item{reactive_input}{(\code{reactive}) The reactive to be verified.}

\item{label}{(\code{character}) Label for identifying problematic reactive inputs in logging.}
}
\value{
\code{reactive()} \code{NULL} or \code{teal_data}.
}
\description{
Function returns the reactive if the current input reactive contains a valid \code{teal_data} object.
Application: In \code{teal} we try to prevent the error from being propagated into modules and instead we replace failing
output with previous input.
}
\keyword{internal}
