% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_filter_manager.R
\name{.filter_module_slices}
\alias{.filter_module_slices}
\title{Filter module slices}
\usage{
.filter_module_slices(module_label, slices)
}
\arguments{
\item{module_label}{(\code{character(1)})
label of the \code{teal_module}}

\item{slices}{(\code{teal_slices})
object containing all filters and mapping}
}
\value{
\code{teal_slices} object containing only filters for the specified module
}
\description{
Filter module slices
}
\keyword{internal}
