% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.0_module_data.R
\name{.inject_signature_in_data}
\alias{.inject_signature_in_data}
\title{Adds signature protection to the datanames in the data}
\usage{
.inject_signature_in_data(data)
}
\arguments{
\item{data}{(\code{teal_data})}
}
\value{
\code{teal_data} with additional code that has signature of the datanames
}
\description{
Adds signature protection to the datanames in the data
}
\keyword{internal}
