% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{.log}
\alias{.log}
\title{A logging function}
\usage{
.log(..., sep = " ", type = "debug")
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
\details{
Deprecated. Use \link[teal.logger:register_logger]{teal.logger::register_logger}.
}
\keyword{internal}
