% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-data-utils.R
\name{.mask_code}
\alias{.mask_code}
\title{substitute inputs in the code}
\usage{
.mask_code(code, args)
}
\arguments{
\item{code}{(\code{language}) code to substitute}

\item{args}{(\code{list}) named list or arguments}
}
\description{
Function replaces symbols in the provided code prefixed with \verb{input$} or \verb{input[["}
by values of the \code{args} argument.
}
