% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.0_filter_panel.R
\name{.observe_active_filter_changed}
\alias{.observe_active_filter_changed}
\title{Trigger only active module when filter is changed}
\usage{
.observe_active_filter_changed(datasets, is_active, active_datanames)
}
\value{
A \code{reactiveVal} which is triggered when filter is changed and this module is selected.
}
\description{
Creates a trigger to limit reactivity between filter-panel and modules. We want to recalculate
only active (currently visible) modules. \code{trigger_data} triggers only if all conditions are met:
\itemize{
\item tab is selected (is_active)
\item when filters are changed (get_filter_expr is different than previous)
}
}
\keyword{internal}
