% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporter_previewer_module.R
\name{.set_reporter_options}
\alias{.set_reporter_options}
\title{Temporary function to server before hard deprecate report_previewer_module}
\usage{
.set_reporter_options(args)
}
\description{
Temporary function to server before hard deprecate report_previewer_module
}
