% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_filter_manager.R
\name{.slices_to_list}
\alias{.slices_to_list}
\title{\code{teal_slices} to list}
\usage{
.slices_to_list(slices)
}
\arguments{
\item{slices}{(\code{teal_slices})}
}
\value{
\code{list}
}
\description{
Wrapper on \code{as.list.teal_slices} that is used internally in \code{teal}
\code{filter_manager} always sets mapping for each module individually even if \code{!module_specific}
This means that we need to set \verb{<mapping>$global_filters} to all active filters when global
filter are used.
}
\keyword{internal}
