% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.smart_rbind}
\alias{.smart_rbind}
\title{Smart \code{rbind}}
\usage{
.smart_rbind(...)
}
\arguments{
\item{...}{(\code{data.frame})}
}
\description{
Combine \code{data.frame} objects which have different columns
}
\keyword{internal}
