% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-ddl-utils.R
\name{.substitute_code}
\alias{.substitute_code}
\title{Substitute symbols in the code}
\usage{
.substitute_code(expr, env)
}
\arguments{
\item{expr}{any syntactically valid \R expression}

\item{env}{an environment or a list object.  Defaults to the
    current evaluation environment.}
}
\description{
Function replaces symbols in the provided code by values of the \code{args} argument.
}
\keyword{internal}
