% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_validate.R
\name{.substitute_template}
\alias{.substitute_template}
\title{Custom substitute function that injects multiple lines to an expression}
\usage{
.substitute_template(template_str, module_server_body, check_calls)
}
\arguments{
\item{template_str}{(\code{character(1)}) The call in the expression to be replaced.}

\item{module_server_body}{(\code{expression}) Any syntactically valid R expression.}

\item{check_calls}{(\code{list}) A list of expressions to be injected.}
}
\value{
An expression with the \code{template_str} replaced by the \code{check_calls}.
}
\description{
It must contain the \code{template_str} on the first level of the expression.
}
\keyword{internal}
