% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_teal_navigation.R
\name{.teal_custom_nav}
\alias{.teal_custom_nav}
\title{Custom Navigation Widget for \code{teal_modules}}
\usage{
.teal_custom_nav(id, modules, modules_ui)
}
\arguments{
\item{id}{(\code{character(1)}) The ID of the navigation widget. Which returns the \code{module_id} of the active module.}

\item{modules}{(\code{list}) A \code{teal_modules} object.}

\item{modules_ui}{(\code{list}) A list of UI elements that should be linked with the \code{module_id} of the.\code{teal_modules}}
}
\description{
Custom Navigation Widget for \code{teal_modules}
}
\keyword{internal}
