% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{.teal_favicon}
\alias{.teal_favicon}
\title{The default favicon for the teal app.}
\format{
An object of class \code{character} of length 1.
}
\usage{
.teal_favicon
}
\description{
The default favicon for the teal app.
}
\keyword{internal}
