% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teal_navbar.R
\name{.teal_navbar}
\alias{.teal_navbar}
\title{Navbar in teal}
\usage{
.teal_navbar(id, nav_items, tab_content)
}
\arguments{
\item{nav_items}{(\code{tagList}) list of elements (e.g. buttons, drop-downs) which will be placed in
the navbar.}

\item{tab_content}{(\code{tagList}) list of panels which will be displayed if respective navigation
button will be clicked.}
}
\description{
Creates a navigation bar with consistent styling
\itemize{
\item \code{.teal_navbar} creates a container for modules navigation in \code{teal}
}
}
