% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{drop_landing}
\alias{drop_landing}
\title{Remove a \code{teal_module_landing} from list of \code{modules}}
\usage{
drop_landing(modules)
}
\arguments{
\item{modules}{\code{teal_modules}}
}
\value{
\code{teal_modules}
}
\description{
Remove a \code{teal_module_landing} from list of \code{modules}
}
\keyword{internal}
