% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{drop_module}
\alias{drop_module}
\title{Remove a specific class from list of \code{modules}}
\usage{
drop_module(modules, class)
}
\arguments{
\item{modules}{\code{teal_modules}}
}
\value{
\code{teal_modules}
}
\description{
Remove a specific class from list of \code{modules}
}
\keyword{internal}
