% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_spec.R
\name{dynamic_filter_spec}
\alias{dynamic_filter_spec}
\alias{dynamic_filter_spec.delayed_data}
\alias{dynamic_filter_spec.default}
\title{Data extract dynamic filter specification}
\usage{
dynamic_filter_spec(
  choices,
  selected = if (is(choices, "delayed_data")) NULL else choices[1],
  multiple = length(selected) > 1,
  label = NULL,
  drop_keys = FALSE
)

\method{dynamic_filter_spec}{delayed_data}(
  choices,
  selected = NULL,
  multiple = length(selected) > 1,
  label = NULL,
  drop_keys = FALSE
)

\method{dynamic_filter_spec}{default}(
  choices,
  selected = choices[1],
  multiple = length(selected) > 1,
  label = NULL,
  drop_keys = FALSE
)
}
\arguments{
\item{choices}{(\code{character} or \code{delayed_data}) \cr
vector of dataset column names available to build dynamic filter
\code{delayed_data} objects can be created via \code{\link{variable_choices}}.}

\item{selected}{(\code{NULL} or named \code{character}) \cr
selected column name out from \code{choices}.}

\item{multiple}{(\code{logical}) \cr
allow multiple filter value selection.}

\item{label}{optional (\code{character}). Define a label
on top of this specific shiny \code{\link[shiny]{selectInput}}.}

\item{drop_keys}{optional, (\code{logical}) whether to drop filter column from the dataset keys,
\code{TRUE} on default}
}
\value{
\code{dynamic_filter_spec} or \code{delayed_dynamic_filter_spec} S3-class object.
}
\description{
\ifelse{html}{\out{<a href='https://www.tidyverse.org/lifecycle/#experimental'><img src='figures/lifecycle-experimental.svg' alt='Experimental lifecycle'></a>}}{\strong{Experimental}}
Set the variables from which dynamically build a filter.
}
\examples{
dynamic_filter_spec(
  choices = c("PARAMCD", "AVISIT"),
  selected = "PARAMCD",
  multiple = TRUE
)

library(random.cdisc.data)
ADRS <- radrs(cached = TRUE)
dynamic_filter_spec(
  choices = variable_choices(ADRS),
  selected = "PARAMCD",
  multiple = TRUE
)

dynamic_filter_spec(
  choices = variable_choices("ADRS"),
  selected = "PARAMCD",
  multiple = TRUE
)
}
\seealso{
filter_spec
}
