% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_code.R
\name{enclosed_with_dataname}
\alias{enclosed_with_dataname}
\title{Get code enclosed within}
\usage{
enclosed_with_dataname(lines, dataname = NULL)
}
\arguments{
\item{lines}{(\code{character}) of preprocessing code.}

\item{dataname}{(\code{character}) metadata for returned lines}
}
\value{
(\code{list}) list of lines and their numbers from certain chunks of code at the specific file.
}
\description{
Extracts lines from code which are enclosed within regexp starts_at and stops_at
}
\keyword{internal}
