% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teal_data_module-eval_code.R
\name{eval_code}
\alias{eval_code}
\alias{eval_code,teal_data_module,character-method}
\alias{eval_code,teal_data_module,language-method}
\alias{eval_code,teal_data_module,expression-method}
\title{Evaluate the code in the qenv environment}
\usage{
\S4method{eval_code}{teal_data_module,character}(object, code)

\S4method{eval_code}{teal_data_module,language}(object, code)

\S4method{eval_code}{teal_data_module,expression}(object, code)
}
\arguments{
\item{object}{(\code{teal_data_module})}

\item{code}{(\code{character} or \code{language}) code to evaluate. Also accepts and stores comments}
}
\value{
Returns a \code{teal_data_module} object.
}
\description{
Given code is evaluated in the \code{qenv} environment of \code{teal_data} reactive defined in \code{teal_data_module}.
}
\examples{
tdm <- teal_data_module(
  ui = function(id) div(id = shiny::NS(id)("div_id")),
  server = function(id) {
    shiny::moduleServer(id, function(input, output, session) {
      shiny::reactive(teal_data(IRIS = iris))
    })
  }
)
eval_code(tdm, "IRIS <- subset(IRIS, Species == 'virginica')")
}
