% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummy_functions.R
\name{example_cdisc_data}
\alias{example_cdisc_data}
\title{Get dummy \code{CDISC} data}
\usage{
example_cdisc_data()
}
\value{
\code{cdisc_data}
}
\description{
Get dummy \code{CDISC} data including \code{ADSL}, \code{ADAE} and \code{ADLB}.
Some NAs are also introduced to stress test.
}
\keyword{internal}
