% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummy_functions.R
\name{example_filter}
\alias{example_filter}
\title{Get dummy filter states to apply initially}
\usage{
example_filter(data)
}
\arguments{
\item{data}{(\code{TealData})}
}
\value{
dummy filter states
}
\description{
This can be used for the argument \code{filter} in \code{\link{srv_teal}}.
}
\keyword{internal}
