% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_module.R
\name{example_module}
\alias{example_module}
\title{An example \code{teal} module}
\usage{
example_module(label = "example teal module")
}
\arguments{
\item{label}{\code{character}, the label of the module}
}
\value{
A \code{teal} module which can be included in the \code{modules} argument to \code{\link[=init]{init()}}.
}
\description{
An example \code{teal} module
}
\examples{
app <- init(
  data = teal_data(
    dataset("IRIS", iris),
    dataset("MTCARS", mtcars)
  ),
  modules = modules(example_module())
)
\dontrun{
shinyApp(app$ui, app$server)
}
}
