% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_relational.R
\name{execute_script_code}
\alias{execute_script_code}
\title{Evaluate script code to modify data}
\usage{
execute_script_code(x, code, vars = list())
}
\arguments{
\item{x}{(\code{RawDataset} or \code{RawDatasetConnector}) object}

\item{code}{(\code{character})\cr
A character string defining the code needed to produce the data set in \code{x}}

\item{vars}{(named \code{list}) additional pre-requisite vars to execute code}
}
\value{
(\code{environment}) which stores modified \code{x}
}
\description{
Evaluate script code to modify data
}
