% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_choices_labels}
\alias{extract_choices_labels}
\title{Extract labels from choices basing on attributes and names}
\usage{
extract_choices_labels(choices, values = NULL)
}
\arguments{
\item{choices}{(\code{list} or \code{vector}) select choices}

\item{values}{optional, choices subset for which labels should be extracted, \code{NULL} for all choices}
}
\value{
(\code{character}) vector with labels
}
\description{
Extract labels from choices basing on attributes and names
}
\keyword{internal}
