% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{extract_landing_module}
\alias{extract_landing_module}
\title{Extract landing popup module}
\usage{
extract_landing_module(modules)
}
\arguments{
\item{modules}{(\code{teal_modules})}
}
\description{
Extract landing popup module
}
\keyword{internal}
