% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{extract_module}
\alias{extract_module}
\title{Extract specific class from a list of \code{modules}}
\usage{
extract_module(modules, class)
}
\arguments{
\item{modules}{\code{teal_modules}}
}
\value{
\code{teal_module_landing}
}
\description{
Extract specific class from a list of \code{modules}
}
\keyword{internal}
