% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{extract_reporter_module}
\alias{extract_reporter_module}
\title{Extract report previewer module}
\usage{
extract_reporter_module(modules)
}
\arguments{
\item{modules}{(\code{teal_modules})}
}
\description{
Extract report previewer module
}
\keyword{internal}
