% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teal_transform_module.R
\name{extract_transformators}
\alias{extract_transformators}
\title{Extract all \code{transformators} from \code{modules}.}
\usage{
extract_transformators(modules)
}
\arguments{
\item{modules}{\code{teal_modules} or \code{teal_module}}
}
\value{
A list of \code{teal_transform_module} nested in the same way as input \code{modules}.
}
\description{
Extract all \code{transformators} from \code{modules}.
}
\keyword{internal}
