% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_extract.R
\name{extracted_data}
\alias{extracted_data}
\title{Extract a single data set from multiple outcomes of data_extractor}
\usage{
extracted_data(...)
}
\arguments{
\item{...}{\code{data.frame} Each input to this function shall be a data.frame that
has the attributes: \code{keys}, \code{dataname}}
}
\description{
This function builds a joined data set of all the data sets coming
 being pasted into it. All columns that are non key columns get renamed into
 \code{<COLUMNNAME>.<DATANAME>} due to the data set they were coming from.
}
