% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules_debugging.R
\name{filter_calls_module}
\alias{filter_calls_module}
\title{Dummy module to show the filter calls generated by the right encoding panel}
\usage{
filter_calls_module(label = "Filter Calls Module", active_datanames = "all")
}
\arguments{
\item{label}{\code{character} label of module}

\item{active_datanames}{\verb{character vector} datanames shown in filter panel;
can be \code{"all"} to mean all available datasets}
}
\description{
The \verb{Show R Code} functionality is not available here because
it is only defined in \code{teal.devel}.
}
\details{
Please do not remove, this is useful for debugging teal without
dependencies and simplifies \code{devtools::load_all} which otherwise fails
and avoids session restarts!
}
\examples{
library(random.cdisc.data)
ADSL <- radsl(cached = TRUE)
# by testing with NA values, we will see whether the filtering really works when
# we add and remove filters
ADSL$SEX[1:150] <- NA
\dontrun{
app <- init(
  data = cdisc_data(
    cdisc_dataset(
      dataname = "ADSL",
      data = ADSL
    ),
    code = "ADSL <- radsl(cached = TRUE)"
  ),
  modules = root_modules(
    teal:::filter_calls_module()
  ),
  header = "Simple teal app"
)
shinyApp(app$ui, app$server)
}
}
