% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_filter_manager.R
\name{filter_manager_module_srv}
\alias{filter_manager_module_srv}
\title{Module specific filter manager}
\usage{
filter_manager_module_srv(
  module_name,
  module_fd,
  slices_map_module,
  slices_global
)
}
\arguments{
\item{module_fd}{(\code{FilteredData})\cr
object to filter data in the teal-module}

\item{slices_global}{(\code{reactiveVal} or \code{teal_slices})\cr
stores a list of all available filters which can be utilized in several ways, for example:
\itemize{
\item to disable/enable specific filter in the module
\item to restore filter saved settings
\item to save current filter settings panel
}}

\item{id}{(\code{character(1)})\cr
\code{shiny} module id.}

\item{slices_map_modul}{(\code{reactiveVal} of \code{character})\cr
\code{id} of the \code{teal_slice} objects used in the module specific \code{FilteredData}.}
}
\value{
shiny module returning NULL
}
\description{
This module compares filters between single \code{FilteredData} settings
and global \code{teal_slices}. Updates appropriate objects \code{module_fd},
\code{slices_map_module}, \code{slices_global} to keep them constistent.
}
\keyword{internal}
