% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_spec.R
\name{filter_spec}
\alias{filter_spec}
\title{filter spec}
\usage{
filter_spec(vars, choices, selected, multiple, label = "Filter",
  sep = " - ")
}
\arguments{
\item{vars}{(\code{character}) Character vector giving the columns to be filtered}

\item{choices}{(\code{character}) Named character vector to define the choices
    of a shiny select input. These shall be filter values of the \code{vars} input
    separated by \code{sep}.

E.g. \code{vars = c("PARAMCD","AVISITN")} and \code{choices = c("CRP - BASELINE","ALT - BASELINE")}
 will lead to a filtering of
 \code{(PARAMCD == "CRP" & AVISITN == "BASELINE") | (PARAMCD == "ALT" & AVISITN == "BASELINE")}.

Please make sure the order is right. The \code{sep} input has to be \code{" - "} in this case.}

\item{selected}{(\code{character}) Named character vector to define the selected
values of a shiny select input (default values). Please check the \code{choices} description for further
details}

\item{multiple}{(\code{logical}) Whether multiple values shall be allowed in the
shiny select input.}

\item{label}{(\code{character}) Label of the Key Filtering Input inside the shiny app.}

\item{sep}{(\code{character}) A separator string to split the \code{choices} or
\code{selected} inputs into the values of the different columns}
}
\value{
\code{filter_choices_spec}-S3-class object
}
\description{
It consists in choices and additionally the variable names for the choices
}
\examples{
filter_spec(
  vars = c("PARAMCD", "AVISIT"),
  sep = " - ",
  choices = c("CRP - BASELINE", "CRP - SCREENING", "ALT - BASELINE"),
  selected = c("CRP - BASELINE"),
  multiple = TRUE
)
}
