% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilteredData.R
\name{filter_state_to_str}
\alias{filter_state_to_str}
\title{Print the state in a nice format}
\usage{
filter_state_to_str(var_type, state, add_na = FALSE)
}
\arguments{
\item{var_type}{\code{character} variable type}

\item{state}{\code{list} with elements corresponding to variable
type; can be \code{filter_state} or \code{filter_info}}

\item{add_na}{\code{logical} whether to display information about NA,
either \code{keep_na} (for \code{filter_state}) or \code{na_count} (for \code{filter_info})}
}
\description{
The \code{keep_na} is not printed.
}
