% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtered_data_wrappers.R
\name{filtered_data_new}
\alias{filtered_data_new}
\alias{filtered_data_new.RelationalData}
\alias{filtered_data_new.CDISCData}
\title{Create a new \code{FilteredData} object}
\usage{
filtered_data_new(x)

\method{filtered_data_new}{RelationalData}(x)

\method{filtered_data_new}{CDISCData}(x)
}
\arguments{
\item{x}{an object that inherits from \code{RelationalData}}
}
\value{
a (\code{CDISCDataset}, \code{CDISCDatasetConnector}) object
}
\description{
\ifelse{html}{\out{<a href='https://www.tidyverse.org/lifecycle/#experimental'><img src='figures/lifecycle-experimental.svg' alt='Experimental lifecycle'></a>}}{\strong{Experimental}}
}
