% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtered_data_wrappers.R
\name{filtered_data_set}
\alias{filtered_data_set}
\alias{filtered_data_set.RelationalData}
\title{Set \code{FilteredData} with data from \code{RelationalData}}
\usage{
filtered_data_set(data, datasets)

\method{filtered_data_set}{RelationalData}(data, datasets)
}
\arguments{
\item{data}{an object that inherits from \code{RelationalData}}

\item{datasets}{an object that inherits from \code{FilteredData}}
}
\value{
modified \code{FilteredData} object
}
\description{
\ifelse{html}{\out{<a href='https://www.tidyverse.org/lifecycle/#experimental'><img src='figures/lifecycle-experimental.svg' alt='Experimental lifecycle'></a>}}{\strong{Experimental}}
}
