% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtered_data_wrappers.R
\name{filtered_data_set_filters}
\alias{filtered_data_set_filters}
\title{Set up \code{FilteredData} filters}
\usage{
filtered_data_set_filters(datasets, filter)
}
\arguments{
\item{datasets}{an object that inherits from \code{FilteredData}}

\item{filter}{(named \code{list}) of filter entries}
}
\description{
Set up \code{FilteredData} filters
}
