% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilteredDataset.R
\name{FilteredDataset}
\alias{FilteredDataset}
\title{\code{FilterStates} R6 class}
\description{
\code{FilteredDataset} is a class which renders/controls \code{FilterStates}(s) for \code{Dataset}.
Each \code{FilteredDataset} contains \code{filter_states} field - a \code{list} which contains one
(\code{data.frame}) or multiple (\code{MultiAssayExperiment}) \code{FilterStates} objects.
Each \code{FilterStates} is responsible for one filter/subset expression applied for specific
components of the \code{Dataset}.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{FilteredDataset$new()}}
\item \href{#method-add_to_eval_env}{\code{FilteredDataset$add_to_eval_env()}}
\item \href{#method-queues_empty}{\code{FilteredDataset$queues_empty()}}
\item \href{#method-get_call}{\code{FilteredDataset$get_call()}}
\item \href{#method-get_data}{\code{FilteredDataset$get_data()}}
\item \href{#method-get_data_reactive}{\code{FilteredDataset$get_data_reactive()}}
\item \href{#method-get_filter_states}{\code{FilteredDataset$get_filter_states()}}
\item \href{#method-get_data_info}{\code{FilteredDataset$get_data_info()}}
\item \href{#method-get_subjects_info}{\code{FilteredDataset$get_subjects_info()}}
\item \href{#method-get_dataname}{\code{FilteredDataset$get_dataname()}}
\item \href{#method-get_dataset}{\code{FilteredDataset$get_dataset()}}
\item \href{#method-get_hash}{\code{FilteredDataset$get_hash()}}
\item \href{#method-get_keys}{\code{FilteredDataset$get_keys()}}
\item \href{#method-get_join_keys}{\code{FilteredDataset$get_join_keys()}}
\item \href{#method-get_varlabels}{\code{FilteredDataset$get_varlabels()}}
\item \href{#method-get_varnames}{\code{FilteredDataset$get_varnames()}}
\item \href{#method-get_filtered_dataname}{\code{FilteredDataset$get_filtered_dataname()}}
\item \href{#method-ui}{\code{FilteredDataset$ui()}}
\item \href{#method-server}{\code{FilteredDataset$server()}}
\item \href{#method-ui_add_filter_state}{\code{FilteredDataset$ui_add_filter_state()}}
\item \href{#method-srv_add_filter_state}{\code{FilteredDataset$srv_add_filter_state()}}
\item \href{#method-clone}{\code{FilteredDataset$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initialize \code{FilteredDataset} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FilteredDataset$new(dataset, join_keys)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dataset}}{(\code{Dataset})\cr
single dataset for which filters are rendered}

\item{\code{join_keys}}{(\code{list})\cr
keys to join this \code{dataset} with the other}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_to_eval_env"></a>}}
\if{latex}{\out{\hypertarget{method-add_to_eval_env}{}}}
\subsection{Method \code{add_to_eval_env()}}{
Add objects to the filter call evaluation environment
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FilteredDataset$add_to_eval_env(name, value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{(\code{character}) object name}

\item{\code{value}}{object value}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-queues_empty"></a>}}
\if{latex}{\out{\hypertarget{method-queues_empty}{}}}
\subsection{Method \code{queues_empty()}}{
Removes all active filter items applied to this dataset
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FilteredDataset$queues_empty()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_call"></a>}}
\if{latex}{\out{\hypertarget{method-get_call}{}}}
\subsection{Method \code{get_call()}}{
Get filter expression

This functions returns filter calls equivalent to selected items
within each of \code{filter_states}. Configuration of the calls is constant and
depends on \code{filter_states} type and order which are set during initialization.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FilteredDataset$get_call()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
filter \code{call} or \code{list} of filter calls
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_data"></a>}}
\if{latex}{\out{\hypertarget{method-get_data}{}}}
\subsection{Method \code{get_data()}}{
Get raw data of this dataset
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FilteredDataset$get_data(filtered)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{filtered}}{(\code{logical(1)})\cr
whether returned data should be filtered or not}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
type of returned object depending on a data stored in
\code{Dataset}. Currently \code{data.frame} or \code{MultiAssayExperiment}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_data_reactive"></a>}}
\if{latex}{\out{\hypertarget{method-get_data_reactive}{}}}
\subsection{Method \code{get_data_reactive()}}{
Get reactive object which returns filtered data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FilteredDataset$get_data_reactive()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
(\code{reactive})
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_filter_states"></a>}}
\if{latex}{\out{\hypertarget{method-get_filter_states}{}}}
\subsection{Method \code{get_filter_states()}}{
Get filter states
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FilteredDataset$get_filter_states(id = character(0))}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{(\code{character(1)}, \code{character(0)})\cr
id of the \code{private$filter_states} list element where \code{FilterStates} is kept.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{FilterStates} or \code{list} of \code{FilterStates} objects.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_data_info"></a>}}
\if{latex}{\out{\hypertarget{method-get_data_info}{}}}
\subsection{Method \code{get_data_info()}}{
Get data info
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FilteredDataset$get_data_info(filtered = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{filtered}}{(\code{logical(1)})\cr
whether \code{data.info} should depend on filtered data.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{integer(1)} number of rows
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_subjects_info"></a>}}
\if{latex}{\out{\hypertarget{method-get_subjects_info}{}}}
\subsection{Method \code{get_subjects_info()}}{
Get subjects info
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FilteredDataset$get_subjects_info(filtered)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{filtered}}{(\code{logical(1)})\cr
whether \code{data.info} should depend on filtered data.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{integer} number of unique subjects
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_dataname"></a>}}
\if{latex}{\out{\hypertarget{method-get_dataname}{}}}
\subsection{Method \code{get_dataname()}}{
Get name of the dataset

Get name of the dataset
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FilteredDataset$get_dataname()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{character(1)} as a name of this dataset
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_dataset"></a>}}
\if{latex}{\out{\hypertarget{method-get_dataset}{}}}
\subsection{Method \code{get_dataset()}}{
Get dataset
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FilteredDataset$get_dataset()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{Dataset}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_hash"></a>}}
\if{latex}{\out{\hypertarget{method-get_hash}{}}}
\subsection{Method \code{get_hash()}}{
Returns the hash of the unfiltered dataset
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FilteredDataset$get_hash()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
(\code{character(1)})
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_keys"></a>}}
\if{latex}{\out{\hypertarget{method-get_keys}{}}}
\subsection{Method \code{get_keys()}}{
Get keys for the dataset
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FilteredDataset$get_keys()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
(\code{character}) keys of dataset
Get join keys to join this dataset with others
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_join_keys"></a>}}
\if{latex}{\out{\hypertarget{method-get_join_keys}{}}}
\subsection{Method \code{get_join_keys()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FilteredDataset$get_join_keys()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{list} of keys
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_varlabels"></a>}}
\if{latex}{\out{\hypertarget{method-get_varlabels}{}}}
\subsection{Method \code{get_varlabels()}}{
Get labels of variables in the data

Variables are the column names of the data.
Either, all labels must have been provided for all variables
in \code{set_data} or \code{NULL}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FilteredDataset$get_varlabels(variables = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{variables}}{(\code{character} vector) variables to get labels for;
if \code{NULL}, for all variables in data}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
(\code{character} or \code{NULL}) variable labels, \code{NULL} if \code{column_labels}
attribute does not exist for the data
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_varnames"></a>}}
\if{latex}{\out{\hypertarget{method-get_varnames}{}}}
\subsection{Method \code{get_varnames()}}{
Get variable names from dataset
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FilteredDataset$get_varnames()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{character}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_filtered_dataname"></a>}}
\if{latex}{\out{\hypertarget{method-get_filtered_dataname}{}}}
\subsection{Method \code{get_filtered_dataname()}}{
Get suffixed dataname
Used when filtering the data to get \verb{<dataname>_FILTERED},
\verb{<dataname>_FILTERED_ALONE} or any other name.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FilteredDataset$get_filtered_dataname(
  dataname = self$get_dataname(),
  suffix = "_FILTERED"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dataname}}{(\code{character(1)}) dataname}

\item{\code{suffix}}{(\code{character(1)}) string to be putted after dataname}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{character(1)}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ui"></a>}}
\if{latex}{\out{\hypertarget{method-ui}{}}}
\subsection{Method \code{ui()}}{
UI module for dataset active filters

UI module containing dataset active filters along with
title and remove button.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FilteredDataset$ui(id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{(\code{character(1)})\cr
identifier of the element - preferably containing dataset name}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
function - shiny UI module
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-server"></a>}}
\if{latex}{\out{\hypertarget{method-server}{}}}
\subsection{Method \code{server()}}{
Server module for a dataset active filters

Server module managing a  active filters.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FilteredDataset$server(input, output, session)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{(\code{shiny})\cr}

\item{\code{output}}{(\code{shiny})\cr}

\item{\code{session}}{(\code{shiny})\cr
single dataset for which filters are rendered}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
function - shiny server module
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ui_add_filter_state"></a>}}
\if{latex}{\out{\hypertarget{method-ui_add_filter_state}{}}}
\subsection{Method \code{ui_add_filter_state()}}{
UI module to add filter variable for this dataset

UI module to add filter variable for this dataset
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FilteredDataset$ui_add_filter_state(id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{(\code{character(1)})\cr
identifier of the element - preferably containing dataset name}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
function - shiny UI module
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-srv_add_filter_state"></a>}}
\if{latex}{\out{\hypertarget{method-srv_add_filter_state}{}}}
\subsection{Method \code{srv_add_filter_state()}}{
Server module to add filter variable for this dataset

Server module to add filter variable for this dataset
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FilteredDataset$srv_add_filter_state(input, output, session)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{(\code{shiny})\cr}

\item{\code{output}}{(\code{shiny})\cr}

\item{\code{session}}{(\code{shiny})\cr}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
function - shiny server module
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FilteredDataset$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
