% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_code.R
\name{find_source_code}
\alias{find_source_code}
\title{Finds code with source call}
\usage{
find_source_code(code)
}
\arguments{
\item{code}{(\code{character}) containing preprocessing code lines.}
}
\description{
Finds code where \code{source()} call is located
}
