% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_code.R
\name{find_source_lines}
\alias{find_source_lines}
\title{Finds lines with source call}
\usage{
find_source_lines(lines)
}
\arguments{
\item{lines}{(\code{character}) of code as separate element in vector}
}
\description{
Finds lines where \code{source()} call is located
}
