% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rcode_utils.R
\name{fold_lines}
\alias{fold_lines}
\title{Fixed line width folding}
\usage{
fold_lines(txt, width = 80, indent_from = NULL)
}
\arguments{
\item{txt}{(\code{character}) text to be adjusted}

\item{width}{(\code{integer}) maximum number of characters in vector}

\item{indent_from}{(\code{character}) character which begins the indent.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
Folds lines longer than specified width.
}
\keyword{internal}
