% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RelationalDatasetConnector_constructors.R
\name{fun_cdisc_dataset_connector}
\alias{fun_cdisc_dataset_connector}
\title{Function \code{CDISC} Dataset Connector}
\usage{
fun_cdisc_dataset_connector(
  dataname,
  func,
  func_args = NULL,
  label = character(0),
  code = character(0),
  script = character(0),
  ...
)
}
\arguments{
\item{dataname}{(\code{character})\cr
A given name for the dataset it may not contain spaces}

\item{func}{(\code{function})\cr
a custom function to obtain dataset.}

\item{func_args}{(\code{list})\cr
additional arguments for (\code{func})\cr.}

\item{label}{(\code{character})\cr
Label to describe the dataset.}

\item{code}{(\code{character})\cr
A character string defining code to modify \code{raw_data} from this dataset. To modify
current dataset code should contain at least one assignment to object defined in \code{dataname}
argument. For example if \code{dataname = ADSL} example code should contain
\verb{ADSL <- <some R code>}. Can't be used simultaneously with \code{script}}

\item{script}{(\code{character})\cr
Alternatively to \code{code} - location of the file containing modification code.
Can't be used simultaneously with \code{script}.}

\item{...}{(\code{optional})\cr
Additional arguments applied to pull function.
In case when this object code depends on the \code{raw_data} from the other
\code{NamedDataset}, \code{NamedDatasetConnector} object(s) or other constant value,
this/these object(s) should be included. Please note that \code{vars}
are included to this object as local \code{vars} and they cannot be modified
within another dataset.}
}
\description{
\ifelse{html}{\out{<a href='https://www.tidyverse.org/lifecycle/#experimental'><img src='figures/lifecycle-experimental.svg' alt='Experimental lifecycle'></a>}}{\strong{Experimental}}
Create a \code{RelationalDatasetConnector} from \code{function} and its arguments
with keys assigned automatically by \code{dataname}.
}
