% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_sample_data.R
\name{generate_sample_data}
\alias{generate_sample_data}
\title{Generate a Sample Dataset}
\usage{
generate_sample_data(type = c("asl", "ars"))
}
\arguments{
\item{type}{case insensitiv, specify which data to generate}
}
\value{
a data.frame
}
\description{
This dataset generator function is useful to produce
}
\examples{

ASL <- generate_sample_data('ASL')
}
