% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_attrs.R
\name{get_attrs}
\alias{get_attrs}
\alias{get_attrs.Dataset}
\title{Get dataset attributes}
\usage{
get_attrs(x)

\method{get_attrs}{Dataset}(x)
}
\arguments{
\item{x}{an object of (\code{Dataset}) class}
}
\value{
named \code{list} of object attributes
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
Get dataset attributes in form of named list.
}
\examples{
# Dataset --------

library(random.cdisc.data)
ADSL <- radsl(cached = TRUE)

x1 <- dataset("ADSL", x = ADSL, label = "custom label")
get_attrs(x1)

x2 <- dataset(
  "ADSL",
  x = ADSL,
  keys = get_cdisc_keys("ADSL"),
  label = "custom label"
)
get_attrs(x2)

# CDISCDataset --------

library(random.cdisc.data)
ADSL <- radsl(cached = TRUE)
x3 <- cdisc_dataset(
  "ADSL",
  x = ADSL,
  keys = get_cdisc_keys("ADSL"),
  label = "custom label"
)
get_attrs(x3)
}
