% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CallableFunction.R
\name{get_binding_name}
\alias{get_binding_name}
\title{Gets the name of the binding}
\usage{
get_binding_name(object, envir)
}
\arguments{
\item{object}{(R object)\cr
any R object}

\item{envir}{(\code{environment})\cr
if origin of the object is known then should be provided for
more precise search}
}
\value{
character
}
\description{
Gets the name of the object by finding its origin.
Depending on type of object function uses different methods
to obtain original location. If no \code{env} is specified then
object is tracked by \code{substitute} along the \code{sys.frames}.
If \code{env} is specified then search is limited to specified
environment.\cr
}
\note{
Raises an error if the object is not found in the environment.
}
\keyword{internal}
