% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cdisc_data.R
\name{get_cdisc_data}
\alias{get_cdisc_data}
\alias{get_cdisc_data.RelationalData}
\title{Get a cdisc data object.}
\usage{
get_cdisc_data(x)

\method{get_cdisc_data}{RelationalData}(x)
}
\arguments{
\item{x}{(\code{object}) of class\link{RawDatasetConnector} or \link{NamedDataset}. If of
class \code{character} will be treated as file to read.}
}
\value{
\code{cdisc_data} object
}
\description{
Get a cdisc data object.
}
\examples{

# RelationalData ----------------
library(random.cdisc.data)
adsl <- cdisc_dataset(dataname = "ADSL", # RelationalDataset
                      data = radsl(cached = TRUE),
                      code = "library(random.cdisc.data)\nADSL <- radsl(cached = TRUE)")

adtte <- cdisc_dataset(dataname = "ADTTE", # RelationalDataset
                       data = radtte(cached = TRUE),
                       code = "library(random.cdisc.data)\nADTTE <- radtte(cached = TRUE)")

rd <- teal:::RelationalData$new(adsl, adtte)
get_cdisc_data(rd)
}
