% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdisc_data.R
\name{get_cdisc_keys}
\alias{get_cdisc_keys}
\title{Function that returns the default keys for a \code{CDISC} dataset by name}
\usage{
get_cdisc_keys(dataname)
}
\arguments{
\item{dataname}{name of the \code{CDISC} dataset}
}
\value{
\code{keys} object
}
\description{
Function that returns the default keys for a \code{CDISC} dataset by name
}
\examples{
get_cdisc_keys("ADSL")
}
