% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_client_timezone}
\alias{get_client_timezone}
\title{Get client timezone}
\usage{
get_client_timezone(ns)
}
\arguments{
\item{ns}{(\code{function}) namespace function passed from the \code{session} object in the \code{shiny} server.
For \code{shiny} modules this will allow for proper name spacing of the registered input.}
}
\value{
\code{NULL}, invisibly.
}
\description{
User timezone in the browser may be different to the one on the server.
This script can be run to register a \code{shiny} input which contains information about the timezone in the browser.
}
\keyword{internal}
