% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chunks.R
\name{get_code_chunk}
\alias{get_code_chunk}
\title{Evaluate a code chunk for global chunks}
\usage{
get_code_chunk(id = NULL, unfiltered_datasets = c(),
  chunks = parent.env(parent.frame())$session$userData$chunks)
}
\arguments{
\item{id}{(\code{character}) ID given for the code chunk}

\item{unfiltered_datasets}{(\code{character}) Name of the dataset that was not filtered.}

\item{chunks}{(\code{chunks}) object of type \code{chunks} that the code chunk is added to. If not
set the default value picks the global variable \code{chunks}.}
}
\value{
The R code stored inside the chunk that can be used to reproduce the value of the evaluation
   of the code chunk.
}
\description{
Evaluate a code chunk for global chunks
}
\references{
chunks
}
