% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_code.R
\name{get_code_single}
\alias{get_code_single}
\title{Get code}
\usage{
get_code_single(file_path, starts_at, stops_at, exclude_comments,
  read_sources)
}
\arguments{
\item{file_path}{(\code{character}) path of the file to be parsed}

\item{starts_at}{(\code{character}) regular expression determining start of the block of code}

\item{stops_at}{(\code{character}) regular expression determining stop  of the block of code}

\item{exclude_comments}{(\code{logical}) whether exclude commented-out lines of code}

\item{read_sources}{(\code{logical}) whether to replace \code{source("path")} with code lines from sourced file.}
}
\value{
code (\code{character}) preprocessing code
}
\description{
Get code from specified file.
}
