% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tdata.R
\name{get_code_tdata}
\alias{get_code_tdata}
\title{Wrapper for \code{get_code.tdata}}
\usage{
get_code_tdata(data)
}
\arguments{
\item{data}{(\code{tdata}) object}
}
\value{
(\code{character}) code used in the \code{tdata} object.
}
\description{
This wrapper is to be used by downstream packages to extract the code of a \code{tdata} object.
}
