% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dataname.R
\name{get_dataname}
\alias{get_dataname}
\alias{get_dataname.DataAbstract}
\alias{get_dataname.DatasetConnector}
\alias{get_dataname.Dataset}
\title{S3 method for getting a dataname(s) of
(\code{DataAbstract}, (\code{DatasetConnector} or
\code{Dataset}) R6 object}
\usage{
get_dataname(x)

\method{get_dataname}{DataAbstract}(x)

\method{get_dataname}{DatasetConnector}(x)

\method{get_dataname}{Dataset}(x)
}
\arguments{
\item{x}{(\code{DataAbstract}, (\code{DatasetConnector} or
\code{Dataset}) R6 object}
}
\value{
dataname (\code{character}) A given name for the dataset(s)
it may not contain spaces
}
\description{
\ifelse{html}{\out{<a href='https://www.tidyverse.org/lifecycle/#experimental'><img src='figures/lifecycle-experimental.svg' alt='Experimental lifecycle'></a>}}{\strong{Experimental}}
}
