% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dataset.R
\name{get_dataset}
\alias{get_dataset}
\alias{get_dataset.RawDatasetConnector}
\title{Get dataset from \code{DatasetConnector}}
\usage{
get_dataset(x)

\method{get_dataset}{RawDatasetConnector}(x)
}
\arguments{
\item{x}{\code{RawDatasetConnector} or \code{RelationalDatasetConnector})}
}
\value{
(\code{RawDataset} or \code{RelationalDataset})
}
\description{
Get dataset from \code{DatasetConnector}
}
\examples{
library(random.cdisc.data)
fun <- callable_function(radsl)
set_args(fun, list(N = 5, seed = 1, cached = TRUE))
x <- raw_dataset_connector(fun)

load_dataset(x)
ds <- get_dataset(x)
}
