% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dataset_label.R
\name{get_dataset_label}
\alias{get_dataset_label}
\alias{get_dataset_label.DatasetConnector}
\alias{get_dataset_label.Dataset}
\title{S3 method for getting a label of
(\code{DatasetConnector} or \code{Dataset}) R6 object}
\usage{
get_dataset_label(x)

\method{get_dataset_label}{DatasetConnector}(x)

\method{get_dataset_label}{Dataset}(x)
}
\arguments{
\item{x}{(\code{DatasetConnector} or \code{Dataset}) R6 object}
}
\value{
label (\code{character}) Label to describe the dataset
}
\description{
\ifelse{html}{\out{<a href='https://www.tidyverse.org/lifecycle/#experimental'><img src='figures/lifecycle-experimental.svg' alt='Experimental lifecycle'></a>}}{\strong{Experimental}}
}
\examples{
fun <- callable_function(data.frame)
fun$set_args(list(c1 = seq_len(10)))

x <- dataset_connector(
  pull_callable = fun,
  dataname = "ADSL",
  label = "My custom label"
)
get_dataset_label(x)
library(random.cdisc.data)
ADSL <- radsl(cached = TRUE)
ADSL_dataset <- dataset(dataname = "ADSL", x = ADSL, label = "My custom label")
get_dataset_label(ADSL_dataset)
}
