% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_selected_columns.R
\name{get_dataset_prefixed_column_names}
\alias{get_dataset_prefixed_column_names}
\title{Derive non-key column names from data}
\usage{
get_dataset_prefixed_column_names(data)
}
\arguments{
\item{data}{(\code{data.frame}) Data with attribute \code{keys} and \code{dataname}}
}
\value{
A named character vector with <variable> ---> <dataname>.<variable>
}
\description{
Derive non-key column names from data
}
