% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rcode_utils.R
\name{get_datasets_code}
\alias{get_datasets_code}
\title{Get datasets code}
\usage{
get_datasets_code(datanames, datasets, hashes)
}
\arguments{
\item{datanames}{(\code{character}) names of datasets to extract code from}

\item{datasets}{(\code{FilteredData}) object}

\item{hashes}{named (\code{list}) of hashes per dataset}
}
\value{
\code{character(3)} containing following elements:
\itemize{
\item code from \code{CodeClass} (data loading code)
\item hash check of loaded objects
}
}
\description{
Get combined code from \code{FilteredData} and from \code{CodeClass} object.
}
\keyword{internal}
