% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummy_functions.R
\name{get_dummy_datasets}
\alias{get_dummy_datasets}
\title{Get a dummy \code{datasets} object with \code{ADSL} data, useful in the examples}
\usage{
get_dummy_datasets()
}
\value{
\code{FilteredData} with \code{ADSL} set
}
\description{
Returns a new \code{R6} object on each invocation, not a singleton.
}
