% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummy_functions.R
\name{get_dummy_filter}
\alias{get_dummy_filter}
\title{Get dummy filter states to apply initially}
\usage{
get_dummy_filter(data)
}
\arguments{
\item{data}{(\code{TealData})}
}
\value{
dummy filter states
}
\description{
This can be used for the argument \code{filter} in \link{srv_teal}.
}
