% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummy_functions.R
\name{get_dummy_filter_states}
\alias{get_dummy_filter_states}
\title{Get dummy filter states to apply initially}
\usage{
get_dummy_filter_states()
}
\value{
dummy filter states
}
\description{
This can be used for the argument \code{filter_states} in
\verb{\link{srv_teal}}.
}
