% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilteredData.R
\name{get_filter_expr}
\alias{get_filter_expr}
\title{Get filter expression for multiple datanames taking into account its order.}
\usage{
get_filter_expr(datasets, datanames = datasets$get_datanames())
}
\arguments{
\item{datasets}{(\code{FilteredData})}

\item{datanames}{(\code{character}) vector of dataset names}
}
\value{
(\code{expression})
}
\description{
To be used in show R code button.
}
