% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rcode_header_and_filters.R
\name{get_filter_txt}
\alias{get_filter_txt}
\title{Generates text for the code to filter datasets}
\usage{
get_filter_txt(datanames, datasets)
}
\arguments{
\item{datanames}{vector with data names}

\item{datasets}{datasets object}
}
\value{
A character string for the code to generate the filtered datasets
}
