% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilterStates.R
\name{get_filterable_varnames}
\alias{get_filterable_varnames}
\title{Gets filterable variable names}
\usage{
get_filterable_varnames(data)
}
\arguments{
\item{data}{(\code{object})\cr
the R object containing elements which class can be checked through \code{vapply} or \code{apply}.}
}
\value{
\code{character} the array of the matched element names
}
\description{
Gets filterable variable names from a given object. The names match variables
of classes in the \code{teal:::.filterable_class} enum.
}
\examples{
df <- data.frame(
  a = letters[1:3],
  b = 1:3,
  c = Sys.Date() + 1:3,
  d = Sys.time() + 1:3,
  z = complex(3)
)
teal:::get_filterable_varnames(df)
}
\keyword{internal}
