% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_key_duplicates.data.frame}
\alias{get_key_duplicates.data.frame}
\title{Creates a short information summary about the duplicate key values in a dataset.}
\usage{
\method{get_key_duplicates}{data.frame}(dataset, keys = NULL)
}
\arguments{
\item{dataset}{\code{data.frame} object}

\item{keys}{\code{character} vector of variable names in \code{dataset} consisting the key
or \code{keys} object, which does have a \code{primary} element with a vector of variable
names in \code{dataset} consisting the key. Optional, default: NULL}
}
\value{
a \code{tibble} with a short information summary
}
\description{
Creates a short information summary about the duplicate key values in a dataset.
}
\details{
When the key argument is provided the function uses it to generate the summary, otherwise
looks for \code{primary_key} attribute of the \code{dataset}. If neither are provided raises an exception.
}
\examples{
df <- as.data.frame(
  list(a = c("a", "a", "b", "b", "c"), b = c(1, 2, 3, 3, 4), c = c(1, 2, 3, 4, 5))
)
res <- get_key_duplicates(df, keys = c("a", "b")) # duplicated keys are in rows 3 and 4
print(res) # outputs a tibble
}
\seealso{
\link{get_key_duplicates} \link{get_key_duplicates_util}
}
