% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_key_duplicates}
\alias{get_key_duplicates}
\title{S3 generic for creating an information summary about the duplicate key values in a dataset}
\usage{
get_key_duplicates(dataset, keys = NULL)
}
\arguments{
\item{dataset}{\code{TealDataset} or \code{dataframe} a dataset, which will be tested}

\item{keys}{\code{character} vector of variable names in \code{dataset} consisting the key
or \code{keys} object, which does have a \code{primary} element with a vector of variable
names in \code{dataset} consisting the key. Optional, default: NULL}
}
\value{
a \code{tibble} with variables consisting the key and \code{row_no} and \code{duplicates_count} columns
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\details{
The information summary provides row numbers and number of duplicates
for each duplicated key value.
}
\examples{
library(scda)

adsl <- synthetic_cdisc_data("latest")$adsl
# create a TealDataset with default keys
rel_adsl <- cdisc_dataset("ADSL", adsl)
get_key_duplicates(rel_adsl)

df <- as.data.frame(
  list(a = c('a', 'a', 'b', 'b', 'c'), b = c(1, 2, 3, 3, 4), c = c(1, 2, 3, 4, 5))
)
res <- get_key_duplicates(df, keys = c('a', 'b')) # duplicated keys are in rows 3 and 4
print(res) # prints a tibble

\dontrun{
get_key_duplicates(df) # raises an exception, because keys are missing with no default
}

}
\seealso{
\itemize{
\item{\code{\link{get_key_duplicates_util}}}
\item{\code{\link{get_key_duplicates.TealDataset}}}
\item{\code{\link{get_key_duplicates.data.frame}}}
}
}
