% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_key_duplicates.RelationalDataset}
\alias{get_key_duplicates.RelationalDataset}
\title{Creates a short information summary about the duplicate primary key values in a dataset}
\usage{
\method{get_key_duplicates}{RelationalDataset}(dataset, keys = NULL)
}
\arguments{
\item{dataset}{a \code{RelationalDataset} object, which will be used to detect duplicated
primary keys}

\item{keys}{\code{character} vector of variable names in \code{dataset} consisting the key
or \code{keys} object, which does have a \code{primary} element with a vector of variable
names in \code{dataset} consisting the key. Optional, default: NULL}
}
\description{
\ifelse{html}{\out{<a href='https://www.tidyverse.org/lifecycle/#experimental'><img src='figures/lifecycle-experimental.svg' alt='Experimental lifecycle'></a>}}{\strong{Experimental}}
}
\details{
S3 method for get_key_duplicates. Uses the public API of
\code{teal.devel::RelationalDataset} to read the primary key and the raw data.

If \code{keys} argument is provided, then checks against that, if it's \code{NULL}, then checks
against the \code{get_keys()$primary} method of the \code{dataset} argument.
}
\examples{
library(random.cdisc.data)

adsl <- radsl(cached = TRUE)
# Keys are automatically assigned, because the name
# is in the recognized ADAM nomenclature ("ADSL")
rel_adsl <- cdisc_dataset("ADSL", adsl)
get_key_duplicates(rel_adsl)

}
\seealso{
\link{get_key_duplicates} \link{get_key_duplicates_util}
}
