% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_key_duplicates_util}
\alias{get_key_duplicates_util}
\title{Creates a duplicate keys information summary.}
\usage{
get_key_duplicates_util(dataframe, keys)
}
\arguments{
\item{dataframe}{dataframe}

\item{keys}{\code{character} list of variable names consisting the key to the \code{dataframe}}
}
\value{
\code{tibble} with a duplicate keys information summary
}
\description{
\ifelse{html}{\out{<a href='https://www.tidyverse.org/lifecycle/#experimental'><img src='figures/lifecycle-experimental.svg' alt='Experimental lifecycle'></a>}}{\strong{Experimental}}
}
\details{
Accepts a list of variable names - \code{keys}, which are treated as the
key to the \code{dataframe} argument. An instance of duplicated key is
defined as two rows, which have the same values in columns defined by \code{keys}.
Per each key value with duplicates returns a row in a \code{tibble}. The return table
has columns corresponding to the variable names passed in \code{keys} and
two additional columns: \code{row_no} and \code{duplicates_count}, which provide
information about row numbers of the original dataframe, which contain duplicated keys
and total duplicates counts.
}
\examples{
df <- as.data.frame(
  list(a = c("a", "a", "b", "b", "c"), b = c(1, 2, 3, 3, 4), c = c(1, 2, 3, 4, 5))
)
res <- teal:::get_key_duplicates_util(df, keys = c("a", "b")) # duplicated keys are in rows 3 and 4
print(res) # outputs a tibble

}
\seealso{
\code{\link{get_key_duplicates}}
}
