% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tdata.R
\name{get_metadata}
\alias{get_metadata}
\alias{get_metadata.tdata}
\alias{get_metadata.default}
\title{Function to get metadata from a \code{tdata} object}
\usage{
get_metadata(data, dataname)

\method{get_metadata}{tdata}(data, dataname)

\method{get_metadata}{default}(data, dataname)
}
\arguments{
\item{data}{(\code{tdata} - object) to extract the data from}

\item{dataname}{(\code{character(1)}) the dataset name whose metadata is requested}
}
\value{
Either list of metadata or NULL if no metadata.
}
\description{
Function to get metadata from a \code{tdata} object
}
