% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{get_module_ids}
\alias{get_module_ids}
\title{Get unique ids}
\usage{
get_module_ids(modules)
}
\arguments{
\item{modules}{(\code{character}) list of \code{teal_module}}
}
\value{
(\code{character}) vector of unique ids
}
\description{
Get unique ids for the module to avoid namespace conflicts.
}
\keyword{internal}
