% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{get_modules_attributes}
\alias{get_modules_attributes}
\title{Retrieve labels from \code{teal_modules}}
\usage{
get_modules_attributes(modules, key)
}
\arguments{
\item{modules}{(\code{teal_modules} or \code{teal_module}) object}
}
\description{
Retrieve labels from \code{teal_modules}
}
\keyword{internal}
