% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/column_functions.R
\name{get_nonkey_col_values}
\alias{get_nonkey_col_values}
\title{Derive non-key column values from data}
\usage{
get_nonkey_col_values(data)
}
\arguments{
\item{data}{(\code{data.frame}) Data with attribute \code{keys} and \code{dataname}}
}
\value{
A character data.frame with just the columns selected that are not
 keys
}
\description{
Derive non-key column values from data
}
