% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_package_file}
\alias{get_package_file}
\title{Function to get a file out of a package}
\usage{
get_package_file(pkg = NULL, file_name = NULL)
}
\arguments{
\item{pkg}{(\code{character})\cr
The name of the package the file should be received from.}

\item{file_name}{(\code{character})\cr
The name of the file to be received or path to it starting from
the base package path.}
}
\value{
The path to the file
}
\description{
Function to get a file out of a package
}
\examples{
teal:::get_package_file("teal", "WORDLIST")
teal:::get_package_file("teal", "cdisc_datasets/cdisc_datasets.yaml")
}
\keyword{internal}
