% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RawDataset.R, R/get_raw_data.R
\name{get_raw_data}
\alias{get_raw_data}
\alias{RawDataset}
\alias{get_raw_data.RawDataset}
\alias{get_raw_data.RawDatasetConnector}
\alias{get_raw_data.RelationalData}
\alias{get_raw_data.cdisc_data}
\title{R6 Class representing a raw data set}
\usage{
get_raw_data(x, dataname = NULL)

\method{get_raw_data}{RawDataset}(x, dataname = NULL)

\method{get_raw_data}{RawDatasetConnector}(x, dataname = NULL)

\method{get_raw_data}{RelationalData}(x, dataname = NULL)

\method{get_raw_data}{cdisc_data}(x, dataname = NULL)
}
\arguments{
\item{x}{(\code{RawDataset, RawDatasetConnector, RelationalData, cdisc_data})\cr object}

\item{dataname}{(\code{character} value)\cr
Name of dataset to return raw data for.}
}
\value{
\code{data.frame} with the raw data inserted into the R6 objects. In case of
\code{RelationalData} or \code{cdisc_data} \code{list} of \code{data.frame} can be returned
if user doesn't specify \code{dataname} (\code{get_raw_data} from all datasets).
}
\description{
Any \code{data.frame} or \code{rtable} object can be
stored inside this object. Some attributes like colnames,
dimension or column names for a specific type will
be automatically derived.
}
\examples{

# RawDatasets
library(random.cdisc.data)
ADSL <- radsl(cached = TRUE)
ADSL_raw <- raw_dataset(x = ADSL)
get_raw_data(ADSL_raw)

ADSL_named <- named_dataset(dataname = "ADSL", x = ADSL)
get_raw_data(ADSL_named)

ADSL_relational <- as_relational(ADSL_raw,
  dataname = "ADSL",
  keys = keys(primary = c("USUBJID", "STUDYID"), foreign = NULL, parent = NULL)
)
get_raw_data(ADSL_relational)

# RawDatasetConnector ---------
library(random.cdisc.data)
dc <- rcd_cdisc_dataset_connector(dataname = "ADSL", fun = radsl, cached = TRUE)
load_dataset(dc)
get_raw_data(dc)

# RelationalData ----------------
library(random.cdisc.data)
adsl <- cdisc_dataset(dataname = "ADSL", # RelationalDataset
                      data = radsl(cached = TRUE),
                      code = "library(random.cdisc.data)\nADSL <- radsl(cached = TRUE)")

adtte <- cdisc_dataset(dataname = "ADTTE", # RelationalDataset
                       data = radtte(cached = TRUE),
                       code = "library(random.cdisc.data)\nADTTE <- radtte(cached = TRUE)")

rd <- teal:::RelationalData$new(adsl, adtte)
get_raw_data(rd)

# RelationalDataConnector --------
rdc <- rcd_cdisc_data(
  rcd_cdisc_dataset_connector("ADSL", radsl, cached = TRUE),
  rcd_cdisc_dataset_connector("ADTTE", radtte, cached = TRUE)
)

get_raw_data(rdc)

\dontrun{
load_datasets(rdc)
get_raw_data(rdc)
}

# DelayedRelationalData --------
rdc2 <- rcd_cdisc_data(
  rcd_cdisc_dataset_connector("ADRS", radrs, cached = TRUE),
  rcd_cdisc_dataset_connector("ADLB", radlb, cached = TRUE)
)

drc <- teal_data(adsl, adtte, rdc2)

get_raw_data(drc)
\dontrun{
load_datasets(drc)
get_raw_data(drc)
}

# cdisc_data ----------------
library(random.cdisc.data)

ADSL <- radsl(cached = TRUE)
ADTTE <- radtte(cached = TRUE)

# Example with keys
cd = cdisc_data(
  cdisc_dataset("ADSL", ADSL, keys = keys(
    primary = c("STUDYID", "USUBJID"),
    foreign = NULL,
    parent = NULL
  )),
  cdisc_dataset("ADTTE", ADTTE, keys = keys(
    primary = c("STUDYID", "USUBJID", "PARAMCD"),
    foreign = c("STUDYID", "USUBJID"),
    parent = "ADSL"),
    code =  "ADTTE <- radtte(ADSL, seed = 123)"
  ),
  code = "ADSL <- radsl(N = 600, seed = 123)",
  check = FALSE
)

get_raw_data(cd, "ADSL")
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{ncol}}{Number of columns}

\item{\code{nrow}}{Number of rows}

\item{\code{dim}}{Dimension \code{c(x, y)}}

\item{\code{colnames}}{The column names of the data}

\item{\code{rownames}}{The rownames of the data}

\item{\code{raw_data}}{The data.frame behind this R6 class}

\item{\code{data}}{The data.frame behind this R6 class}

\item{\code{column_labels}}{for backwards compatibility
will be deprecated in future}

\item{\code{var_names}}{The column names of the data}

\item{\code{row_labels}}{Row labels (can have spaces)}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{RawDataset$new()}}
\item \href{#method-get_raw_data}{\code{RawDataset$get_raw_data()}}
\item \href{#method-get_numeric_colnames}{\code{RawDataset$get_numeric_colnames()}}
\item \href{#method-get_character_colnames}{\code{RawDataset$get_character_colnames()}}
\item \href{#method-get_factor_colnames}{\code{RawDataset$get_factor_colnames()}}
\item \href{#method-get_colnames}{\code{RawDataset$get_colnames()}}
\item \href{#method-get_column_labels}{\code{RawDataset$get_column_labels()}}
\item \href{#method-get_rownames}{\code{RawDataset$get_rownames()}}
\item \href{#method-get_row_labels}{\code{RawDataset$get_row_labels()}}
\item \href{#method-clone}{\code{RawDataset$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
initialize a \code{RawDataset} class object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RawDataset$new(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{(\code{data.frame}) object}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_raw_data"></a>}}
\if{latex}{\out{\hypertarget{method-get_raw_data}{}}}
\subsection{Method \code{get_raw_data()}}{
Derive the raw data frame inside this object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RawDataset$get_raw_data()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{data.frame} or \code{rtable}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_numeric_colnames"></a>}}
\if{latex}{\out{\hypertarget{method-get_numeric_colnames}{}}}
\subsection{Method \code{get_numeric_colnames()}}{
Derive the names of all \code{numeric} columns
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RawDataset$get_numeric_colnames(include_factors = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{include_factors}}{\code{logical} Whether to include
factor variables}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_character_colnames"></a>}}
\if{latex}{\out{\hypertarget{method-get_character_colnames}{}}}
\subsection{Method \code{get_character_colnames()}}{
Derive the names of all \code{character} columns
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RawDataset$get_character_colnames()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_factor_colnames"></a>}}
\if{latex}{\out{\hypertarget{method-get_factor_colnames}{}}}
\subsection{Method \code{get_factor_colnames()}}{
Derive the names of all \code{factor} columns
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RawDataset$get_factor_colnames()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_colnames"></a>}}
\if{latex}{\out{\hypertarget{method-get_colnames}{}}}
\subsection{Method \code{get_colnames()}}{
Derive the column names
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RawDataset$get_colnames()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_column_labels"></a>}}
\if{latex}{\out{\hypertarget{method-get_column_labels}{}}}
\subsection{Method \code{get_column_labels()}}{
Derive the column labels
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RawDataset$get_column_labels()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_rownames"></a>}}
\if{latex}{\out{\hypertarget{method-get_rownames}{}}}
\subsection{Method \code{get_rownames()}}{
Derive the row names
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RawDataset$get_rownames()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_row_labels"></a>}}
\if{latex}{\out{\hypertarget{method-get_row_labels}{}}}
\subsection{Method \code{get_row_labels()}}{
Derive the row labels
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RawDataset$get_row_labels()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RawDataset$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
