% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rcode.R
\name{get_rcode}
\alias{get_rcode}
\title{Returns R Code from a teal module}
\usage{
get_rcode(
  datasets = NULL,
  datanames = if (is.null(datasets)) datasets else datasets$datanames(),
  chunks = teal.code::get_chunks_object(),
  selected_chunk_ids = character(0),
  session = NULL,
  title = NULL,
  description = NULL
)
}
\arguments{
\item{datasets}{(\code{list}) list of \code{FilteredData} available inside the
server function of any \link{teal} module.}

\item{datanames}{(\code{character})\cr
names of datasets which code should be returned for. Due to fact that
\code{teal} filter panel depending on \code{"ADSL"}, code for \code{ADSL}
is always returned even if not specified.}

\item{chunks}{(\code{chunks}) \cr
object of class \code{chunks} that stores code chunks. These code
chunks are used in \link{teal} to enable reproducibility. Normally these chunks
are stored within the \link[shiny:shiny-package]{shiny::shiny-package} session. The default value
can normally be used.}

\item{selected_chunk_ids}{(\code{character} vector)\cr
vector of code chunks to be shown
in the code. If chunk id's are available this can be used to limit the
chunks that appear in the \code{"Show R-Code"} modal. Please only use this
feature if all chunks were set with designated IDs.}

\item{session}{(\code{environment}) deprecated.}

\item{title}{A character title of the module}

\item{description}{A character description of the module with additional
information not reflected in the title}
}
\value{
Return the R Code needed to reproduce a teal module. The \code{\link[=get_rcode_header]{get_rcode_header()}} part allows
to install the module. Additionally if the user filtered data by
teal inherited functions, the code to filter the data is included. If the teal module
is using \code{\link[teal.transform:data_extract_srv]{teal.transform::data_extract_srv()}} the extraction and merging
code will be returned, too.
If code chunks were used, these will also be used to derive module R Code.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
Return the R-code used to create a \link{teal} module analysis. This function
will return all analysis code as a character string. In case of a good setup it will
not only return the code used create the module analysis, but also the code used by
the app author to create the app. The main feature of this function is encapsulating
the R code to merge datasets by \code{\link[teal.transform:merge_datasets]{teal.transform::merge_datasets()}} and all the R code stored inside
code \link[teal.code:chunks]{teal.code::chunks}.
}
\note{
The \code{teal.load_nest_code} option is being used to customize the code that sets correct library paths
with all packages available. If empty (the default), a placeholder string is being used.
}
\examples{
\dontrun{
show_rcode_modal(
  title = "R Code for a Regression Plot",
  rcode = get_rcode(
    datasets = datasets,
    title = title,
    description = description
  )
)
}
}
\references{
\code{\link[=show_rcode_modal]{show_rcode_modal()}}, \code{\link[=get_rcode_header]{get_rcode_header()}}
}
