% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rcode_header.R
\name{get_rcode_data}
\alias{get_rcode_data}
\title{Get R code to generate data of teal module}
\usage{
get_rcode_data(data = NULL, datanames = NULL, datasets = NULL,
  code_data_processing = NULL)
}
\arguments{
\item{data}{optionally either an \code{FilteredData} object, or a named list of
datasets. The data sets need the \code{source} attribute and optionally the
\code{md5sum} attribute. NOTE: Just use either \code{data} or \code{datasets}}

\item{datanames}{(\code{character}) Selection of data sets from datasets or data}

\item{datasets}{optionally either an \code{FilteredData} object, or a named list of
datasets. The data sets need the \code{source} attribute and optionally the
\code{md5sum} attribute. NOTE: Just use either \code{data} or \code{datasets}}

\item{code_data_processing}{(\code{character}) You can hard copy your data preprocessing
and add it to the head instead of using a \code{FilteredData} object}
}
\value{
A string that contains the R code calls that are needed to reproduce
  the data sets that were an input of this function.
}
\description{
Get R code to generate data of teal module
}
\references{
\code{\link{get_rcode_datasets}}, \code{\link{get_rcode_data_list}}
}
\author{
wolfs25
}
