% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rcode_header.R
\name{get_rcode_data_list}
\alias{get_rcode_data_list}
\title{Get R Code for any data list}
\usage{
get_rcode_data_list(datanames = NULL, datasets = NULL,
  code_data_processing = NULL)
}
\arguments{
\item{datanames}{(\code{character}) vector of dataset names. Should always contain the
element \code{ASL}.}

\item{datasets}{(\code{list}) or (\code{FilteredData}) object from
which \code{md5} and source shall be shown}

\item{code_data_processing}{(\code{character}) You can hard copy your data preprocessing
and add it to the head instead of using a \code{FilteredData} object as
the \code{datasets} input}
}
\value{
R code with three segments
\itemize{
  \item{data }{data generation information}
  \item{preprocessing }{If \code{code_data_preprocessing} was added, it will be shown}
  \item{filters }{In case data is filtered data, how this is done}
}
}
\description{
Get R Code for any data list
}
\author{
wolfs25
}
