% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rcode_header.R
\name{get_rcode_git_pkgs}
\alias{get_rcode_git_pkgs}
\title{Derive git package installation Code}
\usage{
get_rcode_git_pkgs(git_repo = c(global = "https://github.com", roche =
  "https://github.roche.com"), git_pkgs = list(global = "Roche/rtables",
  roche = c("NEST/teal", "NEST/random.cdisc.data")), needs_rcd = FALSE)
}
\arguments{
\item{git_repo}{(\code{character}) Character vector with addresses of git repos such as
 \code{https://github.com}. You can also leave \code{git_pkgs} empty and just
hand over a single character string to \code{git_repo} and only this single
repo will be installed. Therefore it needs to be specified like:
\code{https://github.com/hadley/tidyverse}}

\item{git_pkgs}{(\code{list}) of packages to install from the git repositoires}

\item{needs_rcd}{(\code{logical}) Whether to use \code{random.cdisc.data}}
}
\value{
A character vector containing the \code{devtools::install_github} calls for the desired packages
  and the \code{library} calls for the desired packages.
}
\description{
Derive git package installation Code
}
\author{
wolfs25
}
