% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rcode_header_and_filters.R
\name{get_rcode_header}
\alias{get_rcode_header}
\title{Generates header text for analysis items}
\usage{
get_rcode_header(title, description = NULL, libraries = NULL,
  git_repo = NULL, data = NULL)
}
\arguments{
\item{title}{A character title of the module}

\item{description}{A character description of the module with additional
information not reflected in the title}

\item{libraries}{A character vector of the names of libraries required for
the module.}

\item{git_repo}{A character vector of the link to the GitHub repository
containing the R-code for the teal module. Optional argument}

\item{data}{optionally either an FilteredData object, or a named list of
datasets. The data sets need the \code{source} attribute and optionally the
\code{md5sum} attribute.}
}
\value{
A character string for the header text
}
\description{
Generates header text for analysis items
}
\examples{

cat(get_rcode_header(
 title = "Cross-Table",
 description = "A cross-table counts ..."
))

cat(get_rcode_header(
 title = "Cross-Table",
 description = "A cross-table counts ...",
 libraries = c('haven', 'ggplot2', 'dplyr')
))

cat(get_rcode_header(
 title = "Cross-Table",
 description = "A cross-table counts ...",
 libraries = c('haven', 'ggplot2', 'dplyr'),
 git_repo = "http://github.roche.com/NEST/teal/",
 data = list(
   ASL = structure(data.frame(a = 1), source = "haven::read_sas('/opt/BIOSTAT/asl.sas7bdat')"),
   ATE = structure(data.frame(a = 1),
                   source = "haven::read_sas('/opt/BIOSTAT/ate.sas7bdat')",
                   md5sum = "32sdf32fds324")
 )
))

}
