% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rcode_utils.R
\name{get_rcode_libraries}
\alias{get_rcode_libraries}
\title{Generates library calls from current session info}
\usage{
get_rcode_libraries()
}
\value{
Character vector of \verb{library(<package>)} calls.
}
\description{
Function to create multiple library calls out of current session info to ensure reproducible code works.
}
\keyword{internal}
