% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rcode.R
\name{get_rcode_srv}
\alias{get_rcode_srv}
\title{Server part of get R code module}
\usage{
get_rcode_srv(
  id,
  datasets,
  datanames = datasets$datanames(),
  modal_title = "R Code",
  code_header = "Automatically generated R code",
  disable_buttons = reactiveVal(FALSE)
)
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the module's
UI function.}

\item{datasets}{(\code{list}) list of \code{FilteredData} available inside the
server function of any \link{teal} module.}

\item{datanames}{(\code{character})\cr
names of datasets which code should be returned for. Due to fact that
\code{teal} filter panel depending on \code{"ADSL"}, code for \code{ADSL}
is always returned even if not specified.}

\item{modal_title}{optional, (\code{character}) title of the modal}

\item{code_header}{optional, (\code{character}) header inside R}

\item{disable_buttons}{optional, (\code{reactive})
a shiny reactive value. Should be a single boolean value, indicating whether to disable
or enable the show R code and Debug info buttons. Default: \code{reactiveVal(FALSE)}.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
