% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rcode.R
\name{get_rcode_ui}
\alias{get_rcode_ui}
\title{Ui part of get R code module}
\usage{
get_rcode_ui(id)
}
\arguments{
\item{id}{(\code{character}) id of shiny module}
}
\value{
(\code{shiny.tag})
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
