% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_selected_columns.R
\name{get_selected_columns}
\alias{get_selected_columns}
\title{Derive non-key columns from data}
\usage{
get_selected_columns(data)
}
\arguments{
\item{data}{(\code{data.frame}) Data with attribute \code{keys} and \code{dataname}}
}
\value{
A character vector with all non-key variable names as
  \code{<VARIABLE>.<DATANAME>}
}
\description{
Derive non-key columns from data
}
