% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_suppress_deps}
\alias{get_suppress_deps}
\title{Helper function to keep the list of HTML dependencies to be suppressed}
\usage{
get_suppress_deps()
}
\description{
Helper function to keep the list of HTML dependencies to be suppressed
}
