% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_bookmark_manager.R
\name{get_teal_bookmarkable_summary}
\alias{get_teal_bookmarkable_summary}
\title{Get bookmarkable summary list}
\usage{
get_teal_bookmarkable_summary(modules)
}
\arguments{
\item{modules}{\code{teal_modules} object}
}
\value{
named list of the same structure as \code{modules} with \code{TRUE} or \code{FALSE} values indicating
whether the module is bookmarkable.
}
\description{
Get bookmarkable summary list
}
\keyword{internal}
