% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_unique_labels}
\alias{get_unique_labels}
\title{Get unique labels}
\usage{
get_unique_labels(labels)
}
\arguments{
\item{labels}{(\code{character}) vector of labels}
}
\value{
(\code{character}) vector of unique labels
}
\description{
Get unique labels for the modules to avoid namespace conflicts.
}
\keyword{internal}
