% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdisc_data.R
\name{get_variable_labels}
\alias{get_variable_labels}
\title{Function that extract column labels from CDISC dataset}
\usage{
get_variable_labels(data, columns = NULL)
}
\arguments{
\item{data}{(\code{data.frame}) Any CDISC data set}

\item{columns}{(\code{character}) Column names to extract the labels
from}
}
\value{
labels of the columns
}
\description{
Function that extract column labels from CDISC dataset
}
\examples{
library(random.cdisc.data)
ADSL <- radsl(cached = TRUE)

get_variable_labels(ADSL)
get_variable_labels(ADSL, c("AGE", "RACE", "BMRKR1"))
get_variable_labels(ADSL, c("AGE", "RACE", "BMRKR1", "xyz"))
}
