% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{\%is_in\%}
\alias{\%is_in\%}
\title{List element in other list}
\usage{
x \%is_in\% y
}
\arguments{
\item{x}{\code{list} to be matched.}

\item{y}{\code{list} to be matched against.}
}
\value{
\code{logical} vector length of \code{x} denoting if element was found in second list.
}
\description{
\ifelse{html}{\out{<a href='https://www.tidyverse.org/lifecycle/#maturing'><img src='figures/lifecycle-maturing.svg' alt='Maturing lifecycle'></a>}}{\strong{Maturing}}
Checks if \code{x} element matches any of \code{y} element. If one of the arguments is a list then list elements
are treated as whole - in this case list elements can be a vector, so it looks
for equal element in second vector to be matched.
}
