% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{handle_active_datanames}
\alias{handle_active_datanames}
\title{When active_datanames is "all", sets them to all datanames
otherwise, it makes sure that it is a subset of the available datanames}
\usage{
handle_active_datanames(datasets, datanames)
}
\arguments{
\item{datasets}{\code{FilteredData} object}

\item{datanames}{\verb{character vector} datanames to pick}
}
\value{
intersection of \code{datasets$datanames()} and \code{datanames}
}
\description{
When active_datanames is "all", sets them to all datanames
otherwise, it makes sure that it is a subset of the available datanames
}
