% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rcode_header.R
\name{has_source_attribute}
\alias{has_source_attribute}
\title{Check if elements of input list have "source" element}
\usage{
has_source_attribute(x)
}
\arguments{
\item{x}{(\code{list})}
}
\value{
TRUE or FALSE
}
\description{
Check if elements of input list have "source" element
}
