% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/include_css_js.R
\name{include_css_files}
\alias{include_css_files}
\title{Include \code{CSS} files from \verb{/inst/css/} package directory to application header}
\usage{
include_css_files(package = "teal", pattern = "*")
}
\arguments{
\item{package}{(\code{character}) package name}

\item{pattern}{(\code{character}) pattern of files to be included}
}
\value{
HTML code that includes \code{CSS} files
}
\description{
Include \code{CSS} files from \verb{/inst/css/} package directory to application header
}
