% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_code.R
\name{include_source_code}
\alias{include_source_code}
\title{Includes source in code lines}
\usage{
include_source_code(code, exclude_comments, dir)
}
\arguments{
\item{code}{(\code{code}) class object, containing preprocessing code and additional information from parsed files.}

\item{exclude_comments}{(\code{logical}) whether exclude commented-out lines of code. Lines to be excluded
should be ended with \code{# nocode}. For multiple line exclusions one should enclose ignored block of code with
\code{# nocode>} and \code{# <nocode}}

\item{dir}{of the file where source is called from.}
}
\value{
lines of code with source text included
}
\description{
Includes source in code lines
}
