% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_code.R
\name{include_source_lines}
\alias{include_source_lines}
\title{Includes source in code lines}
\usage{
include_source_lines(lines, starts_at, stops_at, exclude_comments, dir)
}
\arguments{
\item{lines}{(\code{character}) of code as seperate element in vector}

\item{starts_at}{(\code{character}) regular expression determining start of the block of code}

\item{stops_at}{(\code{character}) regular expression determining stop  of the block of code}

\item{exclude_comments}{(\code{logical}) whether exclude commented-out lines of code}
}
\value{
lines of code with source text included
}
\description{
Includes source in code lines
}
