% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/include_css_js.R
\name{include_teal_css_js}
\alias{include_teal_css_js}
\title{Code to include teal CSS and Javascript files}
\usage{
include_teal_css_js()
}
\value{
HTML code to include
}
\description{
This is useful when you want to use the same Javascript and CSS files that are
used with the teal application.
This is also useful for running standalone modules in teal with the correct
styles.
Also initializes \code{shinyjs} so you can use it.
}
\examples{
shiny_ui <- tagList(
  teal:::include_teal_css_js(),
  p("Hello")
)
}
