% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{init}
\alias{init}
\title{Create the Server and Ui Function for the Shiny App}
\usage{
init(data, modules, filter = NULL, header = tags$p("title here"),
  footer = tags$p("footer here"))
}
\arguments{
\item{data}{named list with datasets. Dataset names are case sensitive. The
`ASL` data is mandatory.}

\item{modules}{nested list with one list per module with the
following named list elements: \tabular{ll}{ name \tab string with name
shown in menu for the analysis item \cr server \tab required, shiny server
module function, see \code{\link[shiny]{callModule}} for more
information\cr ui \tab required, shiny ui module function, see
\code{\link[shiny]{callModule}} for more information\cr data \tab required,
vector with datasets names that are passed on (filtered) to the server
function\cr options \tab optional, other arguments passed on to the server
function }}

\item{filter}{filter settings. Nested named list, currently with \code{init}
list element.}

\item{header}{object of class `shiny.tag` to be used as the header of the app}

\item{footer}{object of class `shiny.tag` to be used as the footer of the app}
}
\value{
named list with server and ui function
}
\description{
Creates the server and ui part for a teal shiny app
}
\examples{
library(random.cdisc.data)

ASL <- radsl(seed = 1)
ARS <- radrs(ASL, seed = 100)
ATE <- radtte(ASL, seed = 1000)

app <- init(
  data = list(ASL = ASL, ARS = ARS, ATE = ATE),
  modules = root_modules(
    module(
      "data source",
      server = function(input, output, session, datasets) {},
      ui = function(id) div(p("information about data source")),
      filters = NULL
    ),
    module(
      "ASL AGE histogram",
      server = function(input, output, session, datasets) {
        output$hist <- renderPlot(hist(datasets$get_data("ASL")$AGE))
      },
      ui = function(id) {ns <- NS(id); plotOutput(ns('hist'))},
      filters = NULL
    ),
    modules(
      label = "Example datasets (tree module)",
      module(
        "iris",
        server = function(input, output, session, datasets) {output$iris <- renderPlot(plot(iris))},
        ui = function(id) {ns <- NS(id); plotOutput(ns('iris'))},
        filters = NULL
      ),
      module(
        "cars",
        server = function(input, output, session, datasets) {output$cars <- renderPlot(plot(cars))},
        ui = function(id) {ns <- NS(id); plotOutput(ns('cars'))},
        filters = NULL
      )
    )
  ),
  header = tags$h1("Sample App"),
  footer = tags$p("Copyright 2017")
)

\dontrun{
shinyApp(app$ui, app$server)
}
}
