% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilterStates.R
\name{init_filter_states}
\alias{init_filter_states}
\title{Initialize \code{FilterStates} object}
\usage{
init_filter_states(
  data,
  input_dataname,
  output_dataname = input_dataname,
  datalabel = character(0),
  ...
)
}
\arguments{
\item{data}{(\code{data.frame}, \code{MultiAssayExperiment}, \code{SummarizedExperiment}, \code{matrix})\cr
R object which \code{subset} function is applied on.}

\item{input_dataname}{(\code{character(1)} or \code{name} or \code{call})\cr
name of the data used on lhs of the expression
specified to the function argument attached to this \code{FilterStates}.}

\item{output_dataname}{(\code{character(1)} or \code{name} or \code{call})\cr
name of the output data on the lhs of the assignment expression.}

\item{datalabel}{(\code{character(0)} or \code{character(1)})\cr
text label value.}

\item{...}{(optional)
additional arguments for specific classes: keys}
}
\description{
Initialize \code{FilterStates} object
}
\examples{
df <- data.frame(
  character = letters,
  numeric = seq_along(letters),
  date = seq(Sys.Date(), length.out = length(letters), by = "1 day"),
  datetime = seq(Sys.time(), length.out = length(letters), by = "33.33 hours")
)
rf <- teal:::init_filter_states(
  data = df,
  input_dataname = "DF",
  output_dataname = "DF_FILTERED",
  varlabels = c(
    "character variable", "numeric variable", "date variable", "datetime variable"
  )
)
\dontrun{
shinyApp(
  ui = fluidPage(
    actionButton("clear", span(icon("times"), "Remove all filters")),
    rf$ui_add_filter_state(id = "add", data = df),
    rf$ui("active"),
    verbatimTextOutput("expr"),
  ),
  server = function(input, output, session) {
    callModule(
      module = rf$srv_add_filter_state,
      id = "add",
      data = df
    )
    output$expr <- renderText({
      pdeparse(rf$get_call())
    })
    observeEvent(
      input$clear,
      rf$queue_empty()
    )
  }
)
}
}
