% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_ui.R
\name{init_ui}
\alias{init_ui}
\title{Initial user-interface of any teal app}
\usage{
init_ui(datasets, modules)
}
\arguments{
\item{datasets}{\link{FilteredData} object where all datasets
to be used inside the teal app are in}

\item{modules}{nested list with one list per module with the
following named list elements: \tabular{ll}{ name \tab string with name
shown in menu for the analysis item \cr server \tab required, shiny server
module function, see \code{\link[shiny]{callModule}} for more
information\cr ui \tab required, shiny ui module function, see
\code{\link[shiny]{callModule}} for more information\cr data \tab required,
vector with datasets names that are passed on (filtered) to the server
function\cr options \tab optional, other arguments passed on to the server
function }}
}
\value{
A \code{tagList} of The main menu, place holders for filters and
  place holders for the teal modules
}
\description{
Initial user-interface of any teal app
}
