% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choices_selected.R
\name{is.choices_selected}
\alias{is.choices_selected}
\title{Check if an object is a choices_selected class.}
\usage{
is.choices_selected(x)
}
\arguments{
\item{x}{object to check}
}
\description{
\ifelse{html}{\out{<a href='https://www.tidyverse.org/lifecycle/#maturing'><img src='figures/lifecycle-maturing.svg' alt='Maturing lifecycle'></a>}}{\strong{Maturing}}
}
