% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{is_arg_used}
\alias{is_arg_used}
\alias{is_arg_used.default}
\alias{is_arg_used.teal_modules}
\alias{is_arg_used.teal_module}
\alias{is_arg_used.function}
\title{Does the object make use of the \code{arg}}
\usage{
is_arg_used(modules, arg)

\method{is_arg_used}{default}(modules, arg)

\method{is_arg_used}{teal_modules}(modules, arg)

\method{is_arg_used}{teal_module}(modules, arg)

\method{is_arg_used}{`function`}(modules, arg)
}
\arguments{
\item{modules}{(\code{teal_module} or \code{teal_modules}) object}

\item{arg}{(\code{character(1)}) names of the arguments to be checked against formals of \code{teal} modules.}
}
\value{
\code{logical} whether the object makes use of \code{arg}
}
\description{
Does the object make use of the \code{arg}
}
\keyword{internal}
