% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{is_args_used}
\alias{is_args_used}
\alias{is_args_used.default}
\alias{is_args_used.teal_modules}
\alias{is_args_used.teal_module}
\alias{is_args_used.function}
\title{Does the object make use of \code{teal.reporter} reporting}
\usage{
is_args_used(modules, args)

\method{is_args_used}{default}(modules, args)

\method{is_args_used}{teal_modules}(modules, args)

\method{is_args_used}{teal_module}(modules, args)

\method{is_args_used}{`function`}(modules, args)
}
\arguments{
\item{modules}{(\code{teal_module} or \code{teal_modules}) object}

\item{args}{(\code{character}) names of the arguments to be checked against formals of \code{teal} modules.}
}
\value{
\code{logical} whether the object makes use of \code{teal.reporter} reporting
}
\description{
Does the object make use of \code{teal.reporter} reporting
}
\keyword{internal}
