% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CDISCFilteredData.R
\name{is_dag}
\alias{is_dag}
\title{Checks whether a graph is a \verb{Directed Acyclic Graph (DAG)}}
\usage{
is_dag(graph)
}
\arguments{
\item{graph}{(named \code{list}) list with node vector elements}
}
\value{
\code{logical(1)} \code{TRUE} if the graph is a \code{DAG}; \code{FALSE} otherwise
}
\description{
Checks whether a graph is a \verb{Directed Acyclic Graph (DAG)}
}
\keyword{internal}
