% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_html_like}
\alias{is_html_like}
\title{Whether the element is a Shiny HTML element}
\usage{
is_html_like(x)
}
\arguments{
\item{x}{any object}
}
\value{
\code{logical} whether the element is Shiny HTML like
}
\description{
The classes returned by \code{shiny::tag}, \code{shiny::tagList}, \code{shiny::HTML}
are one of \verb{"shiny.tag", "shiny.tag.list", "html"}. This checks if the element
has any of these classes.
}
