% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_pulled.R
\name{is_pulled}
\alias{is_pulled}
\alias{is_pulled.RawDatasetConnector}
\alias{is_pulled.RelationalDataset}
\alias{is_pulled.RelationalDatasetConnector}
\alias{is_pulled.RelationalDataConnector}
\title{Is pulled}
\usage{
is_pulled(x)

\method{is_pulled}{RawDatasetConnector}(x)

\method{is_pulled}{RelationalDataset}(x)

\method{is_pulled}{RelationalDatasetConnector}(x)

\method{is_pulled}{RelationalDataConnector}(x)
}
\arguments{
\item{x}{(\code{object}) of class \link{RawDatasetConnector}, \link{RelationalDataset},
\link{RelationalDatasetConnector}, or \link{RelationalDataConnector}.}
}
\value{
\code{TRUE} if connector has been already pulled, else \code{FALSE}.
}
\description{
S3 method to determine if dataset is pulled (loaded).
}
\examples{
# RawDatasetConnector --------
library(random.cdisc.data)
dc <- rcd_cdisc_dataset_connector(dataname = "ADSL", fun = radsl)

is_pulled(dc)

load_dataset(dc)
is_pulled(dc)
rel_data <- teal:::RelationalDataset$new(
  x = data.frame(x = c(1, 2), y = c("a", "b"), stringsAsFactors = FALSE),
  keys = keys(primary = "y", foreign = NULL, parent = NULL),
  dataname = "XY",
  code = "XY <- data.frame(x = c(1, 2), y = c('aa', 'bb'),
                           stringsAsFactors = FALSE)"
)

is_pulled(rel_data)

# RelationalDatasetConnector ---------
library(random.cdisc.data)
adsl <- rcd_cdisc_dataset_connector(dataname = "ADSL", fun = radsl, cached = TRUE)

is_pulled(adsl)

load_dataset(adsl)
is_pulled(adsl)

# RelationalDataConnector --------
library(random.cdisc.data)
adsl <- rcd_cdisc_dataset_connector(dataname = "ADSL", fun = radsl, cached = TRUE)
adrs <- rcd_cdisc_dataset_connector(dataname = "ADRS", fun = radrs, ADSL = adsl)

rdc <- rcd_cdisc_data(adsl, adrs)

is_pulled(rdc)
\dontrun{
load_datasets(rdc)
is_pulled(rdc)
}
}
