% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{is_reporter_used}
\alias{is_reporter_used}
\alias{is_reporter_used.default}
\alias{is_reporter_used.teal_modules}
\alias{is_reporter_used.teal_module}
\title{Does the object make use of \code{teal.reporter} reporting}
\usage{
is_reporter_used(modules)

\method{is_reporter_used}{default}(modules)

\method{is_reporter_used}{teal_modules}(modules)

\method{is_reporter_used}{teal_module}(modules)
}
\arguments{
\item{modules}{\code{teal_module} or \code{teal_modules} object}
}
\value{
\code{logical} whether the object makes use of \code{teal.reporter} reporting
}
\description{
Does the object make use of \code{teal.reporter} reporting
}
\keyword{internal}
