% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JoinKeys.R
\name{join_keys}
\alias{join_keys}
\title{Create a \code{JoinKeys} out of a list of \code{JoinKeySet} objects}
\usage{
join_keys(...)
}
\arguments{
\item{...}{optional, a \code{JoinKeySet} objects created using the \code{join_key} function.}
}
\value{
\code{JoinKeys}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
}
\details{
Note that join keys are symmetric although the relationship only needs
to be specified once.
}
\examples{
join_keys()
join_keys(
  join_key("dataset_A", "dataset_B", c("col_1" = "col_a")),
  join_key("dataset_A", "dataset_C", c("col_2" = "col_x", "col_3" = "col_y"))
)
join_keys(
  join_key("dataset_A", "dataset_B", c("col_1" = "col_a"))
)
}
