% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tdata.R
\name{join_keys.tdata}
\alias{join_keys.tdata}
\title{Extract \code{join_keys} from \code{tdata}}
\usage{
\method{join_keys}{tdata}(data, ...)
}
\arguments{
\item{data}{(\code{tdata}) object}

\item{...}{Additional arguments (not used)}
}
\description{
Extract \code{join_keys} from \code{tdata}
}
