% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keys.R
\name{keys<-}
\alias{keys<-}
\title{Set object keys}
\usage{
keys(x) <- value
}
\arguments{
\item{x}{any R object}

\item{value}{(\code{character}) with keys}
}
\description{
Set object keys
}
\examples{
library(random.cdisc.data)
ASL <- radsl()

keys(ASL) <- c("USUBJID", "STUDYID")
keys(ASL)
}
