% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keys_filter.R
\name{keys_filter_from_sep}
\alias{keys_filter_from_sep}
\title{Setup of Key Filtering for teal data sets}
\usage{
keys_filter_from_sep(vars, sep, choices, selected, multiple,
  label = "Filter")
}
\arguments{
\item{vars}{(\code{character}) Character vector giving the columns to be filtered}

\item{sep}{(\code{character}) A separator string to split the \code{choices} or
\code{selected} inputs into the values of the different columns}

\item{choices}{(\code{character}) Named character vector to define the choices
    of a shiny select input. These shall be filter values of the \code{vars} input.

E.g. \code{vars = c("PARAMCD","AVISITN")} and \code{choices = c("CRP - BASELINE","ALT - BASELINE")}
 will lead to a filtering of
 \code{(PARAMCD == "CRP" & AVISITN == "BASELINE") | (PARAMCD == "ALT" & AVISITN == "BASELINE")}.

Please make sure the order is right. The \code{sep} input has to be \code{" - "} in this case.}

\item{selected}{(\code{character}) Named character vector to define the selected
values of a shiny select input. Please check the \code{choices} description for further
details}

\item{multiple}{(\code{logical}) Whether multiple values shall be allowed in the
shiny select input.}

\item{label}{(\code{character}) Label of the Key Filtering Input inside the shiny app.}
}
\value{
A list of \code{vars} + a \link{choices_selected} outcome built by the split
	selected and choices inputs.
}
\description{
Setup of Key Filtering for teal data sets
}
