% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keys_filter.R
\name{keys_filtering_spec}
\alias{keys_filtering_spec}
\title{Constructor for \link{KeysFilteringSpec}}
\usage{
keys_filtering_spec(vars, sep, choices, selected, multiple,
  label = "Filter")
}
\arguments{
\item{vars}{(\code{character}) Character vector giving the key columns to be filtered}

\item{sep}{(\code{character}) A separator string to split the \code{choices} or
\code{selected} inputs into the values of the different columns}

\item{choices}{(\code{character}) Named character vector to define the choices
    of a shiny select input. These shall be filter values of the \code{vars} input
    separated by \code{sep}.

E.g. \code{vars = c("PARAMCD","AVISITN")} and \code{choices = c("CRP - BASELINE","ALT - BASELINE")}
 will lead to a filtering of
 \code{(PARAMCD == "CRP" & AVISITN == "BASELINE") | (PARAMCD == "ALT" & AVISITN == "BASELINE")}.

Please make sure the order is right. The \code{sep} input has to be \code{" - "} in this case.}

\item{selected}{(\code{character}) Named character vector to define the selected
values of a shiny select input. Please check the \code{choices} description for further
details}

\item{multiple}{(\code{logical}) Whether multiple values shall be allowed in the
shiny select input.}

\item{label}{(\code{character}) Label of the Key Filtering Input inside the shiny app.}
}
\value{
\code{\link{KeysFilteringSpec}}-class object
}
\description{
Constructor for \link{KeysFilteringSpec}
}
\examples{
keys_filtering_spec(
  vars = c("PARAMCD", "AVISIT"),
  sep = " - ",
  choices = c("CRP - BASELINE", "CRP - SCREENING", "ALT - BASELINE"),
  selected = c("CRP - BASELINE"),
  multiple = TRUE
)
}
