% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{label_to_id}
\alias{label_to_id}
\title{Turn a label into a valid html id, prefix a string}
\usage{
label_to_id(label, prefix = NULL)
}
\arguments{
\item{label}{label of module}

\item{prefix}{\verb{character or NULL} to prepend to label;
\code{NULL} for no prefix}
}
\value{
valid HTML label with invalid characters removed
}
\description{
From both \code{label} and \code{prefix}, remove one trailing and
one leading "\emph{", then convert all non-alphanumeric characters
to "}".
This can be used to create a hierarchy within a Shiny module namespace
itself, e.g. create nested tabs whose ids all live in one namespace,
but where a child tab's name is prefixed with the parent tab's name.
See \verb{\link{ui_nested_tabs}}.
}
\examples{
label_to_id <- teal:::label_to_id
label_to_id("var", prefix = "prefix")
label_to_id("var")
label_to_id("__var___", prefix = "prefix")
label_to_id("__var___", prefix = "_prefix__")
}
