% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{landing_modal}
\alias{landing_modal}
\title{A Landing Page Popup}
\usage{
landing_modal(title = NULL, text = NULL, button = NULL)
}
\arguments{
\item{title, text, button}{Arguments passed to \code{shinyalert::shinyalert}.}
}
\description{
Should be a part of \code{teal.modules.general}
}
