% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/register_logger.R
\name{layout_teal_glue_generator}
\alias{layout_teal_glue_generator}
\title{Generate log layout function using common variables available via glue syntax including shiny session token}
\usage{
layout_teal_glue_generator(layout)
}
\arguments{
\item{layout}{(\code{character(1)}) the log layout. Alongside the standard logging variables
provided by the \code{logging} package (e.g. \code{pid}) the \code{token} variable can be used which will
write the last 8 characters of the shiny session token to the log.}
}
\value{
function taking \code{level} and \code{msg} arguments - keeping the original call creating the generator
in the generator attribute that is returned when calling log_layout for the currently used layout
}
\description{
Generate log layout function using common variables available via glue syntax including shiny session token
}
\details{
this function behaves in the same way as \code{\link[logger:layout_glue_generator]{logger::layout_glue_generator()}}
but allows the shiny session token (last 8 chars) to be included in the logging layout
}
\keyword{internal}
