% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_app_usage.R
\name{line_pkg_log}
\alias{line_pkg_log}
\title{Package metadata to add to log file}
\usage{
line_pkg_log(fields)
}
\arguments{
\item{fields}{package metadata to be retrieved}
}
\description{
Metadata included in logs relates to loaded NEST packages
}
\examples{
\dontrun{
teal.utils:::line_pkg_log(
  pkgs = c("rtables", "tern", "teal"),
  fields = c("Package", "Title", "Version", "RemoteRef")
)
}
}
\keyword{internal}
