% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_app_usage.R
\name{line_usage_log}
\alias{line_usage_log}
\title{app usage data fields to add to log file}
\usage{
line_usage_log(...)
}
\description{
app usage data fields to add to log file
}
\examples{
\dontrun{
teal.utils:::line_usage_log("Oncology", "Tecentriq", "NSCLC", "Exploratory")
}
}
\keyword{internal}
