% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{list_adsl_first}
\alias{list_adsl_first}
\title{Move ADSL to the first index in a vector}
\usage{
list_adsl_first(datanames)
}
\arguments{
\item{datanames}{(\code{character} vector) datanames}
}
\description{
This is useful in the UI as ADSL should always appear first in
any vector of the dataset names.
If it does not contain ADSL, the vector is returned unmodified.
}
\examples{
list_adsl_first <- teal:::list_adsl_first
list_adsl_first(c("A", "B", "ADSL", "C"))
list_adsl_first(c("A", "B", "C"))
}
