% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{list_to_named_list}
\alias{list_to_named_list}
\title{Create a named list from a list.}
\usage{
list_to_named_list(data_list, names_data_list, data_names_call)
}
\arguments{
\item{data_list}{\code{RelationalData} or \code{list} or \code{data.frame}.}

\item{names_data_list}{\code{character} names of \code{data_list} input.}

\item{data_names_call}{\code{call} output of \code{substitute} function of data.}
}
\value{
named list.
}
\description{
Create a named list from a list.
}
\details{
takes the list output of \code{cast_to_list} and returns a named list.
}
