% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_dataset}
\alias{load_dataset}
\alias{load_dataset.RawDataset}
\alias{load_dataset.RawDatasetConnector}
\title{Load data from connection}
\usage{
load_dataset(x, ...)

\method{load_dataset}{RawDataset}(x, ...)

\method{load_dataset}{RawDatasetConnector}(x, args = NULL, try = FALSE, ...)
}
\arguments{
\item{x}{(\code{RawDatasetConnector} or \code{RawDataset})}

\item{...}{not used, only for support of S3}

\item{args}{(\code{NULL} or named \code{list})\cr
additional dynamic arguments passed to function which loads the data.}

\item{try}{(\code{logical}) whether perform function evaluation inside \code{try} clause}
}
\value{
\code{x} with loaded \code{dataset} object
}
\description{
Load data from connection. Function used on \link{RawDatasetConnector} and
\link{RawDataset} to obtain data from connection.
}
\examples{
# RawDataset --------
library(random.cdisc.data)
ADSL <- radsl(cached = TRUE)
ADSL_dataset <- raw_dataset(x = ADSL)

load_dataset(ADSL_dataset)
# RawDatasetConnector --------
library(random.cdisc.data)
adsl <- rcd_cdisc_dataset_connector("ADSL", fun = radsl, cached = TRUE)
load_dataset(adsl)
get_dataset(adsl)

adrs <- rcd_cdisc_dataset_connector("ADRS", fun = radrs, ADSL = adsl)
load_dataset(adrs)
}
