% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_datasets}
\alias{load_datasets}
\alias{load_datasets.RelationalData}
\alias{load_datasets.RelationalDataConnector}
\alias{load_datasets.DelayedRelationalData}
\title{Load datasets}
\usage{
load_datasets(x)

\method{load_datasets}{RelationalData}(x)

\method{load_datasets}{RelationalDataConnector}(x)

\method{load_datasets}{DelayedRelationalData}(x)
}
\arguments{
\item{x}{(\code{object}) of class\link{RawDatasetConnector} or \link{NamedDataset}. If of
class \code{character} will be treated as file to read.}
}
\value{
object of the same class as \code{x}
}
\description{
Load datasets
}
\examples{

# RelationalData ------
library(random.cdisc.data)
adsl <- cdisc_dataset(dataname = "ADSL", # RelationalDataset
                      data = radsl(cached = TRUE),
                      code = "library(random.cdisc.data)\nADSL <- radsl(cached = TRUE)")

adtte <- cdisc_dataset(dataname = "ADTTE", # RelationalDataset
                       data = radtte(cached = TRUE),
                       code = "library(random.cdisc.data)\nADTTE <- radtte(cached = TRUE)")

rd <- teal_data(adsl, adtte)

\dontrun{
# return a warning
load_datasets(rd)
}

# RelationalDataConnector --------
library(random.cdisc.data)
rdc <- rcd_cdisc_data(
  rcd_cdisc_dataset_connector("ADSL", radsl, cached = TRUE),
  rcd_cdisc_dataset_connector("ADTTE", radtte, cached = TRUE)
)

\dontrun{
load_datasets(rdc)
}

# DelayedRelationalData --------
library(random.cdisc.data)
x <- rcd_cdisc_data( # RelationalDataConnector
  rcd_cdisc_dataset_connector("ADSL", radsl, cached = TRUE),
  rcd_cdisc_dataset_connector("ADLB", radlb, cached = TRUE)
)
x2 <- rcd_cdisc_data( # RelationalDataConnector
  rcd_cdisc_dataset_connector("ADRS", radrs, cached = TRUE)
)

tc <- teal_data(x, x2)

\dontrun{
load_datasets(tc)
}
}
