% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_app_usage.R
\name{log_app_usage}
\alias{log_app_usage}
\title{Teal Application Usage Logging.}
\usage{
log_app_usage(
  ta,
  molecule,
  indication,
  anl_type,
  pkg_meta = c("Package", "Title", "Version", "RemoteRef")
)
}
\arguments{
\item{ta}{therapeutic area name. required argument. e.g. \code{"I2ON"}.}

\item{molecule}{molecule name. required argument. e.g. \code{"Lucentis"}.}

\item{indication}{indication. required argument. e.g. \code{"AMD"}.}

\item{anl_type}{analysis type. required argument. e.g. \code{"Interim Analysis"}.}

\item{pkg_meta}{package metadata fields. default values: \code{c("Package", "Title", "Version", "RemoteRef")}.}
}
\description{
Analysis of Teal application utilization is a key component towards ongoing enhancements
of the framework. The utilization logs provide source data for better understanding
frequency and extent of use.
}
\details{
Each Teal application can implement usage logging by calling this
function in the application startup file(s) like app.R. The following data are
captured by default: UNIX ID, System Date, Application Directory, Therapeutic Area, Indication,
Package Name, Package Title, Package Version and Remotes Definition. Capture frequency is
currently per user at session start only.
Suggested values for ta: "Oncology" or "I2ON"
Suggested values for anl_type: "Exploratory", "Interim Analysis", "CSR"
}
\examples{
\dontrun{
log_app_usage(ta = "Oncology", molecule = "Tecentriq", ind = "NSCLC", anl_type = "Exploratory")
readLines(file.path("logs", "utilization.log"), warn = FALSE)
}
}
