% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{log_system_info}
\alias{log_system_info}
\title{Logs the basic information about the session.}
\usage{
log_system_info()
}
\value{
\code{invisible(NULL)}
}
\description{
Logs the basic information about the session.
}
\keyword{internal}
