% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAEDataset.R
\name{mae_dataset}
\alias{mae_dataset}
\title{Public facing constructor for MAEDataset}
\usage{
mae_dataset(
  dataname,
  x,
  label = data_label(x),
  code = character(0),
  vars = list()
)
}
\arguments{
\item{dataname}{(\code{character})\cr
A given name for the dataset it may not contain spaces}

\item{x}{(\code{MultiAssayExperiment})}

\item{label}{(\code{character})\cr
Label to describe the dataset}

\item{code}{(\code{character})\cr
A character string defining the code needed to produce the data set in \code{x}}

\item{vars}{(named \code{list})) \cr
In case when this object code depends on other \code{Dataset} object(s) or
other constant value, this/these object(s) should be included as named
element(s) of the list. For example if this object code needs \code{ADSL}
object we should specify \verb{vars = list(ADSL = <adsl object>)}.
It's recommended to include \code{Dataset} or \code{DatasetConnector} objects to
the \code{vars} list to preserve reproducibility. Please note that \code{vars}
are included to this object as local \code{vars} and they cannot be modified
within another dataset.}
}
\description{
Public facing constructor for MAEDataset
}
\examples{
# Simple example
\dontrun{
library(MultiAssayExperiment)
MAE_dataset <- mae_dataset("MAE", miniACC)
MAE_dataset$get_dataname()
MAE_dataset$get_dataset_label()
MAE_dataset$get_code()
MAE_dataset$get_raw_data()
}
}
