% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teal_data_module.R
\name{make_teal_transform_server}
\alias{make_teal_transform_server}
\title{Make teal_transform_module's server}
\usage{
make_teal_transform_server(expr)
}
\arguments{
\item{expr}{(\code{language})
An R call which will be evaluated within \code{\link[teal.data:teal_data]{teal.data::teal_data}} environment.}
}
\value{
\verb{function(id, data)} returning \code{shiny} module
}
\description{
A factory function to simplify creation of a \code{\link{teal_transform_module}}'s server. Specified \code{expr}
is wrapped in a shiny module function and output can be passed to the \code{server} argument in
\code{\link[=teal_transform_module]{teal_transform_module()}} call. Such server function can be linked with ui and values from the
inputs can be used in the expression. Object names specified in the expression will be substituted
with the value of the respective input (matched by the name) - for example in
\code{expression(graph <- graph + ggtitle(title))} object \code{title} will be replaced with the value of
\code{input$title}.
}
\examples{

teal_transform_module(
  label = "Simplified interactive transform for iris",
  datanames = "iris",
  ui = function(id) {
    ns <- NS(id)
    numericInput(ns("n_rows"), "Subset n rows", value = 6, min = 1, max = 150, step = 1)
  },
  server = make_teal_transform_server(expression(iris <- head(iris, n_rows)))
)

}
