% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_snapshot_manager.R
\name{matrix_to_mapping}
\alias{matrix_to_mapping}
\title{Convert mapping matrix to filter mapping specification.}
\usage{
matrix_to_mapping(mapping_matrix)
}
\arguments{
\item{mapping_matrix}{(\code{data.frame}) of logical vectors where
columns represent modules and row represent \code{teal_slice}s}
}
\value{
Named \code{list} like that in the \code{mapping} attribute of a \code{teal_slices} object.
}
\description{
Transform a mapping matrix, i.e. a data frame that maps each filter state to each module,
to a list specification like the one used in the \code{mapping} attribute of \code{teal_slices}.
Global filters are gathered in one list element.
If a module has no active filters but the global ones, it will not be mentioned in the output.
}
\keyword{internal}
