% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{modify_header}
\alias{modify_header}
\title{Add a Header to \code{teal} Application}
\usage{
modify_header(app, header = tags$p())
}
\arguments{
\item{app}{(\code{environment}) The \code{teal} app environment.}

\item{header}{(\code{shiny.tag} or \code{character(1)}) The header content to set. Defaults to an empty paragraph tag.}
}
\description{
Adds a header to the \code{teal} app.
}
\examples{
app <- init(
  data = teal_data(IRIS = iris),
  modules = modules(example_module())
) |>
  modify_header(
    tags$div(
      h3("Custom header")
    )
  )

if (interactive()) {
  shinyApp(app$ui, app$server)
}
}
