% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{modify_title}
\alias{modify_title}
\title{Add a custom title to the app}
\usage{
modify_title(
  app,
  title = "teal app",
  favicon =
    "https://raw.githubusercontent.com/insightsengineering/hex-stickers/main/PNG/nest.png"
)
}
\arguments{
\item{title}{The title to add}
}
\value{
The modified app object
}
\description{
Add a custom title to the app
}
\examples{
app <- init(
  data = teal_data(IRIS = iris, MTCARS = mtcars),
  modules = modules(example_module())
) |>
  modify_title("Custom title")

shinyApp(app$ui, app$server)
}
