% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{module}
\alias{module}
\title{Create a module with a new shiny page}
\usage{
module(label, server, ui, filters, server_args = NULL, ui_args = NULL)
}
\arguments{
\item{label}{label shown in the navigation for the item}

\item{server}{shiny server module function, see
\code{link[shiny]{callModule}}}

\item{ui}{shiny ui module function, see \code{link[shiny]{callModule}}}

\item{filters}{a vector with datanames that are relevant for the item. The
filter panel will automatically update the shown filters to include only
filters in the listed data sets. \code{NULL} will hide the filter panel,
and the keyowrd \code{'all'} will show the filters of all datasets.}

\item{server_args}{is a named list with additional arguments passed on to the
server function. Note that the \code{FilteredDatasets} object gets
atomatically passed to the server function as arguments \code{datasets}.}

\item{ui_args}{is a named list with additional arguments passed on to the
ui function. The argument \code{'teal_datasets'} will always be
replaced by the \code{FilteredData} object.}
}
\description{
Tab items allows you to add a shiny module to the teal app
}
