% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_filter_manager.R
\name{filter_manager_ui}
\alias{filter_manager_ui}
\alias{filter_manager_srv}
\title{Manage multiple \code{FilteredData} objects}
\usage{
filter_manager_ui(id, is_global)

filter_manager_srv(id, filtered_data_list, filter)
}
\arguments{
\item{id}{(\code{character(1)})\cr
\code{shiny} module id.}

\item{filtered_data_list}{(\code{list} of \code{FilteredData})\cr
Names of the list should be the same as \code{teal_module$label}.}

\item{filter}{(\code{teal_slices})\cr
Specification of initial filter. Filters can be specified using \code{\link[=teal_filters]{teal_filters()}}.
Old way of specifying filters through a list is deprecated and will be removed in the
next release. Please fix your applications to use \code{\link[=teal_filters]{teal_filters()}}.}
}
\description{
Manage multiple \code{FilteredData} objects
}
\details{
This module observes the changes of the filters in each \code{FilteredData} object
and keeps track of all filters used. Map of the filters is kept in so called
\code{slices_map} object where each \code{FilteredData} is linked with its active filters.
This map is represented in the UI as a matrix where rows are ids of the filters and
columns are names of the \code{filtered_data_list} (named after teal modules).
}
\keyword{internal}
