% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{module_labels}
\alias{module_labels}
\title{Retrieve labels from \code{teal_modules}}
\usage{
module_labels(modules)
}
\arguments{
\item{modules}{(\code{teal_modules})}
}
\value{
A \code{list} containing the labels of the modules. If the modules are nested,
the function returns a nested \code{list} of labels.
}
\description{
Retrieve labels from \code{teal_modules}
}
\keyword{internal}
