% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{extract_module}
\alias{extract_module}
\alias{drop_module}
\title{Extract/Remove module(s) of specific class}
\usage{
extract_module(modules, class)

drop_module(modules, class)
}
\arguments{
\item{modules}{(\code{teal_modules})}

\item{class}{The class name of \code{teal_module} to be extracted or dropped.}
}
\value{
\itemize{
\item For \code{extract_module}, a \code{teal_module} of class \code{class} or \code{teal_modules} containing modules of class \code{class}.
\item For \code{drop_module}, the opposite, which is all \code{teal_modules} of  class other than \code{class}.
}

\code{teal_modules}
}
\description{
Given a \code{teal_module} or a \code{teal_modules}, return the elements of the structure according to \code{class}.
}
\keyword{internal}
