% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_session_info.R
\name{module_session_info}
\alias{module_session_info}
\alias{ui_session_info}
\alias{srv_session_info}
\title{\code{teal} user session info module}
\usage{
ui_session_info(id)

srv_session_info(id)
}
\arguments{
\item{id}{(\code{character(1)}) \code{shiny} module instance id.}
}
\value{
\code{NULL} invisibly
}
\description{
Module to display the user session info popup and to download a lockfile. Module is included
when running \code{\link[=init]{init()}} but skipped when using \code{\link{module_teal}}. Please be aware that session info
contains R session information, so the module called multiple times will share the same information.
}
\examples{
ui <- fluidPage(
  ui_session_info("session_info")
)

server <- function(input, output, session) {
  srv_session_info("session_info")
}

if (interactive()) {
  shinyApp(ui, server)
}

}
