% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_teal_data.R
\name{module_teal_data}
\alias{module_teal_data}
\alias{ui_teal_data}
\alias{srv_teal_data}
\alias{ui_validate_reactive_teal_data}
\alias{srv_validate_reactive_teal_data}
\title{Execute and validate \code{teal_data_module}}
\usage{
ui_teal_data(id, data_module)

srv_teal_data(
  id,
  data,
  data_module,
  modules = NULL,
  validate_shiny_silent_error = TRUE
)

ui_validate_reactive_teal_data(id)

srv_validate_reactive_teal_data(
  id,
  data,
  modules = NULL,
  validate_shiny_silent_error = FALSE
)
}
\arguments{
\item{id}{(\code{character(1)}) Module id}

\item{data_module}{(\code{teal_data_module})}

\item{data}{(\verb{reactive teal_data})}

\item{modules}{(\code{teal_modules} or \code{teal_module}) For \code{datanames} validation purpose}

\item{validate_shiny_silent_error}{(\code{logical}) If \code{TRUE}, then \code{shiny.silent.error} is validated and
error message is displayed.
Default is \code{FALSE} to handle empty reactive cycle on \code{init}.}
}
\value{
\code{reactive} \code{teal_data}
}
\description{
This is a low level module to handle \code{teal_data_module} execution and validation.
\code{\link[=teal_transform_module]{teal_transform_module()}} inherits from \code{\link[=teal_data_module]{teal_data_module()}} so it is handled by this module too.
\code{\link[=srv_teal]{srv_teal()}} accepts various \code{data} objects and eventually they are all transformed to \code{reactive}
\code{\link[=teal_data]{teal_data()}} which is a standard data class in whole \code{teal} framework.
}
\section{data validation}{


Executed \code{\link[=teal_data_module]{teal_data_module()}} is validated and output is validated for consistency.
Output \code{data} is invalid if:
\enumerate{
\item \code{\link[=teal_data_module]{teal_data_module()}} is invalid if server doesn't return \code{reactive}. \strong{Immediately crashes an app!}
\item \code{reactive} throws a \code{shiny.error} - happens when module creating \code{\link[=teal_data]{teal_data()}} fails.
\item \code{reactive} returns \code{qenv.error} - happens when \code{\link[=teal_data]{teal_data()}} evaluates a failing code.
\item \code{reactive} object doesn't return \code{\link[=teal_data]{teal_data()}}.
\item \code{\link[=teal_data]{teal_data()}} object lacks any \code{datanames} specified in the \code{modules} argument.
}

\code{teal} (observers in \code{srv_teal}) always waits to render an app until \code{reactive} \code{teal_data} is
returned. If error 2-4 occurs, relevant error message is displayed to app user and after issue is
resolved app will continue to run. \code{teal} guarantees that errors in a data don't crash an app
(except error 1). This is possible by catching the error and returning an empty \code{teal_data} object.
The error message is displayed in the app and the module will eventually get an empty \code{teal_data}.
}

\keyword{internal}
