% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_transform_data.R
\name{module_teal_transform_module}
\alias{module_teal_transform_module}
\alias{ui_teal_transform_module}
\alias{srv_teal_transform_module}
\title{Module to transform \code{reactive} \code{teal_data}}
\usage{
ui_teal_transform_module(id, transforms, class = "well")

srv_teal_transform_module(
  id,
  data,
  transforms,
  modules = NULL,
  is_transformer_failed = reactiveValues()
)
}
\arguments{
\item{id}{(\code{character(1)}) Module id}

\item{data}{(\verb{reactive teal_data})}

\item{modules}{(\code{teal_modules} or \code{teal_module}) For \code{datanames} validation purpose}

\item{is_transformer_failed}{(\code{reactiveValues}) contains \code{logical} flags named after each transformer.
Help to determine if any previous transformer failed, so that following transformers can be disabled
and display a generic failure message.}
}
\value{
\code{reactive} \code{teal_data}
}
\description{
Module calls \code{\link[=teal_transform_module]{teal_transform_module()}} in sequence so that \verb{reactive teal_data} output
from one module is handed over to the following module's input.
}
