% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_transform_data.R
\name{module_transform_teal_data_impl}
\alias{module_transform_teal_data_impl}
\alias{.ui_transform_teal_data}
\alias{.srv_transform_teal_data}
\title{Internal module handling list of \code{\link[=teal_transform_module]{teal_transform_module()}}}
\usage{
.ui_transform_teal_data(id, transformators, class = "well")

.srv_transform_teal_data(
  id,
  data,
  transformators,
  modules = NULL,
  is_transform_failed = reactiveValues()
)
}
\arguments{
\item{id}{(\code{character(1)}) \code{shiny} module instance id.}

\item{transformators}{(\code{list} of \code{teal_transform_module}) decorator modules to apply sequentially
to \code{data}. Each transformator receives the output of the previous one as input.}

\item{class}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} No longer used.}

\item{data}{(\code{teal_data}, \code{teal_data_module}, or \code{reactive} returning \code{teal_data})
The data which application will depend on.}

\item{modules}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} No longer used.}

\item{is_transform_failed}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} No longer used.}
}
\value{
\code{shiny.tag}

\code{reactive} \code{teal_data} object
}
\description{
This module calls \code{\link[=teal_transform_module]{teal_transform_module()}} sequentially by passing output from one step to the next.
Modules have error/warning handling feature and highlight containers when something goes wrong.
}
\keyword{internal}
