% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_validate.R
\docType{data}
\name{module_validate_reactive}
\alias{module_validate_reactive}
\alias{srv_module_check_reactive}
\title{Validate if an argument is a reactive}
\format{
An object of class \code{list} of length 2.
}
\usage{
srv_module_check_reactive(x, types = character(0L), null.ok = FALSE)

module_validate_reactive$server(x, types = character(0L), null.ok = FALSE)
module_validate_reactive$ui(id)
}
\arguments{
\item{x}{(\code{reactive}) A reactive value.}

\item{types}{(\code{character}) A character vector with the types that the reactive.}

\item{null.ok}{(\code{logical}) If \code{TRUE}, the \code{x} argument can be \code{NULL}.}

\item{id}{(\code{character}) The module id.}
}
\value{
A module that validates the reactive value.
}
\description{
Validate if an argument is a reactive
}
\examples{
module_validate_reactive$ui("validate_reactive")

# Show the generated server function
print(module_validate_reactive$server)
}
\seealso{
\code{\link[=module_validate_factory]{module_validate_factory()}}
}
\keyword{datasets}
