% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_validate.R
\docType{data}
\name{module_validate_shinysilenterror}
\alias{module_validate_shinysilenterror}
\alias{srv_module_check_shinysilenterror}
\title{Validate if an argument contains a \code{shiny.silent.error}}
\format{
An object of class \code{list} of length 2.
}
\usage{
srv_module_check_shinysilenterror(x)

module_validate_shinysilenterror$ui(id)
module_validate_shinysilenterror$server(x)
}
\arguments{
\item{x}{(\code{reactive}) A reactive value.}

\item{id}{(\code{character}) The module id.}
}
\value{
A module that validates the reactive value.
}
\description{
Validate if an argument contains a \code{shiny.silent.error}
}
\examples{
module_validate_shinysilenterror$ui("validate_reactive")

# Show the generated server function
print(module_validate_shinysilenterror$server)
}
\seealso{
\code{\link[=module_validate_factory]{module_validate_factory()}}
}
\keyword{datasets}
