% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_wunder_bar.R
\name{module_wunder_bar}
\alias{module_wunder_bar}
\alias{wunder_bar_ui}
\alias{wunder_bar}
\alias{wunder_bar_module}
\alias{wunder_bar_srv}
\title{Manager bar module}
\usage{
wunder_bar_ui(id)

wunder_bar_srv(id, datasets, filter, modules)
}
\arguments{
\item{id}{(\code{character(1)})
module id}

\item{datasets}{(named \code{list} of \code{FilteredData})
object to store filter state and filtered datasets, shared across modules. For more
details see \code{\link[teal.slice:FilteredData]{teal.slice::FilteredData}}. Structure of the list must be the same as structure
of the \code{modules} argument and list names must correspond to the labels in \code{modules}.
When filter is not module-specific then list contains the same object in all elements.}

\item{filter}{(\code{teal_slices})
Specifies the initial filter using \code{\link[=teal_slices]{teal_slices()}}.}

\item{modules}{(\code{teal_modules}) object containing the output modules which
will be displayed in the \code{teal} application. See \code{\link[=modules]{modules()}} and \code{\link[=module]{module()}} for
more details.}
}
\value{
Nothing is returned.
}
\description{
Bar of buttons that open modal dialogs.
}
\details{
Creates a bar of buttons that open modal dialogs where manager modules reside.
Currently contains three modules:
\itemize{
\item \code{\link{module_filter_manager}}
\item \code{\link{module_snapshot_manager}}
\item \code{\link{module_bookmark_manager}}
}

The bar is placed in the \code{teal} app UI, next to the filter panel hamburger.
}
\keyword{internal}
