% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{modules_bookmarkable}
\alias{modules_bookmarkable}
\title{Retrieve \code{teal_bookmarkable} attribute from \code{teal_modules}}
\usage{
modules_bookmarkable(modules)
}
\arguments{
\item{modules}{(\code{teal_modules} or \code{teal_module}) object}
}
\value{
named list of the same structure as \code{modules} with \code{TRUE} or \code{FALSE} values indicating
whether the module is bookmarkable.
}
\description{
Retrieve \code{teal_bookmarkable} attribute from \code{teal_modules}
}
\keyword{internal}
