% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{modules_datasets}
\alias{modules_datasets}
\title{Create filterable data for modules}
\usage{
modules_datasets(
  data,
  modules,
  filters,
  filtered_data_singleton = teal_data_to_filtered_data(data)
)
}
\arguments{
\item{data}{(\code{teal_data})}

\item{modules}{(\code{teal_modules}) object}

\item{filters}{(\code{teal_slices}) object}

\item{filtered_data_singleton}{A result of \code{teal_data_to_filtered_data} applied to \code{data}.}
}
\value{
Returns list of same shape as \code{modules}, containing \code{FilteredData} at every leaf.
If module specific, each leaf contains different instance, otherwise every leaf contains \code{filtered_data_singleton}.
}
\description{
Converts input data to a \code{FilteredData} object(s) to allow filtering before passing data to individual modules.
}
\keyword{internal}
