% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{modules_depth}
\alias{modules_depth}
\title{Get module depth}
\usage{
modules_depth(modules, depth = 0)
}
\arguments{
\item{modules}{\code{teal.module} or \code{teal.modules} object}

\item{depth}{optional, integer determining current depth level}
}
\value{
depth level for given module
}
\description{
Depth starts at 0, so a single \code{teal.module} has depth 0.
Nesting it increases overall depth by 1.
}
\examples{
m <- module(
  "aaa",
  server = function(input, output, session, datasets) {},
  ui = function(id, ...) {},
  filters = 'all'
)
x <- modules(
  "d1",
  modules(
    "d2",
    modules(
      "d3",
      m, m, m
    ),
    m
  ),
  m
)
teal:::modules_depth(x)

x <- modules(
  "a",
  modules(
    "b", m
  ),
  m
)
teal:::modules_depth(x)
}
