% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JoinKeys.R
\name{mutate_join_keys}
\alias{mutate_join_keys}
\alias{mutate_join_keys.JoinKeys}
\alias{mutate_join_keys.RelationalData}
\title{Mutate \code{JoinKeys} with a new values}
\usage{
mutate_join_keys(x, dataset_1, dataset_2, val)

\method{mutate_join_keys}{JoinKeys}(x, dataset_1, dataset_2, val)

\method{mutate_join_keys}{RelationalData}(x, dataset_1, dataset_2, val)
}
\arguments{
\item{x}{(\code{JoinKeys}) object to be modified}

\item{dataset_1}{(\code{character}) one dataset name}

\item{dataset_2}{(\code{character}) other dataset name}

\item{val}{(named \code{character}) column names used to join}
}
\value{
modified \code{JoinKeys} object
}
\description{
\ifelse{html}{\out{<a href='https://www.tidyverse.org/lifecycle/#experimental'><img src='figures/lifecycle-experimental.svg' alt='Experimental lifecycle'></a>}}{\strong{Experimental}}
Mutate \code{JoinKeys} with a new values
}
\examples{
# JoinKeys ----

x <- join_keys(
  join_key("dataset_A", "dataset_B", c("col_1" = "col_a")),
  join_key("dataset_A", "dataset_C", c("col_2" = "col_x", "col_3" = "col_y"))
)
x$get("dataset_A", "dataset_B")

mutate_join_keys(x, "dataset_A", "dataset_B", c("col_1" = "col_10"))
x$get("dataset_A", "dataset_B")
# RelationalData ----

library(random.cdisc.data)

ADSL <- radsl(cached = TRUE)
ADRS <- radrs(cached = TRUE)

x <- cdisc_data(
  cdisc_dataset("ADSL", ADSL),
  cdisc_dataset("ADRS", ADRS)
)
x$get_join_keys()$get("ADSL", "ADRS")

mutate_join_keys(x, "ADSL", "ADRS", c("COLUMN1" = "COLUMN2"))
x$get_join_keys()$get("ADSL", "ADRS")

}
