% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/name_choices.R
\name{name_choices}
\alias{name_choices}
\title{Set the names of a data.frame of vector}
\usage{
name_choices(x, xnames)
}
\arguments{
\item{x}{either data.frame or character vector}

\item{xnames}{vector conatining names to be applied to \code{x}}
}
\value{
a named character vector
}
\description{
Set the names of a data.frame of vector
}
\details{
If either \code{x} or \code{xnames} are factors, they are coerced 
  to character. If the names of \code{x} and \code{xnames} are the same or similar, 
  the unique name is kept. For example if a variable "AGE" should be named 
  "Age" according to \code{xnames}, the vector returned will be named just "Age" 
  instead of "AGE - Age". See examples.
  
  This is function is useful to create the vectors passed on to 
  the arguments of the \code{choices_selected} arguments in \code{teal} modules.
}
\examples{

library(random.cdisc.data)
library(tern)
ADSL <- radsl(N=10, seed = 1)
ADTTE <- radtte(ADSL, seed = 1)

choices1 <- name_choices(ADSL, tern::var_labels(ADSL))
choices2 <- name_choices(ADTTE$PARAMCD, ADTTE$PARAM)

#similar names are not duplicated
choices3 <- name_choices(ADTTE$PARAMCD, ADTTE$PARAMCD)

\dontrun{

library(shiny)

shinyApp(
ui = fluidPage(selectInput("c1", label = "Choices from ADSL",
                choices = choices1,
                selected = choices1[1]),
                 
               selectInput("c2", label = "Choices from ADTTE",
                choices = choices2,
                selected = choices2[1])),
                  
server = function(input, output) {}
)

}
}
