% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/named_choices.R
\name{named_choices}
\alias{named_choices}
\title{Set "<choice>: <label>" type of Names}
\usage{
named_choices(choices, labels)
}
\arguments{
\item{choices}{a character vector}

\item{labels}{vector containing labels to be applied to \code{choices}}
}
\value{
a named character vector
}
\description{
This is often useful for \code{\link[teal]{choices_selected}} as it marks up the dropdown boxes
for \code{\link[shiny]{selectInput}}.
}
\details{
If either \code{choices} or \code{labels} are factors, they are coerced 
  to character.
  
  Duplicated elements from \code{choices} get removed.
}
\examples{

library(random.cdisc.data)
library(tern)
ADSL <- radsl(N=10, seed = 1)
ADTTE <- radtte(ADSL, seed = 1)

choices1 <- named_choices(names(ADSL), tern::var_labels(ADSL))
choices2 <- named_choices(ADTTE$PARAMCD, ADTTE$PARAM)

# if only a subset of variables are of intereste then the following code will work
vl_ADSL <- tern::var_labels(ADSL)
vl_arm <- vl_ADSL[c("ARM", "ARMCD")]
choices3 <- named_choices(names(vl_arm), vl_arm)


\dontrun{

library(shiny)

shinyApp(
  ui = fluidPage(selectInput("c1", label = "Choices from ADSL",
                             choices = choices1,
                             selected = choices1[1]),
                 
                 selectInput("c2", label = "Choices from ADTTE",
                             choices = choices2,
                             selected = choices2[1])),
  
  server = function(input, output) {}
)

}
}
