% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NamedDataset.R
\name{named_dataset}
\alias{named_dataset}
\title{Constructor for \link{NamedDataset} class}
\usage{
named_dataset(
  dataname,
  x,
  code = character(0),
  label = character(0),
  vars = list()
)
}
\arguments{
\item{dataname}{(\code{character})\cr
A given name for the dataset it may not contain spaces}

\item{x}{(\code{data.frame} or \code{rtable}) object}

\item{code}{(\code{character})\cr
A character string defining the code needed to produce the data set in \code{x}}

\item{label}{(\code{character})\cr
Label to describe the dataset}

\item{vars}{(named \code{list})) \cr
In case when this object code depends on the \code{raw_data} from the other
\code{RelationalDataset} object(s) or other constant value,
this/these object(s) should be included as named element of the list.}
}
\value{
\link{NamedDataset} object
}
\description{
\ifelse{html}{\out{<a href='https://www.tidyverse.org/lifecycle/#experimental'><img src='figures/lifecycle-experimental.svg' alt='Experimental lifecycle'></a>}}{\strong{Experimental}}
}
\examples{
library(random.cdisc.data)
ADSL <- radsl(cached = TRUE)
ADSL_dataset <- named_dataset(dataname = "ADSL", x = ADSL)

ADSL_dataset$get_dataname()

ADSL_dataset <- named_dataset(
  dataname = "ADSL",
  x = ADSL,
  label = "AdAM subject-level dataset",
  code = "ADSL <- radsl(cached = TRUE)"
)

ADSL_dataset$get_dataset_label()
ADSL_dataset$get_code()

}
