% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{named_list_to_dataset}
\alias{named_list_to_dataset}
\title{Wrap the contents of list in \code{cdisc_dataset} or \code{dataset}.}
\usage{
named_list_to_dataset(named_list)
}
\arguments{
\item{named_list}{\code{list} returned by \code{list_to_named_list} function.}
}
\value{
list of \code{Dataset}.
}
\description{
Wrap the contents of list in \code{cdisc_dataset} or \code{dataset}.
}
\details{
Takes the output of \code{list_to_named_list} function and wraps the content in
\code{cdisc_dataset} or \code{dataset} to return a list of \code{Dataset}.
}
