% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tdata.R
\name{new_tdata}
\alias{new_tdata}
\title{Create a new \code{tdata} object which contains}
\usage{
new_tdata(data, code = "", join_keys = NULL)
}
\arguments{
\item{data}{A \verb{named list} of \code{data.frames} which optionally can be \code{reactive}.
Inside this object all \code{data.frames} will be made \code{reactive}.}

\item{code}{A \code{character} (or \code{reactive} which evaluates to a \code{character}) containing
the code used to generate the data. This should be \code{reactive} if the code is changing
during a reactive context (e.g. if filtering changes the code). Inside this
object \code{code} will be made reactive}

\item{join_keys}{A \code{teal.data::JoinKeys} object containing relationships between the
datasets.}
}
\value{
A \code{tdata} object
}
\description{
\itemize{
\item \code{data} a \code{reactive} list of data.frames with attributes
i) \code{code} (\code{reactive}) containing code used to generate the data
and ii) join_keys (\code{JoinKeys}) containing the relationships between
the data
}
}
\examples{

data <- new_tdata(
  data = list(iris = iris, mtcars = reactive(mtcars), dd = data.frame(x = 1:10)),
  code = "iris <- iris
    mtcars <- mtcars
    dd <- data.frame(x = 1:10)"
)

# Extract a data.frame
isolate(data[["iris"]]())

}
