% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choices_selected.R
\name{no_selected_as_NULL}
\alias{no_selected_as_NULL}
\title{Check select choices for no choice made}
\usage{
no_selected_as_NULL(x)
}
\arguments{
\item{x}{(\code{character}) Word that shall be checked for
NULL, empty, "--no-selection"}
}
\value{
the word or NULL
}
\description{
\ifelse{html}{\out{<a href='https://www.tidyverse.org/lifecycle/#maturing'><img src='figures/lifecycle-maturing.svg' alt='Maturing lifecycle'></a>}}{\strong{Maturing}}
}
