% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optionalInput.R
\name{optionalSelectInput}
\alias{optionalSelectInput}
\title{Hide, Show Label only or display a pickerInput}
\usage{
optionalSelectInput(inputId, label, choices = NULL, selected = NULL,
  multiple = FALSE, options = list(), ..., label_help = NULL)
}
\arguments{
\item{inputId}{The \code{input} slot that will be used to access the value.}

\item{label}{Display a text in the center of the switch.}

\item{choices}{character vector or \code{NULL}. If \code{choices} is
\code{NULL} no pickerInput widget is displayed and \code{input[[inputId]]}
will be \code{""}. If \code{choices} is of length 1 then a label and
character string will be displayed and the pickerInput widget will be
hidden. If the length of \code{choices} is more than one the pickerInput
element will be displayed.}

\item{selected}{The initially selected value (or multiple values if multiple = TRUE).
If not specified then defaults to the first value for single-select lists
 and no values for multiple select lists.}

\item{multiple}{Is selection of multiple items allowed?}

\item{options}{List of options, see \link{pickerOptions} for all available options.
For limit the number of selections, see example below.}

\item{...}{arguments passed to \code{\link[shinyWidgets]{pickerInput}}}

\item{label_help}{optional an object of class \code{shiny.tag}. E.g. an object
returned by \code{\link[shiny]{helpText}}}
}
\description{
Hidden input widgets are useful to have the \code{input[[inputId]]} variable
on available in the server funciont but no corresponding visual clutter from
input widgets that provide only a single choice.
}
\examples{

optionalSelectInput("xvar", "x variable", 'A', 'A')
optionalSelectInput("xvar", "x variable", LETTERS[1:5], 'A')

}
