% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/header_filters.R
\name{output_header}
\alias{output_header}
\title{Generates header text for analysis items}
\usage{
output_header(title, description, libraries = c("ggplot2", "haven", "dplyr"),
  git_repo = NULL, data = c())
}
\arguments{
\item{title}{A character title of the module}

\item{description}{A character description of the module with additional
information not reflected in the title}

\item{libraries}{A character vector of the names of libraries required for
the module. Defaults to c("ggplot2", "haven", "dplyr")}

\item{git_repo}{A character vector of the link to the GitHub repository
containing the script for the module. Optional argument}

\item{data}{A character vector listing the code for data variable assignments,
one for each dataset and with the file paths to the datasets
(e.g. data = "ASL <- haven::read_sas("/opt/BIOSTAT/qa/cdt7876a/libraries/asl.sas7bdat")").
Note that if any of the dataset paths are not found under /opt/BIOSTAT/...,
then \code{output_header} would not generate the text corresponding to the
code for reading in data, and user would need to manually add it within
the module code. Defaults to c(), or no data lines generated}
}
\value{
A character string for the header text
}
