% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rcode_utils.R
\name{pad}
\alias{pad}
\title{Pads a string}
\usage{
pad(str, pre = NULL, post = "")
}
\arguments{
\item{str}{(\code{character}) vector of lines to be padded}

\item{pre}{(\code{character}) elements to be appended before \code{str}}

\item{post}{(\code{character}) elements to be appended after \code{str}}
}
\description{
Including elements before or after string. If NULL is provided no elements included. Padding in this case means
appending additional element before or after \code{character} vector.
}
\keyword{internal}
