% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_code_chunks.R
\name{parse_code_chunks}
\alias{parse_code_chunks}
\title{Parse Code Chunks}
\usage{
parse_code_chunks(file)
}
\arguments{
\item{file}{path to the file containing the code chunks}
}
\value{
A list of individual code chunks, with each element consisting
  of a character vector of length the number of lines from `file`
  that corresponds to the specific code chunk
}
\description{
Save code chunks between specified start and end tags
  from a code script as a list of individual code chunks, each named
  according to the code chunk name included in the tags
}
\details{
Code chunks should be non-overlapping, start with a valid start tag
  beginning after '#' (e.g. "# @start_chunk1") and end with a valid end tag,
  with non-duplicated code chunk names that adhere to variable naming rules in R.
  Otherwise, the function will throw an error
}
\examples{

\dontrun{

file <- "~/test_r_chunk.R"
file <- "~/test_spaghetti_chunk.R"

parse_code_chunks(file)

}

}

