% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optionalInput.R
\name{picker_options}
\alias{picker_options}
\title{Create \code{choicesOpt} for \code{pickerInput}}
\usage{
picker_options(choices)
}
\arguments{
\item{choices}{(\code{choices_labeled} or \code{character} vector) choices vector}
}
\value{
(\code{list}) to be passed as \code{choicesOpt} argument
}
\description{
Create \code{choicesOpt} for \code{pickerInput}
}
\keyword{internal}
