% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optionalInput.R
\name{picker_options_content}
\alias{picker_options_content}
\title{Optional content for \code{optionalSelectInput}}
\usage{
picker_options_content(var_name, var_label, var_type)
}
\arguments{
\item{var_name}{(atomic vector of \code{character}) variable name}

\item{var_label}{(atomic vector of \code{character})
variable alternative name - for example variable label}

\item{var_type}{(atomic vector of \code{character})
class of the variable.}
}
\value{
(atomic vector of \code{character}) HTML contents with all elements combined
}
\description{
Prepares content to be displayed in \code{optionalSelectInput} with icons and labels
}
\examples{
teal:::picker_options_content(
  var_name = c("SEX", "BRMRKR1"),
  var_label = c("Sex", "Biomarker 1"),
  var_type = c("factor", "numeric")
)
}
\keyword{internal}
