% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pluralize}
\alias{pluralize}
\title{Pluralize a word depending on the size of the input}
\usage{
pluralize(x, singular, plural = NULL)
}
\arguments{
\item{x}{(\code{object}) to check length for plural.}

\item{singular}{(\code{character}) singular form of the word.}

\item{plural}{(optional \code{character}) plural form of the word. If not given an "s"
is added to the singular form.}
}
\value{
A \code{character} that correctly represents the size of the \code{x} argument.
}
\description{
Pluralize a word depending on the size of the input
}
\keyword{internal}
