% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_data.R
\name{prefix_col_names_with_dataset}
\alias{prefix_col_names_with_dataset}
\title{Prefix column names with dataset}
\usage{
prefix_col_names_with_dataset(data, add_rowid = FALSE)
}
\arguments{
\item{data}{(\code{data.frame}) with attributes \code{keys} and \code{dataname}}

\item{add_rowid}{(\code{logical}) Whether to add rowid's for safer comparison. (should
always be set to \code{TRUE} internally)}
}
\description{
Prefix column names with dataset
}
