% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeClass.R
\name{pretty_code_string}
\alias{pretty_code_string}
\title{Format a vector of code into a string}
\usage{
pretty_code_string(code_vector)
}
\arguments{
\item{code_vector}{(\code{character}) vector containing lines of
code to format into a string.}
}
\value{
(\code{character}) string containing the formatted code.
}
\description{
Format a vector of code into a string
}
\keyword{internal}
