% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choices_labeled.R
\name{print.choices_labeled}
\alias{print.choices_labeled}
\title{Print choices_labeled object}
\usage{
\method{print}{choices_labeled}(x, ...)
}
\arguments{
\item{x}{an object used to select a method.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Print choices_labeled object
}
