% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_skeleton.R
\name{project_skeleton}
\alias{project_skeleton}
\title{This function will be called when the user invokes
the New Project wizard using the project template defined in the template file
at:}
\usage{
project_skeleton(path, use_cdisc = TRUE)
}
\description{
inst/rstudio/templates/project/teal_application.dcf
}
\details{
It is exported so that projects can be created outside of teal
}
