% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DatasetConnector_constructors.R
\name{python_cdisc_dataset_connector}
\alias{python_cdisc_dataset_connector}
\title{\code{Python} \code{CDISCDatasetConnector}}
\usage{
python_cdisc_dataset_connector(
  dataname,
  file,
  code,
  object,
  keys = get_cdisc_keys(dataname),
  parent = if (identical(dataname, "ADSL")) character(0L) else "ADSL",
  mutate_code = character(0),
  mutate_script = character(0),
  label = character(0),
  vars = list()
)
}
\arguments{
\item{dataname}{(\code{character})\cr
A given name for the dataset it may not contain spaces}

\item{file}{(\code{character})\cr
Path to the file location containing the python script used to generate the object.}

\item{code}{(\code{character})\cr
string containing the python code to be run using \code{reticulate}. Carefully consider
indentation to follow proper python syntax.}

\item{object}{(\code{character})\cr
name of the object from the python script that is assigned to the dataset to be used.}

\item{keys}{optional, (\code{character})\cr
vector of dataset primary keys column names}

\item{parent}{optional, (\code{character}) \cr
parent dataset name}

\item{mutate_code}{(\code{character})\cr
String containing the code used to mutate the object
after it is produced.}

\item{mutate_script}{(\code{character})\cr
Alternatively to \code{mutate_code} - location of the file containing modification code.
Can't be used simultaneously with \code{mutate_script}.}

\item{label}{(\code{character})\cr
Label to describe the dataset.}

\item{vars}{(named \code{list})) \cr
In case when this object code depends on other \code{Dataset} object(s) or
other constant value, this/these object(s) should be included as named
element(s) of the list. For example if this object code needs \code{ADSL}
object we should specify \verb{vars = list(ADSL = <adsl object>)}.
It's recommended to include \code{Dataset} or \code{DatasetConnector} objects to
the \code{vars} list to preserve reproducibility. Please note that \code{vars}
are included to this object as local \code{vars} and they cannot be modified
within another dataset.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
Create a \code{CDISCDatasetConnector} from \code{.py} file or through python code supplied directly.
}
