% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CallablePythonCode.R
\name{PythonCodeClass}
\alias{PythonCodeClass}
\title{A \code{CallablePythonCode} class of objects}
\description{
\ifelse{html}{\out{<a href='https://www.tidyverse.org/lifecycle/#experimental'><img src='figures/lifecycle-experimental.svg' alt='Experimental lifecycle'></a>}}{\strong{Experimental}}
}
\section{Super class}{
\code{\link[teal:CodeClass]{teal::CodeClass}} -> \code{PythonCodeClass}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-eval}{\code{PythonCodeClass$eval()}}
\item \href{#method-clone}{\code{PythonCodeClass$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="teal" data-topic="CodeClass" data-id="append">}\href{../../teal/html/CodeClass.html#method-append}{\code{teal::CodeClass$append()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="teal" data-topic="CodeClass" data-id="get_code">}\href{../../teal/html/CodeClass.html#method-get_code}{\code{teal::CodeClass$get_code()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="teal" data-topic="CodeClass" data-id="initialize">}\href{../../teal/html/CodeClass.html#method-initialize}{\code{teal::CodeClass$initialize()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="teal" data-topic="CodeClass" data-id="set_code">}\href{../../teal/html/CodeClass.html#method-set_code}{\code{teal::CodeClass$set_code()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-eval"></a>}}
\if{latex}{\out{\hypertarget{method-eval}{}}}
\subsection{Method \code{eval()}}{
Evaluates internal code within global environment
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PythonCodeClass$eval(vars = list(), dataname = NULL, envir = .GlobalEnv)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{vars}}{(named \code{list}) additional pre-requisite vars to execute code}

\item{\code{dataname}}{(\code{character}) name of the data frame object to be returned}

\item{\code{envir}}{(\code{environment}) environment in which code will be evaluated}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{data.frame} containing the mutated dataset
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PythonCodeClass$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
