% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Queue.R
\name{Queue}
\alias{Queue}
\title{R6 Class - A First-In-First-Out Abstract Data Type}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Implements the queue abstract data type. The last element added to this queue is
the last one to be returned from it.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-push}{\code{Queue$push()}}
\item \href{#method-pop}{\code{Queue$pop()}}
\item \href{#method-empty}{\code{Queue$empty()}}
\item \href{#method-size}{\code{Queue$size()}}
\item \href{#method-get}{\code{Queue$get()}}
\item \href{#method-remove}{\code{Queue$remove()}}
\item \href{#method-print}{\code{Queue$print()}}
\item \href{#method-clone}{\code{Queue$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-push"></a>}}
\if{latex}{\out{\hypertarget{method-push}{}}}
\subsection{Method \code{push()}}{
Adds another element to this queue.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Queue$push(new_elements)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{new_elements}}{the elements to be added to this queue}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
invisibly self
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-pop"></a>}}
\if{latex}{\out{\hypertarget{method-pop}{}}}
\subsection{Method \code{pop()}}{
Returns the first added element to this queue and removes it
from this queue.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Queue$pop()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
the first added element to this queue or \code{NULL} if this queue is empty
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-empty"></a>}}
\if{latex}{\out{\hypertarget{method-empty}{}}}
\subsection{Method \code{empty()}}{
Removes all elements from this queue.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Queue$empty()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
invisibly self
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-size"></a>}}
\if{latex}{\out{\hypertarget{method-size}{}}}
\subsection{Method \code{size()}}{
Returns the number of elements in this queue.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Queue$size()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
the number of elements in this queue
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get"></a>}}
\if{latex}{\out{\hypertarget{method-get}{}}}
\subsection{Method \code{get()}}{
Returns an array of elements in this queue. The order of elements is chronological:
the first elements in the returned array is the oldest element added to this queue.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Queue$get(reversed = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{reversed}}{(\code{logical})\cr
if TRUE then returns the First-In-First-Out order; otherwise returns the Last-In-First-Out order.
Default: FALSE}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the array of elements in this queue
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-remove"></a>}}
\if{latex}{\out{\hypertarget{method-remove}{}}}
\subsection{Method \code{remove()}}{
Removes the eldest occurrence of elements from this queue. Relies on implicit
conversions of R types to compare a removed element with elements in this queue.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Queue$remove(elements)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{elements}}{the elements to remove from this queue}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
invisibly self
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
Prints this queue.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Queue$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{the additional arguments to this method, ignored}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
invisibly self
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Queue$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
