% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RawDataset.R
\name{raw_dataset}
\alias{raw_dataset}
\title{Constructor for \link{RawDataset} object}
\usage{
raw_dataset(x)
}
\arguments{
\item{x}{(\code{data.frame} or \code{rtable}) object}
}
\description{
\ifelse{html}{\out{<a href='https://www.tidyverse.org/lifecycle/#experimental'><img src='figures/lifecycle-experimental.svg' alt='Experimental lifecycle'></a>}}{\strong{Experimental}}
}
\examples{
library(random.cdisc.data)
ADSL <- radsl(cached = TRUE)
ADSL_dataset <- raw_dataset(x = ADSL)

ADSL_dataset$get_raw_data()

ADSL_dataset$colnames

}
