% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RelationalDatasetConnector_constructors.R
\name{raw_dataset_connector}
\alias{raw_dataset_connector}
\title{Create \code{RawDatasetConnector} object}
\usage{
raw_dataset_connector(pull_callable)
}
\arguments{
\item{pull_callable}{(\code{CallableFunction})\cr
function with necessary arguments set to fetch data from connection.}
}
\value{
\code{RawDatasetConnector} object
}
\description{
\ifelse{html}{\out{<a href='https://www.tidyverse.org/lifecycle/#experimental'><img src='figures/lifecycle-experimental.svg' alt='Experimental lifecycle'></a>}}{\strong{Experimental}}
Create \link{RawDatasetConnector} object to execute specific call to fetch data
}
\examples{
ds <- raw_dataset_connector(pull_callable = callable_function(data.frame))
set_args(ds, list(x = 1:5, y = letters[1:5], stringsAsFactors = FALSE))
ds$pull()
ds$get_raw_data()
ds$get_code()
}
