% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RawDatasetConnector.R
\name{RawDatasetConnector}
\alias{RawDatasetConnector}
\title{A \code{RawDatasetConnector} class of objects}
\description{
A \code{RawDatasetConnector} class of objects

A \code{RawDatasetConnector} class of objects
}
\details{
Objects of this class store the connection function to fetch a single dataset.
Note that for some specific connection types (e.g. \code{RICE} or \code{SAICE}),
an object of class \code{DataConnection} must be provided.
Data can be pulled via the \code{pull} method and accessed directly
through the \code{dataset} active binding.
Pulled data inherits from the class \link{RelationalDataset}.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{RawDatasetConnector$new()}}
\item \href{#method-set_args}{\code{RawDatasetConnector$set_args()}}
\item \href{#method-get_code}{\code{RawDatasetConnector$get_code()}}
\item \href{#method-get_dataset}{\code{RawDatasetConnector$get_dataset()}}
\item \href{#method-get_pull_fun}{\code{RawDatasetConnector$get_pull_fun()}}
\item \href{#method-get_raw_data}{\code{RawDatasetConnector$get_raw_data()}}
\item \href{#method-pull_dataset}{\code{RawDatasetConnector$pull_dataset()}}
\item \href{#method-clone}{\code{RawDatasetConnector$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
Create a new \code{RawDatasetConnector} object. Set the pulling function
\link{CallableFunction} which returns a \code{data.frame}, e.g. by reading
from a function or creating it on the fly.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RawDatasetConnector$new(pull_fun)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pull_fun}}{(\code{CallableFunction})\cr
function to pull the data.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
new \code{RawDatasetConnector} object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_args"></a>}}
\subsection{Method \code{set_args()}}{
Set arguments to the pulling function
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RawDatasetConnector$set_args(args)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{args}}{(\code{NULL} or named \code{list}) dynamic arguments to function}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
nothing
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_code"></a>}}
\subsection{Method \code{get_code()}}{
Get code to get data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RawDatasetConnector$get_code(deparse = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deparse}}{(\code{logical})\cr
whether return deparsed form of a call}

\item{\code{args}}{(\code{NULL} or named \code{list})\cr
dynamic arguments to function which loads data}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
optionally deparsed \code{call} object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_dataset"></a>}}
\subsection{Method \code{get_dataset()}}{
Get dataset
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RawDatasetConnector$get_dataset()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
dataset (\code{<...>Dataset})
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_pull_fun"></a>}}
\subsection{Method \code{get_pull_fun()}}{
Get pull function
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RawDatasetConnector$get_pull_fun()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
dataset (\code{CallableFunction})
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_raw_data"></a>}}
\subsection{Method \code{get_raw_data()}}{
Get raw data from dataset
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RawDatasetConnector$get_raw_data()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
data (data.frame)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-pull_dataset"></a>}}
\subsection{Method \code{pull_dataset()}}{
Pull the data

Read or create data using \code{pull_fun} specified in the constructor.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RawDatasetConnector$pull_dataset(args = NULL, try = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{args}}{(\code{NULL} or named \code{list})\cr
additional dynamic arguments for pull function. \code{args} can be omitted if \code{pull_fun}
from constructor already contains all necessary arguments to pull data. One can try
to execute \code{pull_fun} directly by \code{x$pull_fun$run()} or to get code using
\code{x$pull_fun$get_code()}. \code{args} specified in pull are used temporary to get data but
not saved in code.}

\item{\code{try}}{(\code{logical}) whether perform function evaluation inside \code{try} clause}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
nothing, in order to get the data please use \code{get_data} method
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RawDatasetConnector$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
