% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RelationalDataConnector.R
\name{rcd_cdisc_data}
\alias{rcd_cdisc_data}
\title{Data connector for \code{random.cdisc.data}}
\usage{
rcd_cdisc_data(..., check = TRUE)
}
\arguments{
\item{...}{(\code{DatasetConnector}) dataset connectors created using \link{rcd_dataset_connector}}

\item{check}{optional, (\code{logical}) whether perform reproducibility check}
}
\value{
An object of class \code{RelationalDataConnector}
}
\description{
Build data connector for \code{random.cdisc.data} functions or datasets
}
\examples{
\dontrun{
library(random.cdisc.data)
x <- rcd_cdisc_data(
  rcd_cdisc_dataset_connector("ADSL", radsl, cached = TRUE),
  rcd_cdisc_dataset_connector("ADLB", radlb, cached = TRUE)
)
app <- init(
  data = x,
  modules = root_modules(
    module(
      "ADSL AGE histogram",
      server = function(input, output, session, datasets) {
        output$hist <- renderPlot(
          hist(datasets$get_dataset("ADSL", filtered = TRUE, reactive = TRUE)$AGE)
        )
      },
      ui = function(id, ...) {ns <- NS(id); plotOutput(ns('hist'))},
      filters = "ADSL"
    )
  ),
  header = tags$h1("Sample App")
)
shinyApp(app$ui, app$server)
}
}
