% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DatasetConnector_constructors.R
\name{rcd_cdisc_dataset_connector}
\alias{rcd_cdisc_dataset_connector}
\title{\code{RCD} \code{CDISCDatasetConnector}}
\usage{
rcd_cdisc_dataset_connector(
  dataname,
  fun,
  keys = get_cdisc_keys(dataname),
  parent = if (identical(dataname, "ADSL")) character(0L) else "ADSL",
  label = character(0),
  code = character(0),
  script = character(0),
  ...
)
}
\arguments{
\item{dataname}{(\code{character})\cr
A given name for the dataset it may not contain spaces}

\item{fun}{(\code{function})\cr
any R function which generates \code{data.frame}, especially functions from
\code{random.cdisc.data} like \code{\link[random.cdisc.data]{radsl}}}

\item{keys}{optional, (\code{character})\cr
vector of dataset primary keys column names}

\item{parent}{optional, (\code{character}) \cr
parent dataset name}

\item{label}{(\code{character})\cr
Label to describe the dataset.}

\item{code}{(\code{character})\cr
A character string defining code to modify \code{raw_data} from this dataset. To modify
current dataset code should contain at least one assignment to object defined in \code{dataname}
argument. For example if \code{dataname = ADSL} example code should contain
\verb{ADSL <- <some R code>}. Can't be used simultaneously with \code{script}}

\item{script}{(\code{character})\cr
Alternatively to \code{code} - location of the file containing modification code.
Can't be used simultaneously with \code{script}.}

\item{...}{(\code{optional})\cr
Additional arguments applied to pull function.
In case when this object code depends on the \code{raw_data} from the other
\code{Dataset}, \code{DatasetConnector} object(s) or other constant value,
this/these object(s) should be included. Please note that \code{vars}
are included to this object as local \code{vars} and they cannot be modified
within another dataset.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\details{
Create a \code{CDISCDatasetConnector} from function in \code{random.cdisc.data}.
}
