% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataConnector.R
\name{rcd_connector}
\alias{rcd_connector}
\title{Dummy random CDISC data connector}
\usage{
rcd_connector()
}
\description{
Dummy random CDISC data connector
}
\examples{
app <- init(
  data = rcd_connector(),
  modules = root_modules(
    module(
      "ADSL AGE histogram",
      server = function(input, output, session, datasets) {
        output$hist <- renderPlot(
          hist(datasets$get_data("ADSL", filtered = TRUE, reactive = TRUE)$AGE)
        )
      },
      ui = function(id, ...) {ns <- NS(id); plotOutput(ns('hist'))},
      filters = "ADSL"
    )
  ),
  filter = NULL,#list(init = list(ADSL = c("AGE"))),
  header = tags$h1("Sample App"),
  footer = tags$p("Copyright 2017")
)
\dontrun{
 shinyApp(app$ui, app$server)
}
}
