% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataConnector.R
\name{rcd_data}
\alias{rcd_data}
\title{Data connector for \code{random.cdisc.data}}
\usage{
rcd_data(..., code = character(0), check = TRUE)
}
\arguments{
\item{...}{(\code{DatasetConnector}) dataset connectors created using \link{rcd_dataset}}

\item{code}{optional, (\code{character}) preprocessing code}

\item{check}{optional, (\code{logical}) whether perform reproducibility check}
}
\value{
An object of class \code{DataConnector}
}
\description{
Build data connector for \code{random.cdisc.data} functions or datasets
}
\examples{
\dontrun{
x <- rcd_data(ADSL = rcd_dataset("ADSL", radsl), ADLB = rcd_dataset("ADLB", radlb))
app <- init(
  data = x,
  modules = root_modules(
    module(
      "ADSL AGE histogram",
      server = function(input, output, session, datasets) {
        output$hist <- renderPlot(
          hist(datasets$get_data("ADSL", filtered = TRUE, reactive = TRUE)$AGE)
        )
      },
      ui = function(id, ...) {ns <- NS(id); plotOutput(ns('hist'))},
      filters = "ADSL"
    )
  ),
  filter = NULL,
  header = tags$h1("Sample App")
)
shinyApp(app$ui, app$server)
}
}
