% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DatasetConnector.R
\name{rcd_dataset}
\alias{rcd_dataset}
\title{Set up connection to \code{random.cdisc.data}}
\usage{
rcd_dataset(dataname, fun, cached = TRUE, seed = 1)
}
\arguments{
\item{dataname}{name of dataset}

\item{fun}{(\code{function}) connection function}

\item{cached}{(\code{logical}) whether to use cached data}

\item{seed}{(\code{numeric}) random seed if used non-cached data}
}
\value{
(\code{DatasetConnector}) type of object
}
\description{
Set up connection to \code{random.cdisc.data}
}
\examples{
library(random.cdisc.data)
x <- rcd_dataset("ADSL", radsl)
x$get_call()
x$get_data()
}
