% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataConnector.R
\name{rds_cdisc_data}
\alias{rds_cdisc_data}
\title{Data connector for \code{.rds} files}
\usage{
rds_cdisc_data(..., code = character(0), check = TRUE)
}
\arguments{
\item{...}{(\code{DatasetConnector}) dataset connectors created using \link{rds_cdisc_dataset}}

\item{code}{optional, (\code{character}) preprocessing code}

\item{check}{optional, (\code{logical}) whether perform reproducibility check}
}
\value{
An object of class \code{DataConnector}
}
\description{
Build data connector for RDS file connections
}
\examples{
\dontrun{
x <- rds_cdisc_dataset("ADSL", "/path/to/file.rds")
app <- init(
  data = rds_cdisc_data(x, check = TRUE),
  modules = root_modules(
    module(
      "ADSL AGE histogram",
      server = function(input, output, session, datasets) {
        output$hist <- renderPlot(
          hist(datasets$get_data("ADSL", filtered = TRUE, reactive = TRUE)$AGE)
        )
      },
      ui = function(id, ...) {ns <- NS(id); plotOutput(ns('hist'))},
      filters = "ADSL"
    )
  ),
  filter = NULL,
  header = tags$h1("Sample App")
)
shinyApp(app$ui, app$server)
}
}
