% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DatasetConnector.R
\name{rds_cdisc_dataset}
\alias{rds_cdisc_dataset}
\title{Set up connection to local \code{rds} file}
\usage{
rds_cdisc_dataset(dataname, file, keys = get_cdisc_keys(dataname))
}
\arguments{
\item{dataname}{name of dataset}

\item{file}{(\code{character}) path to \code{.rds} that contains
a single data.frame stored by \code{saveRDS}}

\item{keys}{list of keys -

Please note that the order of keys is important.

Keys are not matched by name but by order. In case you want to perform:

\code{SELECT ...}
\code{  FROM ADSL}
\code{LEFT JOIN ADTTE on ADTTE.STUDYID = ADSL.STUDYID and}
\code{                                     ADTTE.SUBJECTUNIQUEID = ADSL.USUBJID}

the keys need to be

ADSL \code{primary: USUBJID, STUDYID}

ADTTE \code{primary: PARAMCD, foreign: SUBJECTUNIQUEID, STUDYID, parent: ADSL}

As you can see the names of foreign ADTTE are different to ADSL keys, but the order and length is equal.}
}
\value{
(\code{DatasetConnector}) type of object
}
\description{
Set up connection to local \code{rds} file
}
\examples{
\dontrun{
x <- rds_cdisc_dataset("ADSL", "/path/to/file.rds")
x$get_call()
x$get_data()
}
}
