% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DatasetConnector.R
\name{rds_dataset}
\alias{rds_dataset}
\title{Set up connection to local \code{rds} file}
\usage{
rds_dataset(dataname, file)
}
\arguments{
\item{dataname}{name of dataset}

\item{file}{(\code{character}) file path}
}
\value{
(\code{DatasetConnector}) type of object
}
\description{
Set up connection to local \code{rds} file
}
\examples{
\dontrun{
x <- rds_dataset("ADSL", "/path/to/file.rds")
x$get_call()
x$get_data()
}
}
