% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_code.R
\name{read_lib_names}
\alias{read_lib_names}
\title{Libraries names from code}
\usage{
read_lib_names(code)
}
\arguments{
\item{code}{(\code{character}) containing preprocessing code lines.}
}
\value{
libraries names loaded in preprocessing code
}
\description{
Reads library names from code
}
