% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_code.R
\name{read_script}
\alias{read_script}
\title{Read .R file into character}
\usage{
read_script(file)
}
\arguments{
\item{file}{(\code{character}) File to be parsed into code}
}
\value{
character vector withe the code
}
\description{
Comments will be excluded
}
\examples{
file_example <- tempfile()
writeLines(c("x <- 2", "#second line comment", "x <- x + 2"), file_example)

read_script(file_example)
}
