% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/register_logger.R
\name{register_logger}
\alias{register_logger}
\title{Registers a logger instance in a given logging namespace.}
\usage{
register_logger(namespace = NA_character_)
}
\arguments{
\item{namespace}{(\code{character} or \code{NA}) the name of the logging namespace}
}
\value{
\code{invisible(NULL)}
}
\description{
Registers a logger instance in a given logging namespace.
}
\details{
Creates a new logging namespace specified by the \code{namespace} argument.
Sets the logging level to \code{logger::INFO} and updates the logging layout.
Uses \code{getOption(teal_logging_layout)} to determine the layout of logs and
passes it to \code{\link[logger:layout_glue_generator]{logger::layout_glue_generator}}.
The logs are output to \code{stdout} by default. Check \code{logger} for more information
about layouts and how to use \code{logger}.
}
\note{
It's a thin wrapper around the \code{logger} package.
}
\examples{
register_logger(namespace = "new_namespace")
\dontrun{
logger::log_info("Hello from new_namespace")
}

}
