% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_code_chunks.R
\name{reindent}
\alias{reindent}
\title{Function for adding extra whitespaces at the beginning for indentation}
\usage{
reindent(str, nws)
}
\arguments{
\item{str}{Character string}

\item{nws}{Number of whitespaces to add in front of \code{str}}
}
\description{
Function for adding extra whitespaces at the beginning for indentation
}
