% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RelationalDataConnector.R
\name{relational_data_connector}
\alias{relational_data_connector}
\title{Public facing object constructor for \code{RelationalDataConnector} class.}
\usage{
relational_data_connector(connection, connectors)
}
\arguments{
\item{connection}{(\code{DataConnection})\cr
connection to data source}

\item{connectors}{(\code{list} of \code{DatasetConnector} elements)\cr
list with dataset connectors}
}
\value{
\code{RelationalDataConnector} object
}
\description{
Public facing object constructor for \code{RelationalDataConnector} class.
}
