% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DatasetConnector_constructors.R
\name{relational_dataset_connector}
\alias{relational_dataset_connector}
\alias{rcd_dataset_connector}
\alias{rds_dataset_connector}
\alias{script_dataset_connector}
\alias{rice_dataset_connector}
\alias{rds_cdisc_dataset_connector}
\alias{rcd_cdisc_dataset_connector}
\alias{rice_cdisc_dataset_connector}
\alias{script_cdisc_dataset_connector}
\title{Create a new \code{RelationalDatasetConnector} object}
\usage{
relational_dataset_connector(
  pull_fun,
  dataname,
  keys,
  code = character(0),
  label = character(0),
  ...
)

rcd_dataset_connector(
  dataname,
  fun,
  keys,
  code = character(0),
  script = character(0),
  label = character(0),
  ...
)

rds_dataset_connector(
  dataname,
  file,
  keys,
  code = character(0),
  script = character(0),
  label = character(0),
  ...
)

script_dataset_connector(
  dataname,
  file,
  keys,
  code = character(0),
  script = character(0),
  label = character(0),
  ...
)

rice_dataset_connector(
  dataname,
  path,
  keys,
  code = character(0),
  script = character(0),
  label = character(0),
  ...
)

rds_cdisc_dataset_connector(
  dataname,
  file,
  code = character(0),
  script = character(0),
  label = character(0),
  ...
)

rcd_cdisc_dataset_connector(
  dataname,
  fun,
  code = character(0),
  script = character(0),
  label = character(0),
  ...
)

rice_cdisc_dataset_connector(
  dataname,
  path,
  code = character(0),
  script = character(0),
  label = character(0),
  ...
)

script_cdisc_dataset_connector(
  dataname,
  file,
  keys,
  code = character(0),
  script = character(0),
  label = character(0),
  ...
)
}
\arguments{
\item{pull_fun}{(\code{CallableFunction})\cr
Set the pulling function \link{CallableFunction} to load \code{data.frame}.
\code{dataname} will be used as name of object to be assigned.}

\item{dataname}{(\code{character})\cr
A given name for the dataset it may not contain spaces}

\item{keys}{(\code{keys})\cr
object of S3 class keys containing foreign, primary keys and parent information}

\item{code}{(\code{character})\cr
A character string defining the code needed to produce the data set in \code{x}}

\item{label}{(\code{character})\cr
Label to describe the dataset}

\item{...}{(optional)\cr
additional arguments passed to function loading data.}

\item{fun}{(\code{function})\cr
any R function which generates \code{data.frame}, especially functions from
\code{random.cdisc.data} like \code{\link[random.cdisc.data]{radsl}}}

\item{script}{(\code{character})\cr
file that contains R Code that can be read using \link{read_script}.
Preferred before \code{code} argument}

\item{file}{(\code{character})\cr
path to (\code{.rds} or \code{.R}) that contains \code{data.frame} object or
code to \code{source}}

\item{path}{(\code{character})\cr
path to the file}
}
\value{
new \code{RawDatasetConnector} object
}
\description{
Create \code{RelationalDatasetConnector} from \link{callable_function}.

\code{rcd_dataset_connector} - Create a \code{RelationalDatasetConnector} from any R function.

\code{rds_dataset_connector} - Create a \code{RelationalDatasetConnector} from \code{RDS} file.

\code{rds_dataset_connector} - Create a \code{RelationalDatasetConnector} from \code{RDS} file.

\code{rice_dataset_connector} -
Create a \code{RelationalDatasetConnector} from \code{RICE} dataset.

\code{rds_cdisc_dataset_connector} -
Create a \code{RelationalDatasetConnector} from \code{RDS} file with keys automatically
assigned by \code{dataname}

\code{rcd_cdisc_dataset_connector} -
Create a \code{DatasetConnector} from any R function with keys assigned automatically
by \code{dataname}.

\code{rice_cdisc_dataset_connector} -
Create a \code{RelationalDatasetConnector} from \code{RICE} dataset with keys assigned
automatically by \code{dataname}.

\code{script_cdisc_dataset_connector} -
Create a \code{RelationalDatasetConnector} from \code{script} file with keys assigned
automatically by \code{dataname}.
}
\examples{
library(random.cdisc.data)
x <- rcd_dataset_connector(
  dataname = "ADSL",
  fun = radsl,
  keys = get_cdisc_keys("ADSL"),
  cached = TRUE
)
x$get_code()
load_dataset(x)
get_dataset(x)
x$get_raw_data()

\dontrun{
x <- rds_dataset_connector(
  dataname = "ADSL",
  file = "path/to/file.RDS",
  keys = get_cdisc_keys("ADSL")
)
x$get_code()
}

\dontrun{
x <- script_dataset_connector(
  dataname = "ADSL",
  file = "path/to/script.R",
  keys = get_cdisc_keys("ADSL")
)
x$get_code()
}

x <- rice_dataset_connector(
  dataname = "ADSL",
  path = "/path/to/file.sas7bdat",
  keys = get_cdisc_keys("ADSL")
)
x$get_code()
\dontrun{
load_dataset(x)
get_dataset(x)
x$get_raw_data()
}

}
