% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RelationalDatasetConnector_constructors.R
\name{relational_dataset_connector_file}
\alias{relational_dataset_connector_file}
\title{Load \code{RelationalDatasetConnector} object from a file}
\usage{
relational_dataset_connector_file(path)
}
\arguments{
\item{path}{(\code{character}) string giving the pathname of the file to read from.}
}
\value{
\code{RelationalDatasetConnector} object
}
\description{
\ifelse{html}{\out{<a href='https://www.tidyverse.org/lifecycle/#experimental'><img src='figures/lifecycle-experimental.svg' alt='Experimental lifecycle'></a>}}{\strong{Experimental}}
Please note that the script has to end with a call creating desired object. The error will
be raised otherwise.
}
\examples{
# simple example
file_example <- tempfile(fileext = ".R")
writeLines(
  text = c(
    "library(teal)
     library(random.cdisc.data)

     pull_callable <- callable_function(radsl)
     pull_callable$set_args(list(cached = TRUE))
     relational_dataset_connector(\"ADSL\", pull_callable, get_cdisc_keys(\"ADSL\"))"
  ),
  con = file_example
)
x <- relational_dataset_connector_file(file_example)
get_code(x)
}
