% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RelationalDataset.R
\name{relational_dataset_file}
\alias{relational_dataset_file}
\title{Load \code{RelationalDataset} object from a file}
\usage{
relational_dataset_file(path, code = get_code(path))
}
\arguments{
\item{path}{(\code{character}) string giving the pathname of the file to read from.}

\item{code}{(\code{character}) reproducible code to re-create object}
}
\value{
\code{RelationalDataset} object
}
\description{
Please note that the script has to end with a call creating desired object. The error will be raised otherwise.
}
\examples{
# simple example
file_example <- tempfile(fileext = ".R")
writeLines(
  text = c(
    "library(teal)
     library(random.cdisc.data)

     cdisc_dataset(dataname = \"ADSL\",
                   data = radsl(cached = TRUE),
                   code = \"library(random.cdisc.data)\nADSL <- radsl(cached = TRUE)\")"
  ),
  con = file_example
)
x <- relational_dataset_file(file_example, code = character(0))
get_code(x)

# custom code
file_example <- tempfile(fileext = ".R")
writeLines(
  text = c(
    "library(teal)

     # code>
     library(random.cdisc.data)
     ADSL <- radsl(cached = TRUE)
     ADSL$a1 <- 1
     ADSL$a2 <- 2

     # <code
     cdisc_dataset(dataname = \"ADSL\", data = ADSL)"
  ),
  con = file_example
)
x <- relational_dataset_file(file_example)
get_code(x)
}
