% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RelationalData.R
\name{RelationalData}
\alias{RelationalData}
\title{Manage multiple \code{RelationalDataConnector}, \code{RelationalDatasetConnector}
and \code{RelationalDataset} objects.}
\description{
Class manages \code{RelationalDataConnector}, \code{RelationalDatasetConnector} and
\code{RelationalDataset} objects and aggregate them in one application.
Class also decides whether to launch app before initialize teal application.
}
\examples{
library(random.cdisc.data)
x <- rcd_data( # RelationalDataConnector
  rcd_cdisc_dataset_connector("ADSL", radsl, cached = TRUE),
  rcd_cdisc_dataset_connector("ADLB", radlb, cached = TRUE)
)
x2 <- rcd_data( # RelationalDataConnector
  rcd_cdisc_dataset_connector("ADRS", radrs, cached = TRUE)
)
x3 <- cdisc_dataset(
  dataname = "ADAE", # RelationalDataset
  data = radae(cached = TRUE),
  code = "library(random.cdisc.data)\nADAE <- radae(cached = TRUE)"
)

x4 <- rcd_cdisc_dataset_connector("ADTTE", radtte, cached = TRUE)
tc <- teal:::RelationalData$new(x, x2, x3, x4)
tc$get_datanames()
\dontrun{
tc$get_datasets()
tc$get_dataset("ADAE")
}
tc$get_items()
tc$get_code()
tc$get_code("ADAE")
\dontrun{
tc$launch()
tc$get_datasets()
tc$get_dataset("ADAE")
tc$check()
}

library(random.cdisc.data)
x <- cdisc_dataset(
  dataname = "ADSL", # RelationalDataset
  data = radsl(cached = TRUE),
  code = "library(random.cdisc.data)\nADSL <- radsl(cached = TRUE)"
)

x2 <- rcd_cdisc_dataset_connector("ADTTE", radtte, cached = TRUE)
tc <- teal:::RelationalData$new(x, x2)
\dontrun{
tc$get_datasets()
get_raw_data(tc)
tc$launch()
get_raw_data(tc)
}
}
\section{Super class}{
\code{\link[teal:RelationalDataCollection]{teal::RelationalDataCollection}} -> \code{RelationalData}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{RelationalData$new()}}
\item \href{#method-get_datanames}{\code{RelationalData$get_datanames()}}
\item \href{#method-get_ui}{\code{RelationalData$get_ui()}}
\item \href{#method-get_connectors}{\code{RelationalData$get_connectors()}}
\item \href{#method-get_items}{\code{RelationalData$get_items()}}
\item \href{#method-get_server}{\code{RelationalData$get_server()}}
\item \href{#method-launch}{\code{RelationalData$launch()}}
\item \href{#method-clone}{\code{RelationalData$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="teal" data-topic="RelationalDataCollection" data-id="check">}\href{../../teal/html/RelationalDataCollection.html#method-check}{\code{teal::RelationalDataCollection$check()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="teal" data-topic="RelationalDataCollection" data-id="get_check_result">}\href{../../teal/html/RelationalDataCollection.html#method-get_check_result}{\code{teal::RelationalDataCollection$get_check_result()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="teal" data-topic="RelationalDataCollection" data-id="get_code">}\href{../../teal/html/RelationalDataCollection.html#method-get_code}{\code{teal::RelationalDataCollection$get_code()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="teal" data-topic="RelationalDataCollection" data-id="get_code_class">}\href{../../teal/html/RelationalDataCollection.html#method-get_code_class}{\code{teal::RelationalDataCollection$get_code_class()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="teal" data-topic="RelationalDataCollection" data-id="get_dataset">}\href{../../teal/html/RelationalDataCollection.html#method-get_dataset}{\code{teal::RelationalDataCollection$get_dataset()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="teal" data-topic="RelationalDataCollection" data-id="get_datasets">}\href{../../teal/html/RelationalDataCollection.html#method-get_datasets}{\code{teal::RelationalDataCollection$get_datasets()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="teal" data-topic="RelationalDataCollection" data-id="is_pulled">}\href{../../teal/html/RelationalDataCollection.html#method-is_pulled}{\code{teal::RelationalDataCollection$is_pulled()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="teal" data-topic="RelationalDataCollection" data-id="mutate">}\href{../../teal/html/RelationalDataCollection.html#method-mutate}{\code{teal::RelationalDataCollection$mutate()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="teal" data-topic="RelationalDataCollection" data-id="mutate_dataset">}\href{../../teal/html/RelationalDataCollection.html#method-mutate_dataset}{\code{teal::RelationalDataCollection$mutate_dataset()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="teal" data-topic="RelationalDataCollection" data-id="set_check">}\href{../../teal/html/RelationalDataCollection.html#method-set_check}{\code{teal::RelationalDataCollection$set_check()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="teal" data-topic="RelationalDataCollection" data-id="set_pull_code">}\href{../../teal/html/RelationalDataCollection.html#method-set_pull_code}{\code{teal::RelationalDataCollection$set_pull_code()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RelationalData$new(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{(\code{RelationalDataConnector}, \code{RelationalDataset},
\code{RelationalDatasetConnector}) objects}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_datanames"></a>}}
\if{latex}{\out{\hypertarget{method-get_datanames}{}}}
\subsection{Method \code{get_datanames()}}{
Derive the names of all datasets
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RelationalData$get_datanames()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{character} vector with names
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_ui"></a>}}
\if{latex}{\out{\hypertarget{method-get_ui}{}}}
\subsection{Method \code{get_ui()}}{
Get a shiny-module UI to render the necessary app to
derive \code{RelationalDataConnector} object's data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RelationalData$get_ui(id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{(\code{character}) item ID for the shiny module}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the \code{shiny} \code{ui} function
Get data connectors.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_connectors"></a>}}
\if{latex}{\out{\hypertarget{method-get_connectors}{}}}
\subsection{Method \code{get_connectors()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RelationalData$get_connectors()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{list} with all \code{RelationalDataConnector} objects.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_items"></a>}}
\if{latex}{\out{\hypertarget{method-get_items}{}}}
\subsection{Method \code{get_items()}}{



Get all datasets and all dataset connectors

  name of dataset connector to be returned. If \code{NULL}, all connectors are returned.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RelationalData$get_items(dataname = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dataname}}{(\code{character} value)\cr}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{list} with all datasets and all connectors
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_server"></a>}}
\if{latex}{\out{\hypertarget{method-get_server}{}}}
\subsection{Method \code{get_server()}}{
Get a shiny-module server to render the necessary app to
derive \code{RelationalDataConnector} object's data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RelationalData$get_server()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{shiny} \code{server} module.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-launch"></a>}}
\if{latex}{\out{\hypertarget{method-launch}{}}}
\subsection{Method \code{launch()}}{
Launch an app that allows to run the user-interfaces of all
\code{DataConnector} and \code{DatasetConnector} modules

This piece is mainly used for debugging.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RelationalData$launch()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RelationalData$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
