% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RelationalDataConnector.R
\name{RelationalDataConnector}
\alias{RelationalDataConnector}
\title{Manage multiple and \code{RelationalDatasetConnector} of the same type.}
\description{
Class manages \code{RelationalDatasetConnector} to specify additional dynamic arguments and to
open/close connection.
}
\examples{
library(random.cdisc.data)
adsl <- rcd_cdisc_dataset_connector("ADSL", radsl)
adlb <- rcd_cdisc_dataset_connector("ADLB", radlb, ADSL = adsl)
con <- teal:::rcd_connection()

ui <- function(id) {
  ns <- NS(id)
  tagList(
    numericInput(ns("seed"), "Choose seed", min = 1, max = 1000, value = 1),
    sliderInput(ns("N"), "Choose number of observations", min = 1, max = 400, value = 10)
  )
}

x <- teal:::RelationalDataConnector$new(connection = con, connectors = list(adsl, adlb))

x$set_ui(ui)
x$set_server(function(input, output, session, connectors, connection) {
  lapply(connectors, function(connector) {
    if (get_dataname(connector) == "ADSL") {
      set_args(connector, args = list(seed = input$seed, N = input$N))
    } else {
      set_args(connector, args = list(seed = input$seed))
    }
    connector$pull(try = TRUE)
  })
})
\dontrun{
tc$launch()
tc$get_datasets()
}
}
\section{Super class}{
\code{\link[teal:RelationalData]{teal::RelationalData}} -> \code{RelationalDataConnector}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{RelationalDataConnector$new()}}
\item \href{#method-get_datanames}{\code{RelationalDataConnector$get_datanames()}}
\item \href{#method-get_datasets}{\code{RelationalDataConnector$get_datasets()}}
\item \href{#method-get_dataset}{\code{RelationalDataConnector$get_dataset()}}
\item \href{#method-get_dataset_connectors}{\code{RelationalDataConnector$get_dataset_connectors()}}
\item \href{#method-get_code}{\code{RelationalDataConnector$get_code()}}
\item \href{#method-get_connection}{\code{RelationalDataConnector$get_connection()}}
\item \href{#method-get_all_datasets}{\code{RelationalDataConnector$get_all_datasets()}}
\item \href{#method-get_server}{\code{RelationalDataConnector$get_server()}}
\item \href{#method-get_ui}{\code{RelationalDataConnector$get_ui()}}
\item \href{#method-launch}{\code{RelationalDataConnector$launch()}}
\item \href{#method-pull}{\code{RelationalDataConnector$pull()}}
\item \href{#method-set_args}{\code{RelationalDataConnector$set_args()}}
\item \href{#method-set_check}{\code{RelationalDataConnector$set_check()}}
\item \href{#method-set_ui}{\code{RelationalDataConnector$set_ui()}}
\item \href{#method-set_server}{\code{RelationalDataConnector$set_server()}}
\item \href{#method-is_pulled}{\code{RelationalDataConnector$is_pulled()}}
\item \href{#method-clone}{\code{RelationalDataConnector$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="teal" data-topic="RelationalData" data-id="check">}\href{../../teal/html/RelationalData.html#method-check}{\code{teal::RelationalData$check()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="teal" data-topic="RelationalData" data-id="get_cdisc_data">}\href{../../teal/html/RelationalData.html#method-get_cdisc_data}{\code{teal::RelationalData$get_cdisc_data()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="teal" data-topic="RelationalData" data-id="set_code">}\href{../../teal/html/RelationalData.html#method-set_code}{\code{teal::RelationalData$set_code()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
Create a new \code{RelationalDataConnector} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RelationalDataConnector$new(connection, connectors)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{connection}}{(\code{DataConnection}) connection to data source}

\item{\code{connectors}}{(\code{list} of \code{DatasetConnector} elements) list with dataset connectors}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
new \code{DataConnector} object
Get names of the datasets.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_datanames"></a>}}
\subsection{Method \code{get_datanames()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RelationalDataConnector$get_datanames()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{character} vector with names of all datasets.
Get pulled \code{RelationalDataset} objects
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_datasets"></a>}}
\subsection{Method \code{get_datasets()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RelationalDataConnector$get_datasets()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{list} \code{RelationalDataset(s)} named by dataname.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_dataset"></a>}}
\subsection{Method \code{get_dataset()}}{
Get \code{RelationalDataset} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RelationalDataConnector$get_dataset(dataname)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dataname}}{(\code{character} value)\cr
name of dataset to be returned. If \code{NULL}, all datasets are returned.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{RelationalDataset}.
Get \code{RelationalDatasetConnector} objects
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_dataset_connectors"></a>}}
\subsection{Method \code{get_dataset_connectors()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RelationalDataConnector$get_dataset_connectors()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{list} \code{RelationalDatasetConnector(s)} named by dataname.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_code"></a>}}
\subsection{Method \code{get_code()}}{
Derive the code for all datasets
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RelationalDataConnector$get_code(dataname = NULL, deparse = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dataname}}{(\code{character}) dataname or \code{NULL} for all datasets}

\item{\code{deparse}}{(\code{logical}) whether to return the deparsed form of a call}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{list} of \code{character} containing code
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_connection"></a>}}
\subsection{Method \code{get_connection()}}{
Get connection to data source
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RelationalDataConnector$get_connection()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
connector's connection
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_all_datasets"></a>}}
\subsection{Method \code{get_all_datasets()}}{



Get all datasets and all dataset connectors

  name of dataset connector to be returned. If \code{NULL}, all connectors are returned.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RelationalDataConnector$get_all_datasets(dataname = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dataname}}{(\code{character} value)\cr}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{list} with all datasets and all connectors
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_server"></a>}}
\subsection{Method \code{get_server()}}{

\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RelationalDataConnector$get_server()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
the \code{server} function of the \code{RelationalDataConnector}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_ui"></a>}}
\subsection{Method \code{get_ui()}}{

\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RelationalDataConnector$get_ui(id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{\code{character} shiny element id}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the \code{ui} function of the \code{RelationalDataConnector}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-launch"></a>}}
\subsection{Method \code{launch()}}{
Run simple application that uses its \code{ui} and \code{server} fields to pull data from
connection.

Useful for debugging
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RelationalDataConnector$launch()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
An object that represents the app
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-pull"></a>}}
\subsection{Method \code{pull()}}{
Load data from each \code{RelationalDatasetConnector}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RelationalDataConnector$pull(con_args = NULL, args = NULL, try = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{con_args}}{(\code{NULL} or named \code{list})\cr
 additional dynamic arguments for connection function. \code{args} will be passed to each
\code{RelationalDatasetConnector} object to evaluate \code{CallableFunction} assigned to
this dataset. If \code{args} is null than default set of arguments will be used, otherwise
call will be executed on these arguments only (arguments set before will be ignored).
\code{pull} function doesn't update reproducible call, it's just evaluate function.}

\item{\code{args}}{(\code{NULL} or named \code{list})\cr
additional dynamic arguments to pull dataset. \code{args} will be passed to each
\code{RelationalDatasetConnector} object to evaluate \code{CallableFunction} assigned to
this dataset. If \code{args} is null than default set of arguments will be used, otherwise
call will be executed on these arguments only (arguments set before will be ignored).
\code{pull} function doesn't update reproducible call, it's just evaluate function.}

\item{\code{try}}{(\code{logical} value)\cr
whether perform function evaluation inside \code{try} clause}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
nothing, in order to get the data please use \code{get_data} method
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_args"></a>}}
\subsection{Method \code{set_args()}}{
Set arguments to connection and connectors. Using \code{set_args} will save these
arguments in reproducible code unlike using \code{pull} function.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RelationalDataConnector$set_args(con_args = NULL, args = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{con_args}}{(named \code{list})\cr
arguments values for opening connection.}

\item{\code{args}}{(named \code{list})\cr
arguments values for function to pull data.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
nothing
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_check"></a>}}
\subsection{Method \code{set_check()}}{
Set reproducibility check
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RelationalDataConnector$set_check(check = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{check}}{(\code{logical}) whether to perform reproducibility check}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
nothing
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_ui"></a>}}
\subsection{Method \code{set_ui()}}{
Set connector UI function
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RelationalDataConnector$set_ui(data_input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data_input}}{(\code{function})\cr
shiny module as function. Inputs specified in this \code{ui} are passed to server module
defined by \code{set_server} method.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
nothing
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_server"></a>}}
\subsection{Method \code{set_server()}}{
Set connector server function

This function will be called after submit button will be hit. There is no possibility to
specify some dynamic \code{ui} as \code{server} function is executed after hitting submit
button.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RelationalDataConnector$set_server(data_module)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data_module}}{(\code{function})\cr
A shiny module server function that should load data from all connectors}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
nothing
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-is_pulled"></a>}}
\subsection{Method \code{is_pulled()}}{
Check if dataset has already been pulled.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RelationalDataConnector$is_pulled()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{TRUE} if connector has been already pulled, else \code{FALSE}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RelationalDataConnector$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
