% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RelationalDataset.R
\name{RelationalDataset}
\alias{RelationalDataset}
\title{R6 Class representing a dataset including keys}
\description{
Any \code{data.frame} or \code{rtable} object can be
stored inside this object.

It inherits the \link{NamedDataset} object. This needs the
code that was used to generate this object to be handed over.
Additionally it is possible to label the data set
by handing over a \code{label} argument to the
\code{new} method

Additionally it contains the \code{keys} that allow
this dataset to be merged with other datasets
that are in any relation to this.
}
\examples{
rel_data <- teal:::RelationalDataset$new(
  x = data.frame(x = c(1, 2), y = c("a", "b"), stringsAsFactors = FALSE),
  keys = keys(primary = "y", foreign = NULL, parent = NULL),
  dataname = "XY",
  code = "XY <- data.frame(x = c(1, 2), y = c('aa', 'bb'),
                           stringsAsFactors = FALSE)"
)

rel_data$ncol
rel_data$get_code()
rel_data$get_dataname()
rel_data$get_keys()
rel_data$set_keys(keys(primary = "x", foreign = NULL, parent = NULL))
}
\section{Super classes}{
\code{\link[teal:RawDataset]{teal::RawDataset}} -> \code{\link[teal:NamedDataset]{teal::NamedDataset}} -> \code{RelationalDataset}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{keys}}{(\code{keys}) Derive the keys of the dataset}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{RelationalDataset$new()}}
\item \href{#method-get_keys}{\code{RelationalDataset$get_keys()}}
\item \href{#method-get_datanames}{\code{RelationalDataset$get_datanames()}}
\item \href{#method-set_keys}{\code{RelationalDataset$set_keys()}}
\item \href{#method-is_pulled}{\code{RelationalDataset$is_pulled()}}
\item \href{#method-clone}{\code{RelationalDataset$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="teal" data-topic="RawDataset" data-id="get_character_colnames">}\href{../../teal/html/RawDataset.html#method-get_character_colnames}{\code{teal::RawDataset$get_character_colnames()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="teal" data-topic="RawDataset" data-id="get_colnames">}\href{../../teal/html/RawDataset.html#method-get_colnames}{\code{teal::RawDataset$get_colnames()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="teal" data-topic="RawDataset" data-id="get_column_labels">}\href{../../teal/html/RawDataset.html#method-get_column_labels}{\code{teal::RawDataset$get_column_labels()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="teal" data-topic="RawDataset" data-id="get_factor_colnames">}\href{../../teal/html/RawDataset.html#method-get_factor_colnames}{\code{teal::RawDataset$get_factor_colnames()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="teal" data-topic="RawDataset" data-id="get_numeric_colnames">}\href{../../teal/html/RawDataset.html#method-get_numeric_colnames}{\code{teal::RawDataset$get_numeric_colnames()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="teal" data-topic="RawDataset" data-id="get_raw_data">}\href{../../teal/html/RawDataset.html#method-get_raw_data}{\code{teal::RawDataset$get_raw_data()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="teal" data-topic="RawDataset" data-id="get_row_labels">}\href{../../teal/html/RawDataset.html#method-get_row_labels}{\code{teal::RawDataset$get_row_labels()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="teal" data-topic="RawDataset" data-id="get_rownames">}\href{../../teal/html/RawDataset.html#method-get_rownames}{\code{teal::RawDataset$get_rownames()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="teal" data-topic="NamedDataset" data-id="check">}\href{../../teal/html/NamedDataset.html#method-check}{\code{teal::NamedDataset$check()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="teal" data-topic="NamedDataset" data-id="get_code">}\href{../../teal/html/NamedDataset.html#method-get_code}{\code{teal::NamedDataset$get_code()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="teal" data-topic="NamedDataset" data-id="get_dataname">}\href{../../teal/html/NamedDataset.html#method-get_dataname}{\code{teal::NamedDataset$get_dataname()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="teal" data-topic="NamedDataset" data-id="get_dataset_label">}\href{../../teal/html/NamedDataset.html#method-get_dataset_label}{\code{teal::NamedDataset$get_dataset_label()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="teal" data-topic="NamedDataset" data-id="set_code">}\href{../../teal/html/NamedDataset.html#method-set_code}{\code{teal::NamedDataset$set_code()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="teal" data-topic="NamedDataset" data-id="set_dataname">}\href{../../teal/html/NamedDataset.html#method-set_dataname}{\code{teal::NamedDataset$set_dataname()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="teal" data-topic="NamedDataset" data-id="set_dataset_label">}\href{../../teal/html/NamedDataset.html#method-set_dataset_label}{\code{teal::NamedDataset$set_dataset_label()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
initialize a \code{RelationalDataset} class object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RelationalDataset$new(
  x,
  dataname,
  keys,
  code = character(0),
  label = character(0)
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{(\code{data.frame})}

\item{\code{dataname}}{(\code{character}) A given name for the dataset
it may not contain spaces}

\item{\code{keys}}{(\code{keys}) object of S3 class keys containing
foreign, primary keys and parent information}

\item{\code{code}}{(\code{character}) A character string defining the code
needed to produce the data set in \code{x}}

\item{\code{label}}{(\code{character}) Label to describe the dataset}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_keys"></a>}}
\subsection{Method \code{get_keys()}}{
Derive the keys
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RelationalDataset$get_keys()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_datanames"></a>}}
\if{latex}{\out{\hypertarget{method-get_datanames}{}}}
\subsection{Method \code{get_datanames()}}{
Derive the dataname
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RelationalDataset$get_datanames()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_keys"></a>}}
\subsection{Method \code{set_keys()}}{
Set new keys
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RelationalDataset$set_keys(keys)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{keys}}{(\code{keys}) Set the keys}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-is_pulled"></a>}}
\subsection{Method \code{is_pulled()}}{
Check if dataset has already been pulled.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RelationalDataset$is_pulled()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{TRUE} if dataset has been already pulled, else \code{FALSE}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RelationalDataset$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
