% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chunks.R
\name{remove_chunk}
\alias{remove_chunk}
\title{Remove a chunk from the current environment}
\usage{
remove_chunk(id = NULL,
  chunks = parent.env(parent.frame())$session$userData$chunks)
}
\arguments{
\item{id}{(\code{character}) ID given for the code chunk}

\item{chunks}{(\code{chunks}) object of type \code{chunks} that the code chunk is removed from. If not
set the default value picks the global variable \code{chunks}.}
}
\value{
The chunks without the chunk
}
\description{
Remove a chunk from the current environment
}
\references{
chunks
}
