% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tdata.R
\name{req_tdata}
\alias{req_tdata}
\title{Function to call data elements of a \code{tdata} object}
\usage{
req_tdata(data)
}
\arguments{
\item{data}{\code{tdata} - object to extract the data from}
}
\description{
Function to call data elements of a \code{tdata} object
}
