% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules_debugging.R
\name{reset_filters_module}
\alias{reset_filters_module}
\title{Reset filters for the specified datasets}
\usage{
reset_filters_module(label = "Reset Filters Module", active_datanames = "all")
}
\arguments{
\item{label}{\code{character} label of module}

\item{active_datanames}{\verb{character vector} datanames shown in filter panel;
can be \code{"all"} to mean all available datasets}
}
\description{
The module presents a group of checkboxes to select the datasets for which to reset
all filters.
}
