% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resolve_delayed.R
\name{resolve_delayed_expr}
\alias{resolve_delayed_expr}
\title{Resolve expression after delayed data are loaded}
\usage{
resolve_delayed_expr(x, ds, is_value_choices)
}
\arguments{
\item{x}{(\code{function}) Function that is applied on dataset.
It must take only a single argument "data" and return character vector with columns / values.}

\item{ds}{(\code{data.frame}) \code{TealDataset} on which the function is applied to.}

\item{is_value_choices}{(\code{logical}) Determines which check of the returned value will be applied.}
}
\value{
Character vector - result of calling function \code{x} on dataset \code{ds}.
}
\description{
Resolve expression after delayed data are loaded
}
\examples{
\dontrun{
# get only possible factor variables from mtcars dataset
resolve_delayed_expr(
  function(data) {
    idx <- vapply(data, function(x) is.numeric(x) && length(unique(x)) <= 6, logical(1))
    colnames(data)[idx]
  },
  ds = mtcars,
  is_value_choices = FALSE
)
}
}
\keyword{internal}
