% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{resolve_modules_datanames}
\alias{resolve_modules_datanames}
\title{Resolve the datanames for the modules}
\usage{
resolve_modules_datanames(modules, datanames, join_keys)
}
\arguments{
\item{modules}{(\code{teal_modules}) object}

\item{datanames}{(\code{character}) datanames available in the \code{data} object}

\item{join_keys}{(\code{JoinKeys}) object}
}
\description{
Modifies \code{module$datanames} to include parent datanames (taken from join_keys).
When \code{datanames} is set to \code{"all"} it is replaced with all available datanames.
}
\keyword{internal}
