% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resolve_delayed.R
\name{resolve_teal_args}
\alias{resolve_teal_args}
\title{Handles teal arguments that are only available through delayed loading}
\usage{
resolve_teal_args(args, datasets)
}
\arguments{
\item{args}{\code{list} arguments to evaluate by passing them the datasets}

\item{datasets}{\code{datasets} datasets used to set \code{args}}
}
\description{
Handles teal arguments that are only available through delayed loading
}
