% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataConnection.R
\name{rice_connection}
\alias{rice_connection}
\title{Open connection to \code{rice}}
\usage{
rice_connection()
}
\value{
\code{DataConnection} type of object
}
\description{
Open connection to \code{rice}
}
