% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RelationalDataConnector_constructors.R
\name{rice_data}
\alias{rice_data}
\title{\code{RelationalDataConnector} connector for \code{RICE}}
\usage{
rice_data(..., connection = rice_connection(), additional_ui = NULL)
}
\arguments{
\item{...}{(\code{DatasetConnector} objects)\cr
dataset connectors created using \code{rice_dataset_connector}}

\item{connection}{(\code{DataConnection}) object returned from \code{rice_connection}.}

\item{additional_ui}{(\code{shiny.tag})\cr
additional user interface to be visible over login panel}
}
\value{
An object of class \code{RelationalDataConnector}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#defunct}{\figure{lifecycle-defunct.svg}{options: alt='[Defunct]'}}}{\strong{[Defunct]}}
Build data connector for \code{RICE} datasets
}
