% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DatasetConnector.R
\name{rice_dataset}
\alias{rice_dataset}
\title{Set up connection to local \code{RICE} dataset}
\usage{
rice_dataset(
  dataname,
  path,
  keys = get_cdisc_keys(dataname),
  labels = list(dataset_label = NULL, column_labels = NULL)
)
}
\arguments{
\item{dataname}{name of dataset}

\item{path}{(\code{character}) file path}

\item{keys}{list of keys -

Please note that the order of keys is important.

Keys are not matched by name but by order. In case you want to perform:

\code{SELECT ...}
\code{  FROM ADSL}
\code{LEFT JOIN ADTTE on ADTTE.STUDYID = ADSL.STUDYID and}
\code{                                     ADTTE.SUBJECTUNIQUEID = ADSL.USUBJID}

the keys need to be

ADSL \code{primary: USUBJID, STUDYID}

ADTTE \code{primary: PARAMCD, foreign: SUBJECTUNIQUEID, STUDYID, parent: ADSL}

As you can see the names of foreign ADTTE are different to ADSL keys, but the order and length is equal.}

\item{labels}{list of labels}
}
\value{
(\code{DatasetConnector}) type of object
}
\description{
Set up connection to local \code{RICE} dataset
}
\examples{
x <- rice_dataset("ADSL", "/path/to/ADSL")
x$get_call()
\dontrun{
x$get_data()
}
}
