% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TealDataConnection.R
\name{ricepass_connection}
\alias{ricepass_connection}
\title{Superseded (moved into separate package): Open connection to \code{entimICE} via \code{ricepass}}
\usage{
ricepass_connection()
}
\value{
(\code{TealDataConnection}) type of object
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#superseded}{\figure{lifecycle-superseded.svg}{options: alt='[Superseded]'}}}{\strong{[Superseded]}}
}
