% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/include_css_js.R
\name{run_js_file}
\alias{run_js_file}
\title{Run \code{JS} file from \code{/inst/js/} package directory}
\usage{
run_js_file(file, package = "teal")
}
\arguments{
\item{file}{(\code{character}) vector of file names}

\item{package}{(\code{character}) package name}
}
\description{
Run \code{JS} file from \code{/inst/js/} package directory
}
