% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/include_css_js.R
\name{run_js_files}
\alias{run_js_files}
\title{Run \code{JS} file from \verb{/inst/js/} package directory}
\usage{
run_js_files(files, package = "teal")
}
\arguments{
\item{files}{(\code{character}) vector of filenames}

\item{package}{(\code{character}) package name}
}
\description{
This is triggered from the server to execute on the client
rather than triggered directly on the client.
Unlike \code{include_js_files} which includes Javascript functions,
the \code{run_js} actually executes Javascript functions.
}
