% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TealDatasetConnector_constructors.R
\name{scda_dataset_connector}
\alias{scda_dataset_connector}
\alias{scda_cdisc_dataset_connector}
\title{\code{scda} \code{TealDatasetConnector}}
\usage{
scda_dataset_connector(
  dataname,
  scda_dataname = tolower(dataname),
  scda_name = "latest",
  keys = character(0),
  label = character(0),
  code = character(0),
  script = character(0),
  metadata = list(type = "scda", version = scda_name)
)

scda_cdisc_dataset_connector(
  dataname,
  scda_dataname = tolower(dataname),
  scda_name = "latest",
  keys = get_cdisc_keys(dataname),
  parent = if (identical(dataname, "ADSL")) character(0L) else "ADSL",
  label = character(0),
  code = character(0),
  script = character(0),
  metadata = list(type = "scda", version = scda_name)
)
}
\arguments{
\item{dataname}{(\code{character})\cr
A given name for the dataset it may not contain spaces}

\item{scda_dataname}{(\code{character}) which scda dataset to use (e.g. "adsl").}

\item{scda_name}{(\code{character}) which version of scda data to take, default "latest".}

\item{keys}{optional, (\code{character})\cr
vector of dataset primary keys column names}

\item{label}{(\code{character})\cr
Label to describe the dataset.}

\item{code}{(\code{character})\cr
A character string defining code to modify \code{raw_data} from this dataset. To modify
current dataset code should contain at least one assignment to object defined in \code{dataname}
argument. For example if \code{dataname = ADSL} example code should contain
\verb{ADSL <- <some R code>}. Can't be used simultaneously with \code{script}}

\item{script}{(\code{character})\cr
Alternatively to \code{code} - location of the file containing modification code.
Can't be used simultaneously with \code{script}.}

\item{metadata}{(named \code{list}, \code{NULL} or \code{CallableFunction}) \cr
Field containing either the metadata about the dataset (each element of the list
should be atomic and length one) or a \code{CallableFuntion} to pull the metadata
from a connection. This should return a \code{list} or an object which can be
converted to a list with \code{as.list}.}

\item{parent}{optional, (\code{character}) \cr
parent dataset name}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
}
\details{
Create a \code{TealDatasetConnector} for dataset in \code{scda}

Create a \code{CDISCTealDatasetConnector} from \code{scda} data
}
\examples{
library(scda)
x <- scda_dataset_connector(
  dataname = "ADSL", scda_dataname = "adsl",
)
x$get_code()
load_dataset(x)
get_dataset(x)
get_dataset(x)$get_metadata()
x$get_raw_data()

metadata_fun <- callable_function(function(a) list(type = a))
metadata_fun$set_args(args = list(a = "scda"))
y <- scda_dataset_connector(
  dataname = "ADSL", scda_dataname = "adsl",
  metadata = metadata_fun
)
load_dataset(y)
get_dataset(y)$get_metadata()
}
