% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{see_in_file}
\alias{see_in_file}
\title{See a function's code in a temporary file
This is more handy when you want to search for variables in the code a function
rather than doing so in the console.}
\usage{
see_in_file(f)
}
\arguments{
\item{f}{function or object that is printed and whose output is shown in the file}
}
\description{
Debugging function
}
\details{
The file is created in the temporary directory of the R session and
will be deleted once the R session exits.
}
\examples{
\dontrun{
teal:::see_in_file(factor)
}
}
