% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_session_info.R
\name{session_info}
\alias{session_info}
\alias{ui_session_info}
\alias{srv_session_info}
\title{\code{teal} user session info module}
\usage{
ui_session_info(id)

srv_session_info(id)
}
\arguments{
\item{id}{(\code{character}) Optionally,
a string specifying the \code{shiny} module id in cases it is used as a \code{shiny} module
rather than a standalone \code{shiny} app. This is a legacy feature.}
}
\value{
\code{NULL} invisibly
}
\description{
Module to display the user session info popup and teal lockfile download.
}
\examples{
ui <- fluidPage(
  ui_session_info("session_info")
)

server <- function(input, output, session) {
  srv_session_info("session_info")
}

if (interactive()) {
  shinyApp(ui, server)
}

}
