% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_args.R
\name{set_args}
\alias{set_args}
\alias{set_args.CallableFunction}
\alias{set_args.RawDatasetConnector}
\title{Set arguments of a \code{CallableFunction}}
\usage{
set_args(x, args)

\method{set_args}{CallableFunction}(x, args)

\method{set_args}{RawDatasetConnector}(x, args)
}
\arguments{
\item{x}{\code{CallableFunction} or \code{RawDatasetConnector})}

\item{args}{(\code{NULL} or named \code{list}) dynamic arguments to function}
}
\value{
nothing
}
\description{
\ifelse{html}{\out{<a href='https://www.tidyverse.org/lifecycle/#experimental'><img src='figures/lifecycle-experimental.svg' alt='Experimental lifecycle'></a>}}{\strong{Experimental}}
Set arguments of a \code{CallableFunction}
}
\examples{
library(random.cdisc.data)
fun <- callable_function(radsl)
set_args(fun, list(N = 5, seed = 1, cached = TRUE))
ds <- raw_dataset_connector(pull_callable = callable_function(data.frame))
set_args(ds, list(x = 1:5, y = letters[1:5]))
}
