% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_filter_state.R
\name{set_filter_state}
\alias{set_filter_state}
\alias{set_filter_state.default}
\alias{set_filter_state.default_filter}
\alias{set_filter_state.list}
\title{Sets the filter state in \code{FilterState} object}
\usage{
set_filter_state(x, filter_state)

\method{set_filter_state}{default}(x, filter_state)

\method{set_filter_state}{default_filter}(x, filter_state)

\method{set_filter_state}{list}(x, filter_state)
}
\arguments{
\item{x}{(\code{list},\code{vector}, \code{default_filter})\cr
values of the variable used in filter. Depending on the \code{FilterState} type
list should contain:
\itemize{
\item{\code{selected}}{ defines initial selection. See notes for more details}
\item{\code{keep_na} (\code{logical})}{ defines whether to keep or remove \code{NA} values}
\item{\code{keep_inf} (\code{logical})}{ defines whether to keep or remove \code{Inf} values}
}
If \code{vector} is provided then \code{keep_na} and \code{keep_inf} can be specified
adding \code{NA} and \code{Inf} to the selection vector.}

\item{filter_state}{(\code{FilterState})\cr}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
Sets values of the selection in \code{FilterState} object
}
\note{
The value of the \code{x$selected} needs to be modified according to the type
of the passed \code{filter_state}.
}
\examples{
filter_state <- teal:::RangeFilterState$new(
  c(1:10, NA, Inf),
  varname = "x"
)

teal:::set_filter_state(
  list(selected = c(1, 2), keep_na = FALSE, keep_inf = TRUE),
  filter_state
)
teal:::set_filter_state(c(1, 2, Inf), filter_state)
teal:::set_filter_state(default_filter(), filter_state)
}
\seealso{
\itemize{
\item{\code{\link{LogicalFilterState}}}
\item{\code{\link{ChoicesFilterState}}}
\item{\code{\link{RangeFilterState}}}
\item{\code{\link{DateFilterState}}}
\item{\code{\link{DatetimeFilterState}}}
}
}
