% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_keys.R
\name{set_keys}
\alias{set_keys}
\alias{set_keys.Dataset}
\alias{set_keys.DatasetConnector}
\alias{set_keys.DataAbstract}
\title{Set dataset primary keys}
\usage{
set_keys(x, ...)

\method{set_keys}{Dataset}(x, keys, ...)

\method{set_keys}{DatasetConnector}(x, keys, ...)

\method{set_keys}{DataAbstract}(x, dataname, keys, ...)
}
\arguments{
\item{x}{an object of \code{Dataset} or \code{DatasetConnector} class}

\item{...}{not used, only for support of S3}

\item{keys}{optional, (\code{character}) vector with primary keys}

\item{dataname}{(\code{character}) name of dataset for which set the keys}
}
\value{
(\code{character}) vector of column names
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
Set dataset primary keys
}
\examples{
# Dataset --------

library(scda)
set_keys(
  dataset(
    "ADSL",
    synthetic_cdisc_data("latest")$adsl,
    code = "ADSL <- synthetic_cdisc_data(\"latest\")$adsl"
  ),
  keys = get_cdisc_keys("ADSL")
)
# DatasetConnector --------

library(scda)
pull_fun_adsl <- callable_function(
  function() {synthetic_cdisc_data("latest")$adsl}
)
set_keys(
  dataset_connector(
    "ADSL",
    pull_fun_adsl
  ),
  keys = get_cdisc_keys("ADSL")
)
# RelationalData --------

set_keys(
  teal_data(
    dataset("x", data.frame(x1 = 1:10, y1 = 11:20), keys = "x1"),
    dataset("y", data.frame(x2 = 1:10, y2 = 11:20), keys = "x2")
  ),
  "x",
  c("x1", "y1")
)
}
