% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_labels_df}
\alias{set_labels_df}
\title{Set labels of a data.frame}
\usage{
set_labels_df(df, labels)
}
\arguments{
\item{df}{\code{data.frame}}

\item{labels}{named \verb{character vector} of labels,
must all be column names present in \code{df}}
}
\value{
\code{data.frame} with labels set
}
\description{
Set labels of a data.frame
}
\examples{
df1 <- teal:::set_labels_df(
  data.frame(a = 1, b = 2, c = 3, d = 4),
  c(a = "l1", b = "l2", c = "l3")
)
teal:::get_labels_df(df1)
}
