% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilteredData.R
\name{set_single_filter_state}
\alias{set_single_filter_state}
\title{Set the filter state for a single variable}
\usage{
set_single_filter_state(datasets, dataname, varname, state)
}
\arguments{
\item{datasets}{(\code{FilteredData})}

\item{dataname}{(\code{character}) dataname}

\item{varname}{(\code{character}) variable name}

\item{state}{(\code{list}) new state for the variable, can be
\code{default_filter()}}
}
\description{
See \code{FilteredData$set_filter_state}.
}
