% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_state}
\alias{set_state}
\title{Set state of \code{FilterState}}
\usage{
set_state(x, value, is_reactive = shiny::isRunning())
}
\arguments{
\item{x}{(\code{vector})\cr
values of the variable used in filter}

\item{value}{(named \code{list})\cr
see \code{set_state} method in \link{FilterState} and \code{filter} argument in the \code{\link[=init]{init()}}}

\item{is_reactive}{(\code{logical(1)})\cr
\itemize{
\item \code{TRUE} to change \code{reactive} fields which triggers observers in the \code{FilterState}
\item \code{FALSE} to change the state directly.
}}
}
\value{
invisible \code{NULL}
}
\description{
Set state of \code{FilterState}. Function can change states in \link{FilterState} in two ways, by:
\itemize{
\item changing \code{reactive} state fields which triggers observers in the \code{FilterState}.
\item change state directly.
}

For more, please see section "Modifying state" in \link{FilterState}
}
\keyword{internal}
