% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_state_manager.R
\name{setdiff_teal_grab}
\alias{setdiff_teal_grab}
\title{Compare \code{teal_grab} objects.}
\usage{
setdiff_teal_grab(x, y)
}
\arguments{
\item{x, y}{\code{teal_grab} objects}
}
\value{
\code{teal_grab} or \code{NULL}, if the difference is empty.
}
\description{
Performs a set difference adapted for the \code{teal_grab} class. Returns NULL if the difference is empty.
}
\keyword{internal}
