% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_entimice_setup.R
\name{shiny_entimice_setup}
\alias{shiny_entimice_setup}
\title{Create page with SAICE setup}
\usage{
shiny_entimice_setup(app_url = NULL)
}
\arguments{
\item{app_url}{TODO}
}
\description{
TODO
}
\examples{
\dontrun{
shiny_entimice_setup()
}

}
