% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showRCodeModal.R
\name{show_r_code_modal}
\alias{show_r_code_modal}
\title{Show R Code Modal}
\usage{
show_r_code_modal(title, rcode)
}
\arguments{
\item{title}{modal title}

\item{rcode}{character string with R code}
}
\description{
Show R Code Modal
}
