% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_rcode_modal.R
\name{show_rcode_modal}
\alias{show_rcode_modal}
\title{Show R Code Modal}
\usage{
show_rcode_modal(title = NULL, rcode, session = getDefaultReactiveDomain())
}
\arguments{
\item{title}{(\code{character(1)})\cr
Title of the modal, displayed in the first comment of the R-code.}

\item{rcode}{(\code{character})\cr
vector with R code to show inside the modal. You can use \code{\link[teal.code:get_code]{teal.code::get_code()}} to derive this R
code inside a module.}

\item{session}{(\code{ShinySession} optional)\cr
\code{shiny} Session object, if missing then \code{\link[shiny:domains]{shiny::getDefaultReactiveDomain()}} is used.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
Use the \code{\link[shiny:showModal]{shiny::showModal()}} function to show the R code inside.
}
\references{
\code{\link[shiny:showModal]{shiny::showModal()}} \code{\link[teal.code:get_code]{teal.code::get_code()}}
}
