% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showRCodeModal.R
\name{showRCodeModal}
\alias{showRCodeModal}
\title{Show R Code Modal}
\usage{
showRCodeModal(title, rcode)
}
\arguments{
\item{title}{modal title}

\item{rcode}{character string with R code}
}
\description{
Show R Code Modal
}
