% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataConnection.R
\name{snowflake_connection_function}
\alias{snowflake_connection_function}
\title{Helper function to connect to \code{Snowflake}}
\usage{
snowflake_connection_function(
  username = askpass::askpass("Please enter your username"),
  password = askpass::askpass("Please enter your password"),
  role,
  database,
  schema,
  warehouse,
  server = "roche_pd.eu-central-1.snowflakecomputing.com",
  port = 443,
  driver = "SnowflakeDSIIDriver"
)
}
\arguments{
\item{username}{The username used to collect the auth token to connect to snowflake.}

\item{password}{The password used to collect the auth token to connect to snowflake}

\item{role}{The user role used to connect to \code{Snowflake}.}

\item{database}{The \code{Snowflake} database to connect to.}

\item{schema}{The \code{Snowflake} schema to connect to.}

\item{warehouse}{The \code{Snowflake} warehouse to connect to.}

\item{server}{The \code{Snowflake}server to connect to.}

\item{port}{The port to connect to the \code{Snowflake} instance.}

\item{driver}{The driver to use to connect to the \code{Snowflake} instance.}
}
\description{
This is used by \code{snowflake_connection} and does not need to be called directly
}
