% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TealDatasetConnector_constructors.R
\name{snowflake_dataset_connector}
\alias{snowflake_dataset_connector}
\alias{snowflake_cdisc_dataset_connector}
\title{\code{Snowflake} \code{TealDatasetConnector}}
\usage{
snowflake_dataset_connector(
  dataname,
  sql_query,
  keys = character(0),
  label = character(0),
  code = character(0),
  script = character(0),
  ...
)

snowflake_cdisc_dataset_connector(
  dataname,
  sql_query,
  keys = get_cdisc_keys(dataname),
  parent = if (identical(dataname, "ADSL")) character(0L) else "ADSL",
  label = character(0),
  code = character(0),
  script = character(0),
  ...
)
}
\arguments{
\item{dataname}{(\code{character})\cr
A given name for the dataset it may not contain spaces}

\item{sql_query}{(\code{character}) SQL statement to extract data from snowflake}

\item{keys}{optional, (\code{character})\cr
vector of dataset primary keys column names}

\item{label}{(\code{character})\cr
Label to describe the dataset.}

\item{code}{(\code{character})\cr
A character string defining code to modify \code{raw_data} from this dataset. To modify
current dataset code should contain at least one assignment to object defined in \code{dataname}
argument. For example if \code{dataname = ADSL} example code should contain
\verb{ADSL <- <some R code>}. Can't be used simultaneously with \code{script}}

\item{script}{(\code{character})\cr
Alternatively to \code{code} - location of the file containing modification code.
Can't be used simultaneously with \code{script}.}

\item{...}{Additional arguments applied to pull function.
In case when this object code depends on the \code{raw_data} from the other
\code{TealDataset}, \code{TealDatasetConnector} object(s) or other constant value,
this/these object(s) should be included. Please note that \code{vars}
are included to this object as local \code{vars} and they cannot be modified
within another dataset.}

\item{parent}{optional, (\code{character}) \cr
parent dataset name}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#defunct}{\figure{lifecycle-defunct.svg}{options: alt='[Defunct]'}}}{\strong{[Defunct]}}

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#defunct}{\figure{lifecycle-defunct.svg}{options: alt='[Defunct]'}}}{\strong{[Defunct]}}
}
\details{
Create a \code{TealDatasetConnector} from \code{Snowflake}.

Create a \code{CDISCTealDatasetConnector} from \code{Snowflake} with keys and parent name assigned
automatically by \code{dataname}.
}
