% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_add_filter_variable.R
\name{srv_add_filter_variable}
\alias{srv_add_filter_variable}
\title{Server function to add a filter variable to \code{datasets}}
\usage{
srv_add_filter_variable(
  input,
  output,
  session,
  datasets,
  dataname,
  omit_vars = function() c()
)
}
\arguments{
\item{input}{\verb{Shiny input object}}

\item{output}{\verb{Shiny output object}}

\item{session}{\verb{Shiny session object}}

\item{datasets}{\code{FilteredData} object to store filter state and filtered
datasets, shared across modules}

\item{dataname}{name of dataset whose columns should be filtered}

\item{omit_vars}{\verb{function / reactive returning a character vector} variables that are
not available for filtering}
}
\description{
For the return value, note that the currently selected input will be
reset in the UI for the user to select another variable.
So, it will be \code{NULL} again in the next reactive flush.
}
