% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_filter_items.R
\name{srv_filter_items}
\alias{srv_filter_items}
\title{Server function to display filters for a dataset}
\usage{
srv_filter_items(input, output, session, datasets, dataname)
}
\arguments{
\item{input}{\verb{Shiny input object}}

\item{output}{\verb{Shiny output object}}

\item{session}{\verb{Shiny session object}}

\item{datasets}{\code{FilteredData} object to store filter state and filtered
datasets, shared across modules}

\item{dataname}{\code{character} dataname to show filters for}
}
\value{
\code{reactive} returning a named list, names being the shown variables
and values being the observers for the modules associated to that variable
(by calling a Shiny submodule for each variable). This is useful for dynamic
UI generation, so observers can be properly canceled. Note that currently, not
all observers defined in this module are returned (the observer to remove all
filters at once is not returned)
}
\description{
Server function to display filters for a dataset
}
