% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_filtered_data_overview.R
\name{srv_filtered_data_overview}
\alias{srv_filtered_data_overview}
\title{Server function to display the number of patients in the filtered and unfiltered
data}
\usage{
srv_filtered_data_overview(
  input,
  output,
  session,
  datasets,
  datanames = function() "all"
)
}
\arguments{
\item{input}{\verb{Shiny input object}}

\item{output}{\verb{Shiny output object}}

\item{session}{\verb{Shiny session object}}

\item{datasets}{\code{FilteredData} object to store filter state and filtered
datasets, shared across modules}

\item{datanames}{\verb{function / reactive returning a character vector} datanames
to show information for; if \code{"all"}, takes all datanames, if \code{NULL}}
}
\description{
Server function to display the number of patients in the filtered and unfiltered
data
}
