% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_nested_tabs.R
\name{srv_nested_tabs}
\alias{srv_nested_tabs}
\title{Server function that returns currently active module}
\usage{
srv_nested_tabs(input, output, session, datasets, modules)
}
\arguments{
\item{input}{\verb{Shiny input object}}

\item{output}{\verb{Shiny output object}}

\item{session}{\verb{Shiny session object}}

\item{datasets}{\code{FilteredData} object to store filter state and filtered
datasets, shared across modules}

\item{modules}{\code{teal_module} or \code{teal_modules} object, the latter
can be used for nested tabs, see \code{\link{ui_nested_tabs}}}
}
\value{
\code{reactive} which returns the active module that corresponds to the selected tab
}
\description{
Server function that returns currently active module
}
