% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_single_filter_item.R
\name{srv_single_filter_item}
\alias{srv_single_filter_item}
\title{Server function to filter for a single variable}
\usage{
srv_single_filter_item(
  input,
  output,
  session,
  datasets,
  dataname,
  varname,
  drop_levels
)
}
\arguments{
\item{input}{\verb{Shiny input object}}

\item{output}{\verb{Shiny output object}}

\item{session}{\verb{Shiny session object}}

\item{datasets}{\code{FilteredData} object to store filter state and filtered
datasets, shared across modules}

\item{dataname}{\code{character} dataname}

\item{varname}{\code{character} variable within \code{dataname} to filter}

\item{drop_levels}{\code{reactive} the reactive logical flag indicating whether to drop unused levels of variables
in the \code{dataname} dataset}
}
\value{
\code{reactive} which returns \code{list(observers = ...)} with registered observers.
}
\description{
Regarding the return value: The \code{observers} are returned so they can be canceled
when the module is removed.
}
