% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_teal.R
\name{srv_teal}
\alias{srv_teal}
\title{Server function corresponding to teal}
\usage{
srv_teal(id, modules, raw_data, filter = list())
}
\arguments{
\item{id}{(\code{character})\cr
module id to embed it, if provided,
the server function must be called with \code{\link[shiny:moduleServer]{shiny::moduleServer()}};
See the vignette for an example. However, \code{\link[=ui_teal_with_splash]{ui_teal_with_splash()}}
is then preferred to this function.}

\item{modules}{(\code{list} or \code{teal_modules})\cr
nested list of \code{teal_modules} or \code{module} objects. See \code{\link[=modules]{modules()}} and \code{\link[=module]{module()}} for
more details.}

\item{raw_data}{(\code{reactive})\cr
returns the \code{TealData}, only evaluated once, \code{NULL} value is ignored}

\item{filter}{(\code{list})\cr
You can define filters that show when the app starts.
The general pattern is:
\code{list(iris = list(Species = ..., Sepal.Length = ...), mtcars = ...)}.
An example is:
\code{list(iris = list(Species = c("setosa", "versicolor")))}.
More generally, the filters for the variable, e.g. \code{Species} can be
specified as follows:
\code{list(Species = c("setosa", "versicolor"))},
\code{list(Species = list(c("setosa", "versicolor"), keep_na = TRUE))},
or equivalently with:
\code{list(Species = c("setosa", "versicolor", NA))},
or for the default filter (not very restrictive):
\code{list(Species = default_filter())}

Instead of \code{choices} above, use the following names:
\itemize{
\item \code{numerical}: \code{range}
\item \code{factor}: \code{choices}
\item \code{logical}: \code{logical}
A general example is:
\code{list( iris = list(Sepal.Length = default_filter(), Species = c("M", NA)), mtcars = list(mpg = default_filter())   )}
Ignored if the app is restored from a bookmarked state.
}}
}
\value{
\code{reactive} which returns the currently active module
}
\description{
It evaluates the \code{raw_data} (delayed data mechanism) and creates the
\code{datasets} object that is shared across modules.
Once it is ready and non-\code{NULL}, the splash screen is replaced by the
main teal UI that depends on the data.
The currently active tab is tracked and the right filter panel
updates the displayed datasets to filter for according to the active datanames
of the tab.
The initially displayed filter states can be provided, bookmarked filter
states always take precedence over them.
}
\details{
For more doc, see \link{ui_teal}.
}
