% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_state_manager.R
\name{state_manager_module}
\alias{state_manager_module}
\alias{state_manager_ui}
\alias{grab}
\alias{grab_manager}
\alias{state_manager}
\alias{state_manager_srv}
\title{App state management.}
\usage{
state_manager_ui(id)

state_manager_srv(id)
}
\arguments{
\item{id}{(\code{character(1)}) \code{shiny} module id}
}
\value{
Nothing is returned.
}
\description{
Capture and restore the global (app) input state.
}
\details{
This is a work in progress.
}
\seealso{
\code{\link{app_state_grab}}, \code{\link{app_state_store}}, \code{\link{app_state_restore}}
}
\author{
Aleksander Chlebowski
}
\keyword{internal}
