% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stop_shiny.R
\name{stop_shiny}
\alias{stop_shiny}
\title{Stop app.R evaluation with a Jumbotron Shiny App}
\usage{
stop_shiny(title, body)
}
\arguments{
\item{title}{string with title}

\item{body}{either a string or an object of class \code{shiny.tag}}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
This will start a separate shiny app with a message
}
\examples{
\dontrun{
stop_shiny("No data access", "You do not seem to have data access")
stop_shiny("No data access", tagList(
  tags$h2("Hello World"), tags$p("paragraph"),
  tags$footer("Please contact", tags$a(href = "mailto:john@doe.com", "John Doe"))
))
}
}
