% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabs.R
\name{tabs}
\alias{tabs}
\title{Create a collection of \code{tab_item} and \code{tabs_item} object}
\usage{
tabs(...)
}
\arguments{
\item{...}{\code{\link{tab_item}} and \code{\link{tabs_item}} object}
}
\value{
object of class \code{teal_tabs}
}
\description{
Tabs collects a tree of \code{\link{tab_item}} and \code{\link{tabs_item}}
objects. This is useful to define the navigation structure of a teal app.
}
