% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabs.R
\name{tabs_item}
\alias{tabs_item}
\title{A parent navigation item}
\usage{
tabs_item(label, tabs)
}
\arguments{
\item{label}{label shown in the navigation for the item}

\item{tabs}{an object that is returend from the \code{\link{tabs}} function.
Note currently teal only suppors trees of depth 2, hence this tabs object
can not contain any other \code{tabs_item} objects.}
}
\description{
This is a labeled parent item in order to be able to create a navigation
tree.
}
