% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tdata.R
\name{tdata}
\alias{tdata}
\alias{new_tdata}
\alias{tdata2env}
\alias{get_code_tdata}
\alias{join_keys.tdata}
\alias{get_metadata}
\alias{as_tdata}
\title{Create a \code{tdata} object}
\usage{
new_tdata(...)

tdata2env(...)

get_code_tdata(...)

\method{join_keys}{tdata}(...)

get_metadata(...)

as_tdata(...)
}
\arguments{
\item{...}{ignored}
}
\value{
nothing
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#superseded}{\figure{lifecycle-superseded.svg}{options: alt='[Superseded]'}}}{\strong{[Superseded]}}

Recent changes in \code{teal} cause modules to fail because modules expect a \code{tdata} object
to be passed to the \code{data} argument but instead they receive a \code{teal_data} object,
which is additionally wrapped in a reactive expression in the server functions.
In order to easily adapt such modules without a proper refactor,
use this function to downgrade the \code{data} argument.
}
