% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tdata.R
\name{tdata2env}
\alias{tdata2env}
\title{Function to convert a \code{tdata} object to a list
Any \code{reactives} inside \code{tdata} are evaluated}
\usage{
tdata2env(data)
}
\arguments{
\item{`data`}{a \code{tdata} object}
}
\value{
a list
}
\description{
Function to convert a \code{tdata} object to a list
Any \code{reactives} inside \code{tdata} are evaluated
}
\examples{

td <- new_tdata(
  data = list(iris = iris, mtcars = reactive(mtcars)),
  code = "iris <- iris
    mtcars = mtcars"
)

my_list <- isolate(tdata2env(td))

}
