% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tdata.R
\name{.tdata_upgrade}
\alias{.tdata_upgrade}
\alias{.tdata_downgrade}
\title{Upgrade or downgrade data objects.}
\usage{
.tdata_upgrade(x)

.tdata_downgrade(x)
}
\arguments{
\item{`x`}{data object, either \code{tdata} or \code{teal_data}; reactive expressions are handled, see \code{Details}}
}
\value{
Object of class \code{tdata} for \code{.tdata_upgrade} and \code{tdata} for \code{.tdata_downgrade}.
}
\description{
Convert between \code{teal::tdata} and \code{teal.data::teal_data} classes.
}
\details{
Functions to switch between data classes during deprecation of the \code{tdata} class defined in \code{teal}.
\code{.tdata_upgrade} converts object of class \code{tdata} as defined in \code{teal} package
(list of reactive expressions, each containing one data set, with a \code{code} attribute)
to object of class \code{teal_data} as defined in package \code{teal.data}
(extension of class \code{qenv} as defined in \code{teal.code}).
\code{.tdata_downgrade} does the reverse.

Note the \code{metadata} attribute is discarded.

In apps and modules data objects passed to upgrade/downgrade may be reactive expressions.
For convenience, these functions accept reactive expressions2 and run on their values.
}
\keyword{internal}
