% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teal.R
\docType{package}
\name{teal-package}
\alias{teal}
\alias{teal-package}
\title{\code{teal}: Interactive exploration of clinical trials data}
\description{
The \code{teal} package provides a \code{shiny} based framework for creating an
interactive data analysis environment.
}
\details{
To learn mode about the package, visit the \href{https://pharmaverse.github.io/teal/latest-tag/}{project website}
or read the \code{\link[=init]{init()}} manual page.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://pharmaverse.github.io/teal/}
  \item \url{https://github.com/pharmaverse/teal/}
  \item Report bugs at \url{https://github.com/pharmaverse/teal/issues}
}

}
\author{
\strong{Maintainer}: Dony Unardi \email{unardid@gene.com}

Authors:
\itemize{
  \item Dony Unardi \email{unardid@gene.com}
  \item Dawid Kaledkowski \email{dawid.kaledkowski@roche.com} (\href{https://orcid.org/0000-0001-9533-457X}{ORCID})
  \item Pawel Rucki \email{pawel.rucki@roche.com}
  \item Aleksander Chlebowski \email{aleksander.chlebowski@contractors.roche.com} (\href{https://orcid.org/0000-0001-5018-6294}{ORCID})
  \item Andre Verissimo \email{andre.verissimo@roche.com} (\href{https://orcid.org/0000-0002-2212-339X}{ORCID})
  \item Kartikeya Kirar \email{kartikeya.kirar@businesspartner.roche.com}
  \item Vedha Viyash \email{vedha.viyash@roche.com}
  \item Marcin Kosinski \email{marcin.kosinski.mk1@roche.com}
  \item Adrian Waddell \email{adrian.waddell@gene.com}
  \item Nikolas Burkoff
  \item Mahmoud Hallal
  \item Maciej Nasinski
  \item Konrad Pagacz
  \item Junlue Zhao
  \item Tadeusz Lewandowski
}

Other contributors:
\itemize{
  \item Nina Qi \email{qi.ting@gene.com} [reviewer]
  \item Chendi Liao [reviewer]
  \item F. Hoffmann-La Roche AG [copyright holder, funder]
  \item Maximilian Mordig [contributor]
}

}
\keyword{internal}
