% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{teal.devel-package}
\alias{teal.devel}
\alias{teal.devel-package}
\title{Tools for Developers of Teal Modules}
\description{
Utility functions for teal module developers.
}
\author{
\strong{Maintainer}: Adrian Waddell \email{adrian.waddell@roche.com}

}
