% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teal_data.R
\name{teal_data}
\alias{teal_data}
\title{Teal data}
\usage{
teal_data(...)
}
\arguments{
\item{...}{(\code{RelationalDataConnector}, \code{RelationalDataset} or
\code{RelationalDatasetConnector}) elements to include in \code{RelationalData} object}
}
\value{
\code{RelationalData}
}
\description{
Universal function to pass data to teal application
}
\examples{
# RelationalData
library(random.cdisc.data)
x1 <- relational_dataset(
  x = radsl(cached = TRUE),
  dataname = "ADSL",
  keys = get_cdisc_keys("ADSL"),
  code = "ADSL <- radsl(cached = TRUE)",
  label = "ADTTE dataset"
)

x2 <- relational_dataset(
  x = radtte(cached = TRUE),
  dataname = "ADTTE",
  keys = get_cdisc_keys("ADTTE"),
  code = "ADTTE <- radtte(cached = TRUE)",
  label = "ADTTE dataset"
)

data <- cdisc_data(x1, x2)
get_raw_data(data)

# RelationalData with connectors
x3 <- rcd_data( # RelationalDataConnector
  rcd_cdisc_dataset_connector("ADSL", radsl, cached = TRUE),
  rcd_cdisc_dataset_connector("ADLB", radlb, cached = TRUE)
)

x4 <- rcd_data( # RelationalDataConnector
  rcd_cdisc_dataset_connector("ADRS", radrs, cached = TRUE)
)

data_list <- teal_data(x3, x2, x4)
\dontrun{
data_list$launch()
get_raw_data(data_list)
}
}
