% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{teal_data_datanames}
\alias{teal_data_datanames}
\title{Wrapper on \code{teal.data::datanames}}
\usage{
teal_data_datanames(data)
}
\arguments{
\item{data}{(\code{teal_data})}
}
\value{
\code{character}
}
\description{
Special function used in internals of \code{teal} to return names of datasets even if \code{datanames}
has not been set.
}
\keyword{internal}
