% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teal_data.R
\name{teal_data_file}
\alias{teal_data_file}
\title{Load \code{RelationalData} object from a file}
\usage{
teal_data_file(path, code = get_code(path))
}
\arguments{
\item{path}{A (\code{connection}) or a (\code{character})\cr
string giving the pathname of the file or URL to read from. "" indicates the connection \code{stdin}.}

\item{code}{(\code{character})\cr
reproducible code to re-create object}
}
\value{
\code{RelationalData} object
}
\description{
\ifelse{html}{\out{<a href='https://www.tidyverse.org/lifecycle/#experimental'><img src='figures/lifecycle-experimental.svg' alt='Experimental lifecycle'></a>}}{\strong{Experimental}}
Please note that the script has to end with a call creating desired object. The error will be raised otherwise.
}
\examples{
# simple example
file_example <- tempfile(fileext = ".R")
writeLines(
  text = c(
    "library(teal)
     library(random.cdisc.data)

     adsl <- cdisc_dataset(dataname = \"ADSL\",
                           x = radsl(cached = TRUE),
                           code = \"library(random.cdisc.data)\nADSL <- radsl(cached = TRUE)\")

     adtte <- cdisc_dataset(dataname = \"ADTTE\",
                            x = radtte(cached = TRUE),
                            code = \"library(random.cdisc.data)\nADTTE <- radtte(cached = TRUE)\")

     cdisc_data(adsl, adtte)"
  ),
  con = file_example
)
x <- teal_data_file(file_example, code = character(0))
get_code(x)

# exaample with custom datasets
file_example <- tempfile(fileext = ".R")
writeLines(
  text = c(
    "library(teal)
    library(random.cdisc.data)

    # code>
    ADSL <- radsl(cached = TRUE)
    ADTTE <- radtte(cached = TRUE)

    ADSL$n <- nrow(ADTTE)
    ADTTE$n <- nrow(ADSL)
    # <code

    ADSL <- cdisc_dataset(dataname = \"ADSL\", x = ADSL)
    ADTTE <- cdisc_dataset(dataname = \"ADTTE\", x = ADTTE)

    cdisc_data(ADSL, ADTTE)"
  ),
  con = file_example
)
x <- teal_data_file(file_example)
get_code(x)
}
