% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{teal_data_ls}
\alias{teal_data_ls}
\title{Retrieve the dataset names from a \code{teal_data} object}
\usage{
teal_data_ls(data)
}
\arguments{
\item{data}{(\code{teal_data})}
}
\value{
\code{character}
}
\description{
This function extracts the names of datasets stored within a \code{teal_data} object
excluding those that represent raw datasets indicated by name ending with \verb{_raw}.
}
\keyword{internal}
