% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{teal_data_to_filtered_data}
\alias{teal_data_to_filtered_data}
\title{Create a \code{FilteredData}}
\usage{
teal_data_to_filtered_data(x, datanames = teal.data::datanames(x))
}
\arguments{
\item{x}{(\code{teal_data}) object}
}
\value{
(\code{FilteredData}) object
}
\description{
Create a \code{FilteredData} object from a \code{teal_data} object
}
\keyword{internal}
