% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teal_data_utils.R
\name{teal_data_utilities}
\alias{teal_data_utilities}
\alias{.append_evaluated_code}
\alias{.append_modified_data}
\alias{.subset_teal_data}
\alias{.teal_data_datanames}
\alias{.teal_data_ls}
\title{\code{teal_data} utils}
\usage{
.append_evaluated_code(data, code)

.append_modified_data(data, objects)

.subset_teal_data(data, datanames)

.teal_data_datanames(data)

.teal_data_ls(data)
}
\arguments{
\item{data}{(\code{teal_data})}

\item{code}{(\code{character}) code to append to \code{data@code}}

\item{objects}{(\code{list}) objects to append to \code{data@env}}

\item{datanames}{(\code{character}) names of the datasets}
}
\value{
modified \code{teal_data}
}
\description{
In \code{teal} we need to recreate the \code{teal_data} object due to two operations:integer.max
\itemize{
\item we need to append filter-data code and objects which have been evaluated in \code{FilteredData} and
we want to avoid double-evaluation.
\item we need to subset \code{teal_data} to \code{datanames} used by the module, to shorten obtainable R-code
}
}
\details{
Due to above recreation of \code{teal_data} object can't be done simply by using public
\code{teal.code} and \code{teal.data} methods.
}
