% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-transform_module.R
\name{teal_transform}
\alias{teal_transform}
\title{Transform module for \code{teal_data}}
\usage{
teal_transform(data, ui, server)
}
\arguments{
\item{data}{\code{teal_data} object}

\item{ui}{(\verb{function(id)}) function to create UI}

\item{server}{(\verb{function(id, data)}) \code{shiny} server
which returns \code{teal_data} object wrapped in \code{reactive}.}
}
\description{
Function creates object of class \code{teal_trnasform_module} which allows
\code{teal} app developer to transform freely \code{teal_data} object passed to \code{data} argument in
\code{\link[=init]{init()}}. This helps in case when app developer wants to use \code{teal} app
where \code{data} can be influenced by app user. For example, app developer can create
\code{teal} app which allows user to connect to database and then use data from this database.
}
