% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TealDataConnection.R
\name{teradata_connection}
\alias{teradata_connection}
\title{Superseded (moved into separate package) Open connection to \code{Teradata}}
\usage{
teradata_connection(
  open_args = list(),
  close_args = list(),
  ping_args = list()
)
}
\arguments{
\item{open_args}{optional, named (\code{list}) of additional parameters for the connection's
\code{RocheTeradata::connect_teradata} open function. Please note that the \code{type}
argument will be overwritten with \code{ODBC}.}

\item{close_args}{optional, named (\code{list}) of additional parameters for the connection's
\code{\link[DBI:dbDisconnect]{DBI::dbDisconnect()}} close function.}

\item{ping_args}{optional, named (\code{list}) of additional parameters for the connection's
\code{\link[DBI:dbIsValid]{DBI::dbIsValid()}} ping function.}
}
\value{
(\code{TealDataConnection}) type of object
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#superseded}{\figure{lifecycle-superseded.svg}{options: alt='[Superseded]'}}}{\strong{[Superseded]}}
}
