% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RelationalDataConnector_constructors.R
\name{teradata_data}
\alias{teradata_data}
\title{\code{RelationalDataConnector} connector for \code{TERADATA}}
\usage{
teradata_data(..., connection = teradata_connection())
}
\arguments{
\item{...}{(\code{DatasetConnector}) dataset connectors created using \code{\link{teradata_dataset_connector}}}

\item{connection}{(\code{DataConnection}) object returned from \code{teradata_connection}.}
}
\value{
An object of class \code{RelationalDataConnector}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
Build data connector for \code{TERADATA} functions or datasets
}
\examples{
\dontrun{
x <- teradata_data(
  teradata_cdisc_dataset_connector("ADSL", "ADSL_table"),
  teradata_cdisc_dataset_connector("ADAE", "ADAE_table"),
  connection = teradata_connection(open_args = list(datalab = "my_custom_datalab"))
)

# add the data to a proper teal app
app <- init(
  data = teal_data(x),
  modules = root_modules(
    module(
      "ADSL AGE histogram",
      server = function(input, output, session, datasets) {
        output$hist <- renderPlot(
          hist(datasets$get_data("ADSL", filtered = TRUE)$AGE)
        )
      },
      ui = function(id, ...) {ns <- NS(id); plotOutput(ns('hist'))},
      filters = "ADSL"
    )
  ),
  header = tags$h1("Sample App")
)
shinyApp(app$ui, app$server)
}
}
