% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DatasetConnector_constructors.R
\name{teradata_dataset_connector}
\alias{teradata_dataset_connector}
\alias{teradata_cdisc_dataset_connector}
\title{\code{Teradata} \code{DatasetConnector}}
\usage{
teradata_dataset_connector(
  dataname,
  table,
  keys = character(0),
  label = character(0),
  code = character(0),
  script = character(0),
  ...
)

teradata_cdisc_dataset_connector(
  dataname,
  table,
  keys = get_cdisc_keys(dataname),
  parent = if (identical(dataname, "ADSL")) character(0L) else "ADSL",
  label = character(0),
  code = character(0),
  script = character(0),
  ...
)
}
\arguments{
\item{dataname}{(\code{character})\cr
A given name for the dataset it may not contain spaces}

\item{table}{(\code{character}) table name}

\item{keys}{optional, (\code{character})\cr
vector of dataset primary keys column names}

\item{label}{(\code{character})\cr
Label to describe the dataset.}

\item{code}{(\code{character})\cr
A character string defining code to modify \code{raw_data} from this dataset. To modify
current dataset code should contain at least one assignment to object defined in \code{dataname}
argument. For example if \code{dataname = ADSL} example code should contain
\verb{ADSL <- <some R code>}. Can't be used simultaneously with \code{script}}

\item{script}{(\code{character})\cr
Alternatively to \code{code} - location of the file containing modification code.
Can't be used simultaneously with \code{script}.}

\item{...}{Additional arguments applied to pull function.
In case when this object code depends on the \code{raw_data} from the other
\code{Dataset}, \code{DatasetConnector} object(s) or other constant value,
this/these object(s) should be included. Please note that \code{vars}
are included to this object as local \code{vars} and they cannot be modified
within another dataset.}

\item{parent}{optional, (\code{character}) \cr
parent dataset name}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\details{
Create a \code{DatasetConnector} from \code{Teradata}.

Create a \code{CDISCDatasetConnector} from \code{Teradata} with keys and parent name assigned
automatically by \code{dataname}.
}
