% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeClass.R
\name{text_to_call}
\alias{text_to_call}
\title{Create call from string}
\usage{
text_to_call(x)
}
\arguments{
\item{x}{(\code{character}) string containing the code.}
}
\value{
(\code{call}) object.
}
\description{
Create call from string
}
\keyword{internal}
