% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_nest.R
\name{theme_nest}
\alias{theme_nest}
\title{Returns a custom \code{NEST} \code{ggplot2} theme}
\usage{
theme_nest(...)
}
\arguments{
\item{...}{(\code{list}) additional arguments to the \code{ggplot2} theme function}
}
\description{
Returns a custom \code{NEST} \code{ggplot2} theme
}
\examples{
plot <- ggplot2::ggplot(iris, ggplot2::aes(x = Sepal.Length, y = Sepal.Width)) +
  ggplot2::geom_point() +
  theme_nest()
}
