% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_made_up.R
\name{tm_made_up}
\alias{tm_made_up}
\title{Made Up Plot module}
\usage{
tm_made_up(label = "Regression Analysis", dataname = NULL, response,
  regressor, facetting = NULL, pre_output = NULL, post_output = NULL)
}
\arguments{
\item{label}{character}

\item{dataname}{character}

\item{response}{(\code{data_extract})}

\item{regressor}{(\code{data_extract})}

\item{facetting}{(\code{data_extract})}

\item{pre_output}{optional, \code{shiny.tag} with text placed before the
output to put the output into context. For example the
\code{\link[shiny]{helpText}} elements are useful.}

\item{post_output}{optional, \code{shiny.tag} with text placed before the
output to put the output into context. For example the
\code{\link[shiny]{helpText}} elements are useful.}
}
\description{
Made Up Plot module
}
