% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_scatterplot.R
\name{tm_scatterplot}
\alias{tm_scatterplot}
\title{Create a simple cross-table}
\usage{
tm_scatterplot(label, dataname, xvar, yvar, xvar_choices = xvar,
  yvar_choices = yvar, color_by = NULL, color_by_choices = color_by,
  pre_output = NULL, post_output = NULL, plot_height = 600)
}
\arguments{
\item{label}{menu label}

\item{pre_output}{html tags appended below the output}

\item{post_output}{html tags appended after the output}
}
\description{
Create a table with the \code{\link{table}[base]} function
}
\examples{

\dontrun{

x <- teal::init(
  data = list(ASL = generate_sample_data('ASL'),
              AAE = generate_sample_data('AAE')),
  tabs(
     tm_data_table(),
     tm_variable_browser(),
     tm_scatterplot("Scatterplot Choices",
                    dataname = 'AAE',
                    xvar = 'AESDY', yvar = 'AEEDY',
                    color_by = "none", color_by_choices = c("none", "AESTMF", "ANLFL")
     ),
     tm_scatterplot("Scatterplot No Color Choices",
                    dataname = 'ASL',
                    xvar = 'AGE', yvar = 'TRTDUR'
     )
  )
)

shinyApp(x$ui, x$server)


}
}
