% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_splom.R
\name{tm_splom_wide_data}
\alias{tm_splom_wide_data}
\title{Scatterplot Matrix Teal Module}
\usage{
tm_splom_wide_data(label)
}
\description{
Accepts Vertical and Horizontal Data
}
\examples{
anl_hor <- data.frame(
  PARAMCD = rep(letters[1:5], each = 40),
  AVAL = rnorm(40*5)
)


anl_ver <- as.data.frame(replicate(5, rnorm(40), simplify=FALSE))
names(anl_ver) <- letters[1:5]

anl_ver


}
