% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_table.R
\name{tm_table}
\alias{tm_table}
\title{Create a simple cross-table}
\usage{
tm_table(label, dataname, xvar, yvar, xvar_choices = xvar,
  yvar_choices = yvar, useNA = c("ifany", "no", "always"),
  pre_output = NULL, post_output = NULL)
}
\arguments{
\item{label}{menu label}

\item{pre_output}{html tags appended below the output}

\item{post_output}{html tags appended after the output}
}
\description{
Create a table with the \code{\link{table}[base]} function
}
\examples{

\dontrun{

x <- teal::init(
  data = list(ASL = generate_sample_data('ASL')),
  root_modules(
     tm_data_table(),
     tm_variable_browser(),
     tm_table("Table Choices", 'ASL', xvar = 'SEX', yvar = 'RACE',
              xvar_choices = c('SEX', 'RACE', 'STUDYID'),
              yvar_choices = c('RACE', 'SAFFL')),
     tm_table("Table No Choices", 'ASL', 'SEX', 'RACE',
              pre_output = helpText("Titles"),
              post_output = helpText("Footnotes"))
  )
)

shinyApp(x$ui, x$server)

}
}
