% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_variable_browser.R
\name{tm_variable_browser}
\alias{tm_variable_browser}
\title{Variable Browser Teal Module}
\usage{
tm_variable_browser(label = "variable browser")
}
\arguments{
\item{label}{label shown in the navigation for the item}
}
\description{
The variable browser provides a table with variable names and labels and a
plot the visualizes the content of a particular variable.
}
