% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{to_relational_data.DatasetConnector}
\alias{to_relational_data.DatasetConnector}
\title{DatasetConnector input method for \code{\link{to_relational_data}}}
\usage{
\method{to_relational_data}{DatasetConnector}(data, data_call = NULL)
}
\arguments{
\item{data}{\code{RelationalData}, \code{Dataset}, \code{data.frame}, \code{list} or \code{function} returning a named list.}

\item{data_call}{\code{call} output of \code{substitute(data)}.}
}
\description{
DatasetConnector input method for \code{\link{to_relational_data}}
}
\seealso{
\code{\link{to_relational_data}}
}
