% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_relational_data.R
\name{to_relational_data.list}
\alias{to_relational_data.list}
\title{list input method for \code{\link{to_relational_data}}}
\usage{
\method{to_relational_data}{list}(data, data_call = NULL)
}
\arguments{
\item{data}{\code{RelationalData}, \code{Dataset}, \code{data.frame}, \code{list} or \code{function} returning a named list.}

\item{data_call}{\code{call} output of \code{substitute(data)}.}
}
\description{
list input method for \code{\link{to_relational_data}}
}
\seealso{
\code{\link{to_relational_data}}
}
