% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_relational_data.R
\name{to_relational_data}
\alias{to_relational_data}
\title{S3 generic for \code{to_relational_data} function.}
\usage{
to_relational_data(data)
}
\arguments{
\item{data}{\code{TealDataset}, \code{TealDatasetConnector}, \code{data.frame}, \code{list} or \code{function} returning a named list.}
}
\value{
list of \code{TealData} objects
}
\description{
Takes the input of data argument and translates them to relational data objects.
}
\note{
\code{to_relational_data} should only be used inside \code{init} call to guarantee correct behavior.
}
\keyword{internal}
