% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CDISCFilteredData.R
\name{topological_sort}
\alias{topological_sort}
\title{Topological graph sort}
\usage{
topological_sort(graph)
}
\arguments{
\item{graph}{(named \code{list}) list with node vector elements}
}
\description{
graph is a list which for each node contains a vector of child nodes
in the returned list, parents appear before their children.
}
\examples{
correct_orders <- list(
  list("A", "B", "D", "C"),
  list("A", "B", "C", "D"),
  list("A", "B", "C", "D")
)
stopifnot(identical(
  teal:::topological_sort(list(A = c(), B = c("A"), C = c("B"), D = c("A"))),
  correct_orders[[1]]
))
stopifnot(identical(
  teal:::topological_sort(list(D = c("A"), A = c(), B = c("A"), C = c("B"))),
  correct_orders[[2]]
))
stopifnot(identical(
  teal:::topological_sort(list(D = c("A"), B = c("A"), C = c("B"), A = c())),
  correct_orders[[3]]
))
}
