% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{toString.teal_module}
\alias{toString.teal_module}
\title{Convert \code{teal_module} to a string}
\usage{
\method{toString}{teal_module}(x, indent = 0, ...)
}
\arguments{
\item{x}{\code{teal_module}}

\item{indent}{(\code{integer}) indent level;
each submodule is indented one level more}

\item{...}{ignored}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
}
