% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules_debugging.R
\name{trigger_after_first_cycle}
\alias{trigger_after_first_cycle}
\title{Activates the reactive expression after the first reactive flush of Shiny,
This is useful when the UI does not exist yet and we want to wait for it
Note that this triggers the first time due to the timer and from then on due to
the expression being evaluated.}
\usage{
trigger_after_first_cycle(expr)
}
\arguments{
\item{expr}{\verb{function or reactive}}
}
\value{
\code{reactive}
}
\description{
Note: Usage of this function should be avoided as it is a workaround.
}
\details{
\preformatted{# This ensures that the input is evaluated in the second reactive flash
# (and possibly later as well)
get_name <- reactiveVal("hello")
get_name_ui_exists <- trigger_after_first_cycle(get_name)
observeEvent(
  get_name_ui_exists(),
  updateInput(session, "initially_inexistent_input", value = get_name_ui_exists())
)
}
}
\keyword{internal}
