% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataConnector.R
\name{ui_connectors}
\alias{ui_connectors}
\title{Creates UI from \code{DatasetConnector} objects}
\usage{
ui_connectors(type, connectors)
}
\arguments{
\item{type}{\code{character} giving the type of connection.}

\item{connectors}{\code{list} of \code{DatasetConnector} objects.}
}
\value{
\code{shiny.tag} UI describing the connectors
}
\description{
Creates UI from \code{DatasetConnector} objects
}
