% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_filter_overview.R
\name{ui_filter_overview}
\alias{ui_filter_overview}
\title{Creates the UI for the module showing counts for each dataset
contrasting the filtered to the full unfiltered dataset}
\usage{
ui_filter_overview(id)
}
\arguments{
\item{id}{module id}
}
\description{
Per dataset, it displays
the number of rows/observations in each dataset,
the number of unique subjects.
}
\examples{
datasets <- teal:::get_dummy_datasets()
app <- shinyApp(ui = function() {
  tagList(
    teal:::include_teal_css_js(),
    teal:::ui_filter_overview("filter_panel")
  )
}, server = function(input, output, session) {
  callModule(
    teal:::srv_filter_overview, "filter_panel", datasets
  )
})
\dontrun{
runApp(app)
}
}
