% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_filter_panel.R
\name{ui_filter_panel}
\alias{ui_filter_panel}
\title{Module for the right filter panel in the teal app
with a filter overview panel and a filter variable panel.}
\usage{
ui_filter_panel(id, datasets, datanames)
}
\arguments{
\item{id}{(\code{character(1)})\cr
module id}

\item{datasets}{(\code{FilteredData})\cr
object which stores datasets and manages filters.}

\item{datanames}{(\code{character})\cr
datanames to create empty UIs for (which will be populated in the server)}
}
\description{
This panel contains info about the number of observations left in
the (active) datasets and allows to filter the datasets.
}
