% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_filtered_data_overview.R
\name{ui_filtered_data_overview}
\alias{ui_filtered_data_overview}
\title{Creates the UI for the module showing counts for each dataset
contrasting the filtered to the full unfiltered dataset}
\usage{
ui_filtered_data_overview(id)
}
\arguments{
\item{id}{module id}
}
\description{
Per dataset, it displays
the number of rows/observations in each dataset,
the number of unique subjects.
}
\examples{
datasets <- get_dummy_datasets()
app <- shinyApp(ui = function() {
  tagList(
    teal:::include_teal_css_js(),
    teal:::ui_filtered_data_overview("filter_panel")
  )
}, server = function(input, output, session) {
  callModule(
    teal:::srv_filtered_data_overview, "filter_panel", datasets
  )
})
\dontrun{
  runApp(app)
}
}
