% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_snapshot_manager.R
\name{unfold_mapping}
\alias{unfold_mapping}
\title{Explicitly enumerate global filters.}
\usage{
unfold_mapping(mapping, module_names)
}
\arguments{
\item{mapping}{(named \code{list}) as stored in mapping parameter of \code{teal_slices}}

\item{module_names}{(\code{character}) vector containing names of all modules in the app}
}
\value{
A \code{named_list} with one element per module, each element containing all filters applied to that module.
}
\description{
Transform module mapping such that global filters are explicitly specified for every module.
}
\keyword{internal}
