% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{update_default_dataname}
\alias{update_default_dataname}
\title{Retrieve and Update Default Dataset Names from a Data Object}
\usage{
update_default_dataname(data)
}
\arguments{
\item{tealData}{(\code{teal_data}) object}
}
\value{
A \code{teal_data} object.
}
\description{
This function extracts dataset names from a \code{teal_data} object. If the object has no predefined
dataset names.
}
\keyword{internal}
