% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{use_reporter}
\alias{use_reporter}
\alias{use_reporter.default}
\alias{use_reporter.teal_modules}
\alias{use_reporter.teal_module}
\title{Does the object make use of \code{teal.reporter} reporting}
\usage{
use_reporter(modules)

\method{use_reporter}{default}(modules)

\method{use_reporter}{teal_modules}(modules)

\method{use_reporter}{teal_module}(modules)
}
\arguments{
\item{modules}{\code{teal_module} or \code{teal_modules} object}
}
\value{
\code{logical} whether the object makes use of \code{teal.reporter} reporting
}
\description{
Does the object make use of \code{teal.reporter} reporting
}
\keyword{internal}
