% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_app_title_tag}
\alias{validate_app_title_tag}
\title{Function for validating the title parameter of \code{teal::init}}
\usage{
validate_app_title_tag(shiny_tag)
}
\arguments{
\item{shiny_tag}{(\code{shiny.tag}) Object to validate for a valid title.}
}
\description{
Checks if the input of the title from \code{teal::init} will create a valid title and favicon tag.
}
\keyword{internal}
