% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.0_validate_reactive_teal_data.R
\name{validate_datanames}
\alias{validate_datanames}
\alias{ui_validate_datanames}
\alias{srv_validate_datanames}
\title{Validate reactive \code{teal_data} datanames}
\usage{
ui_validate_datanames(id)

srv_validate_datanames(id, modules, data)
}
\arguments{
\item{id}{(\code{character}) optional
string specifying the \code{shiny} module id in cases it is used as a \code{shiny} module
rather than a standalone \code{shiny} app. This is a legacy feature.}

\item{modules}{(\code{list} or \code{teal_modules} or \code{teal_module})
nested list of \code{teal_modules} or \code{teal_module} objects or a single
\code{teal_modules} or \code{teal_module} object. These are the specific output modules which
will be displayed in the \code{teal} application. See \code{\link[=modules]{modules()}} and \code{\link[=module]{module()}} for
more details.}

\item{data}{(\code{teal_data}, \code{teal_data_module} or \code{reactive} returning \code{teal_data})}

\item{validate_shiny_silent_error}{(\code{logical}) If \code{TRUE}, then \code{shiny.silent.error} is validated and
error message is displayed.
Default is \code{FALSE} to handle empty reactive cycle on init.}
}
\value{
(\code{reactive} returning \code{teal_data})
}
\description{
Validate reactive \code{teal_data} datanames
}
\keyword{internal}
