% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choices_labeled.R
\name{value_choices}
\alias{value_choices}
\title{Wrapper on \code{\link{choices_labeled}} to label variable values basing on other variable values}
\usage{
value_choices(data, var_choices, var_label, subset = NULL, sep = " - ")
}
\arguments{
\item{data}{(\code{data.frame}) data to extract labels from}

\item{var_choices}{(\code{character}) vector with choices column names}

\item{var_label}{(\code{character}) vector with labels column names}

\item{subset}{(\code{vector}) vector with values to subset}

\item{sep}{(\code{character}) separator used in case of multiple column names}
}
\value{
named character vector
}
\description{
Wrapper on \code{\link{choices_labeled}} to label variable values basing on other variable values
}
\examples{
library(random.cdisc.data)
ADRS <- radrs(cached = TRUE)

value_choices(cadrs, "PARAMCD", "PARAM", subset = c("BESRSPI", "INVET"))
value_choices(cadrs, c("PARAMCD", "ARMCD"), c("PARAM", "ARM"))
value_choices(cadrs, c("PARAMCD", "ARMCD"), c("PARAM", "ARM"),
  subset = c("BESRSPI - ARM A", "INVET - ARM A", "OVRINV - ARM A"))
value_choices(cadrs, c("PARAMCD", "ARMCD"), c("PARAM", "ARM"), sep = " --- ")
}
