% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_label.R
\name{var_relabel}
\alias{var_relabel}
\title{Copy and Change Variable Labels of a \code{data.frame}}
\usage{
var_relabel(x, ...)
}
\arguments{
\item{x}{(\code{data.frame}) object with columns}

\item{...}{name-value pairs, where name corresponds to a variable name in
\code{x} and the value to the new variable label}
}
\value{
a copy of \code{x} with changed labels according to \code{...}
}
\description{
Relabels a subset of the variables.
}
\examples{
x <- var_relabel(iris, Sepal.Length = "Sepal Length of iris flower")
variable_labels(x)

}
