% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_variable_browser.R
\name{variable_browser_item}
\alias{variable_browser_item}
\title{Create a variable browser item that plots a variable summary}
\usage{
variable_browser_item(label = "variable browser")
}
\arguments{
\item{label}{label shown in the navigation for the item}
}
\description{
The variable browser provides a table with variable names and labels and a
plot the visualizes the content of a particular variable.
}
