% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choices_labeled.R
\name{variable_choices}
\alias{variable_choices}
\title{Wrapper on \code{\link{choices_labeled}} to label variables basing on existing labels in data}
\usage{
variable_choices(data, subset = NULL)
}
\arguments{
\item{data}{(\code{data.frame}) data to extract labels from}

\item{subset}{(\code{character}) vector of column names}
}
\value{
named character vector with additional attributes
}
\description{
Wrapper on \code{\link{choices_labeled}} to label variables basing on existing labels in data
}
\examples{
library(random.cdisc.data)
ADRS <- radrs(cached = TRUE)

variable_choices(ADRS)
variable_choices(ADRS, subset = c("PARAM", "PARAMCD"))
}
