% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_label.R
\name{variable_labels<-}
\alias{variable_labels<-}
\title{Sets the column labels of a \code{data.frame}.}
\usage{
variable_labels(x) <- value
}
\arguments{
\item{x}{(\code{data.frame}) object with columns}

\item{value}{(\code{charater}) labels}
}
\value{
\code{x}
}
\description{
Sets the column labels of a \code{data.frame}.
}
\examples{
variable_labels(iris) <- colnames(iris)

}
