% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choices_icons.R
\name{variable_type_icons}
\alias{variable_type_icons}
\title{Get icons to represent variable types in dataset}
\usage{
variable_type_icons(columns, data)
}
\arguments{
\item{columns}{(atomic vector of \code{character}) columns to take from data}

\item{data}{(\code{data.frame}) data to determine variable types from}
}
\value{
(atomic vector of \code{character}) vector of HTML icons corresponding to
  data type in each column, NULL item if in unknown category
}
\description{
Get icons to represent variable types in dataset
}
\examples{
teal:::variable_type_icons(
  c("x"),
  data.frame(
    x = 1:3, y = factor(c("a", "b", "a")), z = c("h1", "h2", "h3"),
    stringsAsFactors = FALSE
  )
)

teal:::variable_type_icons(
  c("x", "z"),
  data.frame(x = 1:3, y = factor(c("a", "b", "a")), z = c("h1", "h2", "h3"),
    stringsAsFactors = FALSE)
)
teal:::variable_type_icons(character(0), data.frame(x = 1:3))
}
