% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choices_labeled.R
\name{variable_types}
\alias{variable_types}
\title{Get classes selected columns from dataset}
\usage{
variable_types(data, columns = NULL)
}
\arguments{
\item{data}{(\code{data.frame}) data to determine variable types from}

\item{columns}{(atomic vector of \code{character} or \code{NULL}) column names chosen chosen from \code{data},
\code{NULL} for all data columns}
}
\value{
(atomic vector of \code{character}) classes of \code{columns} from provided \code{data}
}
\description{
Get classes selected columns from dataset
}
\examples{
teal:::variable_types(
  data.frame(
    x = 1:3, y = factor(c("a", "b", "a")), z = c("h1", "h2", "h3"),
    stringsAsFactors = FALSE
  ),
  "x"
)

teal:::variable_types(
  data.frame(
    x = 1:3, y = factor(c("a", "b", "a")), z = c("h1", "h2", "h3"),
    stringsAsFactors = FALSE
  ),
  c("x", "z")
)

teal:::variable_types(
  data.frame(
    x = 1:3, y = factor(c("a", "b", "a")), z = c("h1", "h2", "h3"),
    stringsAsFactors = FALSE
  )
)
}
\keyword{internal}
