% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{whiteSmallWell}
\alias{whiteSmallWell}
\title{Add Class Small Well and overflow-x property to HTML output element}
\usage{
whiteSmallWell(...)
}
\arguments{
\item{...}{other arguments to pass to tag object's div attributes.}
}
\value{
An HTML output element with class Small Well and overflow-x property
}
\description{
Add Class Small Well and overflow-x property to HTML output element
}
\details{
\code{whiteSmallWell} is intended to be used with \code{\link{uiOutput}}.
The overflow-x property is set to auto so that a scroll bar is added
when the content overlows at the left and right edges of the output window.
For example, this is useful for displaying wide tables.
}
\examples{

whiteSmallWell(htmlOutput("summary"))

}
