% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{add_custom_server}
\alias{add_custom_server}
\title{Add a Custom Server Logic to \code{teal} Application}
\usage{
add_custom_server(app, custom_server)
}
\arguments{
\item{app}{(\code{list}) The \code{teal} app environment.}

\item{custom_server}{(\verb{function(input, output, session)}) The custom server function to set.}
}
\description{
Adds a custom server function to the \code{teal} app. This function can define additional server logic.
}
\examples{
app <- init(
  data = teal_data(IRIS = iris),
  modules = modules(example_module())
) |>
  add_custom_server(function(input, output, session) {
    print("injected server logic")
  })

if (interactive()) {
  shinyApp(app$ui, app$server)
}
}
