% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{add_custom_server}
\alias{add_custom_server}
\title{Add a Custom Server Function to a \code{teal} App}
\usage{
add_custom_server(app, custom_server)
}
\arguments{
\item{app}{(\code{environment}) The \code{teal} app environment.}

\item{custom_server}{(\code{function}) The custom server function to set.}
}
\value{
The modified \code{teal} app environment, invisibly.
}
\description{
Adds a custom server function to the \code{teal} app. This function can define additional server logic.
}
\examples{
app <- list2env(list(custom_server = NULL))
custom_server <- function(input, output, session) {
  message("Custom server logic here")
}
app <- add_custom_server(app, custom_server)
}
