% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{add_footer}
\alias{add_footer}
\title{Add a Footer to a \code{teal} App}
\usage{
add_footer(app, footer = tags$p())
}
\arguments{
\item{app}{(\code{environment}) The \code{teal} app environment.}

\item{footer}{(\code{shiny.tag} or \code{character(1)}) The footer content to set. Defaults to an empty paragraph tag.}
}
\value{
The modified \code{teal} app environment, invisibly.
}
\description{
Adds a footer to the \code{teal} app.
}
\examples{
app <- list2env(list(footer = NULL))
app <- add_footer(app, tags$p("App Footer"))
}
