% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{add_header}
\alias{add_header}
\title{Add a Header to a \code{teal} App}
\usage{
add_header(messages, header = "")
}
\arguments{
\item{header}{(\code{shiny.tag} or \code{character(1)}) The header content to set. Defaults to an empty paragraph tag.}

\item{app}{(\code{environment}) The \code{teal} app environment.}
}
\value{
The modified \code{teal} app environment, invisibly.
}
\description{
Adds a header to the \code{teal} app.
}
\examples{
app <- list2env(list(header = NULL))
app <- add_header(app, tags$h1("App Header"))
}
