% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teal_modifiers.R
\name{add_landing_modal}
\alias{add_landing_modal}
\title{Add a Landing Popup to \code{teal} Application}
\usage{
add_landing_modal(
  x,
  title = NULL,
  content = NULL,
  footer = modalButton("Accept"),
  ...
)
}
\arguments{
\item{x}{(\code{teal_app}) A \code{teal_app} object created using the \code{init} function.}

\item{title}{An optional title for the dialog.}

\item{footer}{UI for footer. Use \code{NULL} for no footer.}

\item{...}{Additional arguments to \code{\link[shiny:modalDialog]{shiny::modalDialog()}}.}
}
\description{
Adds a landing popup to the \code{teal} app. This popup will be shown when the app starts.
The dialog must be closed by the app user to proceed to the main application.
}
\examples{
app <- init(
  data = teal_data(IRIS = iris, MTCARS = mtcars),
  modules = modules(example_module())
) |>
  add_landing_modal(
    title = "Welcome",
    content = "This is a landing popup.",
    buttons = modalButton("Accept")
  )

if (interactive()) {
  shinyApp(app$ui, app$server)
}
}
