% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{add_landing_popup}
\alias{add_landing_popup}
\title{Add a Landing Popup to a \code{teal} App}
\usage{
add_landing_popup(
  app,
  id = "landingpopup",
  label = "Landing Popup",
  title = NULL,
  content = NULL,
  buttons = modalButton("Accept")
)
}
\arguments{
\item{app}{(\code{environment}) The \code{teal} app environment.}

\item{id}{(\code{character(1)}) The ID of the modal dialog.}

\item{label}{(\code{character(1)}) Label of the module.}

\item{title}{(\code{character(1)}) Text to be displayed as popup title.}

\item{content}{(\code{character(1)}, \code{shiny.tag} or \code{shiny.tag.list}) with the content of the popup.
Passed to \code{...} of \code{shiny::modalDialog}. See examples.}

\item{buttons}{(\code{shiny.tag} or \code{shiny.tag.list}) Typically a \code{modalButton} or \code{actionButton}. See examples.}
}
\description{
Adds a landing popup to the \code{teal} app. This popup will be shown when the app starts.

This modifier is used to display a popup dialog when the application starts.
The dialog blocks access to the application and must be closed with a button before the application can be viewed.
}
\examples{
app <- init(
  data = teal_data(IRIS = iris, MTCARS = mtcars),
  modules = modules(example_module())
) |>
  add_landing_popup(
    title = "Welcome",
    content = "This is a landing popup.",
    buttons = modalButton("Accept")
  )

shinyApp(app$ui, app$server)
}
