% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selected_choices.R
\name{add_no_selected_choices}
\alias{add_no_selected_choices}
\title{Add empty choice to choices selected}
\usage{
add_no_selected_choices(x, multiple = FALSE)
}
\arguments{
\item{x}{(\code{choices_selected}) output}

\item{multiple}{(\code{logical}) whether multiple selections are allowed or not}
}
\description{
Add empty choice to choices selected
}
