% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{add_title}
\alias{add_title}
\title{Add a custom title to the app}
\usage{
add_title(app, title = build_app_title())
}
\arguments{
\item{title}{The title to add}
}
\value{
The modified app object
}
\description{
Add a custom title to the app
}
\examples{
app <- init(
  data = teal_data(IRIS = iris, MTCARS = mtcars),
  modules = modules(example_module())
) |>
  add_title("Custom title")

shinyApp(app$ui, app$server)
}
