% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/after.R
\name{after}
\alias{after}
\title{Executes modifications to the result of a module}
\usage{
after(
  x,
  ui = function(id, elem) elem,
  server = function(input, output, session, data) data,
  ...
)
}
\arguments{
\item{x}{(\code{teal_module}) A teal module.}

\item{ui}{(\verb{function(id, elem, ...)}) function to receive output (\code{shiny.tag}) from \code{x$ui}.}

\item{server}{(\verb{function(input, output, session, data, ...)}) function to receive output data from \code{x$server}.}

\item{...}{additional argument passed to \code{ui} and \code{server} by matching their formals names.}
}
\value{
A \code{teal_module} object with the modifications.
New element ids are under \code{wrapper} namespace, old elements' ids are on the \code{wrapped} namesepace.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\details{
Primarily used to modify the output object of module to change the containing
report.
}
\examples{
library("teal.reporter")
app <- init(
  data = teal_data(IRIS = iris, MTCARS = mtcars),
  modules = example_module() |>
    after(
      ui = function(id, elem) {
        ns <- NS(id)
        check_box <- checkboxInput(ns("src"), "Include R Code in the report", TRUE)
        htmltools::tagAppendChild(elem, check_box, .cssSelector = ".standard-layout .sidebar .sidebar-content")
      },
      server = function(input, output, session, data) {
        teal_card(data) <- c(teal_card(data), teal_card("Modification"))
        if (!input$`wrapper-src`) {
          teal_card(data) <- Filter(function(x) !inherits(x, "code_chunk"), teal_card(data))
        }
        data
      }
    )
)

runApp(app)
}
\seealso{
To modify just the output see \code{\link{teal_transform_module}}.
}
