% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/after.R
\name{after}
\alias{after}
\title{Executes modifications to the result of a module}
\usage{
after(
  x,
  ui = function(id, elem) elem,
  server = function(input, output, session, data) data,
  ...
)
}
\arguments{
\item{x}{(\code{teal_module}).}

\item{ui}{(\verb{function(id, elem, ...)}) function to receive output (\code{shiny.tag}) from \code{x$ui}}

\item{server}{(\verb{function(input, output, session, data, ...)}) function to receive output data from \code{x$server}}

\item{...}{additional argument passed to \code{ui} and \code{server} by matching their formals names.}
}
\value{
A \code{teal_report} object with the result of the server function.
}
\description{
Primarily used to modify the output object of module to change the containing
report.
}
