% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_choices.R
\name{all_choices}
\alias{all_choices}
\title{An S3 structure representing the selection of all
possible choices in a \code{filter_spec}, \code{select_spec} or \code{choices_selected} object.}
\usage{
all_choices()
}
\value{
\code{all_choices} object
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
# Both structures are semantically identical
filter_spec(
  vars = c("selected_variable"),
  choices = c("value1", "value2"),
  selected = c("value1", "value2")
)

filter_spec(
  vars = c("selected_variable"),
  choices = c("value1", "value2"),
  selected = all_choices()
)

choices_selected(choices = letters, selected = letters)
choices_selected(choices = letters, selected = all_choices())
}
