% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_state_manager.R
\name{app_state_grab}
\alias{app_state_grab}
\title{Grab selection state (value) of all input items in the app.}
\usage{
app_state_grab()
}
\value{
Object of class \code{teal_grab}, which is a list of lists,
each of which has two elements, one named "id" and the other "value".
}
\description{
Grab selection state (value) of all input items in the app.
}
\seealso{
\code{\link{app_state_store}}, \code{\link{app_state_restore}}, \code{\link{state_manager_module}}
}
\keyword{internal}
