% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{append_module}
\alias{append_module}
\title{Append a \code{teal_module} to \code{children} of a \code{teal_modules} object}
\usage{
append_module(modules, module)
}
\arguments{
\item{modules}{(\code{teal_modules})}

\item{module}{(\code{teal_module}) object to be appended onto the children of \code{modules}}
}
\value{
A \code{teal_modules} object with \code{module} appended.
}
\description{
Append a \code{teal_module} to \code{children} of a \code{teal_modules} object
}
\keyword{internal}
