% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_relational.R
\name{as_relational}
\alias{as_relational}
\title{Convert a \verb{Named<...>} object to a \verb{Relational<...>} object}
\usage{
as_relational(x, keys)
}
\arguments{
\item{x}{an object of (\code{NamedDataset} or \code{NamedDatasetConnector}) class}

\item{keys}{optional, (\code{character})\cr
vector with primary keys}
}
\value{
(\code{RelationalDataset} or \code{RelationalDatasetConnector}) object
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#soft-deprecated}{\figure{lifecycle-soft-deprecated.svg}{options: alt='[Soft-deprecated]'}}}{\strong{[Soft-deprecated]}}
}
\details{
Convert a \verb{Named<...>} to a \verb{Relational<...>}
}
