% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_utills.R
\name{call_condition_range_posixct}
\alias{call_condition_range_posixct}
\title{\code{POSIXct} range condition call}
\usage{
call_condition_range_posixct(varname, range, timezone = Sys.timezone())
}
\arguments{
\item{varname}{(\code{name} or \code{character(1)})\cr
name of the variable}

\item{range}{(\code{POSIXct})\cr
range of the variable. Be aware that output
uses truncated range format \code{"\%Y-\%m-\%d \%H:\%M:\%S"}, which means that
some precision might be lost.}

\item{timezone}{(\code{character(1)})\cr
specifies the time zone to be used for the conversion.
By default \code{Sys.timezone()} is used.}
}
\value{
a \code{call}
}
\description{
Compose \code{POSIXct} range condition call from inputs.
}
\examples{
call_condition_range_posixct(
  varname = as.name("datetime"),
  range = c(Sys.time(), Sys.time() + 1),
  timezone = "UTC"
)
}
\keyword{internal}
