% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CallableFunction.R
\name{callable_function}
\alias{callable_function}
\title{Create \code{CallableFunction} object}
\usage{
callable_function(fun)
}
\arguments{
\item{fun}{(\code{function})\cr
any R function, directly by name or \code{character} string.}
}
\value{
\code{CallableFunction} object
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
Create \code{\link{CallableFunction}} object to execute specific function and get reproducible
call.
}
\examples{
cf <- callable_function(fun = stats::median)
cf$set_args(list(x = 1:10, na.rm = FALSE))
cf$run()
cf$get_call()
}
