% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CallableFunction.R
\docType{data}
\name{CallableFunction}
\alias{CallableFunction}
\title{A \code{CallableFunction} class of objects}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
CallableFunction
}
\arguments{
\item{fun}{(\code{function}) function to set up}

\item{return}{(\code{logical}) whether to return an object}

\item{args}{(\code{NULL} or named \code{list}) dynamic arguments to function}

\item{try}{(\code{logical}) whether perform function evaluation inside \code{try} clause}

\item{deparse}{(\code{logical}) whether to return a deparsed version of call}

\item{args}{(\code{NULL} or named \code{list}) dynamic arguments to function}

\item{args}{(\code{NULL} or named \code{list}) dynamic arguments to function}

\item{name}{(\code{character}) argument name}

\item{value}{argument value}
}
\value{
new \code{CallableFunction} object

nothing or output from function depending on \code{return} argument

\code{call} or \code{character} depending on \code{deparse} argument

nothing

nothing

nothing
}
\description{
Create a new \code{CallableFunction} object

Run function

Get function call with substituted arguments in \code{args}.
These arguments will not be stored in the object.

Refresh call with function name and saved arguments

Set up function arguments

Set up single function argument with value
}
\details{
Object that stores function name with it's arguments. Methods to get call and run it.
}
\keyword{datasets}
