% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CallableFunction.R
\name{CallableFunction}
\alias{CallableFunction}
\title{A \code{CallableFunction} class of objects

Object that stores function name with its arguments. Methods to get call and run it.}
\description{
A \code{CallableFunction} class of objects

Object that stores function name with its arguments. Methods to get call and run it.

A \code{CallableFunction} class of objects

Object that stores function name with its arguments. Methods to get call and run it.
}
\keyword{internal}
\section{Super class}{
\code{\link[teal:Callable]{teal::Callable}} -> \code{CallableFunction}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{CallableFunction$new()}}
\item \href{#method-get_args}{\code{CallableFunction$get_args()}}
\item \href{#method-get_call}{\code{CallableFunction$get_call()}}
\item \href{#method-set_args}{\code{CallableFunction$set_args()}}
\item \href{#method-set_arg_value}{\code{CallableFunction$set_arg_value()}}
\item \href{#method-clone}{\code{CallableFunction$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="teal" data-topic="Callable" data-id="assign_to_env">}\href{../../teal/html/Callable.html#method-assign_to_env}{\code{teal::Callable$assign_to_env()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="teal" data-topic="Callable" data-id="get_error_message">}\href{../../teal/html/Callable.html#method-get_error_message}{\code{teal::Callable$get_error_message()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="teal" data-topic="Callable" data-id="is_failed">}\href{../../teal/html/Callable.html#method-is_failed}{\code{teal::Callable$is_failed()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="teal" data-topic="Callable" data-id="run">}\href{../../teal/html/Callable.html#method-run}{\code{teal::Callable$run()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new \code{CallableFunction} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CallableFunction$new(fun, env = new.env(parent = parent.env(globalenv())))}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{fun}}{(\code{function})\cr
function to be evaluated in class. Function should be named}

\item{\code{env}}{(\code{environment})\cr
environment where function will be evaluated}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
new \code{CallableFunction} object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_args"></a>}}
\if{latex}{\out{\hypertarget{method-get_args}{}}}
\subsection{Method \code{get_args()}}{
get the arguments a function gets called with
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CallableFunction$get_args()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
arguments the function gets called with
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_call"></a>}}
\if{latex}{\out{\hypertarget{method-get_call}{}}}
\subsection{Method \code{get_call()}}{
Get function call with substituted arguments in \code{args}.
These arguments will not be stored in the object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CallableFunction$get_call(deparse = TRUE, args = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deparse}}{(\code{logical} value)\cr
whether to return a deparsed version of call}

\item{\code{args}}{(\code{NULL} or named \code{list})\cr
dynamic arguments to function}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{call} or \code{character} depending on \code{deparse} argument
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_args"></a>}}
\if{latex}{\out{\hypertarget{method-set_args}{}}}
\subsection{Method \code{set_args()}}{
Set up function arguments
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CallableFunction$set_args(args)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{args}}{(\code{NULL} or named \code{list})\cr
function arguments to be stored persistently in the object. Setting \code{args} doesn't
remove other \code{args}, only create new of modify previous of the same name.
To clean arguments specify \code{args = NULL}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
(\code{self}) invisibly for chaining.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_arg_value"></a>}}
\if{latex}{\out{\hypertarget{method-set_arg_value}{}}}
\subsection{Method \code{set_arg_value()}}{
Set up single function argument with value
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CallableFunction$set_arg_value(name, value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{(\code{character}) argument name}

\item{\code{value}}{argument value}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
(\code{self}) invisibly for chaining.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CallableFunction$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
