% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CDISC_data.R
\name{CDISC_data}
\alias{CDISC_data}
\title{CDISC data check}
\usage{
CDISC_data(...)
}
\arguments{
\item{...}{datasets constructed with \link{data_for_teal}}
}
\value{
a list of the input data sets
}
\description{
CDISC data check
}
\details{
This function checks if there were keys added to all data
	sets that shall be analyzed inside a teal app.
}
