% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdisc_data.R
\name{cdisc_data}
\alias{cdisc_data}
\title{Data input for teal app}
\usage{
cdisc_data(ADSL, ..., code = "", check = FALSE)
}
\arguments{
\item{ADSL}{ADSL dataset object}

\item{...}{other datasets objects}

\item{code}{(\code{character}) preprocessing code.}

\item{check}{(\code{logical}) reproducibility check - whether evaluated preprocessing code gives the same objects
as provided in arguments. Check is run only if flag is true and preprocessing code is not empty.}
}
\value{
a list of the input data sets
}
\description{
Function passes datasets to teal application with option to read preprocessing code and reproducibility checking.
}
\details{
This function checks if there were keys added to all data
  sets that shall be analyzed inside a teal app.
}
\examples{
library(random.cdisc.data)

ADSL <-  suppressWarnings(radsl(N = 600, seed = 123))
ADTTE <- radtte(ADSL, event.descr = c("STUDYID", "USUBJID", "PARAMCD"), seed = 123)

cdisc_data(
  cdisc_dataset("ADSL", ADSL),
  cdisc_dataset("ADTTE", ADTTE)),
  code = 'ADSL <- radsl(N = 600, seed = 123)
          ADTTE <- radtte(ADSL, event.descr = c("STUDYID", "USUBJID", "PARAMCD"), seed = 123)')
}
