% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CDISCDataConnector.R
\name{cdisc_data_connector}
\alias{cdisc_data_connector}
\title{Public facing object constructor for \code{CDISCDataConnector} class.}
\usage{
cdisc_data_connector(connection, connectors)
}
\arguments{
\item{connection}{(\code{DataConnection})\cr
connection to data source}

\item{connectors}{(\code{list} of \code{DatasetConnector} elements)\cr
list with dataset connectors}
}
\value{
\code{CDISCDataConnector} object
}
\description{
Public facing object constructor for \code{CDISCDataConnector} class.
}
