% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CDISCDataConnector.R
\name{cdisc_data_connector}
\alias{cdisc_data_connector}
\title{Public facing object constructor for \code{CDISCDataConnector} class.}
\usage{
cdisc_data_connector(connection, connectors)
}
\arguments{
\item{connection}{(\code{DataConnection})\cr
connection to data source}

\item{connectors}{(\code{list} of \code{DatasetConnector} elements)\cr
list with dataset connectors}
}
\value{
\code{CDISCDataConnector} object
}
\description{
Public facing object constructor for \code{CDISCDataConnector} class.
}
\examples{
\dontrun{
adsl_cf <- CallableFunction$new(function() as.data.frame(as.list(setNames(nm = get_cdisc_keys("ADSL")))))
adae_cf <- CallableFunction$new(function() as.data.frame(as.list(setNames(nm = get_cdisc_keys("ADAE")))))
adsl <- CDISCDatasetConnector$new("ADSL", adsl_cf, keys = get_cdisc_keys("ADSL"), parent = character(0))
adae <- CDISCDatasetConnector$new("ADAE", adae_cf, keys = get_cdisc_keys("ADAE"), parent = "ADSL")
data <- cdisc_data_connector(
  connection = data_connection(open_fun = CallableFunction$new(function() "open function")),
  connectors = list(adsl, adae)
)
}
}
