% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdisc_data.R
\name{cdisc_data_file}
\alias{cdisc_data_file}
\title{Load \code{cdisc_data} object from a file}
\usage{
cdisc_data_file(x)
}
\arguments{
\item{x}{A (\code{connection}) or a (\code{character}) string giving the pathname
of the file or URL to read from. "" indicates the connection \code{stdin}.}
}
\value{
\code{cdisc_data} object if file returns a \code{cdisc_data}
  object.
}
\description{
Load \code{cdisc_data} object from a file
}
\examples{
file_example <- tempfile(fileext = ".R")
writeLines(
  text = c(
    "library(random.cdisc.data)

     ADSL <- radsl(cached = TRUE)
     ADTTE <- radtte(ADSL, cached = TRUE)

     cdisc_data(
          cdisc_dataset(\"ADSL\", ADSL), cdisc_dataset(\"ADTTE\", ADTTE),
          code = \"ADSL <- radsl(cached = TRUE)
                  ADTTE <- radtte(ADSL, cached = TRUE)\",
          check = FALSE
     )"
  ),
  con = file_example
)

cdisc_data_file(file_example)

}
