% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CDISCTealData.R
\name{cdisc_data_file}
\alias{cdisc_data_file}
\title{Load \code{CDISCTealData} object from a file}
\usage{
cdisc_data_file(path, code = get_code(path))
}
\arguments{
\item{path}{A (\code{connection}) or a (\code{character})\cr
string giving the pathname of the file or URL to read from. "" indicates the connection \code{stdin}.}

\item{code}{(\code{character})\cr
reproducible code to re-create object}
}
\value{
\code{CDISCTealData} object
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
file_example <- tempfile(fileext = ".R")
writeLines(
  text = c(
    "library(scda)

     # code>
     ADSL <- synthetic_cdisc_data('latest')$adsl
     ADTTE <- synthetic_cdisc_data('latest')$adtte

     cdisc_data(
          cdisc_dataset(\"ADSL\", ADSL), cdisc_dataset(\"ADTTE\", ADTTE),
          code = \"ADSL <- synthetic_cdisc_data('latest')$adsl
                  ADTTE <- synthetic_cdisc_data('latest')$adtte\",
          check = FALSE
     )
     # <code"
  ),
  con = file_example
)

cdisc_data_file(file_example)
}
