% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CDISCDataset.R
\name{cdisc_dataset}
\alias{cdisc_dataset}
\title{Create a new object of \code{CDISCDataset} class}
\usage{
cdisc_dataset(
  dataname,
  x,
  keys = get_cdisc_keys(dataname),
  parent = if (identical(dataname, "ADSL")) character(0) else "ADSL",
  label = data_label(x),
  code = character(0),
  vars = list()
)
}
\arguments{
\item{dataname}{(\code{character})\cr
A given name for the dataset it may not contain spaces}

\item{x}{(\code{data.frame})\cr}

\item{keys}{optional, (\code{character})\cr
vector with primary keys}

\item{parent}{optional, (\code{character}) \cr
parent dataset name}

\item{label}{(\code{character})\cr
Label to describe the dataset}

\item{code}{(\code{character})\cr
A character string defining the code needed to produce the data set in \code{x}}

\item{vars}{(named \code{list})) \cr
In case when this object code depends on other \code{Dataset} object(s) or
other constant value, this/these object(s) should be included as named
element(s) of the list. For example if this object code needs \code{ADSL}
object we should specify \verb{vars = list(ADSL = <adsl object>)}.
It's recommended to include \code{Dataset} or \code{DatasetConnector} objects to
the \code{vars} list to preserve reproducibility. Please note that \code{vars}
are included to this object as local \code{vars} and they cannot be modified
within another dataset.}
}
\value{
(\code{CDISCDataset}) a dataset with connected metadata
}
\description{
\ifelse{html}{\out{<a href='https://www.tidyverse.org/lifecycle/#experimental'><img src='figures/lifecycle-experimental.svg' alt='Experimental lifecycle'></a>}}{\strong{Experimental}}
Function that creates \code{CDISCDataset} object
}
\examples{
library(random.cdisc.data)

ADSL <- radsl(cached = TRUE)

cdisc_dataset("ADSL", ADSL)
}
