% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdisc_data.R
\name{cdisc_dataset}
\alias{cdisc_dataset}
\title{Data input for teal app}
\usage{
cdisc_dataset(dataname, data, keys = get_cdisc_keys(dataname),
  labels = get_cdisc_labels(data))
}
\arguments{
\item{dataname}{name of dataset}

\item{data}{data}

\item{labels}{}
}
\value{
a dataset with connected metadata
}
\description{
Function that creates CDISC dataset object
}
\examples{
library(random.cdisc.data)

ASL <-  suppressWarnings(radsl(N = 600, seed = 123))

cdisc_dataset("ASL", ASL)

}
