% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdisc_data.R
\name{cdisc_dataset}
\alias{cdisc_dataset}
\title{Data input for teal app}
\usage{
cdisc_dataset(dataname, data, keys = get_cdisc_keys(dataname))
}
\arguments{
\item{dataname}{name of dataset}

\item{data}{data}

\item{keys}{list of keys -

Please note that the order of keys is important.

Keys are not matched by name but by order. In case you want to perform:

\code{SELECT ...}
\code{  FROM ADSL}
\code{LEFT JOIN ADTTE on ADTTE.STUDYID = ADSL.STUDYID and}
\code{                                     ADTTE.SUBJECTUNIQUEID = ADSL.USUBJID}

the keys need to be

ADSL \code{primary: USUBJID, STUDYID}

ADTTE \code{primary: PARAMCD, foreign: SUBJECTUNIQUEID, STUDYID, parent: ADSL}

As you can see the names of foreign ADTTE are different to ADSL keys, but the order and length is equal.}
}
\value{
a dataset with connected metadata
}
\description{
Function that creates CDISC dataset object
}
\details{
List of implemented cdisc datasets:

\itemize{
  \item ADSL
  \item ADTTE
  \item ADAE
  \item ADLB
  \item ADCM
  \item ADRS
  \item ADVS
}
}
\examples{
library(random.cdisc.data)

ADSL <-  suppressWarnings(radsl(N = 600, seed = 123))

cdisc_dataset("ADSL", ADSL)
}
