% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdisc_data.R
\name{cdisc_dataset}
\alias{cdisc_dataset}
\title{Data input for teal app}
\usage{
cdisc_dataset(
  dataname,
  data,
  keys = get_cdisc_keys(dataname),
  label = data_label(data),
  code = character(0),
  vars = list()
)
}
\arguments{
\item{dataname}{(\code{character} value)\cr
name of dataset.}

\item{data}{(\code{data.frame})\cr
data must contain fields defined in keys.}

\item{keys}{(\code{keys})\cr
see \code{\link{keys}}}

\item{label}{(\code{character} value)\cr
a dataset label}

\item{code}{(\code{character} value)\cr
code to reproduce \code{data}}

\item{vars}{(named \code{list})) \cr
  In case when this object code depends on the \code{raw_data} from the other
  \code{RelationalDataset} object(s) or other constant value,
  this/these object(s) should be included as named element of the list.

Please note that the order of keys is important.

Keys are not matched by name but by order. In case you want to perform:

\code{SELECT ...}
\code{  FROM ADSL}
\code{LEFT JOIN ADTTE on ADTTE.STUDYID = ADSL.STUDYID and}
\code{                                     ADTTE.SUBJECTUNIQUEID = ADSL.USUBJID}

the keys need to be

ADSL \code{primary: USUBJID, STUDYID}

ADTTE \code{primary: PARAMCD, foreign: SUBJECT, UNIQUEID, STUDYID, parent: ADSL}

As you can see the names of foreign ADTTE are different to ADSL keys, but the order and length is equal.}
}
\value{
a dataset with connected metadata
}
\description{
Function that creates CDISC dataset object
}
\examples{
library(random.cdisc.data)

ADSL <- radsl(cached = TRUE)

cdisc_dataset("ADSL", ADSL)
}
