% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdisc_data.R
\name{cdisc_dataset_file}
\alias{cdisc_dataset_file}
\title{Load \code{CDISC} \code{RelationalDataset} object from a file}
\usage{
cdisc_dataset_file(path, code = get_code(path))
}
\arguments{
\item{path}{(\code{character}) string giving the pathname of the file to read from.}

\item{code}{(\code{character}) reproducible code to re-create object}
}
\value{
\code{RelationalDataset} object
}
\description{
\ifelse{html}{\out{<a href='https://www.tidyverse.org/lifecycle/#maturing'><img src='figures/lifecycle-maturing.svg' alt='Maturing lifecycle'></a>}}{\strong{Maturing}}
Please note that the script has to end with a call creating desired object. The error will be raised otherwise.
}
\examples{
# simple example
file_example <- tempfile(fileext = ".R")
writeLines(
  text = c(
    "library(teal)
     library(random.cdisc.data)

     cdisc_dataset(dataname = \"ADSL\",
                   data = radsl(cached = TRUE),
                   code = \"library(random.cdisc.data)\nADSL <- radsl(cached = TRUE)\")"
  ),
  con = file_example
)
x <- cdisc_dataset_file(file_example, code = character(0))
get_code(x)

# custom code
file_example <- tempfile(fileext = ".R")
writeLines(
  text = c(
    "library(teal)

     # code>
     library(random.cdisc.data)
     adsl <- radsl(cached = TRUE)
     adsl$a1 <- 1
     adsl$a2 <- 2

     # <code
     cdisc_dataset(dataname = \"ADSL\", data = adsl)"
  ),
  con = file_example
)
x <- cdisc_dataset_file(file_example)
get_code(x)
}
