% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CDISCDataset.R
\name{cdisc_dataset_file}
\alias{cdisc_dataset_file}
\title{Load \code{CDISCDataset} object from a file}
\usage{
cdisc_dataset_file(path, code = get_code(path))
}
\arguments{
\item{path}{(\code{character}) string giving the pathname of the file to read from.}

\item{code}{(\code{character}) reproducible code to re-create object}
}
\value{
(\code{CDISCDataset}) object
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
Please note that the script has to end with a call creating desired object. The error will be raised otherwise.
}
\examples{
# simple example
file_example <- tempfile(fileext = ".R")
writeLines(
  text = c(
    "library(teal)
     library(random.cdisc.data)

     cdisc_dataset(dataname = \"ADSL\",
                   x = radsl(cached = TRUE),
                   code = \"library(random.cdisc.data)\nADSL <- radsl(cached = TRUE)\")"
  ),
  con = file_example
)
x <- cdisc_dataset_file(file_example, code = character(0))
get_code(x)
}
