% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CDISCDataset.R
\name{cdisc_dataset_file}
\alias{cdisc_dataset_file}
\title{Load \code{CDISCDataset} object from a file}
\usage{
cdisc_dataset_file(path, code = get_code(path))
}
\arguments{
\item{path}{(\code{character}) string giving the pathname of the file to read from.}

\item{code}{(\code{character}) reproducible code to re-create object}
}
\value{
(\code{CDISCDataset}) object
}
\description{
\ifelse{html}{\out{<a href='https://www.tidyverse.org/lifecycle/#experimental'><img src='figures/lifecycle-experimental.svg' alt='Experimental lifecycle'></a>}}{\strong{Experimental}}
Please note that the script has to end with a call creating desired object. The error will be raised otherwise.
}
\examples{
# simple example
file_example <- tempfile(fileext = ".R")
writeLines(
  text = c(
    "library(teal)
     library(scda)

     cdisc_dataset(dataname = \"ADSL\",
                   x = synthetic_cdisc_data('latest')$adsl,
                   code = \"library(scda)\nADSL <- synthetic_cdisc_data('latest')$adsl\")"
  ),
  con = file_example
)
x <- cdisc_dataset_file(file_example, code = character(0))
get_code(x)
}
