% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TealDataConnection.R
\name{cdse_connection}
\alias{cdse_connection}
\title{Superseded (moved into separate package): Open connection to \code{CDSE}}
\usage{
cdse_connection(env = "prod")
}
\arguments{
\item{env}{optional, \code{CDSE} environment name.}
}
\value{
(\code{TealDataConnection}) type of object
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#superseded}{\figure{lifecycle-superseded.svg}{options: alt='[Superseded]'}}}{\strong{[Superseded]}}
}
