% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TealDataConnector_constructors.R
\name{cdse_data}
\alias{cdse_data}
\title{Superseded (moved into separate package): \code{TealDataConnector} connector for \code{CDSE}}
\usage{
cdse_data(..., connection = cdse_connection())
}
\arguments{
\item{...}{(\code{TealDatasetConnector} objects)\cr
dataset connectors created using \link{cdse_dataset_connector}}

\item{connection}{(\code{TealDataConnection}) object returned from \code{cdse_connection}.}
}
\value{
An object of class \code{TealDataConnector}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#superseded}{\figure{lifecycle-superseded.svg}{options: alt='[Superseded]'}}}{\strong{[Superseded]}}
Build data connector for \code{CDSE} datasets
}
