% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RelationalDataConnector_constructors.R
\name{cdse_data}
\alias{cdse_data}
\title{\code{RelationalDataConnector} connector for \code{CDSE}}
\usage{
cdse_data(..., connection = cdse_connection())
}
\arguments{
\item{...}{(\code{DatasetConnector} objects)\cr
dataset connectors created using \code{\link{cdse_dataset_connector}}}

\item{connection}{(\code{DataConnection}) object returned from \code{cdse_connection}.}
}
\value{
An object of class \code{RelatonalDataConnector}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
Build data connector for \code{CDSE} datasets
}
\examples{

\dontrun{
x <- cdse_data(
  cdse_cdisc_dataset_connector("ADSL", "cid1234567890"),
  cdse_cdisc_dataset_connector("ADLB", "cid0987654321")
)
app <- init(
  data = cdisc_data(x),
  modules = root_modules(
    module(
      "ADSL AGE histogram",
      server = function(input, output, session, datasets) {
        output$hist <- renderPlot({
          hist(datasets$get_data("ADSL", filtered = TRUE)$AGE)
        })
      },
      ui = function(id, ...) {ns <- NS(id); plotOutput(ns('hist'))},
      filters = "ADSL"
    )
  ),
  header = tags$h1("Sample App")
)
shinyApp(app$ui, app$server)
}
}
