% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decorators-helpers.R
\name{check_decorators}
\alias{check_decorators}
\alias{assert_decorators}
\title{Check decorators list}
\usage{
check_decorators(x, names = NULL)

assert_decorators(x, names = NULL, .var.name = checkmate::vname(x), add = NULL)
}
\arguments{
\item{x}{(named \code{list}) of \code{\link[=teal_transform_module]{teal_transform_module()}} objects, or nested lists thereof.}

\item{names}{(\code{character}) optional vector of valid output names. When provided, all names
in \code{x} must be one of these names, and names must be unique.}

\item{.var.name}{[\code{character(1)}]\cr
The custom name for \code{x} as passed to any \code{assert*} function.
Defaults to a heuristic name lookup.}

\item{add}{If an \code{AssertCollection} is provided, the error message is stored in it.
If \code{NULL}, an exception is raised if res is not \code{TRUE}.}
}
\value{
\code{TRUE} if valid, otherwise a \code{character(1)} string describing the problem.
}
\description{
Check if a decorators list is valid and matches the expected output names.
}
\examples{
decorator <- teal_transform_module(server = function(id, data) data)
check_decorators(decorator)
check_decorators(list(all = decorator))
check_decorators(list(all = decorator, output = decorator))
check_decorators(list(all = decorator, output = list(decorator, decorator)))
}
\seealso{
\code{\link[=module_transform_data]{module_transform_data()}}
}
