% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choices_labeled.R
\name{choices_labeled}
\alias{choices_labeled}
\title{Set "\verb{<choice>:<label>}" type of Names}
\usage{
choices_labeled(choices, labels, subset = NULL, types = NULL)
}
\arguments{
\item{choices}{a character / numeric / logical vector}

\item{labels}{character vector containing labels to be applied to \code{choices}. If \code{NA} then
"Label Missing" will be used.}

\item{subset}{a vector that is a subset of \code{choices}. This is useful if
only a few variables need to be named. If this argument is used, the returned vector will
match its order.}

\item{types}{vector containing the types of the columns to be used for applying the appropriate
icons to the \code{\link[teal]{choices_selected}} drop down box}
}
\value{
a named character vector
}
\description{
\ifelse{html}{\out{<a href='https://www.tidyverse.org/lifecycle/#maturing'><img src='figures/lifecycle-maturing.svg' alt='Maturing lifecycle'></a>}}{\strong{Maturing}}
This is often useful for \code{\link[teal]{choices_selected}} as it marks up the dropdown boxes
for \code{\link[shiny]{selectInput}}.
}
\details{
If either \code{choices} or \code{labels} are factors, they are coerced to character.
Duplicated elements from \code{choices} get removed.
}
\examples{
library(scda)

ADSL <- synthetic_cdisc_data("latest")$adsl
ADTTE <- synthetic_cdisc_data("latest")$adtte
choices1 <- choices_labeled(names(ADSL), rtables::var_labels(ADSL))
choices2 <- choices_labeled(ADTTE$PARAMCD, ADTTE$PARAM)
# if only a subset of variables are needed, use subset argument
choices3 <- choices_labeled(names(ADSL), rtables::var_labels(ADSL), subset = c("ARMCD", "ARM"))
\dontrun{
shinyApp(
  ui = fluidPage(
    selectInput("c1",
      label = "Choices from ADSL",
      choices = choices1,
      selected = choices1[1]
    ),
    selectInput("c2",
      label = "Choices from ADTTE",
      choices = choices2,
      selected = choices2[1]
    ),
    selectInput("c3",
      label = "Arm choices from ADSL",
      choices = choices3,
      selected = choices3[1]
    )
  ),
  server = function(input, output) {}
)
}
}
