% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selected_choices.R
\name{choices_selected}
\alias{choices_selected}
\title{Choices function}
\usage{
choices_selected(choices, selected = choices[1])
}
\arguments{
\item{choices}{vector of possible choices}

\item{selected}{vector of pre-selected options, first element of \code{choices} if blank}
}
\description{
Choices function
}
\examples{

choices_selected(
   choices = setNames(LETTERS[1:5], paste("Letter", LETTERS[1:5])),
   selected = "X"
)

}
