% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choices_selected.R
\name{choices_selected}
\alias{choices_selected}
\title{Choices selected}
\usage{
choices_selected(choices, selected, multiple, show = FALSE,
  label = NULL)
}
\arguments{
\item{choices}{(\code{character}) Named character vector to define the choices
of a shiny select input.}

\item{selected}{(\code{character}) Named character vector to define the selected
values of a shiny select input.}

\item{multiple}{(\code{logical}) Whether multiple values shall be allowed in the
shiny select input.}

\item{show}{(\code{logical}) (optional) \link{DataExtractSpec} specific feature to
hide the choices selected in case they are not needed.}

\item{label}{(\code{logical}) (optional) \link{DataExtractSpec} specific feature to
show a different label on top of this specific \link{shiny}{selectInput}.}
}
\value{
A \code{list} of all input values. The function double checks the \code{choices}
  and \code{selected} inputs.
}
\description{
Choices selected
}
