% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DatasetConnector_constructors.R
\name{code_cdisc_dataset_connector}
\alias{code_cdisc_dataset_connector}
\title{Code \code{CDISCDatasetConnector}}
\usage{
code_cdisc_dataset_connector(
  dataname,
  code,
  keys = get_cdisc_keys(dataname),
  parent = if (identical(dataname, "ADSL")) character(0L) else "ADSL",
  label = character(0),
  mutate_code = character(0),
  ...
)
}
\arguments{
\item{dataname}{(\code{character})\cr
A given name for the dataset it may not contain spaces}

\item{code}{(\code{character})\cr
String containing the code to produce the object.
The code must end in a call to the object.}

\item{keys}{optional, (\code{character})\cr
vector of dataset primary keys column names}

\item{parent}{optional, (\code{character}) \cr
parent dataset name}

\item{label}{(\code{character})\cr
Label to describe the dataset.}

\item{mutate_code}{(\code{character})\cr
String containing the code used to mutate the object
after it is produced.}

\item{...}{(\code{optional})\cr
Additional arguments applied to pull function.
In case when this object code depends on the \code{raw_data} from the other
\code{Dataset}, \code{DatasetConnector} object(s) or other constant value,
this/these object(s) should be included. Please note that \code{vars}
are included to this object as local \code{vars} and they cannot be modified
within another dataset.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\details{
Create a \code{CDISCDatasetConnector} from a string of code with keys
assigned automatically by \code{dataname}.
}
