% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TealDatasetConnector_constructors.R
\name{code_dataset_connector}
\alias{code_dataset_connector}
\alias{code_cdisc_dataset_connector}
\title{Code \code{TealDatasetConnector}}
\usage{
code_dataset_connector(
  dataname,
  code,
  keys = character(0),
  label = character(0),
  mutate_code = character(0),
  mutate_script = character(0),
  metadata = NULL,
  ...
)

code_cdisc_dataset_connector(
  dataname,
  code,
  keys = get_cdisc_keys(dataname),
  parent = if (identical(dataname, "ADSL")) character(0L) else "ADSL",
  label = character(0),
  mutate_code = character(0),
  metadata = NULL,
  ...
)
}
\arguments{
\item{dataname}{(\code{character})\cr
A given name for the dataset it may not contain spaces}

\item{code}{(\code{character})\cr
String containing the code to produce the object.
The code must end in a call to the object.}

\item{keys}{optional, (\code{character})\cr
vector of dataset primary keys column names}

\item{label}{(\code{character})\cr
Label to describe the dataset.}

\item{mutate_code}{(\code{character})\cr
String containing the code used to mutate the object
after it is produced.}

\item{mutate_script}{(\code{character})\cr
Alternatively to \code{mutate_code} - location of the file containing modification code.
Can't be used simultaneously with \code{mutate_script}.}

\item{metadata}{(named \code{list}, \code{NULL} or \code{CallableFunction}) \cr
Field containing either the metadata about the dataset (each element of the list
should be atomic and length one) or a \code{CallableFuntion} to pull the metadata
from a connection. This should return a \code{list} or an object which can be
converted to a list with \code{as.list}.}

\item{...}{Additional arguments applied to pull function.
In case when this object code depends on the \code{raw_data} from the other
\code{TealDataset}, \code{TealDatasetConnector} object(s) or other constant value,
this/these object(s) should be included. Please note that \code{vars}
are included to this object as local \code{vars} and they cannot be modified
within another dataset.}

\item{parent}{optional, (\code{character}) \cr
parent dataset name}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
}
\details{
Create a \code{TealDatasetConnector} from a string of code.

Create a \code{CDISCTealDatasetConnector} from a string of code with keys
assigned automatically by \code{dataname}.
}
\examples{
library(scda)
x <- code_dataset_connector(
  dataname = "ADSL",
  keys = get_cdisc_keys("ADSL"),
  code = "ADSL <- synthetic_cdisc_data(\"latest\")$adsl; ADSL"
)

x$get_code()

mutate_dataset(x, code = "ADSL$new_variable <- 1")
x$get_code()

file_example <- tempfile(fileext = ".R")
writeLines(
  text = c(
    "seed <- 1; ADSL <- radsl(cached = TRUE, seed = seed)\nADSL"
  ),
  con = file_example
)

y <- code_dataset_connector(
  dataname = "ADSL",
  keys = get_cdisc_keys("ADSL"),
  code = paste0(readLines(file_example), collapse = "\n")
)
}
