% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_code.R
\name{code_exclude}
\alias{code_exclude}
\title{Exclude from code}
\usage{
code_exclude(lines, exclude_comments)
}
\arguments{
\item{lines}{(\code{character}) of code as seperate element in vector}

\item{exclude_comments}{(\code{logical}) whether exclude commented-out lines of code}
}
\description{
Excludes lines from code. It is possible to exclude one line ended by \code{# nocode}
}
