% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_code.R
\name{code_exclude}
\alias{code_exclude}
\title{Exclude from code}
\usage{
code_exclude(lines, exclude_comments, file_path)
}
\arguments{
\item{lines}{(\code{character}) of preprocessing code.}

\item{exclude_comments}{(\code{logical}) whether exclude commented-out lines of code. Lines to be excluded
should be ended with \verb{# nocode}. For multiple line exclusions one should enclose ignored block of code with
\verb{# nocode>} and \verb{# <nocode}}

\item{file_path}{(\code{character}) path or URL address of the file to be parsed}
}
\description{
Excludes lines from code. It is possible to exclude one line ended by \verb{# nocode}
}
\keyword{internal}
