% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{code_from_script}
\alias{code_from_script}
\title{Get code from script}
\usage{
code_from_script(code, script, dataname = NULL)
}
\arguments{
\item{code}{(\code{character})\cr
an R code to be evaluated or a \code{PythonCodeClass} created using \link{python_code}.}

\item{script}{(\code{character})\cr
Alternatively to \code{code} - location of the file containing modification code.
Can't be used simultaneously with \code{script}.}

\item{dataname}{(\code{character})\cr
A given name for the dataset it may not contain spaces}
}
\value{
code (\code{character})
}
\description{
Get code from script. Switches between \code{code} and \code{script} arguments
to return non-empty one to pass it further to constructors.
}
\keyword{internal}
