% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_app_id}
\alias{create_app_id}
\title{Application ID}
\usage{
create_app_id(data, modules)
}
\arguments{
\item{data}{(\code{teal_data} or \code{teal_data_module}) as accepted by \code{init}}

\item{modules}{(\code{teal_modules}) object as accepted by \code{init}}
}
\value{
A single character string.
}
\description{
Creates App ID used to match filter snapshots to application.
}
\details{
Calculate app ID that will be used to stamp filter state snapshots.
App ID is a hash of the app's data and modules.
See "transferring snapshots" section in ?snapshot.
}
\keyword{internal}
