% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilterStates.R
\name{data_choices_labeled}
\alias{data_choices_labeled}
\title{Returns a \code{choices_labeled} object}
\usage{
data_choices_labeled(
  data,
  choices,
  varlabels = character(0),
  keys = character(0)
)
}
\arguments{
\item{data}{(\code{data.frame}, \code{DFrame}, \code{list})\cr
where labels can be taken from in case when \code{varlabels} is not specified.
\code{data} must be specified if \code{varlabels} is not specified.}

\item{choices}{(\code{character})\cr
the array of chosen variables}

\item{varlabels}{(\code{character})\cr
the labels of variables in data}

\item{keys}{(\code{character})\cr
the names of the key columns in data}
}
\value{
\code{character(0)} if choices are empty; a \code{choices_labeled} object otherwise
}
\description{
Returns a \code{choices_labeled} object
}
\keyword{internal}
