% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_extract.R
\name{data_extract_spec}
\alias{data_extract_spec}
\title{Constructor for \link{DataExtractSpec}}
\usage{
data_extract_spec(...)
}
\arguments{
\item{...}{params of \link{DataExtractSpec}}
}
\description{
Constructor for \link{DataExtractSpec}
}
