% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{data_for_teal}
\alias{data_for_teal}
\title{Constructor for teal data set}
\usage{
data_for_teal(df, keys, source = NULL)
}
\arguments{
\item{df}{(\code{data.frame}) data to be handled inside}

\item{keys}{(\code{character}) vector telling which columns of the \code{ds} are
key variables.}

\item{source}{(\code{character}) string defining how the data was constructed. Important
for reproducibility.}
}
\value{
(\code{data.frame}) with additional attributes keys and source.
}
\description{
Constructor for teal data set
}
