% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_label.R
\name{data_label}
\alias{data_label}
\title{Get dataset label attribute}
\usage{
data_label(data)
}
\arguments{
\item{data}{\code{data.frame} from which attribute is extracted}
}
\value{
(\code{character}) label or \code{NULL} if it's missing
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
}
\examples{
library(scda)
data_label(synthetic_cdisc_data("latest")$adsl)
}
