% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{data_table_item}
\alias{data_table_item}
\title{Deprecated: Data Table Viewer Teal Module}
\usage{
data_table_item(...)
}
\arguments{
\item{...}{arguments passed on to \code{\link{tm_data_table}}}
}
\description{
Please use the \code{\link{tm_data_table}} function instead.
}
