% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_extract_spec.R
\docType{class}
\name{DataExtractSpec}
\alias{DataExtractSpec}
\title{Data Extract input for teal modules}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
DataExtractSpec
}
\description{
The Data Extract input can be used to filter and select columns from a data
set to be used in teal. Please use the constructor function \link{data_extract_spec}
to set it up.
}
\details{
From this function's output a \code{\link{data_extract_input}} can
be constructed. This input can be read by a \link{data_extract_module} module.
}
\section{Fields}{

\describe{
\item{\code{dataname}}{(\code{character}) Data set to be extracted and selected}

\item{\code{keys_filtering}}{(\code{filter_choices_spec}-S3-class) Setup of the dataset filtering}

\item{\code{columns}}{(\code{\link{choices_selected}}) Columns to be selected from the input dataset}
}}

\section{Initialize}{

\describe{
  \item{via new}{ \code{DataExtractSpec$new(dataname, keys_filtering = NULL, columns = NULL)} }
  \item{via constructor}{
   \code{\link{data_extract_spec}(dataname, keys_filtering = NULL, columns = NULL)}
  }
}
}

\keyword{data}
