% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdisc_data.R
\name{dataset}
\alias{dataset}
\title{Data input for teal app}
\usage{
dataset(dataname, data, keys = list(primary = NULL, foreign = NULL,
  parent = NULL), labels = list(dataset_label = NULL, column_labels =
  NULL))
}
\arguments{
\item{dataname}{name of dataset}

\item{data}{data}

\item{keys}{list of keys}

\item{labels}{list of labels}
}
\value{
a dataset with connected metadata
}
\description{
Abstract function that creates dataset object with connected metadata.
}
\examples{
library(random.cdisc.data)

ADSL <-  suppressWarnings(radsl(N = 600, seed = 123))

dataset("ADSL", ADSL)
}
