% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TealDataset.R
\name{dataset}
\alias{dataset}
\alias{dataset.data.frame}
\title{Constructor for \code{\link{TealDataset}} class}
\usage{
dataset(
  dataname,
  x,
  keys = character(0),
  label = data_label(x),
  code = character(0),
  vars = list()
)

\method{dataset}{data.frame}(
  dataname,
  x,
  keys = character(0),
  label = data_label(x),
  code = character(0),
  vars = list()
)
}
\arguments{
\item{dataname}{(\code{character})\cr
A given name for the dataset it may not contain spaces}

\item{x}{(\code{data.frame} or \code{MultiAssayExperiment})\cr}

\item{keys}{optional, (\code{character})\cr
vector with primary keys}

\item{label}{(\code{character})\cr
Label to describe the dataset}

\item{code}{(\code{character})\cr
A character string defining the code needed to produce the data set in \code{x}}

\item{vars}{(named \code{list})) \cr
In case when this object code depends on other \code{TealDataset} object(s) or
other constant value, this/these object(s) should be included as named
element(s) of the list. For example if this object code needs \code{ADSL}
object we should specify \verb{vars = list(ADSL = <adsl object>)}.
It's recommended to include \code{TealDataset} or \code{TealDatasetConnector} objects to
the \code{vars} list to preserve reproducibility. Please note that \code{vars}
are included to this object as local \code{vars} and they cannot be modified
within another dataset.}
}
\value{
\code{\link{TealDataset}} object
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
}
\examples{
# Simple example
dataset("iris", iris)

# Example with more arguments
library(scda)
ADSL <- synthetic_cdisc_data("latest")$adsl
ADSL_dataset <- dataset(dataname = "ADSL", x = ADSL)

ADSL_dataset$get_dataname()

ADSL_dataset <- dataset(
  dataname = "ADSL",
  x = ADSL,
  label = "AdAM subject-level dataset",
  code = "ADSL <- synthetic_cdisc_data(\"latest\")$adsl"
)

ADSL_dataset$get_dataset_label()
ADSL_dataset$get_code()
}
