% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdisc_data.R
\name{dataset}
\alias{dataset}
\title{Data input for teal app}
\usage{
dataset(
  dataname,
  data,
  keys = teal::keys(primary = NULL, foreign = NULL, parent = NULL)
)
}
\arguments{
\item{dataname}{name of dataset}

\item{data}{data}

\item{keys}{(\code{keys}) object -

Please note that the order of keys is important.

Keys are not matched by name but by order. In case you want to perform:

\code{SELECT ...}
\code{  FROM ADSL}
\code{LEFT JOIN ADTTE on ADTTE.STUDYID = ADSL.STUDYID and}
\code{                                     ADTTE.SUBJECTUNIQUEID = ADSL.USUBJID}

the keys need to be

ADSL \code{primary: USUBJID, STUDYID}

ADTTE \code{primary: PARAMCD, foreign: SUBJECT, UNIQUEID, STUDYID, parent: ADSL}

As you can see the names of foreign ADTTE are different to ADSL keys, but the order and length is equal.}
}
\value{
a dataset with connected metadata
}
\description{
Abstract function that creates dataset object with connected metadata.
}
\examples{
dataset("iris", iris)
}
