% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAETealDataset.R, R/TealDataset.R
\name{dataset.MultiAssayExperiment}
\alias{dataset.MultiAssayExperiment}
\alias{dataset}
\alias{dataset.data.frame}
\title{S3 method to construct an \code{MAETealDataset} object from \code{MultiAssayExperiment}}
\usage{
\method{dataset}{MultiAssayExperiment}(
  dataname,
  x,
  keys = character(0),
  label = data_label(x),
  code = character(0),
  vars = list(),
  metadata = NULL
)

dataset(
  dataname,
  x,
  keys = character(0),
  label = data_label(x),
  code = character(0),
  vars = list(),
  metadata = NULL
)

\method{dataset}{data.frame}(
  dataname,
  x,
  keys = character(0),
  label = data_label(x),
  code = character(0),
  vars = list(),
  metadata = NULL
)
}
\arguments{
\item{dataname}{(\code{character})\cr
A given name for the dataset it may not contain spaces}

\item{x}{(\code{data.frame} or \code{MultiAssayExperiment})\cr}

\item{keys}{optional, (\code{character})\cr
vector with primary keys}

\item{label}{(\code{character})\cr
Label to describe the dataset}

\item{code}{(\code{character})\cr
A character string defining the code needed to produce the data set in \code{x}}

\item{vars}{(named \code{list})) \cr
In case when this object code depends on other \code{TealDataset} object(s) or
other constant value, this/these object(s) should be included as named
element(s) of the list. For example if this object code needs \code{ADSL}
object we should specify \verb{vars = list(ADSL = <adsl object>)}.
It's recommended to include \code{TealDataset} or \code{TealDatasetConnector} objects to
the \code{vars} list to preserve reproducibility. Please note that \code{vars}
are included to this object as local \code{vars} and they cannot be modified
within another dataset.}

\item{metadata}{(named \code{list} or \code{NULL}) \cr
Field containing metadata about the dataset. Each element of the list
should be atomic and length one.}
}
\value{
\code{\link{TealDataset}} object
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
}
\examples{
# Simple example
mae_d <- dataset(
  "MAE",
  MultiAssayExperiment::miniACC,
  keys = c("STUDYID", "USUBJID"),
  metadata = list(type = "example")
)
mae_d$get_dataname()
mae_d$get_dataset_label()
mae_d$get_metadata()
mae_d$get_code()
mae_d$get_raw_data()
# Simple example
dataset("iris", iris)

# Example with more arguments
library(scda)
ADSL <- synthetic_cdisc_data("latest")$adsl
ADSL_dataset <- dataset(dataname = "ADSL", x = ADSL)

ADSL_dataset$get_dataname()

ADSL_dataset <- dataset(
  dataname = "ADSL",
  x = ADSL,
  label = "AdAM subject-level dataset",
  code = "ADSL <- synthetic_cdisc_data(\"latest\")$adsl",
  metadata = list(type = "synthetic data")
)
ADSL_dataset$get_metadata()
ADSL_dataset$get_dataset_label()
ADSL_dataset$get_code()
}
