% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dataset.R
\name{Dataset}
\alias{Dataset}
\title{R6 Class representing a dataset with its attributes}
\description{
Any \code{data.frame} object can be stored inside this object.
Some attributes like colnames, dimension or column names for a specific type will
be automatically derived.
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{ncol}}{Number of columns}

\item{\code{nrow}}{Number of rows}

\item{\code{dim}}{Dimension \code{c(x, y)}}

\item{\code{colnames}}{The column names of the data}

\item{\code{rownames}}{The rownames of the data}

\item{\code{raw_data}}{The data.frame behind this R6 class}

\item{\code{data}}{The data.frame behind this R6 class}

\item{\code{column_labels}}{for backwards compatibility
will be deprecated in future}

\item{\code{var_names}}{The column names of the data}

\item{\code{row_labels}}{Row labels (can have spaces)}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Dataset$new()}}
\item \href{#method-recreate}{\code{Dataset$recreate()}}
\item \href{#method-get_attrs}{\code{Dataset$get_attrs()}}
\item \href{#method-get_raw_data}{\code{Dataset$get_raw_data()}}
\item \href{#method-get_numeric_colnames}{\code{Dataset$get_numeric_colnames()}}
\item \href{#method-get_character_colnames}{\code{Dataset$get_character_colnames()}}
\item \href{#method-get_factor_colnames}{\code{Dataset$get_factor_colnames()}}
\item \href{#method-get_colnames}{\code{Dataset$get_colnames()}}
\item \href{#method-get_column_labels}{\code{Dataset$get_column_labels()}}
\item \href{#method-get_rownames}{\code{Dataset$get_rownames()}}
\item \href{#method-get_row_labels}{\code{Dataset$get_row_labels()}}
\item \href{#method-get_dataname}{\code{Dataset$get_dataname()}}
\item \href{#method-get_datanames}{\code{Dataset$get_datanames()}}
\item \href{#method-get_dataset_label}{\code{Dataset$get_dataset_label()}}
\item \href{#method-get_keys}{\code{Dataset$get_keys()}}
\item \href{#method-set_dataset_label}{\code{Dataset$set_dataset_label()}}
\item \href{#method-set_keys}{\code{Dataset$set_keys()}}
\item \href{#method-get_code}{\code{Dataset$get_code()}}
\item \href{#method-get_code_class}{\code{Dataset$get_code_class()}}
\item \href{#method-mutate}{\code{Dataset$mutate()}}
\item \href{#method-set_vars}{\code{Dataset$set_vars()}}
\item \href{#method-set_code}{\code{Dataset$set_code()}}
\item \href{#method-check}{\code{Dataset$check()}}
\item \href{#method-check_keys}{\code{Dataset$check_keys()}}
\item \href{#method-is_pulled}{\code{Dataset$is_pulled()}}
\item \href{#method-clone}{\code{Dataset$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new object of \code{Dataset} class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dataset$new(
  dataname,
  x,
  keys = character(0),
  code = character(0),
  label = character(0),
  vars = list()
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dataname}}{(\code{character})\cr
A given name for the dataset it may not contain spaces}

\item{\code{x}}{(\code{data.frame})\cr}

\item{\code{keys}}{optional, (\code{character})\cr
vector with primary keys}

\item{\code{code}}{(\code{character})\cr
A character string defining the code needed to produce the data set in \code{x}.
\code{initialize()} and \code{recreate()} accept code as \code{CodeClass} also
which is needed to preserve the code uniqueness and correct order.}

\item{\code{label}}{(\code{character})\cr
Label to describe the dataset}

\item{\code{vars}}{(named \code{list})) \cr
In case when this object code depends on the \code{raw_data} from the other
\code{Dataset} object(s) or other constant value,
this/these object(s) should be included as named element of the list.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-recreate"></a>}}
\if{latex}{\out{\hypertarget{method-recreate}{}}}
\subsection{Method \code{recreate()}}{
Recreate a dataset with its current attributes
This is useful way to have access to class initialize method basing on class object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dataset$recreate(
  dataname = self$get_dataname(),
  x = self$get_raw_data(),
  keys = self$get_keys(),
  code = self$get_code_class(),
  label = self$get_dataset_label(),
  vars = list()
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dataname}}{(\code{character})\cr
A given name for the dataset it may not contain spaces}

\item{\code{x}}{(\code{data.frame})\cr}

\item{\code{keys}}{optional, (\code{character})\cr
vector with primary keys}

\item{\code{code}}{(\code{character})\cr
A character string defining the code needed to produce the data set in \code{x}.
\code{initialize()} and \code{recreate()} accept code as \code{CodeClass} also
which is needed to preserve the code uniqueness and correct order.}

\item{\code{label}}{(\code{character})\cr
Label to describe the dataset}

\item{\code{vars}}{(named \code{list})) \cr
In case when this object code depends on the \code{raw_data} from the other
\code{Dataset} object(s) or other constant value,
this/these object(s) should be included as named element of the list.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a new object of \code{Dataset} class
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_attrs"></a>}}
\if{latex}{\out{\hypertarget{method-get_attrs}{}}}
\subsection{Method \code{get_attrs()}}{
Get all dataset attributes
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dataset$get_attrs()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
(named \code{list}) with dataset attributes
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_raw_data"></a>}}
\if{latex}{\out{\hypertarget{method-get_raw_data}{}}}
\subsection{Method \code{get_raw_data()}}{
Derive the raw data frame inside this object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dataset$get_raw_data()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{data.frame} or \code{rtable}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_numeric_colnames"></a>}}
\if{latex}{\out{\hypertarget{method-get_numeric_colnames}{}}}
\subsection{Method \code{get_numeric_colnames()}}{
Derive the names of all \code{numeric} columns
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dataset$get_numeric_colnames()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{character} vector.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_character_colnames"></a>}}
\if{latex}{\out{\hypertarget{method-get_character_colnames}{}}}
\subsection{Method \code{get_character_colnames()}}{
Derive the names of all \code{character} columns
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dataset$get_character_colnames()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{character} vector.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_factor_colnames"></a>}}
\if{latex}{\out{\hypertarget{method-get_factor_colnames}{}}}
\subsection{Method \code{get_factor_colnames()}}{
Derive the names of all \code{factor} columns
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dataset$get_factor_colnames()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{character} vector.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_colnames"></a>}}
\if{latex}{\out{\hypertarget{method-get_colnames}{}}}
\subsection{Method \code{get_colnames()}}{
Derive the column names
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dataset$get_colnames()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{character} vector.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_column_labels"></a>}}
\if{latex}{\out{\hypertarget{method-get_column_labels}{}}}
\subsection{Method \code{get_column_labels()}}{
Derive the column labels
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dataset$get_column_labels()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{character} vector.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_rownames"></a>}}
\if{latex}{\out{\hypertarget{method-get_rownames}{}}}
\subsection{Method \code{get_rownames()}}{
Derive the row names
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dataset$get_rownames()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{character} vector.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_row_labels"></a>}}
\if{latex}{\out{\hypertarget{method-get_row_labels}{}}}
\subsection{Method \code{get_row_labels()}}{
Derive the row labels
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dataset$get_row_labels()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{character} vector.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_dataname"></a>}}
\if{latex}{\out{\hypertarget{method-get_dataname}{}}}
\subsection{Method \code{get_dataname()}}{
Derive the \code{name} which was formerly called \code{dataname}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dataset$get_dataname()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{character} name of the dataset
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_datanames"></a>}}
\if{latex}{\out{\hypertarget{method-get_datanames}{}}}
\subsection{Method \code{get_datanames()}}{
Derive the dataname
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dataset$get_datanames()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{character} name of the dataset
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_dataset_label"></a>}}
\if{latex}{\out{\hypertarget{method-get_dataset_label}{}}}
\subsection{Method \code{get_dataset_label()}}{
Derive the \code{label} which was former called \code{datalabel}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dataset$get_dataset_label()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{character} label of the dataset
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_keys"></a>}}
\if{latex}{\out{\hypertarget{method-get_keys}{}}}
\subsection{Method \code{get_keys()}}{
Get primary keys of dataset
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dataset$get_keys()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
(\code{character} vector) with dataset primary keys
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_dataset_label"></a>}}
\if{latex}{\out{\hypertarget{method-set_dataset_label}{}}}
\subsection{Method \code{set_dataset_label()}}{
Set the label for the dataset
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dataset$set_dataset_label(label)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{label}}{(\code{character})\cr
Label to describe the dataset}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
(\code{self}) invisibly for chaining
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_keys"></a>}}
\if{latex}{\out{\hypertarget{method-set_keys}{}}}
\subsection{Method \code{set_keys()}}{
Set new keys
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dataset$set_keys(keys)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{keys}}{optional, (\code{character})\cr
vector with primary keys}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
(\code{self}) invisibly for chaining.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_code"></a>}}
\if{latex}{\out{\hypertarget{method-get_code}{}}}
\subsection{Method \code{get_code()}}{
Get code to get data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dataset$get_code(deparse = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deparse}}{(\code{logical}) whether return deparsed form of a call}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
optionally deparsed \code{call} object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_code_class"></a>}}
\if{latex}{\out{\hypertarget{method-get_code_class}{}}}
\subsection{Method \code{get_code_class()}}{
Get internal \code{CodeClass} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dataset$get_code_class()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\verb{\code{CodeClass}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-mutate"></a>}}
\if{latex}{\out{\hypertarget{method-mutate}{}}}
\subsection{Method \code{mutate()}}{
Mutate dataset by code

Either code or script must be provided, but not both.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dataset$mutate(code, vars = list())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{code}}{(\code{character})\cr
A character string defining the code needed to produce the data set in \code{x}.
\code{initialize()} and \code{recreate()} accept code as \code{CodeClass} also
which is needed to preserve the code uniqueness and correct order.}

\item{\code{vars}}{(named \code{list})) \cr
In case when this object code depends on the \code{raw_data} from the other
\code{Dataset} object(s) or other constant value,
this/these object(s) should be included as named element of the list.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
(\code{self}) invisibly for chaining
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_vars"></a>}}
\if{latex}{\out{\hypertarget{method-set_vars}{}}}
\subsection{Method \code{set_vars()}}{
Adds variables which code depends on
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dataset$set_vars(vars)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{vars}}{(named \code{list})) \cr
In case when this object code depends on the \code{raw_data} from the other
\code{Dataset} object(s) or other constant value,
this/these object(s) should be included as named element of the list.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
(\code{self}) invisibly for chaining
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_code"></a>}}
\if{latex}{\out{\hypertarget{method-set_code}{}}}
\subsection{Method \code{set_code()}}{
Sets reproducible code
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dataset$set_code(code)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{code}}{(\code{character})\cr
A character string defining the code needed to produce the data set in \code{x}.
\code{initialize()} and \code{recreate()} accept code as \code{CodeClass} also
which is needed to preserve the code uniqueness and correct order.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
(\code{self}) invisibly for chaining
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-check"></a>}}
\if{latex}{\out{\hypertarget{method-check}{}}}
\subsection{Method \code{check()}}{
Check to determine if the raw data is reproducible from the \code{get_code()} code.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dataset$check()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{TRUE} if the dataset generated from evaluating the
\code{get_code()} code is identical to the raw data, else \code{FALSE}.
Check if keys has been specified correctly for dataset. Set of \code{keys}
should distinguish unique rows or be \code{character(0)}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-check_keys"></a>}}
\if{latex}{\out{\hypertarget{method-check_keys}{}}}
\subsection{Method \code{check_keys()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dataset$check_keys(keys = private$.keys)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{keys}}{optional, (\code{character})\cr
vector with primary keys}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{TRUE} if dataset has been already pulled, else \code{FALSE}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-is_pulled"></a>}}
\if{latex}{\out{\hypertarget{method-is_pulled}{}}}
\subsection{Method \code{is_pulled()}}{
Check if dataset has already been pulled.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dataset$is_pulled()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{TRUE} if dataset has been already pulled, else \code{FALSE}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dataset$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
