% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TealDataset.R
\name{dataset_file}
\alias{dataset_file}
\title{Load \code{TealDataset} object from a file}
\usage{
dataset_file(path, code = get_code(path))
}
\arguments{
\item{path}{(\code{character}) string giving the pathname of the file to read from.}

\item{code}{(\code{character}) reproducible code to re-create object}
}
\value{
\code{TealDataset} object
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
Please note that the script has to end with a call creating desired object. The error will be raised otherwise.
}
\examples{
# simple example
file_example <- tempfile(fileext = ".R")
writeLines(
  text = c(
    "library(teal)
     dataset(dataname = \"iris\",
             x = iris,
             code = \"iris\")"
  ),
  con = file_example
)
x <- dataset_file(file_example, code = character(0))
get_code(x)

# custom code
file_example <- tempfile(fileext = ".R")
writeLines(
  text = c(
    "library(teal)

     # code>
     x <- iris
     x$a1 <- 1
     x$a2 <- 2

     # <code
     dataset(dataname = \"iris_mod\", x = x)"
  ),
  con = file_example
)
x <- dataset_file(file_example)
get_code(x)
}
