% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dataset.R
\name{dataset_file}
\alias{dataset_file}
\alias{named_dataset_file}
\alias{relational_dataset_file}
\title{Load \code{Dataset} object from a file}
\usage{
dataset_file(path, code = get_code(path))

named_dataset_file(path, code = get_code(path))

relational_dataset_file(path, code = get_code(path))
}
\arguments{
\item{path}{(\code{character}) string giving the pathname of the file to read from.}

\item{code}{(\code{character}) reproducible code to re-create object}
}
\value{
\code{Dataset} object
}
\description{
\ifelse{html}{\out{<a href='https://www.tidyverse.org/lifecycle/#experimental'><img src='figures/lifecycle-experimental.svg' alt='Experimental lifecycle'></a>}}{\strong{Experimental}}
Please note that the script has to end with a call creating desired object. The error will be raised otherwise.

\ifelse{html}{\figure{lifecycle-soft-deprecated.svg}{options: alt='Soft-deprecated lifecycle'}}{\strong{Soft-deprecated}}

\ifelse{html}{\figure{lifecycle-soft-deprecated.svg}{options: alt='Soft-deprecated lifecycle'}}{\strong{Soft-deprecated}}
}
\examples{
# simple example
file_example <- tempfile(fileext = ".R")
writeLines(
  text = c(
    "library(teal)
     dataset(dataname = \"iris\",
             x = iris,
             code = \"iris\")"
  ),
  con = file_example
)
x <- dataset_file(file_example, code = character(0))
get_code(x)

# custom code
file_example <- tempfile(fileext = ".R")
writeLines(
  text = c(
    "library(teal)

     # code>
     x <- iris
     x$a1 <- 1
     x$a2 <- 2

     # <code
     dataset(dataname = \"iris_mod\", x = x)"
  ),
  con = file_example
)
x <- dataset_file(file_example)
get_code(x)
}
