% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataConnection.R
\name{datasetdb_connection}
\alias{datasetdb_connection}
\title{Open connection to \code{DataSetDB}}
\usage{
datasetdb_connection()
}
\value{
(\code{DataConnection}) type of object
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
