% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TealDataConnection.R
\name{datasetdb_connection}
\alias{datasetdb_connection}
\title{Open connection to \code{DataSetDB}}
\usage{
datasetdb_connection()
}
\value{
(\code{TealDataConnection}) type of object
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#superseded}{\figure{lifecycle-superseded.svg}{options: alt='[Superseded]'}}}{\strong{[Superseded]}}
}
