% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TealDataConnector_constructors.R
\name{datasetdb_data}
\alias{datasetdb_data}
\title{\code{TealDataConnector} connector for \code{DataSetDB}}
\usage{
datasetdb_data(..., connection = datasetdb_connection())
}
\arguments{
\item{...}{(\code{TealDatasetConnector} objects)\cr
dataset connectors created using \code{\link{datasetdb_dataset_connector}}}

\item{connection}{(\code{TealDataConnection}) object returned from \code{datasetdb_connection}.}
}
\value{
An object of class \code{TealDataConnector}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#defunct}{\figure{lifecycle-defunct.svg}{options: alt='[Defunct]'}}}{\strong{[Defunct]}}
Build data connector for \code{DataSetDB} datasets
}
