% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RelationalDataConnector_constructors.R
\name{datasetdb_data}
\alias{datasetdb_data}
\title{\code{RelationalDataConnector} connector for \code{DataSetDB}}
\usage{
datasetdb_data(..., connection = datasetdb_connection())
}
\arguments{
\item{...}{(\code{DatasetConnector} objects)\cr
dataset connectors created using \code{\link{datasetdb_dataset_connector}}}

\item{connection}{(\code{DataConnection}) object returned from \code{datasetdb_connection}.}
}
\value{
An object of class \code{RelatonalDataConnector}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
Build data connector for \code{DataSetDB} datasets
}
\examples{

if (all(c("dsassembly", "gdbauth") \%in\% installed.packages())) {
  x <- datasetdb_data(
    datasetdb_dataset_connector("MAE", id = "DS000000267")
  )
  app <- init(
    data = teal_data(x),
    modules = root_modules(
      module(
        "dummy module",
        server = function(input, output, session, datasets) {
          output$str <- renderText({
            paste0(capture.output(str(
              datasets$get_data("MAE", filtered = TRUE)
            )), collapse = "\n")
          })
        },
        ui = function(id, ...) {ns <- NS(id); verbatimTextOutput(ns('str'))},
        filters = "MAE"
      )
    ),
    header = tags$h1("Sample App")
  )
}
\dontrun{
shinyApp(app$ui, app$server)
}

}
