% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TealDatasetConnector_constructors.R
\name{datasetdb_dataset_connector}
\alias{datasetdb_dataset_connector}
\title{\code{DataSetDB} \code{TealDatasetConnector}}
\usage{
datasetdb_dataset_connector(
  dataname,
  id,
  keys = character(0),
  label = character(0),
  code = character(0),
  script = character(0),
  ...
)
}
\arguments{
\item{dataname}{(\code{character})\cr
A given name for the dataset it may not contain spaces}

\item{id}{(\code{character})\cr
identifier of the dataset in \code{DataSetDB}}

\item{keys}{optional, (\code{character})\cr
vector of dataset primary keys column names}

\item{label}{(\code{character})\cr
Label to describe the dataset.}

\item{code}{(\code{character})\cr
A character string defining code to modify \code{raw_data} from this dataset. To modify
current dataset code should contain at least one assignment to object defined in \code{dataname}
argument. For example if \code{dataname = ADSL} example code should contain
\verb{ADSL <- <some R code>}. Can't be used simultaneously with \code{script}}

\item{script}{(\code{character})\cr
Alternatively to \code{code} - location of the file containing modification code.
Can't be used simultaneously with \code{script}.}

\item{...}{(\code{optional})\cr
additional arguments applied to pull function}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#defunct}{\figure{lifecycle-defunct.svg}{options: alt='[Defunct]'}}}{\strong{[Defunct]}}
}
\details{
Create a \code{TealDatasetConnector} from \code{DataSetDB}.
}
