% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-ddl-utils.R
\name{ddl}
\alias{ddl}
\title{Convenience wrapper for ddl}
\usage{
ddl(code, env_mask, ui, server)
}
\arguments{
\item{code}{(\code{character} or \code{language})}

\item{env_mask}{(\verb{named list})}

\item{ui}{(\verb{function(id)}) function to create UI}

\item{server}{(\verb{function(id)}) \code{shiny} server which returns \code{teal_data} object wrapped in}
}
\description{
Convenience wrapper for ddl
}
