% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddl.R
\name{ddl_run}
\alias{ddl_run}
\title{Run code and mask inputs}
\usage{
ddl_run(input = list(), code, input_mask, datanames, join_keys)
}
\arguments{
\item{input}{(\code{list}) containing inputs to be used in the \code{code}}

\item{code}{(\code{language}) code to be executed}

\item{input_mask}{(\code{list}) containing inputs to be masked in the \code{code}}

\item{datanames}{(\code{character}) names of the objects to be created from the code evaluation}

\item{join_keys}{(\code{join_keys}) object}
}
\value{
\code{teal_data} object
}
\description{
Function runs the \code{code}, masks the \code{code} and creates \code{teal_data} object.
}
