% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkmate.R
\name{decorate_err_msg}
\alias{decorate_err_msg}
\title{Capture error and decorate error message.}
\usage{
decorate_err_msg(x, pre = character(0), post = character(0))
}
\arguments{
\item{x}{object to evaluate}

\item{pre}{(\code{character(1)}) A string to prepend to error message}

\item{post}{(\code{character(1)}) A string to append to error message}
}
\value{
\code{x} if no error, otherwise throws error with decorated message
}
\description{
Capture error and decorate error message.
}
\keyword{internal}
