% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_filter.R
\name{default_filter}
\alias{default_filter}
\title{Refers the default filter state}
\usage{
default_filter()
}
\value{
\code{default_filter} a default filter object
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
You can use it to refer to the variable's default filter state,
which will be set when \code{FilteredData$set_data} is called.
}
\details{
This is a simple wrapper around an S3 class.
}
\examples{
list() # test printing
}
