% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilteredData.R
\name{default_filter}
\alias{default_filter}
\title{Refer to the default filter state}
\usage{
default_filter()
}
\description{
You can use it to refer to the variable's default filter state,
which will be set when \code{FilteredData::set_data} is called.
It can be used together with \code{teal::init}.
}
\details{
This is a simple wrapper around an S3 class.
}
\examples{
default_filter() # test printing
}
