% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_teal_reporter.R
\name{disable_report}
\alias{disable_report}
\title{Disable the report for a \code{teal_module}}
\usage{
disable_report(x)
}
\arguments{
\item{x}{(\code{teal_module}) a \code{teal_module} object.}
}
\value{
modified data object that indicates that it should disable the reporter functionality.
}
\description{
Convenience function that disables the user's ability to add the module
to the report previewer.
}
\examples{
app <- init(
  data = within(teal_data(), iris <- iris),
  modules = modules(
    example_module(label = "example teal module") |> disable_report()
  )
)
if (interactive()) {
  shinyApp(app$ui, app$server)
}
}
