% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_source_code.R
\name{disable_src}
\alias{disable_src}
\title{Disable the "Show R Code" global button in the UI}
\usage{
disable_src(x)
}
\arguments{
\item{x}{(\code{teal_module}) a \code{teal_module} object.}
}
\value{
modified data object that indicates that it should not show the "Show R Code"
button in the UI.
}
\description{
Convenience function that disables the user's ability to see the code of the module.
}
\details{
This is equivalent to setting the attribute \code{teal.enable_src} to \code{FALSE} on the
data object returned by the module.
}
\examples{
app <- init(
  data = within(teal_data(), iris <- iris),
  modules = modules(
    example_module(label = "example teal module") |> disable_src()
  )
)
if (interactive()) {
  shinyApp(app$ui, app$server)
}
}
