% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.0_module_teal_transform.R
\name{.fallback_on_failure}
\alias{.fallback_on_failure}
\title{Fallback on failure}
\usage{
.fallback_on_failure(data_current, data_previous, label)
}
\arguments{
\item{data_current}{(\code{reactive}) Current data}

\item{data_previous}{(\code{reactive}) Previous data}
}
\value{
\code{reactive} \code{teal_data}
}
\description{
Function returns the previous data if the current data is invalid. In \code{teal} we try to prevent the error
from being thrown and instead we replace failing transform module data output with data input from the
previous module (or from previous \code{teal} reactive tree elements).
}
