% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_teal_data.R
\name{.fallback_on_failure}
\alias{.fallback_on_failure}
\title{Fallback on failure}
\usage{
.fallback_on_failure(this, that, label)
}
\arguments{
\item{this}{(\code{reactive}) Current reactive.}

\item{that}{(\code{reactive}) Previous reactive.}

\item{label}{(\code{character}) Label for identifying problematic \code{teal_data_module} transform in logging.}
}
\value{
\code{reactive} \code{teal_data}
}
\description{
Function returns the previous reactive if the current reactive is invalid (throws error or returns NULL).
Application: In \code{teal} we try to prevent the error from being thrown and instead we replace failing
transform module data output with data input from the previous module (or from previous \code{teal} reactive
tree elements).
}
\keyword{internal}
