% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.0_module_data.R
\name{.get_hashes_code}
\alias{.get_hashes_code}
\title{Get code that tests the integrity of the reproducible data}
\usage{
.get_hashes_code(data, datanames = teal_data_datanames(data))
}
\arguments{
\item{datanames}{(\code{character}) names of datasets}

\item{datasets}{(\code{FilteredData}) object holding the data}
}
\value{
A character vector with the code lines.
}
\description{
Get code that tests the integrity of the reproducible data
}
\keyword{internal}
