% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.0_module_teal_modules.R
\name{.get_hashes_code}
\alias{.get_hashes_code}
\title{Get code that tests the integrity of the reproducible data}
\usage{
.get_hashes_code(datasets, datanames)
}
\arguments{
\item{datasets}{(\code{FilteredData}) object holding the data}

\item{datanames}{(\code{character}) names of datasets}
}
\value{
A character vector with the code lines.
}
\description{
Get code that tests the integrity of the reproducible data
}
\keyword{internal}
