% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_init_data.R
\name{.get_hashes_code}
\alias{.get_hashes_code}
\title{Get code that tests the integrity of the reproducible data}
\usage{
.get_hashes_code(data, datanames = names(data))
}
\arguments{
\item{data}{(\code{teal_data}) object holding the data}

\item{datanames}{(\code{character}) names of \code{datasets}}
}
\value{
A character vector with the code lines.
}
\description{
Get code that tests the integrity of the reproducible data
}
\keyword{internal}
