% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporter_previewer_module.R
\name{.get_reporter_options}
\alias{.get_reporter_options}
\title{Temporary function to handle server_args of the report_previewer_module before its hard
deprecation.}
\usage{
.get_reporter_options(args)
}
\arguments{
\item{args}{(\code{list})}
}
\description{
Temporary function to handle server_args of the report_previewer_module before its hard
deprecation.
}
\keyword{internal}
