% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{.log}
\alias{.log}
\title{A logging function}
\usage{
.log(..., sep = " ", type = "debug")
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#superseded}{\figure{lifecycle-superseded.svg}{options: alt='[Superseded]'}}}{\strong{[Superseded]}}
}
\details{
Superseded by \link[teal.logger:register_logger]{teal.logger::register_logger}.
}
\keyword{internal}
