% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-data-utils.R
\name{.substitute_code}
\alias{.substitute_code}
\title{Substitute symbols in the code}
\usage{
.substitute_code(code, args)
}
\arguments{
\item{code}{(\code{language}) code to substitute}

\item{args}{(\code{list}) named list or arguments}
}
\description{
Function replaces symbols in the provided code by values of the \code{args} argument.
}
\keyword{internal}
