% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_code.R
\name{enclosed_with}
\alias{enclosed_with}
\title{Get lines enclosed in}
\usage{
enclosed_with(lines, starts_at, stops_at)
}
\arguments{
\item{lines}{(\code{character}) of code as seperate element in vector}

\item{starts_at}{(\code{character}) regular expression determining start of the block of code}

\item{stops_at}{(\code{character}) regular expression determining stop  of the block of code}
}
\description{
Extracts lines from code which are enclosed within regexp starts_at and stops_at
}
