% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_code.R
\name{enclosed_with}
\alias{enclosed_with}
\title{Get code enclosed within}
\usage{
enclosed_with(lines)
}
\arguments{
\item{lines}{(\code{character}) of preprocessing code.}
}
\value{
(\code{character}) subset of lines which start and end with preprocessing
start and stop tags.
}
\description{
Extracts lines from code which are enclosed within regexp starts_at and stops_at
}
\keyword{internal}
