% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-data-utils.R
\name{eval_and_mask}
\alias{eval_and_mask}
\title{Function runs the \code{code}, masks the \code{code} and creates \code{teal_data} object.}
\usage{
eval_and_mask(data, code, input = list(), input_mask = list())
}
\arguments{
\item{code}{(optional \code{character} or \code{language})\cr
Object containing (defused) syntactically valid R expression to be executed in the shiny session.
Shouldn't be specified when \code{expr} is specified.}

\item{input}{(\code{list}) containing inputs to be used in the \code{code}}

\item{input_mask}{(optional \verb{named list})\cr
arguments to be substituted in the \code{code}. These (named) list elements are going to replace
symbols in the code prefixed with \verb{input$} or \verb{input[["}. Typically \code{input_mask} is used
to mask username or password with \code{list(password = quote(askpass::askpass()))}.
See section \verb{code masking} for more details.}
}
\value{
\code{teal_data} object
}
\description{
Function runs the \code{code}, masks the \code{code} and creates \code{teal_data} object.
}
