% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teal_data_module.R
\name{eval_code}
\alias{eval_code}
\alias{eval_code,teal_data_module,language-method}
\alias{eval_code,teal_data_module,expression-method}
\title{Evaluate the code in the \code{qenv} environment}
\usage{
\S4method{eval_code}{teal_data_module,character}(object, code)

\S4method{eval_code}{teal_data_module,language}(object, code)

\S4method{eval_code}{teal_data_module,expression}(object, code)
}
\arguments{
\item{object}{(\code{qenv})}

\item{code}{(\code{character} or \code{language}) code to evaluate. Also accepts and stores comments}
}
\value{
\code{qenv} object.
}
\description{
Given code is evaluated in the \code{qenv} environment and appended to the \code{code} slot. This means
that state of the environment is always a result of the stored code (if \code{qenv} was initialized)
with reproducible code.
}
\examples{
q1 <- new_qenv(env = list2env(list(a = 1)), code = quote(a <- 1))
q2 <- eval_code(q1, quote(library(checkmate)))
q3 <- eval_code(q2, quote(assert_number(a)))
}
