% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummy_functions.R
\name{example_module}
\alias{example_module}
\title{An example \code{teal} module}
\usage{
example_module(label = "example teal module", filters = "all")
}
\arguments{
\item{label}{(\code{character(1)}) Label shown in the navigation item for the module. Any label possible except
\code{"global_filters"} - read more in \code{mapping} argument of \link{teal_slices}.}

\item{filters}{(\code{character}) Deprecated. Use \code{datanames} instead.}
}
\value{
A \code{teal} module which can be included in the \code{modules} argument to \code{\link[=init]{init()}}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
app <- init(
  data = teal_data(
    dataset("IRIS", iris),
    dataset("MTCARS", mtcars)
  ),
  modules = example_module()
)
if (interactive()) {
  shinyApp(app$ui, app$server)
}
}
