% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummy_functions.R
\name{example_module}
\alias{example_module}
\title{An example \code{teal} module}
\usage{
example_module(label = "example teal module", datanames = "all")
}
\arguments{
\item{label}{(\code{character(1)}) Label shown in the navigation item for the module or module group.
For \code{modules()} defaults to \code{"root"}.
Note that for \code{modules()} \code{label} comes after \code{...}, so it must be passed as a named argument,
otherwise it will be captured by \code{...}.

\code{"global_filters"} is reserved - see \code{mapping} argument of \code{\link[=teal_slices]{teal_slices()}}.

\code{"Report previewer"} is reserved - see \code{\link[=reporter_previewer_module]{reporter_previewer_module()}}.}

\item{datanames}{(\code{character}) A vector with \code{datanames} that are relevant for the item. The
filter panel will automatically update the shown filters to include only
filters in the listed datasets. \code{NULL} will hide the filter panel,
and the keyword \code{'all'} will show filters of all datasets. \code{datanames} also determines
a subset of datasets which are appended to the \code{data} argument in server function.}
}
\value{
A \code{teal} module which can be included in the \code{modules} argument to \code{\link[=init]{init()}}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
app <- init(
  data = teal_data(IRIS = iris, MTCARS = mtcars),
  modules = example_module()
)
if (interactive()) {
  shinyApp(app$ui, app$server)
}
}
