% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummy_functions.R
\name{example_module}
\alias{example_module}
\title{An example \code{teal} module}
\usage{
example_module(
  label = "example teal module",
  datanames = "all",
  transformers = list()
)
}
\arguments{
\item{label}{(\code{character(1)}) Label shown in the navigation item for the module or module group.
For \code{modules()} defaults to \code{"root"}. See \code{Details}.}

\item{datanames}{(\code{character}) Names of the datasets that are relevant for the item. The
filter panel will only display filters for specified \code{datanames}. The keyword \code{"all"} will show
filters of all datasets. \code{NULL} will hide the filter panel. \code{datanames} also determines a subset
of datasets which are appended to the \code{data} argument in server function.}

\item{transformers}{(\code{list} of \code{teal_data_module}) that will be applied to transform the data.
Each transform module UI will appear in the \code{teal} application, unless the \code{custom_ui} attribute is set on the list.
If so, the module developer is responsible to display the UI in the module itself. \code{datanames} of the \code{transformers}
will be added to the \code{datanames}.

When the transformation does not have sufficient input data, the resulting data will fallback
to the last successful transform or, in case there are none, to the filtered data.}
}
\value{
A \code{teal} module which can be included in the \code{modules} argument to \code{\link[=init]{init()}}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
app <- init(
  data = teal_data(IRIS = iris, MTCARS = mtcars),
  modules = example_module()
)
if (interactive()) {
  shinyApp(app$ui, app$server)
}
}
