% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummy_functions.R
\name{example_modules}
\alias{example_modules}
\title{Get example modules.}
\usage{
example_modules(datanames = c("ADSL", "ADTTE"))
}
\arguments{
\item{datanames}{(\code{character})\cr
names of the datasets to be used in the example modules. Possible choices are \code{ADSL}, \code{ADTTE}.}
}
\value{
\code{teal_modules}
}
\description{
Creates an example hierarchy of \code{teal_modules} from which a \code{teal} app can be created.
}
\keyword{internal}
