% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{extract_landing}
\alias{extract_landing}
\title{Extract a \code{landing_module} from list of \code{modules}}
\usage{
extract_landing(modules)
}
\arguments{
\item{modules}{\code{teal_modules}}
}
\value{
\code{landing_module}
}
\description{
Extract a \code{landing_module} from list of \code{modules}
}
\keyword{internal}
