% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teal_data_module.R
\name{extract_transformers}
\alias{extract_transformers}
\title{Extract all \code{transformers} from \code{modules}.}
\usage{
extract_transformers(modules)
}
\arguments{
\item{modules}{\code{teal_modules} or \code{teal_module}}
}
\value{
A list of \code{transformers} nested the same way as input \code{modules}.
}
\description{
Extract all \code{transformers} from \code{modules}.
}
\keyword{internal}
