% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules_debugging.R
\name{filter_calls_module}
\alias{filter_calls_module}
\title{Dummy module to show the filter calls generated by the right encoding panel}
\usage{
filter_calls_module(label = "Filter Calls Module")
}
\arguments{
\item{label}{\code{character} label of module}
}
\description{
Please do not remove, this is useful for debugging teal without
dependencies and simplifies \verb{\link[devtools]\{load_all\}} which otherwise fails
and avoids session restarts!
}
\examples{
app <- init(
  data = list(iris = iris, mtcars = mtcars),
  modules = teal:::filter_calls_module(),
  header = "Simple teal app"
)
if (interactive()) {
  shinyApp(app$ui, app$server)
}
}
\keyword{internal}
