% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_filter_manager.R
\name{filter_manager_module_srv}
\alias{filter_manager_module_srv}
\title{Module specific filter manager}
\usage{
filter_manager_module_srv(id, module_fd, slices_global)
}
\arguments{
\item{id}{(\code{character(1)})\cr
\code{shiny} module id.}

\item{module_fd}{(\code{FilteredData})\cr
object to filter data in the teal-module}

\item{slices_global}{(\code{reactiveVal} or \code{teal_slices})\cr
stores a list of all available filters which can be utilized in several ways, for example:
\itemize{
\item to disable/enable specific filter in the module
\item to restore filter saved settings
\item to save current filter settings panel
}}
}
\value{
shiny module returning NULL
}
\description{
This module compares filters between single \code{FilteredData} settings and global \code{teal_slices}.
Updates appropriate objects \code{module_fd}, \code{slices_global} to keep them consistent.
}
\keyword{internal}
