% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.0_filter_panel.R
\name{filter_panel}
\alias{filter_panel}
\alias{ui_filter_panel}
\alias{srv_filter_panel}
\alias{.make_teal_data}
\title{Filter panel module in teal}
\usage{
ui_filter_panel(id)

srv_filter_panel(id, datasets, active_datanames, data_rv, is_active)

.make_teal_data(modules, data, datasets = NULL, datanames)
}
\arguments{
\item{active_datanames}{(\code{reactive} returning \code{character}) this module's datanames}
}
\description{
Reactive filter panel module in teal
}
\keyword{internal}
