% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilteredData.R
\docType{data}
\name{FilteredData}
\alias{FilteredData}
\title{Class to encapsulate filtered data sets}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
FilteredData
}
\description{
Class to encapsulate filtered data sets
}
\details{
This class encapsulates data import and filtering.
  Once a dataname was specified it won't be deleted at runtime to keep
  reactive relations intact. If a dataset is not needed anymore then you can
  set it to NULL with load_data.
  Every data set is also filtered with ADSL, hence ADSL is given by default (NULL)

load_data
1 load data
2 specify filer_info
3 apply filters
}
\examples{
\dontrun{
path <- "/opt/BIOSTAT/qa/cdt7876a/libraries/asl.sas7bdat"

x <- FilteredData$new()

x$load_data(path)

x$datanames()

x$list_data_info("ASL")
x$get_filter_info("ASL")

df <- x$get_data("ASL")

x$get_filter_info("ASL")[['USUBJID']]$type

x$set_filter("ASL", list(AGE=c(3,5), SEX=c('M', 'F')))
}
}
\keyword{datasets}
