% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_code.R
\name{find_source_code}
\alias{find_source_code}
\title{Finds lines of code with source call}
\usage{
find_source_code(lines)
}
\arguments{
\item{lines}{(\code{character}) of preprocessing code.}
}
\description{
Finds lines in preprocessing code where \code{source()} call is located
}
\keyword{internal}
