% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TealDatasetConnector_constructors.R
\name{fun_dataset_connector}
\alias{fun_dataset_connector}
\alias{fun_cdisc_dataset_connector}
\title{Function Dataset Connector}
\usage{
fun_dataset_connector(
  dataname,
  fun,
  fun_args = NULL,
  keys = character(0),
  label = character(0),
  code = character(0),
  script = character(0),
  func_name = substitute(fun),
  metadata = NULL,
  ...
)

fun_cdisc_dataset_connector(
  dataname,
  fun,
  fun_args = NULL,
  keys = get_cdisc_keys(dataname),
  parent = if (identical(dataname, "ADSL")) character(0L) else "ADSL",
  label = character(0),
  code = character(0),
  script = character(0),
  func_name = substitute(fun),
  metadata = NULL,
  ...
)
}
\arguments{
\item{dataname}{(\code{character})\cr
A given name for the dataset it may not contain spaces}

\item{fun}{(\code{function})\cr
a custom function to obtain dataset.}

\item{fun_args}{(\code{list})\cr
additional arguments for (\code{func}).}

\item{keys}{optional, (\code{character})\cr
vector of dataset primary keys column names}

\item{label}{(\code{character})\cr
Label to describe the dataset.}

\item{code}{(\code{character})\cr
A character string defining code to modify \code{raw_data} from this dataset. To modify
current dataset code should contain at least one assignment to object defined in \code{dataname}
argument. For example if \code{dataname = ADSL} example code should contain
\verb{ADSL <- <some R code>}. Can't be used simultaneously with \code{script}}

\item{script}{(\code{character})\cr
Alternatively to \code{code} - location of the file containing modification code.
Can't be used simultaneously with \code{script}.}

\item{func_name}{(\code{name})\cr
for internal purposes, please keep it default}

\item{metadata}{(named \code{list}, \code{NULL} or \code{CallableFunction}) \cr
Field containing either the metadata about the dataset (each element of the list
should be atomic and length one) or a \code{CallableFuntion} to pull the metadata
from a connection. This should return a \code{list} or an object which can be
converted to a list with \code{as.list}.}

\item{...}{Additional arguments applied to pull function.
In case when this object code depends on the \code{raw_data} from the other
\code{TealDataset}, \code{TealDatasetConnector} object(s) or other constant value,
this/these object(s) should be included. Please note that \code{vars}
are included to this object as local \code{vars} and they cannot be modified
within another dataset.}

\item{parent}{optional, (\code{character}) \cr
parent dataset name}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
}
\details{
Create a \code{TealDatasetConnector} from \code{function} and its arguments.

Create a \code{CDISCTealDatasetConnector} from \code{function} and its arguments
with keys and parent name assigned automatically by \code{dataname}.
}
\examples{
my_data <- function(...) {
  data.frame(
    ID = paste0("ABC_", seq_len(10)),
    var1 = rnorm(n = 10),
    var2 = rnorm(n = 10),
    var3 = rnorm(n = 10)
  )
}
y <- fun_dataset_connector(
  dataname = "XYZ",
  fun = my_data
)

y$get_code()

y$pull()

get_raw_data(y)
}
