% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_label.R
\name{get_cdisc_keys}
\alias{get_cdisc_keys}
\title{Function that returns the default keys for a \code{CDISC} dataset by name}
\usage{
get_cdisc_keys(dataname)
}
\arguments{
\item{dataname}{name of the \code{CDISC} dataset}
}
\value{
\code{keys} object
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#maturing}{\figure{lifecycle-maturing.svg}{options: alt='[Maturing]'}}}{\strong{[Maturing]}}
}
\examples{
get_cdisc_keys("ADSL")
}
