% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdisc_data.R
\name{get_cdisc_keys}
\alias{get_cdisc_keys}
\title{Function that returns the default keys for a \code{CDISC} dataset by name}
\usage{
get_cdisc_keys(dataname)
}
\arguments{
\item{dataname}{name of the \code{CDISC} dataset}
}
\value{
\code{keys} object
}
\description{
\ifelse{html}{\out{<a href='https://www.tidyverse.org/lifecycle/#maturing'><img src='figures/lifecycle-maturing.svg' alt='Maturing lifecycle'></a>}}{\strong{Maturing}}
}
\examples{
get_cdisc_keys("ADSL")
}
