% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdisc_data.R
\name{get_cdisc_keys}
\alias{get_cdisc_keys}
\title{Function that creates keys object}
\usage{
get_cdisc_keys(dataname)
}
\arguments{
\item{dataname}{name of dataset}
}
\value{
keys
}
\description{
Function that creates keys object
}
\examples{

get_cdisc_keys("ADSL")
}
