% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_code.R
\name{get_code}
\alias{get_code}
\title{Get code}
\usage{
get_code(files_path, exclude_comments = TRUE, read_sources = TRUE)
}
\arguments{
\item{files_path}{(\code{character}) (optional) vector of files path to be read for preprocessing. Code from
multiple files is joined together.}

\item{exclude_comments}{(\code{logical}) whether exclude commented-out lines of code. Lines to be excluded
should be ended with \code{# nocode}. For multiple line exclusions one should enclose ignored block of code with
\code{# nocode>} and \code{# <nocode}}

\item{read_sources}{(\code{logical}) whether to replace \code{source("path")} with code lines from sourced file.
If \code{read_sources = TRUE} changing working directory inside preprocessing is not allowed.}
}
\value{
(\code{character}) code of import and preparation of data for teal application.
}
\description{
Reads code from specified files, including code from source. Method reads code without
\code{library()} or \code{require()} calls. Function created for teal app, but can be used with any file.
}
