% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_code.R
\name{get_code}
\alias{get_code}
\title{Get code}
\usage{
get_code(files_path, starts_at = "# @start_code",
  stops_at = "# @end_code", exclude_comments = TRUE,
  read_sources = TRUE)
}
\arguments{
\item{files_path}{(\code{character}) (optional) vector of files path, if \code{NULL} then current file is used}

\item{starts_at}{(\code{character}) regular expression determining start of the block of code}

\item{stops_at}{(\code{character}) regular expression determining stop  of the block of code}

\item{exclude_comments}{(\code{logical}) whether exclude commented-out lines of code}

\item{read_sources}{(\code{logical}) whether to replace \code{source("path")} with code lines from sourced file.}
}
\value{
(\code{character}) code of import and preparation of data for teal application.
}
\description{
Reads code from specified files, including code from source. Method reads code without
\code{library()} or \code{require()} calls. Function created for teal app, but can be used with any file.
}
