% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_code.R
\name{get_code_single}
\alias{get_code_single}
\title{Get code}
\usage{
get_code_single(file_path, read_sources)
}
\arguments{
\item{file_path}{(\code{character}) path of the file to be parsed}

\item{read_sources}{(\code{logical}) whether to replace \code{source("path")} with code lines from sourced file.
If \code{read_sources = TRUE} changing working directory inside preprocessing is not allowed.}
}
\value{
lines (\code{character}) of preprocessing code
}
\description{
Get code from specified file.
}
