% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_code.R
\name{get_code_single}
\alias{get_code_single}
\title{Get code}
\usage{
get_code_single(file_path, starts_at, stops_at, exclude_comments,
  read_sources)
}
\arguments{
\item{file_path}{(\code{character}) path of the file to be parsed}

\item{starts_at}{(\code{character}) regexp determining start of preprocessing code. Function returns code from
line where \code{starts_at} is detected. If no regexp found, function reads from first line.}

\item{stops_at}{(\code{character}) regexp determining stop  of preprocessing code. Function returns code until line
where \code{stops_at} is detected. If no regexp found, function reads to the end of the file. \code{stops_at} is
important to be specified, because otherwise it can include \code{teal::init()} which mustn't be included.}

\item{exclude_comments}{(\code{logical}) whether exclude commented-out lines of code. Lines to be excluded
should be ended with \code{# nocode}. For multiple line exclusions one should enclose ignored block of code with
\code{# nocode>} and \code{# <nocode}}

\item{read_sources}{(\code{logical}) whether to replace \code{source("path")} with code lines from sourced file.
If \code{read_sources = TRUE} changing working directory inside preprocessing is not allowed.}
}
\value{
code (\code{character}) preprocessing code
}
\description{
Get code from specified file.
}
