% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dataname.R
\name{get_dataname}
\alias{get_dataname}
\alias{get_dataname.RelationalDataConnector}
\alias{get_dataname.NamedDatasetConnector}
\alias{get_dataname.NamedDataset}
\title{S3 method for getting a dataname(s) of
(\code{RelationalDataConnector}, (\code{RelationalDatasetConnector} or
\code{NamedDataset}) R6 object}
\usage{
get_dataname(x)

\method{get_dataname}{RelationalDataConnector}(x)

\method{get_dataname}{NamedDatasetConnector}(x)

\method{get_dataname}{NamedDataset}(x)
}
\arguments{
\item{x}{(\code{RelationalDataConnector}, (\code{RelationalDatasetConnector} or
\code{NamedDataset}) R6 object}
}
\value{
dataname (\code{character}) A given name for the dataset(s)
  it may not contain spaces
}
\description{
S3 method for getting a dataname(s) of
(\code{RelationalDataConnector}, (\code{RelationalDatasetConnector} or
\code{NamedDataset}) R6 object
}
\examples{
library(random.cdisc.data)
fun <- callable_function(data.frame)
fun$set_args(list(n = 5, seed = 1, cached = TRUE))

x <- relational_dataset_connector(
 pull_fun = fun,
 dataname = "ADSL",
 keys = get_cdisc_keys("ADSL")
)
get_dataname(x)
library(random.cdisc.data)
ADSL <- radsl(cached = TRUE)
ADSL_dataset <- named_dataset(dataname = "ADSL", x = ADSL)
get_dataname(ADSL_dataset)
}
