% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dataname.R
\name{get_dataname}
\alias{get_dataname}
\alias{get_dataname.RelationalData}
\alias{get_dataname.RelationalDataConnector}
\alias{get_dataname.NamedDatasetConnector}
\alias{get_dataname.NamedDataset}
\title{S3 method for getting a dataname(s) of
(\code{RelationalDataConnector}, (\code{RelationalDatasetConnector} or
\code{NamedDataset}) R6 object}
\usage{
get_dataname(x)

\method{get_dataname}{RelationalData}(x)

\method{get_dataname}{RelationalDataConnector}(x)

\method{get_dataname}{NamedDatasetConnector}(x)

\method{get_dataname}{NamedDataset}(x)
}
\arguments{
\item{x}{(\code{RelationalDataConnector}, (\code{RelationalDatasetConnector} or
\code{NamedDataset}) R6 object}
}
\value{
dataname (\code{character}) A given name for the dataset(s)
  it may not contain spaces
}
\description{
S3 method for getting a dataname(s) of
(\code{RelationalDataConnector}, (\code{RelationalDatasetConnector} or
\code{NamedDataset}) R6 object
}
