% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddl.R
\name{get_dataname.ddl}
\alias{get_dataname.ddl}
\title{Get datanames from \code{ddl}}
\usage{
\method{get_dataname}{ddl}(x)
}
\arguments{
\item{x}{(\code{ddl}) object}
}
\description{
Get datanames from \code{ddl}
}
