% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddl-class.R
\name{get_dataname.ddl}
\alias{get_dataname.ddl}
\title{Get data names from \code{ddl}}
\usage{
\method{get_dataname}{ddl}(x)
}
\arguments{
\item{x}{(\code{ddl}) object}
}
\description{
Get data names from \code{ddl}
}
