% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dataname.R
\name{get_dataname}
\alias{get_dataname}
\alias{get_dataname.TealDataAbstract}
\alias{get_dataname.TealDatasetConnector}
\alias{get_dataname.TealDataset}
\title{S3 method for getting a dataname(s) of
(\code{TealDataAbstract}, (\code{TealDatasetConnector} or
\code{TealDataset}) R6 object}
\usage{
get_dataname(x)

\method{get_dataname}{TealDataAbstract}(x)

\method{get_dataname}{TealDatasetConnector}(x)

\method{get_dataname}{TealDataset}(x)
}
\arguments{
\item{x}{(\code{TealDataAbstract}, \code{TealDatasetConnector} or
\code{TealDataset}) object}
}
\value{
dataname (\code{character}) A given name for the dataset(s)
it may not contain spaces
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
}
