% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dataset.R
\name{get_dataset}
\alias{get_dataset}
\alias{get_dataset.RawDatasetConnector}
\alias{get_dataset.RelationalData}
\alias{get_dataset.cdisc_data}
\title{Get dataset from \code{DatasetConnector}}
\usage{
get_dataset(x, dataname)

\method{get_dataset}{RawDatasetConnector}(x, dataname = NULL)

\method{get_dataset}{RelationalData}(x, dataname = NULL)

\method{get_dataset}{cdisc_data}(x, dataname = NULL)
}
\arguments{
\item{x}{\code{RawDatasetConnector} or \code{RelationalDatasetConnector})}

\item{dataname}{\code{character} a name of dataset to be retrieved}
}
\value{
(\code{RawDataset} or \code{RelationalDataset})
}
\description{
Get dataset from \code{DatasetConnector}
}
\examples{

# RawDatasetConnector --------
library(random.cdisc.data)
dc <- rcd_cdisc_dataset_connector(dataname = "ADSL", fun = radsl)

load_dataset(dc)
get_dataset(dc)

# RelationalData --------
library(random.cdisc.data)
adsl <- cdisc_dataset(dataname = "ADSL", # RelationalDataset
                      data = radsl(cached = TRUE),
                      code = "library(random.cdisc.data)\nADSL <- radsl(cached = TRUE)")

adae <- cdisc_dataset(dataname = "ADAE", # RelationalDataset
                       data = radae(cached = TRUE),
                       code = "library(random.cdisc.data)\nADTTE <- radae(cached = TRUE)")

rd <- teal:::RelationalData$new(adsl, adae)
get_dataset(rd, "ADSL")

# RelationalDataConnector --------
rdc <- rcd_data(
  rcd_cdisc_dataset_connector("ADSL", radsl, cached = TRUE),
  rcd_cdisc_dataset_connector("ADTTE", radtte, cached = TRUE)
)

get_dataset(rd, dataname = "ADSL")
\dontrun{
load_datasets(rdc)
get_datasets(rdc)
get_dataset(rdc, dataname = "ADSL")
}

# RelationalDataList --------
drc <- teal_data(rdc, adae)

get_dataset(drc, "ADSL")

# cdisc_data --------
library(random.cdisc.data)

ADSL <- radsl(cached = TRUE)
ADTTE <- radtte(cached = TRUE)

# Example with keys
cd <- cdisc_data(
  cdisc_dataset("ADSL", ADSL, keys = keys(
    primary = c("STUDYID", "USUBJID"),
    foreign = NULL,
    parent = NULL
  )),
  cdisc_dataset("ADTTE", ADTTE, keys = keys(
    primary = c("STUDYID", "USUBJID", "PARAMCD"),
    foreign = c("STUDYID", "USUBJID"),
    parent = "ADSL"),
    code =  "ADTTE <- radtte(ADSL, seed = 123)"
  ),
  code = "ADSL <- radsl(N = 600, seed = 123)",
  check = FALSE
)

get_dataset(cd, "ADSL")
}
