% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dataset.R
\name{get_dataset}
\alias{get_dataset}
\alias{get_dataset.DatasetConnector}
\alias{get_dataset.Dataset}
\alias{get_dataset.DataAbstract}
\title{Get dataset from \code{DatasetConnector}}
\usage{
get_dataset(x, dataname)

\method{get_dataset}{DatasetConnector}(x, dataname = NULL)

\method{get_dataset}{Dataset}(x, dataname = NULL)

\method{get_dataset}{DataAbstract}(x, dataname = NULL)
}
\arguments{
\item{x}{(\code{DatasetConnector} or \code{DatasetConnector} or \code{DataAbstract})}

\item{dataname}{\code{character} a name of dataset to be retrieved}
}
\value{
(\code{Dataset})
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Get dataset from \code{DatasetConnector}
}
\examples{

# DatasetConnector --------
library(random.cdisc.data)
ADSL <- radsl(cached = TRUE)

dc <- rcd_dataset_connector(dataname = "ADAE", fun = radae,
                            ADSL = ADSL, max_n_aes = 2L)

load_dataset(dc)
get_dataset(dc)

# Dataset --------
library(random.cdisc.data)
ADSL <- radsl(cached = TRUE)
x <- dataset("ADSL", ADSL)

get_dataset(x)

# RelationalData  (not containing connectors) --------
library(random.cdisc.data)
adsl <- cdisc_dataset(dataname = "ADSL",
                      x = radsl(cached = TRUE),
                      code = "library(random.cdisc.data)\nADSL <- radsl(cached = TRUE)")

adae <- cdisc_dataset(dataname = "ADAE",
                       x = radae(cached = TRUE),
                       code = "library(random.cdisc.data)\nADTTE <- radae(cached = TRUE)")

rd <- teal:::RelationalData$new(adsl, adae)
get_dataset(rd, "ADSL")

# RelationalDataConnector --------
rdc <- rcd_data(
  rcd_cdisc_dataset_connector("ADSL", radsl, cached = TRUE),
  rcd_cdisc_dataset_connector("ADTTE", radtte, cached = TRUE)
)

\dontrun{
get_dataset(rdc, dataname = "ADSL")
load_datasets(rdc)
get_datasets(rdc)
get_dataset(rdc, dataname = "ADSL")
}

# RelationalData (containing connectors) --------
rd <- cdisc_data(rdc, adae)

\dontrun{
get_dataset(rd, "ADSL")
}
}
