% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dataset.R
\name{get_dataset}
\alias{get_dataset}
\alias{get_dataset.DatasetConnector}
\alias{get_dataset.Dataset}
\alias{get_dataset.DataAbstract}
\title{Get dataset from \code{DatasetConnector}}
\usage{
get_dataset(x, dataname)

\method{get_dataset}{DatasetConnector}(x, dataname = NULL)

\method{get_dataset}{Dataset}(x, dataname = NULL)

\method{get_dataset}{DataAbstract}(x, dataname = NULL)
}
\arguments{
\item{x}{(\code{DatasetConnector} or \code{DatasetConnector} or \code{DataAbstract})}

\item{dataname}{\code{character} a name of dataset to be retrieved}
}
\value{
(\code{Dataset})
}
\description{
\ifelse{html}{\out{<a href='https://www.tidyverse.org/lifecycle/#experimental'><img src='figures/lifecycle-experimental.svg' alt='Experimental lifecycle'></a>}}{\strong{Experimental}}

Get dataset from \code{DatasetConnector}
}
\examples{

# DatasetConnector --------
library(scda)
pull_fun_adae <- callable_function(
  function() {synthetic_cdisc_data("latest")$adae}
)

ADSL <- synthetic_cdisc_data("latest")$adsl

dc <- dataset_connector(dataname = "ADAE", pull_callable = pull_fun_adae,
                            keys = get_cdisc_keys("ADSL"))
load_dataset(dc)
get_dataset(dc)

# Dataset --------
library(scda)
ADSL <- synthetic_cdisc_data("latest")$adsl
x <- dataset("ADSL", ADSL)

get_dataset(x)

# RelationalData  (not containing connectors) --------
library(scda)
library(random.cdisc.data)
adsl <- cdisc_dataset(dataname = "ADSL",
                      x = synthetic_cdisc_data("latest")$adsl,
                      code = "library(scda)\nADSL <- synthetic_cdisc_data(\"latest\")$adsl")

adae <- cdisc_dataset(dataname = "ADAE",
                       x = synthetic_cdisc_data("latest")$adsl,
                       code = "library(scda)\nADTTE <- synthetic_cdisc_data(\"latest\")$adsl")

rd <- teal:::RelationalData$new(adsl, adae)
get_dataset(rd, "ADSL")

# RelationalDataConnector --------
rdc <- rcd_data(
  rcd_cdisc_dataset_connector("ADSL", radsl, cached = TRUE),
  rcd_cdisc_dataset_connector("ADTTE", radtte, cached = TRUE)
)

\dontrun{
get_dataset(rdc, dataname = "ADSL")
load_datasets(rdc)
get_datasets(rdc)
get_dataset(rdc, dataname = "ADSL")
}

# RelationalData (containing connectors) --------
rd <- cdisc_data(rdc, adae)

\dontrun{
get_dataset(rd, "ADSL")
}
}
