% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dataset.R
\name{get_dataset}
\alias{get_dataset}
\alias{get_dataset.TealDatasetConnector}
\alias{get_dataset.TealDataset}
\alias{get_dataset.TealDataAbstract}
\title{Get dataset from \code{TealDatasetConnector}}
\usage{
get_dataset(x, dataname)

\method{get_dataset}{TealDatasetConnector}(x, dataname = NULL)

\method{get_dataset}{TealDataset}(x, dataname = NULL)

\method{get_dataset}{TealDataAbstract}(x, dataname = NULL)
}
\arguments{
\item{x}{(\code{TealDatasetConnector} or \code{TealDatasetConnector} or \code{TealDataAbstract})}

\item{dataname}{(\code{character}) a name of dataset to be retrieved}
}
\value{
(\code{TealDataset})
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Get dataset from \code{TealDatasetConnector}
}
\details{
See \code{help(TealDataConnector)} and \code{help(TealData)} for more complex examples.
}
\examples{

# TealDatasetConnector --------
library(scda)
pull_fun_adae <-  teal.data::callable_function(
  function() {
    synthetic_cdisc_data("latest")$adae
  }
)

ADSL <- synthetic_cdisc_data("latest")$adsl

dc <- dataset_connector(
  dataname = "ADAE", pull_callable = pull_fun_adae,
  keys = teal.data::get_cdisc_keys("ADSL")
)
load_dataset(dc)
get_dataset(dc)

# TealDataset --------
library(scda)
ADSL <- synthetic_cdisc_data("latest")$adsl
x <- dataset("ADSL", ADSL)

get_dataset(x)

# TealData  (not containing connectors) --------
library(scda)
adsl <- cdisc_dataset(
  dataname = "ADSL",
  x = synthetic_cdisc_data("latest")$adsl,
  code = "library(scda)\nADSL <- synthetic_cdisc_data(\"latest\")$adsl"
)

adae <- cdisc_dataset(
  dataname = "ADAE",
  x = synthetic_cdisc_data("latest")$adsl,
  code = "library(scda)\nADTTE <- synthetic_cdisc_data(\"latest\")$adsl"
)

rd <- teal:::TealData$new(adsl, adae)
get_dataset(rd, "ADSL")
}
