% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dataset.R
\name{get_dataset}
\alias{get_dataset}
\alias{get_dataset.RawDatasetConnector}
\alias{get_dataset.RawDataset}
\alias{get_dataset.RelationalDataCollection}
\title{Get dataset from \code{DatasetConnector}}
\usage{
get_dataset(x, dataname)

\method{get_dataset}{RawDatasetConnector}(x, dataname = NULL)

\method{get_dataset}{RawDataset}(x, dataname = NULL)

\method{get_dataset}{RelationalDataCollection}(x, dataname = NULL)
}
\arguments{
\item{x}{\code{RawDatasetConnector} or \code{RelationalDatasetConnector})}

\item{dataname}{\code{character} a name of dataset to be retrieved}
}
\value{
(\code{RawDataset} or \code{RelationalDataset})
}
\description{
\ifelse{html}{\out{<a href='https://www.tidyverse.org/lifecycle/#experimental'><img src='figures/lifecycle-experimental.svg' alt='Experimental lifecycle'></a>}}{\strong{Experimental}}

Get dataset from \code{DatasetConnector}
}
\examples{

# RawDatasetConnector --------
library(random.cdisc.data)
ADSL <- radsl(cached = TRUE)

dc <- rcd_cdisc_dataset_connector(dataname = "ADAE", fun = radae,
                                  ADSL = ADSL, max_n_aes = 2L)

load_dataset(dc)
get_dataset(dc)

# RawDataset --------
library(random.cdisc.data)
ADSL <- radsl(cached = TRUE)
rd <- raw_dataset(ADSL)

get_dataset(rd)

# RelationalData  (not containing connectors) --------
library(random.cdisc.data)
adsl <- cdisc_dataset(dataname = "ADSL", # RelationalDataset
                      data = radsl(cached = TRUE),
                      code = "library(random.cdisc.data)\nADSL <- radsl(cached = TRUE)")

adae <- cdisc_dataset(dataname = "ADAE", # RelationalDataset
                       data = radae(cached = TRUE),
                       code = "library(random.cdisc.data)\nADTTE <- radae(cached = TRUE)")

rd <- teal:::RelationalData$new(adsl, adae)
get_dataset(rd, "ADSL")

# RelationalDataConnector --------
rdc <- rcd_data(
  rcd_cdisc_dataset_connector("ADSL", radsl, cached = TRUE),
  rcd_cdisc_dataset_connector("ADTTE", radtte, cached = TRUE)
)

\dontrun{
get_dataset(rdc, dataname = "ADSL")
load_datasets(rdc)
get_datasets(rdc)
get_dataset(rdc, dataname = "ADSL")
}

# RelationalData (containing connectors) --------
rd <- cdisc_data(rdc, adae)

\dontrun{
get_dataset(rd, "ADSL")
}
}
