% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dataset_label.R
\name{get_dataset_label}
\alias{get_dataset_label}
\alias{get_dataset_label.TealDatasetConnector}
\alias{get_dataset_label.TealDataset}
\title{S3 method for getting a label of
(\code{TealDatasetConnector} or \code{TealDataset}) R6 object}
\usage{
get_dataset_label(x)

\method{get_dataset_label}{TealDatasetConnector}(x)

\method{get_dataset_label}{TealDataset}(x)
}
\arguments{
\item{x}{(\code{TealDatasetConnector} or \code{TealDataset}) R6 object}
}
\value{
label (\code{character}) Label to describe the dataset
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
}
\examples{
fun <- callable_function(data.frame)
fun$set_args(list(c1 = seq_len(10)))

x <- dataset_connector(
  pull_callable = fun,
  dataname = "ADSL",
  label = "My custom label"
)
get_dataset_label(x)
library(scda)
ADSL <- synthetic_cdisc_data("latest")$adsl
ADSL_dataset <- dataset(dataname = "ADSL", x = ADSL, label = "My custom label")
get_dataset_label(ADSL_dataset)
}
