% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_datasets.R
\name{get_datasets}
\alias{get_datasets}
\alias{get_datasets.RelationalDataCollection}
\alias{get_datasets.RawDatasetConnector}
\alias{get_datasets.RawDataset}
\title{Get a \code{\link{RelationalDataset}} objects.}
\usage{
get_datasets(x)

\method{get_datasets}{RelationalDataCollection}(x)

\method{get_datasets}{RawDatasetConnector}(x)

\method{get_datasets}{RawDataset}(x)
}
\arguments{
\item{x}{(\code{\link{RelationalData}})\cr
object containing datasets.}
}
\value{
\code{list} or \code{RelationalDataset} objects
}
\description{
\ifelse{html}{\out{<a href='https://www.tidyverse.org/lifecycle/#experimental'><img src='figures/lifecycle-experimental.svg' alt='Experimental lifecycle'></a>}}{\strong{Experimental}}
}
\examples{

# RelationalData --------
library(random.cdisc.data)
adsl <- cdisc_dataset(dataname = "ADSL", # RelationalDataset
                      data = radsl(cached = TRUE),
                      code = "library(random.cdisc.data)\nADSL <- radsl(cached = TRUE)")

adae <- cdisc_dataset(dataname = "ADAE", # RelationalDataset
                       data = radae(cached = TRUE),
                       code = "library(random.cdisc.data)\nADTTE <- radae(cached = TRUE)")

rd <- cdisc_data(adsl, adae)
get_datasets(rd)

# RelationalDataConnector --------
rdc <- rcd_data(
  rcd_cdisc_dataset_connector("ADSL", radsl, cached = TRUE),
  rcd_cdisc_dataset_connector("ADTTE", radtte, cached = TRUE)
)

\dontrun{
load_datasets(rdc)
get_datasets(rdc)
}

# RelationalData --------
drc <- cdisc_data(rdc, adae)
\dontrun{
get_datasets(drc)
}

# RawDatasetConnector --------
library(random.cdisc.data)
rdc <- rcd_cdisc_dataset_connector("ADSL", radsl, cached = TRUE)

\dontrun{
load_datasets(rdc)
get_datasets(rdc)
}

# RawDataset --------
library(random.cdisc.data)
adsl <- cdisc_dataset(dataname = "ADSL",
                      data = radsl(cached = TRUE),
                      code = "library(random.cdisc.data)\nADSL <- radsl(cached = TRUE)")

get_datasets(adsl)
}
