% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_datasets.R
\name{get_datasets}
\alias{get_datasets}
\alias{get_datasets.RelationalData}
\title{Get a \code{\link{RelationalDataset}} objects.}
\usage{
get_datasets(x)

\method{get_datasets}{RelationalData}(x)
}
\arguments{
\item{x}{(\code{\link{RelationalData}})\cr
object containing datasets.}
}
\value{
\code{list} or \code{RelationalDataset} objects
}
\description{
Get a \code{\link{RelationalDataset}} objects.
}
\examples{

# RelationalData --------
library(random.cdisc.data)
adsl <- cdisc_dataset(dataname = "ADSL", # RelationalDataset
                      data = radsl(cached = TRUE),
                      code = "library(random.cdisc.data)\nADSL <- radsl(cached = TRUE)")

adtte <- cdisc_dataset(dataname = "ADTTE", # RelationalDataset
                       data = radtte(cached = TRUE),
                       code = "library(random.cdisc.data)\nADTTE <- radtte(cached = TRUE)")

rd <- teal_data(adsl, adtte)
get_datasets(rd)

# RelationalDataConnector --------
rdc <- rcd_cdisc_data(
  rcd_cdisc_dataset_connector("ADSL", radsl, cached = TRUE),
  rcd_cdisc_dataset_connector("ADTTE", radtte, cached = TRUE)
)

get_datasets(rdc)

\dontrun{
load_datasets(rdc)
get_datasets(rdc)
}

# DelayedRelationalData --------
drc <- teal_data(adsl, adtte, rdc)

get_datasets(drc)
}
