% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_datasets.R
\name{get_datasets}
\alias{get_datasets}
\alias{get_datasets.DataAbstract}
\alias{get_datasets.DatasetConnector}
\alias{get_datasets.Dataset}
\title{Get a \code{\link{Dataset}} objects.}
\usage{
get_datasets(x)

\method{get_datasets}{DataAbstract}(x)

\method{get_datasets}{DatasetConnector}(x)

\method{get_datasets}{Dataset}(x)
}
\arguments{
\item{x}{(\code{\link{RelationalData}})\cr
object containing datasets.}
}
\value{
\code{list} or \code{Dataset} objects
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{

# RelationalData --------
library(random.cdisc.data)
adsl <- cdisc_dataset(dataname = "ADSL",
                      x = radsl(cached = TRUE),
                      code = "library(random.cdisc.data)\nADSL <- radsl(cached = TRUE)")

adae <- cdisc_dataset(dataname = "ADAE",
                       x = radae(cached = TRUE),
                       code = "library(random.cdisc.data)\nADTTE <- radae(cached = TRUE)")

rd <- cdisc_data(adsl, adae)
get_datasets(rd)

# RelationalDataConnector --------
rdc <- rcd_data(
  rcd_cdisc_dataset_connector("ADSL", radsl, cached = TRUE),
  rcd_cdisc_dataset_connector("ADTTE", radtte, cached = TRUE)
)

\dontrun{
load_datasets(rdc)
get_datasets(rdc)
}

# RelationalData --------
drc <- cdisc_data(rdc, adae)
\dontrun{
get_datasets(drc)
}

# DatasetConnector --------
library(random.cdisc.data)
rdc <- rcd_cdisc_dataset_connector("ADSL", radsl, cached = TRUE)

\dontrun{
load_datasets(rdc)
get_datasets(rdc)
}

# Dataset --------
library(random.cdisc.data)
adsl <- cdisc_dataset(dataname = "ADSL",
                      x = radsl(cached = TRUE),
                      code = "library(random.cdisc.data)\nADSL <- radsl(cached = TRUE)")

get_datasets(adsl)
}
