% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_datasets.R
\name{get_datasets}
\alias{get_datasets}
\alias{get_datasets.TealDataAbstract}
\alias{get_datasets.TealDatasetConnector}
\alias{get_datasets.TealDataset}
\title{Get a \code{\link{TealDataset}} objects.}
\usage{
get_datasets(x)

\method{get_datasets}{TealDataAbstract}(x)

\method{get_datasets}{TealDatasetConnector}(x)

\method{get_datasets}{TealDataset}(x)
}
\arguments{
\item{x}{(\code{\link{TealData}})\cr
object containing datasets.}
}
\value{
\code{list} or \code{TealDataset} objects
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
}
\examples{

# TealData --------
library(scda)
adsl <- cdisc_dataset(
  dataname = "ADSL",
  x = synthetic_cdisc_data("latest")$adsl,
  code = "library(scda)\nADSL <- synthetic_cdisc_data(\"latest\")$adsl"
)

adae <- cdisc_dataset(
  dataname = "ADAE",
  x = synthetic_cdisc_data("latest")$adae,
  code = "library(scda)\nADTTE <- synthetic_cdisc_data(\"latest\")$adae"
)

rd <- cdisc_data(adsl, adae)
get_datasets(rd)

# TealDataConnector --------
adsl_cf <- callable_function(function() synthetic_cdisc_data("latest")$adsl)
adsl <- cdisc_dataset_connector(
  dataname = "ADSL",
  pull_callable = adsl_cf,
  keys = get_cdisc_keys("ADSL")
)
adlb_cf <- callable_function(function() synthetic_cdisc_data("latest")$adlb)
adlb <- cdisc_dataset_connector(
  dataname = "ADLB",
  pull_callable = adlb_cf,
  keys = get_cdisc_keys("ADLB")
)

rdc <- relational_data_connector(
  connection = data_connection(),
  connectors = list(adsl, adlb)
)
\dontrun{
load_datasets(rdc)
get_datasets(rdc)
}

# TealData --------
drc <- cdisc_data(rdc, adae)
\dontrun{
get_datasets(drc)
}

# TealDatasetConnector --------
library(scda)
adsl_cf <- callable_function(
  function() {
    synthetic_cdisc_data("latest")$adsl
  }
)
rdc <- cdisc_dataset_connector("ADSL", adsl_cf, keys = get_cdisc_keys("ADSL"))
\dontrun{
load_datasets(rdc)
get_datasets(rdc)
}

# TealDataset --------
library(scda)
adsl <- cdisc_dataset(
  dataname = "ADSL",
  x = synthetic_cdisc_data("latest")$adsl,
  code = "library(scda)\nADSL <- synthetic_cdisc_data(\"latest\")$adsl"
)

get_datasets(adsl)
}
