% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rcode_utils.R
\name{get_datasets_code}
\alias{get_datasets_code}
\title{Get datasets code}
\usage{
get_datasets_code(datanames, datasets, hashes)
}
\arguments{
\item{datanames}{(\code{character}) names of datasets to extract code from}

\item{datasets}{(\code{FilteredData}) object}

\item{hashes}{named (\code{list}) of hashes per dataset}
}
\value{
Character string concatenated from the following elements:
\itemize{
\item data pre-processing code (from \code{data} argument in \code{init})
\item hash check of loaded objects
\item filter code (if any)
}
}
\description{
Retrieve complete code to create, verify, and filter a dataset.
}
\keyword{internal}
