% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummy_functions.R
\name{example_datasets}
\alias{example_datasets}
\title{Get a dummy \code{datasets} object with \code{ADSL} data, useful in the examples}
\usage{
example_datasets()
}
\value{
\code{FilteredData} with \code{ADSL} set
}
\description{
Returns a new \code{R6} object on each invocation, not a singleton.
}
\keyword{internal}
