% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummy_functions.R
\name{get_dummy_modules}
\alias{get_dummy_modules}
\title{Get dummy modules}
\usage{
get_dummy_modules()
}
\value{
\code{teal_modules}
}
\description{
Create an example hierarchy of \code{teal_modules} from which
a teal app can be created.
}
\keyword{internal}
