% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilteredData.R
\name{get_filter_expr}
\alias{get_filter_expr}
\title{Gets filter expression for multiple datanames taking into account its order.}
\usage{
get_filter_expr(datasets, datanames = datasets$datanames())
}
\arguments{
\item{datasets}{(\code{FilteredData})}

\item{datanames}{(\code{character}) vector of dataset names}
}
\value{
(\code{expression})
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
To be used in show R code button.
}
