% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/header_filters.R
\name{get_filter_txt}
\alias{get_filter_txt}
\title{Generates text for the code to filter datasets}
\usage{
get_filter_txt(datasets)
}
\arguments{
\item{datasets}{datasets used for the teal module}
}
\value{
A character string for the code to generate the filtered datasets
}
