% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilterStates.R
\name{get_filterable_varnames}
\alias{get_filterable_varnames}
\title{Get filterable variable names}
\usage{
get_filterable_varnames(data)
}
\arguments{
\item{data}{(\code{list}, \code{matrix})\cr
R object containing columns which class can be checked through \code{vapply} or \code{apply}.}
}
\value{
\code{character} vector
}
\description{
Get filterable variable names from given data which class
matches \code{.filterable_class}
}
\examples{
df <- data.frame(
  a = letters[1:3],
  b = 1:3,
  c = Sys.Date() + 1:3,
  d = Sys.time() + 1:3,
  z = complex(3)
)
teal:::get_filterable_varnames(df)
}
