% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-module.R, R/tdata.R
\name{get_join_keys.ddl}
\alias{get_join_keys.ddl}
\alias{get_join_keys.tdata}
\title{Extract \code{JoinKeys} from \code{tdata}}
\usage{
\method{get_join_keys}{ddl}(data)

\method{get_join_keys}{tdata}(data)
}
\arguments{
\item{data}{(\code{tdata}) object}
}
\description{
Extract \code{JoinKeys} from \code{tdata}
}
\keyword{internal}
