% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tdata.R
\name{get_join_keys}
\alias{get_join_keys}
\alias{get_join_keys.tdata}
\alias{get_join_keys.default}
\title{Temporary generic method until we support \code{tdata}. Since \code{teal.data} has deprecated the method.}
\usage{
get_join_keys(data)

\method{get_join_keys}{tdata}(data)

\method{get_join_keys}{default}(data)
}
\arguments{
\item{data}{object to extract the join keys from}
}
\description{
Temporary generic method until we support \code{tdata}. Since \code{teal.data} has deprecated the method.

Extract \code{JoinKeys} from \code{tdata}
}
\keyword{internal}
