% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tdata.R
\name{get_join_keys}
\alias{get_join_keys}
\alias{get_join_keys.tdata}
\alias{get_join_keys.default}
\title{Function to get join keys from a \code{tdata} object}
\usage{
get_join_keys(data)

\method{get_join_keys}{tdata}(data)

\method{get_join_keys}{default}(data)
}
\arguments{
\item{data}{\code{tdata} - object to extract the join keys}
}
\value{
Either \code{JoinKeys} object or \code{NULL} if no join keys
}
\description{
Function to get join keys from a \code{tdata} object
}
