% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_keys.R
\name{get_keys}
\alias{get_keys}
\alias{get_keys.Dataset}
\alias{get_keys.DatasetConnector}
\alias{get_keys.DataAbstract}
\title{Get dataset primary keys}
\usage{
get_keys(x, ...)

\method{get_keys}{Dataset}(x, ...)

\method{get_keys}{DatasetConnector}(x, ...)

\method{get_keys}{DataAbstract}(x, dataname, ...)
}
\arguments{
\item{x}{an object of \code{Dataset} or \code{DatasetConnector} class}

\item{...}{not used, only for support of S3}

\item{dataname}{(\code{character}) name of dataset to return keys for}
}
\value{
(\code{character}) vector of column names
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
Get dataset primary keys
}
\examples{
# Dataset --------

library(scda)
get_keys(
  dataset(
    "ADSL",
    synthetic_cdisc_data("latest")$adsl,
    keys = get_cdisc_keys("ADSL"),
    code = "ADSL <- synthetic_cdisc_data(\"latest\")$adsl"
  )
)
# DatasetConnector --------

library(scda)
pull_fun_adsl <- callable_function(
  function() {synthetic_cdisc_data("latest")$adsl}
)
get_keys(
  dataset_connector(
    "ADSL",
    pull_fun_adsl,
    keys = get_cdisc_keys("ADSL"),
  )
)
# RelationalData --------

get_keys(
  teal_data(
    dataset("x", data.frame(x1 = 1:10, y1 = 11:20), keys = "x1"),
    dataset("y", data.frame(x2 = 1:10, y2 = 11:20), keys = "x2")
  ),
  "x"
)
}
