% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_keys.R
\name{get_keys}
\alias{get_keys}
\alias{get_keys.TealDataset}
\alias{get_keys.TealDatasetConnector}
\alias{get_keys.TealDataAbstract}
\title{Get dataset primary keys}
\usage{
get_keys(x, ...)

\method{get_keys}{TealDataset}(x, ...)

\method{get_keys}{TealDatasetConnector}(x, ...)

\method{get_keys}{TealDataAbstract}(x, dataname, ...)
}
\arguments{
\item{x}{an object of \code{TealDataset} or \code{TealDatasetConnector} class}

\item{...}{not used, only for support of S3}

\item{dataname}{(\code{character}) name of dataset to return keys for}
}
\value{
(\code{character}) vector of column names
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
Get dataset primary keys
}
\examples{
# TealDataset --------

library(scda)
get_keys(
  dataset(
    "ADSL",
    synthetic_cdisc_data("latest")$adsl,
    keys = get_cdisc_keys("ADSL"),
    code = "ADSL <- synthetic_cdisc_data(\"latest\")$adsl"
  )
)
# TealDatasetConnector --------

library(scda)
pull_fun_adsl <- callable_function(
  function() {
    synthetic_cdisc_data("latest")$adsl
  }
)
get_keys(
  dataset_connector(
    "ADSL",
    pull_fun_adsl,
    keys = get_cdisc_keys("ADSL"),
  )
)
# TealData --------

get_keys(
  teal_data(
    dataset("x", data.frame(x1 = 1:10, y1 = 11:20), keys = "x1"),
    dataset("y", data.frame(x2 = 1:10, y2 = 11:20), keys = "x2")
  ),
  "x"
)
}
