% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_label.R
\name{get_labels}
\alias{get_labels}
\title{Function that extract labels from CDISC dataset}
\usage{
get_labels(data, fill = TRUE)
}
\arguments{
\item{data}{(\code{data.frame}) table to extract the labels from}

\item{fill}{boolean in case the \code{label} attribute does not exist if
\code{TRUE} the variable names is returned, otherwise \code{NA}}
}
\value{
labels
}
\description{
\ifelse{html}{\out{<a href='https://www.tidyverse.org/lifecycle/#maturing'><img src='figures/lifecycle-maturing.svg' alt='Maturing lifecycle'></a>}}{\strong{Maturing}}
}
\examples{
library(random.cdisc.data)

ADSL <-  radsl(cached = TRUE)

get_labels(ADSL)
}
