% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdisc_data.R
\name{get_labels}
\alias{get_labels}
\title{Function that extract labels from CDISC dataset}
\usage{
get_labels(data)
}
\arguments{
\item{data}{data}
}
\value{
labels
}
\description{
Function that extract labels from CDISC dataset
}
\examples{
library(random.cdisc.data)

ADSL <-  suppressWarnings(radsl(N = 600, seed = 123))

get_labels(ADSL)

}
