% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RawDataset.R, R/get_raw_data.R
\name{get_raw_data}
\alias{get_raw_data}
\alias{RawDataset}
\alias{get_raw_data.RawDataset}
\alias{get_raw_data.RawDatasetConnector}
\title{R6 Class representing a raw data set}
\usage{
get_raw_data(x)

\method{get_raw_data}{RawDataset}(x)

\method{get_raw_data}{RawDatasetConnector}(x)
}
\arguments{
\item{x}{(\link{RawDataset} or \link{RawDatasetConnector}) object}
}
\value{
\code{data.frame} with the raw data inserted into the
  R6 objects
}
\description{
Any \code{data.frame} or \code{rtable} object can be
stored inside this object. Some attributes like colnames,
dimension or column names for a specific type will
be automatically derived.
}
\examples{

library(random.cdisc.data)
ADSL <- radsl(cached = TRUE)
ADSL_raw <- raw_dataset(x = ADSL)
get_raw_data(ADSL_raw)

ADSL_named <- named_dataset(dataname = "ADSL", x = ADSL)
get_raw_data(ADSL_named)

ADSL_relational <- as_relational(ADSL_raw,
  dataname = "ADSL",
  keys = keys(primary = c("USUBJID", "STUDYID"), foreign = NULL, parent = NULL)
)
get_raw_data(ADSL_relational)
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{ncol}}{Number of columns}

\item{\code{nrow}}{Number of rows}

\item{\code{dim}}{Dimension \code{c(x, y)}}

\item{\code{colnames}}{The column names of the data}

\item{\code{rownames}}{The rownames of the data}

\item{\code{raw_data}}{The data.frame behind this R6 class}

\item{\code{data}}{The data.frame behind this R6 class}

\item{\code{column_labels}}{for backwards compatibility
will be deprecated in future}

\item{\code{var_names}}{The column names of the data}

\item{\code{row_labels}}{Row labels (can have spaces)}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{RawDataset$new()}}
\item \href{#method-get_raw_data}{\code{RawDataset$get_raw_data()}}
\item \href{#method-get_numeric_colnames}{\code{RawDataset$get_numeric_colnames()}}
\item \href{#method-get_character_colnames}{\code{RawDataset$get_character_colnames()}}
\item \href{#method-get_factor_colnames}{\code{RawDataset$get_factor_colnames()}}
\item \href{#method-get_colnames}{\code{RawDataset$get_colnames()}}
\item \href{#method-get_column_labels}{\code{RawDataset$get_column_labels()}}
\item \href{#method-get_rownames}{\code{RawDataset$get_rownames()}}
\item \href{#method-get_row_labels}{\code{RawDataset$get_row_labels()}}
\item \href{#method-clone}{\code{RawDataset$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
initialize a \code{RawDataset} class object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RawDataset$new(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{(\code{data.frame}) object}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_raw_data"></a>}}
\subsection{Method \code{get_raw_data()}}{
Derive the raw data frame inside this object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RawDataset$get_raw_data()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{data.frame} or \code{rtable}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_numeric_colnames"></a>}}
\subsection{Method \code{get_numeric_colnames()}}{
Derive the names of all \code{numeric} columns
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RawDataset$get_numeric_colnames(include_factors = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{include_factors}}{\code{logical} Whether to include
factor variables}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_character_colnames"></a>}}
\subsection{Method \code{get_character_colnames()}}{
Derive the names of all \code{character} columns
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RawDataset$get_character_colnames()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_factor_colnames"></a>}}
\subsection{Method \code{get_factor_colnames()}}{
Derive the names of all \code{factor} columns
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RawDataset$get_factor_colnames()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_colnames"></a>}}
\subsection{Method \code{get_colnames()}}{
Derive the column names
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RawDataset$get_colnames()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_column_labels"></a>}}
\subsection{Method \code{get_column_labels()}}{
Derive the column labels
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RawDataset$get_column_labels()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_rownames"></a>}}
\subsection{Method \code{get_rownames()}}{
Derive the row names
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RawDataset$get_rownames()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_row_labels"></a>}}
\subsection{Method \code{get_row_labels()}}{
Derive the row labels
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RawDataset$get_row_labels()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RawDataset$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
