% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_data.R
\name{get_raw_data}
\alias{get_raw_data}
\alias{get_raw_data.Dataset}
\alias{get_raw_data.DatasetConnector}
\alias{get_raw_data.DataAbstract}
\title{Retrieve raw data}
\usage{
get_raw_data(x, dataname = NULL)

\method{get_raw_data}{Dataset}(x, dataname = NULL)

\method{get_raw_data}{DatasetConnector}(x, dataname = NULL)

\method{get_raw_data}{DataAbstract}(x, dataname = NULL)
}
\arguments{
\item{x}{(\code{Dataset}, \code{DatasetConnector}, \code{DataAbstract})\cr
object}

\item{dataname}{(\code{character})\cr
Name of dataset to return raw data for.}
}
\value{
\code{data.frame} with the raw data inserted into the R6 objects. In case of
\code{DataAbstract}, \code{list} of \code{data.frame} can be returned
if user doesn't specify \code{dataname} - (\code{get_raw_data} from all datasets).
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{

# Dataset ---------
library(random.cdisc.data)
ADSL <- radsl(cached = TRUE)

x <- dataset(dataname = "ADSL", x = ADSL)
get_raw_data(x)

# DatasetConnector ---------
library(random.cdisc.data)
dc <- rcd_cdisc_dataset_connector(dataname = "ADSL", fun = radsl, cached = TRUE)
load_dataset(dc)
get_raw_data(dc)

# RelationalData ----------------
library(random.cdisc.data)
adsl <- cdisc_dataset(dataname = "ADSL",
                      x = radsl(cached = TRUE),
                      code = "library(random.cdisc.data)\nADSL <- radsl(cached = TRUE)")

adtte <- cdisc_dataset(dataname = "ADTTE",
                       x = radtte(cached = TRUE),
                       code = "library(random.cdisc.data)\nADTTE <- radtte(cached = TRUE)")

rd <- teal:::RelationalData$new(adsl, adtte)
get_raw_data(rd)

# RelationalDataConnector --------
rdc <- rcd_data(
  rcd_cdisc_dataset_connector("ADSL", radsl, cached = TRUE),
  rcd_cdisc_dataset_connector("ADTTE", radtte, cached = TRUE)
)

\dontrun{
load_datasets(rdc)
get_raw_data(rdc)
}

# RelationalData (with connectors) --------
rdc2 <- rcd_data(
  rcd_cdisc_dataset_connector("ADRS", radrs, cached = TRUE),
  rcd_cdisc_dataset_connector("ADLB", radlb, cached = TRUE)
)

drc <- cdisc_data(adsl, adtte, rdc2)
\dontrun{
get_raw_data(drc)
}
}
