% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_data.R
\name{get_raw_data}
\alias{get_raw_data}
\alias{get_raw_data.TealDataset}
\alias{get_raw_data.TealDatasetConnector}
\alias{get_raw_data.TealDataAbstract}
\title{Retrieve raw data}
\usage{
get_raw_data(x, dataname = NULL)

\method{get_raw_data}{TealDataset}(x, dataname = NULL)

\method{get_raw_data}{TealDatasetConnector}(x, dataname = NULL)

\method{get_raw_data}{TealDataAbstract}(x, dataname = NULL)
}
\arguments{
\item{x}{(\code{TealDataset}, \code{TealDatasetConnector}, \code{TealDataAbstract})\cr
object}

\item{dataname}{(\code{character})\cr
Name of dataset to return raw data for.}
}
\value{
\code{data.frame} with the raw data inserted into the R6 objects. In case of
\code{TealDataAbstract}, \code{list} of \code{data.frame} can be returned
if user doesn't specify \code{dataname} - (\code{get_raw_data} from all datasets).
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
}
\examples{

# TealDataset ---------
library(scda)
ADSL <- synthetic_cdisc_data("latest")$adsl

x <- dataset(dataname = "ADSL", x = ADSL)
get_raw_data(x)

# TealDatasetConnector ---------
library(scda)
pull_fun_adsl <- callable_function(
  function() {
    synthetic_cdisc_data("latest")$adsl
  }
)
dc <- dataset_connector("ADSL", pull_fun_adsl)
load_dataset(dc)
get_raw_data(dc)

# TealData ----------------
library(scda)
adsl <- cdisc_dataset(
  dataname = "ADSL",
  x = synthetic_cdisc_data("latest")$adsl,
  code = "library(scda)\nADSL <- synthetic_cdisc_data(\"latest\")$adsl"
)

adtte <- cdisc_dataset(
  dataname = "ADTTE",
  x = synthetic_cdisc_data("latest")$adtte,
  code = "library(scda)\nADTTE <- synthetic_cdisc_data(\"latest\")$adtte"
)

rd <- teal:::TealData$new(adsl, adtte)
get_raw_data(rd)

# TealDataConnector --------
adsl_cf <- callable_function(function() synthetic_cdisc_data("latest")$adsl)
adsl <- cdisc_dataset_connector(
  dataname = "ADSL",
  pull_callable = adsl_cf,
  keys = get_cdisc_keys("ADSL")
)
adlb_cf <- callable_function(function() synthetic_cdisc_data("latest")$adlb)
adlb <- cdisc_dataset_connector(
  dataname = "ADLB",
  pull_callable = adlb_cf,
  keys = get_cdisc_keys("ADLB")
)

rdc <- relational_data_connector(
  connection = data_connection(),
  connectors = list(adsl, adlb)
)
\dontrun{
load_datasets(rdc)
get_raw_data(rdc)
}

# TealData (with connectors) --------
drc <- cdisc_data(rdc)
\dontrun{
get_raw_data(drc)
}
}
