% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rcode_utils.R
\name{get_rcode_header}
\alias{get_rcode_header}
\title{Generates header text for analysis items}
\usage{
get_rcode_header(title = NULL, description = NULL)
}
\arguments{
\item{title}{A character title of the module}

\item{description}{A character description of the module with additional
information not reflected in the title}
}
\value{
A character string for the header text
}
\description{
Generates header text for analysis items
}
\author{
Sebastian Wolf
}
\keyword{internal}
