% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_label.R
\name{get_variable_labels}
\alias{get_variable_labels}
\title{Extracts column labels from CDISC dataset}
\usage{
get_variable_labels(data, columns = NULL, fill = TRUE)
}
\arguments{
\item{data}{(\code{data.frame}) any CDISC data set}

\item{columns}{optional, (\code{character}) column names to extract the labels from. If (\code{NULL}) then all
columns are being used.}

\item{fill}{(\code{logical(1)}) if \code{TRUE}, the function will return variable names for columns with non-existent labels;
otherwise will return \code{NA} for them}
}
\value{
labels of the columns
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#superseded}{\figure{lifecycle-superseded.svg}{options: alt='[Superseded]'}}}{\strong{[Superseded]}}
}
