% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdisc_data.R
\name{get_variable_labels}
\alias{get_variable_labels}
\title{Function that extract column labels from CDISC dataset}
\usage{
get_variable_labels(data, columns = NULL, fill = TRUE)
}
\arguments{
\item{data}{(\code{data.frame}) any CDISC data set}

\item{columns}{optional, (\code{character}) column names to extract the labels from. If (\code{NULL}) then all
columns are being used.}

\item{fill}{boolean in case the \code{label} attribute does not exist if
\code{TRUE} the variable names is returned, otherwise \code{NA}}
}
\value{
labels of the columns
}
\description{
\ifelse{html}{\out{<a href='https://www.tidyverse.org/lifecycle/#maturing'><img src='figures/lifecycle-maturing.svg' alt='Maturing lifecycle'></a>}}{\strong{Maturing}}
}
\examples{
library(random.cdisc.data)
ADSL <- radsl(cached = TRUE)

get_variable_labels(ADSL)
get_variable_labels(ADSL, c("AGE", "RACE", "BMRKR1"))
get_variable_labels(ADSL, c("AGE", "RACE", "BMRKR1", "xyz"))

ADSL$NEW_COL <- 1
get_variable_labels(ADSL, c("AGE", "RACE", "BMRKR1", "NEW_COL"))
get_variable_labels(ADSL, c("AGE", "RACE", "BMRKR1", "NEW_COL"), fill = FALSE)
}
