% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rcode_header.R
\name{git_pkgs_check_rcd}
\alias{git_pkgs_check_rcd}
\title{Add random cdisc data package}
\usage{
git_pkgs_check_rcd(git_repo = c(), git_pkgs = list(),
  needs_rcd = TRUE)
}
\arguments{
\item{git_repo}{character vector of git repositories}

\item{git_pkgs}{list of git laying r packages}
}
\value{
\code{git_repo} and \code{git_pkgs} with \code{random.cdisc.data} added in case
  it was not in them before. Both are setup in one single list that
  is returned.
}
\description{
Add random cdisc data package
}
\author{
wolfs25
}
