% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/include_css_js.R
\name{include_css_files}
\alias{include_css_files}
\title{Include \code{CSS} files from \verb{/inst/css/} package directory to application header}
\usage{
include_css_files(pattern = "*")
}
\arguments{
\item{pattern}{(\code{character}) pattern of files to be included}
}
\value{
HTML code that includes \code{CSS} files
}
\description{
\code{system.file} should not be used to access files in other packages, it does
not work with \code{devtools}. Therefore, we redefine this method in each package
as needed. Thus, we do not export this method
}
\keyword{internal}
