% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/include_css_js.R
\name{include_js_files}
\alias{include_js_files}
\title{Include \code{JS} files from \verb{/inst/js/} package directory to application header}
\usage{
include_js_files(pattern = NULL, except = NULL)
}
\arguments{
\item{pattern}{(\code{character}) pattern of files to be included, passed to \code{system.file}}

\item{except}{(\code{character}) vector of basename filenames to be excluded}
}
\value{
HTML code that includes \code{JS} files
}
\description{
\code{system.file} should not be used to access files in other packages, it does
not work with \code{devtools}. Therefore, we redefine this method in each package
as needed. Thus, we do not export this method
}
\keyword{internal}
