% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_code.R
\name{include_source_code}
\alias{include_source_code}
\title{Includes source in preprocessing code lines}
\usage{
include_source_code(lines, dir = NULL)
}
\arguments{
\item{lines}{(\code{character}) of preprocessing code.}

\item{dir}{of the file where source is called from.}
}
\value{
lines of code with source text included
}
\description{
Includes source in preprocessing code lines
}
\keyword{internal}
