% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/include_css_js.R
\name{include_teal_css_js}
\alias{include_teal_css_js}
\title{Code to include \code{teal} \code{CSS} and \code{JavaScript} files}
\usage{
include_teal_css_js()
}
\value{
A \code{shiny.tag.list}.
}
\description{
This is useful when you want to use the same \code{JavaScript} and \code{CSS} files that are
used with the \code{teal} application.
This is also useful for running standalone modules in \code{teal} with the correct
styles.
Also initializes \code{shinyjs} so you can use it.
}
\details{
Simply add \code{include_teal_css_js()} as one of the UI elements.
}
\keyword{internal}
