% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{init}
\alias{init}
\title{Create the Server and UI Function For the Shiny App}
\usage{
init(
  data,
  modules,
  title = NULL,
  filter = list(),
  header = tags$p("Add Title Here"),
  footer = tags$p("Add Footer Here"),
  id = character(0)
)
}
\arguments{
\item{data}{(\code{RelationalData}) R6 object where \code{cdisc_data} or \code{teal_data} returns such a one.}

\item{modules}{nested list with one list per module with the
following named list elements:
\tabular{ll}{
\cr name \tab string with name shown in menu for the analysis item
\cr server \tab required, shiny server module function, see
\verb{\link[shiny]\{callModule\}} for more information
\cr ui \tab required, shiny ui module function, see
\verb{\link[shiny]\{callModule\}} for more information. Note, due to the app's initialization by teal,
this \code{ui} function is actually part of the \code{server} function of the shiny app which means code which must
be included in the shiny \code{ui} function (such as using \code{htmltools::htmlDependency}) should not be placed here
but should instead be placed in the \code{header} argument to \code{teal::init} as that is included in the shiny \code{ui}
function.
\cr data \tab required, vector with datasets names that are passed
on (filtered) to the server function
\cr options \tab optional, other arguments passed on to the server
function
}}

\item{title}{(\code{NULL} or \code{character}) The browser window title (defaults to the host URL of the page).}

\item{filter}{(\code{list}) You can define filters that show when
the app starts.
The general pattern is:
\code{list(iris = list(Species = ..., Sepal.Length = ...), mtcars = ...)}.
An example is:
\code{list(iris = list(Species = c("setosa", "versicolor")))}.
More generally, the filters for the variable, e.g. \code{Species} can be
specified as follows:
\code{list(Species = c("setosa", "versicolor"))},
\verb{list(Species = c("setosa", "versicolor"), keep_na = TRUE))},
or equivalently with:
\code{list(Species = c("setosa", "versicolor", NA))},
or for the default filter (not very restrictive):
\code{list(Species = default_filter())}

Instead of \code{choices} above, use the following names:
\itemize{
\item \code{numerical}: \code{range}
\item \code{factor}: \code{choices}
\item \code{logical}: \code{logical}
A general example is:
\code{list( iris = list(Sepal.Length = default_filter(), Species = c("M", NA)), mtcars = list(mpg = default_filter())   )}
Ignored if the app is restored from a bookmarked state.
}}

\item{header}{(\code{character} or \code{shiny.tag}) the header of the app. Note shiny code placed here (and in the footer
argument) will be placed in the app's \code{ui} function so code which needs to be placed in the \code{ui} function
(such as loading css via \code{htmltools::htmlDependency}) should be included here.}

\item{footer}{(\code{character} or \code{shiny.tag}) the footer of the app}

\item{id}{(\code{character}) module id to embed it, if provided,
the server function must be called with \code{callModule};
See the vignette for an example. However, \code{\link{ui_teal_with_splash}}
is then preferred to this function.}
}
\value{
named list with server and ui function
}
\description{
\ifelse{html}{\out{<a href='https://www.tidyverse.org/lifecycle/#maturing'><img src='figures/lifecycle-maturing.svg' alt='Maturing lifecycle'></a>}}{\strong{Maturing}}
End-users: This is the most important function for you to start a
teal app that is composed out of teal modules.

\strong{Notes for developers}:
This is a wrapper function around the \code{module_teal.R} functions. Unless you are
an end-user, don't use this function, but instead this module.
}
\examples{
library(random.cdisc.data)

ADSL <- radsl(cached = TRUE)

options(teal_logging = FALSE)

app <- init(
  data = cdisc_data(
    cdisc_dataset("ADSL", ADSL),
    code = "ADSL <- radsl(seed = 1)"
  ),
  modules = root_modules(
    module(
      "data source",
      server = function(input, output, session, datasets) {},
      ui = function(id, ...) div(p("information about data source")),
      filters = "all"
    ),
    module(
      "ADSL AGE histogram",
      server = function(input, output, session, datasets) {
        output$hist <- renderPlot(
          hist(datasets$get_data("ADSL", filtered = TRUE)$AGE)
        )
      },
      ui = function(id, ...) {
        ns <- NS(id)
        plotOutput(ns("hist"))
      },
      filters = "ADSL"
    )
  ),
  title = "App title",
  filter = list(ADSL = list(AGE = default_filter())),
  header = tags$h1("Sample App"),
  footer = tags$p("Copyright 2017 - 2020")
)
\dontrun{
shinyApp(app$ui, app$server)
# or: to also work with bookmarking
bookmarkableShinyApp(app$ui, app$server)
}

# See the vignette for an example how to embed this app as a module
# into a larger application
}
