% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{init}
\alias{init}
\title{Create the Server and Ui Function for the Shiny App}
\usage{
init(data, modules, filter = NULL, header = tags$p("title here"),
  footer = tags$p("footer here"))
}
\arguments{
\item{data}{named list with datasets. Dataset names are case sensitive. The
`ASL` data is mandatory.}

\item{modules}{nested list with one list per module with the
following named list elements: \tabular{ll}{ name \tab string with name
shown in menu for the analysis item \cr server \tab required, shiny server
module function, see \code{\link[shiny]{callModule}} for more
information\cr ui \tab required, shiny ui module function, see
\code{\link[shiny]{callModule}} for more information\cr data \tab required,
vector with datasets names that are passed on (filtered) to the server
function\cr options \tab optional, other arguments passed on to the server
function }}

\item{filter}{(\code{list}) You can pre-define filters for
  datasets inside this argument. Therefore you need to handover an
  \code{init} list. Please provide a named list inside \code{init}
  that contains the names of the datasets. E.g. for filtering
  the dataset \code{ASL} use \code{list(init = list(ASL = ...))}.
  For each datasets you need to provide a vector with column names that are
  relevant for the item. You can specify an ASL filtering for the
  columns \code{SEX} and \code{BAGE} by:

  \code{filter = list(init = list(ASL = c("SEX", "BAGE")))}}

\item{header}{object of class `shiny.tag` to be used as the header of the app}

\item{footer}{object of class `shiny.tag` to be used as the footer of the app}
}
\value{
named list with server and ui function
}
\description{
Creates the server and ui part for a teal shiny app
}
\examples{
library(random.cdisc.data)

ASL <- radsl(seed = 1)

app <- init(
  data = cdisc_data(
    ASL = ASL,
    code = "
      ASL <- radsl(seed = 1)
    "
  ),
  modules = root_modules(
    module(
      "data source",
      server = function(input, output, session, datasets) {},
      ui = function(id) div(p("information about data source")),
      filters = NULL
    ),
    module(
      "ASL AGE histogram",
      server = function(input, output, session, datasets) {
        output$hist <- renderPlot(
           hist(datasets$get_data("ASL", filtered = TRUE, reactive = TRUE)$AGE)
        )
      },
      ui = function(id) {ns <- NS(id); plotOutput(ns('hist'))},
      filters = "ASL"
    )
  ),
  filter = list(init = list(ASL = c("AGE"))),
  header = tags$h1("Sample App"),
  footer = tags$p("Copyright 2017")
)

\dontrun{
shinyApp(app$ui, app$server)
}
}
