% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilterState.R
\name{init_filter_state}
\alias{init_filter_state}
\title{Initializes \code{FilterState}}
\usage{
init_filter_state(
  x,
  varname,
  varlabel = if_null(attr(x, "label"), character(0)),
  input_dataname = NULL,
  extract_type = character(0)
)
}
\arguments{
\item{x}{(\code{vector})\cr
values of the variable used in filter}

\item{varname}{(\code{character(1)}, \code{name})\cr
name of the variable}

\item{varlabel}{(\code{character(1)})\cr
label of the variable (optional).}

\item{input_dataname}{(\code{name} or \code{call})\cr
name of dataset where \code{x} is taken from. Must be specified if \code{extract_type} argument
is not empty.}

\item{extract_type}{(\code{character(0)}, \code{character(1)})\cr
whether condition calls should be prefixed by dataname. Possible values:
\itemize{
\item{\code{character(0)} (default)}{ \code{varname} in the condition call will not be prefixed}
\item{\code{"list"}}{ \code{varname} in the condition call will be returned as \verb{<input_dataname>$<varname>}}
\item{\code{"matrix"}}{ \code{varname} in the condition call will be returned as \verb{<input_dataname>[, <varname>]}}
}}
}
\value{
\code{FilterState} object
}
\description{
Initializes \code{FilterState} depending on a variable class.\cr
}
\examples{
filter_state <- teal:::RangeFilterState$new(
  c(1:10, NA, Inf),
  varname = "x",
  varlabel = "Pretty name",
  input_dataname = as.name("dataname"),
  extract_type = "matrix"
)

filter_state$get_varname()
filter_state$get_varlabel()
isolate(filter_state$get_call())

\dontrun{
shinyApp(
  ui = fluidPage(
    isolate(filter_state$ui(id = "app")),
    verbatimTextOutput("call")
  ),
  server = function(input, output, session) {
    callModule(filter_state$server, "app")

    output$call <- renderText(
      pdeparse(filter_state$get_call(), collapse = "\n")
    )
  }
)
}
}
