% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilteredDataset.R
\name{init_filtered_dataset}
\alias{init_filtered_dataset}
\title{Initialize \code{FilteredDataset}}
\usage{
init_filtered_dataset(dataset, join_keys)
}
\arguments{
\item{dataset}{(\code{Dataset})\cr}

\item{join_keys}{(\verb{named list})\cr
named list of join keys to merge this dataset with
it's relatives.}
}
\description{
\code{FilteredDataset} contains \code{Dataset}
}
\examples{
library(scda)
adsl <- cdisc_dataset("ADSL", synthetic_cdisc_data("latest")$adsl)
adtte <- cdisc_dataset("ADTTE", synthetic_cdisc_data("latest")$adtte)
data <- cdisc_data(adsl, adtte)

filtered_dataset <- teal:::init_filtered_dataset(
  dataset = adtte,
  join_keys = data$get_join_keys()$get("ADTTE")
)

\dontrun{
shinyApp(
  ui = fluidPage(
    actionButton("clear", span(icon("times"), "Remove all filters")),
    filtered_dataset$ui_add_filter_state(id = "add"),
    filtered_dataset$ui("dataset"),
    verbatimTextOutput("call"),
    tableOutput("tbl")
  ),
  server = function(input, output, session) {
    callModule(
      filtered_dataset$srv_add_filter_state,
      id = "add"
    )

    callModule(filtered_dataset$server, id = "dataset")

    output$call <- renderText({
      paste(
        vapply(filtered_dataset$get_call(), pdeparse, character(1)),
        collapse = "\n"
      )
    })

    output$tbl <- renderTable({
      filtered_dataset$get_data(filtered = TRUE)
    })

    observeEvent(
      input$clear,
      filtered_dataset$queues_empty()
    )
  }
)
}
}
