% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilteredDataset.R
\name{init_filtered_dataset}
\alias{init_filtered_dataset}
\title{Initializes \code{FilteredDataset}}
\usage{
init_filtered_dataset(dataset)
}
\arguments{
\item{dataset}{(\code{Dataset})\cr}
}
\description{
\code{FilteredDataset} contains \code{Dataset}
}
\examples{
iris_d <- dataset("iris", iris)
iris_fd <- teal:::init_filtered_dataset(iris_d)

library(scda)
adsl_d <- cdisc_dataset("ADSL", synthetic_cdisc_data("latest")$adsl)
adsl_fd <- teal:::init_filtered_dataset(adsl_d)

library(MultiAssayExperiment)
MAE_d <- dataset("MAE", miniACC)

\dontrun{
shinyApp(
  ui = fluidPage(
    actionButton("clear", span(icon("times"), "Remove all filters")),
    adsl_fd$ui_add_filter_state(id = "add"),
    adsl_fd$ui("dataset"),
    verbatimTextOutput("call"),
    tableOutput("tbl")
  ),
  server = function(input, output, session) {
    adsl_fd$srv_add_filter_state(
      adsl_fd$srv_add_filter_state,
      id = "add"
    )

    adsl_fd$server(id = "dataset")

    output$call <- renderText({
      paste(
        vapply(adsl_fd$get_call(), pdeparse, character(1)),
        collapse = "\n"
      )
    })

    output$tbl <- renderTable({
      adsl_fd$get_data(filtered = TRUE)
    })

    observeEvent(
      input$clear,
      adsl_fd$queues_empty()
    )
  }
)
}
}
