% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choices_selected.R
\name{is.choices_selected}
\alias{is.choices_selected}
\title{Check if an object is a choices_selected class.}
\usage{
is.choices_selected(x)
}
\arguments{
\item{x}{object to check}
}
\description{
Check if an object is a choices_selected class.
}
