% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_pulled.R
\name{is_pulled}
\alias{is_pulled}
\alias{is_pulled.RawDatasetConnector}
\alias{is_pulled.RawDataset}
\alias{is_pulled.RelationalDataConnector}
\alias{is_pulled.RelationalData}
\title{Is pulled}
\usage{
is_pulled(x)

\method{is_pulled}{RawDatasetConnector}(x)

\method{is_pulled}{RawDataset}(x)

\method{is_pulled}{RelationalDataConnector}(x)

\method{is_pulled}{RelationalData}(x)
}
\arguments{
\item{x}{(\code{object}) of class \link{RawDatasetConnector}, \link{RelationalDataset},
\link{RelationalDatasetConnector}, or \link{RelationalDataConnector}.}
}
\value{
\code{TRUE} if connector has been already pulled, else \code{FALSE}.
}
\description{
S3 method to determine if dataset is pulled (loaded).
}
\examples{
# RawDatasetConnector --------
ds <- raw_dataset_connector(pull_fun = callable_function(data.frame))
set_args(ds, list(x = 1:5, y = letters[1:5]))

is_pulled(ds)

load_dataset(ds)
is_pulled(ds)
rel_data <- teal:::RelationalDataset$new(
  x = data.frame(x = c(1, 2), y = c("a", "b"), stringsAsFactors = FALSE),
  keys = keys(primary = "y", foreign = NULL, parent = NULL),
  dataname = "XY",
  code = "XY <- data.frame(x = c(1, 2), y = c('aa', 'bb'),
                           stringsAsFactors = FALSE)"
)

is_pulled(rel_data)

# RelationalDataConnector --------
library(random.cdisc.data)
adsl <- rcd_cdisc_dataset_connector(dataname = "ADSL", fun = radsl, cached = TRUE)
adrs <- rcd_cdisc_dataset_connector(dataname = "ADRS", fun = radrs, ADSL = adsl)

rdc <- rcd_data(adsl, adrs)

is_pulled(rdc)
\dontrun{
load_datasets(rdc)
is_pulled(rdc)
}
}
