% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_pulled.R
\name{is_pulled}
\alias{is_pulled}
\alias{is_pulled.DatasetConnector}
\alias{is_pulled.Dataset}
\alias{is_pulled.DataAbstract}
\title{Is pulled}
\usage{
is_pulled(x)

\method{is_pulled}{DatasetConnector}(x)

\method{is_pulled}{Dataset}(x)

\method{is_pulled}{DataAbstract}(x)
}
\arguments{
\item{x}{(\code{object}) of class \code{\link{DatasetConnector}}, \code{\link{Dataset}} or \code{\link{DataAbstract}}}
}
\value{
(\code{logical}) \code{TRUE} if connector has been already pulled, else \code{FALSE}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
S3 method to determine if dataset is pulled (loaded).
}
\examples{
# DatasetConnector --------
library(random.cdisc.data)
x <- rcd_dataset_connector("ADSL", radsl)

is_pulled(x)

load_dataset(x)
is_pulled(x)
# Dataset --------
x <- dataset(
  dataname = "XY",
  x = data.frame(x = c(1, 2), y = c("a", "b"), stringsAsFactors = FALSE),
  keys = "y",
  code = "XY <- data.frame(x = c(1, 2), y = c('aa', 'bb'),
                           stringsAsFactors = FALSE)"
)

is_pulled(x)
# RelationalData --------
library(random.cdisc.data)
x1 <- dataset(
  x = radsl(cached = TRUE),
  dataname = "ADSL",
  keys = get_cdisc_keys("ADSL"),
  code = "ADSL <- radsl(cached = TRUE)",
  label = "ADTTE dataset"
)

x2 <- dataset(
  x = radtte(cached = TRUE),
  dataname = "ADTTE",
  keys = get_cdisc_keys("ADTTE"),
  code = "ADTTE <- radtte(cached = TRUE)",
  label = "ADTTE dataset"
)

rd <- teal_data(x1, x2)
is_pulled(rd)

# RelationalDataConnector --------
library(random.cdisc.data)
adsl <- rcd_cdisc_dataset_connector(dataname = "ADSL", fun = radsl, cached = TRUE)
adrs <- rcd_cdisc_dataset_connector(dataname = "ADRS", fun = radrs, ADSL = adsl)

rdc <- rcd_data(adsl, adrs)

is_pulled(rdc)
\dontrun{
load_datasets(rdc)
is_pulled(rdc)
}
}
