% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdisc_data.R
\name{keys}
\alias{keys}
\title{Function that returns a keys object}
\usage{
keys(primary, foreign, parent)
}
\arguments{
\item{primary}{vector of primary key values}

\item{foreign}{vector of foreign key values}

\item{parent}{string that indicates parent dataset}
}
\value{
keys
}
\description{
\ifelse{html}{\out{<a href='https://www.tidyverse.org/lifecycle/#maturing'><img src='figures/lifecycle-maturing.svg' alt='Maturing lifecycle'></a>}}{\strong{Maturing}}
}
\examples{

keys(primary = c("STUDYID"), foreign = c("USUBJID"), "ADSL")
}
