% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_label.R
\name{keys}
\alias{keys}
\title{Function that returns a keys object}
\usage{
keys(primary, foreign, parent)
}
\arguments{
\item{primary}{vector of primary key values}

\item{foreign}{vector of foreign key values}

\item{parent}{(\code{character}) indicating parent dataname}
}
\value{
keys
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#soft-deprecated}{\figure{lifecycle-soft-deprecated.svg}{options: alt='[Soft-deprecated]'}}}{\strong{[Soft-deprecated]}}
}
