% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keys.R
\name{keys}
\alias{keys}
\title{Get object keys}
\usage{
keys(x)
}
\arguments{
\item{x}{any R object}
}
\value{
keys attribute from given object
}
\description{
Get object keys
}
\examples{
library(random.cdisc.data)
ASL <- radsl()

keys(ASL) <- c("USUBJID", "STUDYID")
keys(ASL)
}
