% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keys_filter.R
\name{keys_filtering_spec}
\alias{keys_filtering_spec}
\title{Constructor for \link{KeysFilteringSpec}}
\usage{
keys_filtering_spec(...)
}
\arguments{
\item{...}{params of \link{KeysFilteringSpec}}
}
\description{
Constructor for \link{KeysFilteringSpec}
}
