% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keys_filter.R
\docType{data}
\name{KeysFilteringSpec}
\alias{KeysFilteringSpec}
\title{Setup of Key Filtering for teal data sets}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
KeysFilteringSpec
}
\value{
A list of \code{vars} + a \link{choices_selected} outcome built by the split
	selected and choices inputs.
}
\description{
Setup of Key Filtering for teal data sets
}
\section{Fields}{

\describe{
\item{\code{vars}}{\code{character} The variables that shall be filtered with this specification}

\item{\code{choice_spec}}{specification of choices to select with fields \code{choices},
\code{selected}, \code{multiple} and \code{label}}
}}

\section{Arguments}{

\describe{
 \item{\link{keys_filtering_spec}}{ Please see \link{keys_filtering_spec} for detailed argument descriptions}
}
}

\section{Initialize}{

\describe{
   \item{via new}{ \code{KeysFilteringSpec$new(vars, sep, choices, selected, multiple, label = "Filter")} }
  \item{via constructor}{
   \code{\link{keys_filtering_spec}(vars, sep, choices, selected, multiple, label = "Filter")}
  }
}
}

\keyword{data}
