% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landing_popup_module.R
\name{landing_popup_module}
\alias{landing_popup_module}
\title{Landing popup module}
\usage{
landing_popup_module(
  label = "Landing Popup",
  title = NULL,
  content = NULL,
  buttons = modalButton("Accept")
)
}
\arguments{
\item{label}{(\code{character(1)}) Label of the module.}

\item{title}{(\code{character(1)}) Text to be displayed as popup title.}

\item{content}{(\code{character(1)}, \code{shiny.tag} or \code{shiny.tag.list}) with the content of the popup.
Passed to \code{...} of \code{shiny::modalDialog}. See examples.}

\item{buttons}{(\code{shiny.tag} or \code{shiny.tag.list}) Typically a \code{modalButton} or \code{actionButton}. See examples.}
}
\value{
A \code{teal_module} (extended with \code{teal_landing_module} class) to be used in \code{teal} applications.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} Creates a landing welcome popup for \code{teal} applications.

This module is used to display a popup dialog when the application starts.
The dialog blocks access to the application and must be closed with a button before the application can be viewed.
This function is deprecated, please use \code{add_landing_modal()} on the teal app object instead.
}
