% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_app_usage.R
\name{line_pkg_log}
\alias{line_pkg_log}
\title{package metadata to add to log file}
\usage{
line_pkg_log(pkgs, fields)
}
\arguments{
\item{pkgs}{package names}

\item{fields}{package metadata to be retrieved}
}
\description{
package metadata to add to log file
}
\examples{
\dontrun{
teal.utils:::line_pkg_log(pkgs = c("rtables", "tern", "teal"), 
fields = c("Package", "Title", "Version", "RemoteRef"))
}
}
