% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_dataset}
\alias{load_dataset}
\alias{load_dataset.RawDatasetConnector}
\title{Load data from connection}
\usage{
load_dataset(x, args, try)

\method{load_dataset}{RawDatasetConnector}(x, args = NULL, try = FALSE)
}
\arguments{
\item{x}{(\code{RawDatasetConnector} or \code{RelationalDatasetConnection})}

\item{args}{(\code{NULL} or named \code{list})\cr
additional dynamic arguments passed to function which loads the data.}

\item{try}{(\code{logical}) whether perform function evaluation inside \code{try} clause}
}
\value{
\code{x} with loaded \code{dataset} object
}
\description{
Load data from connection. Function used on \link{RawDatasetConnector} and
\link{RelationalDatasetConnector} to obtain data from connection.
}
\examples{
library(random.cdisc.data)
# for RawDatasetConnector
fun <- callable_function(radsl)
fun$set_args(list(N = 5, seed = 1, cached = TRUE))
x <- raw_dataset_connector(fun)
load_dataset(x)

# for RelationalDatasetConnector
\dontrun{
x2 <- as_relational(x, dataname = "ADSL")
load_dataset(x2)
}
}
