% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_dataset.R
\name{load_dataset}
\alias{load_dataset}
\alias{load_dataset.Dataset}
\alias{load_dataset.DatasetConnector}
\title{Load data from connection}
\usage{
load_dataset(x, ...)

\method{load_dataset}{Dataset}(x, ...)

\method{load_dataset}{DatasetConnector}(x, args = NULL, try = FALSE, conn = NULL, ...)
}
\arguments{
\item{x}{(\code{DatasetConnector} or \code{Dataset})}

\item{...}{not used, only for support of S3}

\item{args}{(\code{NULL} or named \code{list})\cr
additional dynamic arguments passed to function which loads the data.}

\item{try}{(\code{logical}) whether perform function evaluation inside \code{try} clause}

\item{conn}{Optional (\code{DataConnection}) object required to pull the data.}
}
\value{
\code{x} with loaded \code{dataset} object
}
\description{
\ifelse{html}{\out{<a href='https://www.tidyverse.org/lifecycle/#experimental'><img src='figures/lifecycle-experimental.svg' alt='Experimental lifecycle'></a>}}{\strong{Experimental}}
Load data from connection. Function used on \code{\link{DatasetConnector}} and
\code{\link{Dataset}} to obtain data from connection.
}
\examples{

# Dataset --------
library(scda)
ADSL <- synthetic_cdisc_data("latest")$adsl
ADSL_dataset <- dataset("ADSL", x = ADSL)

load_dataset(ADSL_dataset)

# DatasetConnector --------
library(scda)
pull_fun_adsl <- callable_function(
  function() {synthetic_cdisc_data("latest")$adsl}
)
adsl <- dataset_connector("ADSL", pull_fun_adsl)
load_dataset(adsl)
get_dataset(adsl)

pull_fun_adae <- callable_function(
  function() {synthetic_cdisc_data("latest")$adae}
)
adae <- dataset_connector("ADAE", pull_fun_adae)
load_dataset(adae)
}
