% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_dataset}
\alias{load_dataset}
\alias{load_dataset.RawDatasetConnector}
\alias{load_dataset.RelationalDatasetConnector}
\title{Load data from connection}
\usage{
load_dataset(x, args, try)

\method{load_dataset}{RawDatasetConnector}(x, args = NULL, try = FALSE)

\method{load_dataset}{RelationalDatasetConnector}(x, args = NULL, try = FALSE)
}
\arguments{
\item{x}{(\code{RawDatasetConnector} or \code{RelationalDatasetConnection})}

\item{args}{(\code{NULL} or named \code{list})\cr
additional dynamic arguments passed to function which loads the data.}

\item{try}{(\code{logical}) whether perform function evaluation inside \code{try} clause}
}
\value{
\code{x} with loaded \code{dataset} object
}
\description{
Load data from connection. Function used on \link{RawDatasetConnector} and
\link{RelationalDatasetConnector} to obtain data from connection.
}
\examples{


# RelationalDatasetConnector ---------
library(random.cdisc.data)
adsl <- rcd_cdisc_dataset_connector(dataname = "ADSL", fun = radsl, cached = TRUE)
load_dataset(adsl)
get_dataset(adsl)

adrs <- rcd_cdisc_dataset_connector(dataname = "ADRS", fun = radrs, ADSL = adsl)
load_dataset(adrs)
}
