% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_dataset.R
\name{load_datasets}
\alias{load_datasets}
\alias{load_datasets.Dataset}
\alias{load_datasets.DatasetConnector}
\alias{load_datasets.RelationalDataConnector}
\alias{load_datasets.RelationalData}
\title{Load datasets}
\usage{
load_datasets(x, ...)

\method{load_datasets}{Dataset}(x, ...)

\method{load_datasets}{DatasetConnector}(x, args = NULL, try = FALSE, ...)

\method{load_datasets}{RelationalDataConnector}(x, ...)

\method{load_datasets}{RelationalData}(x, ...)
}
\arguments{
\item{x}{(\code{object} of class \code{\link{RelationalData}}, \code{\link{Dataset}} or
\code{\link{DatasetConnector}})}

\item{...}{(not used)\cr
only for support of S3}

\item{args}{(\code{NULL} or named \code{list})\cr
additional dynamic arguments passed to function which loads the data. Applicable only on
\code{\link{DatasetConnector}})}

\item{try}{(\code{logical})\cr
whether perform function evaluation inside \code{try} clause. Applicable only on
\code{\link{DatasetConnector}})}
}
\value{
If executed in the interactive session shiny app is opened to load the data. If executed in
shiny application - it returns shiny server module.
}
\description{
\ifelse{html}{\out{<a href='https://www.tidyverse.org/lifecycle/#experimental'><img src='figures/lifecycle-experimental.svg' alt='Experimental lifecycle'></a>}}{\strong{Experimental}}
}
\examples{

# Dataset ------
library(scda)
ADSL <- synthetic_cdisc_data("latest")$adsl
x <- dataset("ADSL", x = ADSL)

load_datasets(x)

# DatasetConnector ------
library(scda)
pull_fun_adsl <- callable_function(
  function() {synthetic_cdisc_data("latest")$adsl}
)
adsl <- dataset_connector("ADSL", pull_fun_adsl)
load_datasets(adsl)
get_dataset(adsl)

pull_fun_adae <- callable_function(
  function() {synthetic_cdisc_data("latest")$adae}
)
adae <- dataset_connector("ADAE", pull_fun_adae)
load_datasets(adae)

# RelationalDataConnector --------
library(scda)
adsl_cf <- callable_function(
  function() {synthetic_cdisc_data("latest")$adsl}
)
adsl <- cdisc_dataset_connector(dataname = "ADSL",
                                pull_callable = adsl_cf,
                                keys = get_cdisc_keys("ADSL"))
adrs_cf <- callable_function(
  function() {synthetic_cdisc_data("latest")$adrs}
)
adrs <- cdisc_dataset_connector(dataname = "ADRS",
                                pull_callable = adrs_cf,
                                keys = get_cdisc_keys("ADRS"))

rdc <- cdisc_data(adsl, adrs)

\dontrun{
load_datasets(rdc)
}

# RelationalData --------
library(scda)
adsl_cf <- callable_function(
  function() {synthetic_cdisc_data("latest")$adsl}
)
adsl <- cdisc_dataset_connector(dataname = "ADSL",
                                pull_callable = adsl_cf,
                                keys = get_cdisc_keys("ADSL"))
adlb_cf <- callable_function(
  function() {synthetic_cdisc_data("latest")$adlb}
)
adlb <- cdisc_dataset_connector(dataname = "ADLB",
                                pull_callable = adlb_cf,
                                keys = get_cdisc_keys("ADLB"))
adrs_cf <- callable_function(
  function() {synthetic_cdisc_data("latest")$adrs}
)
adrs <- cdisc_dataset_connector(dataname = "ADRS",
                                pull_callable = adrs_cf,
                                keys = get_cdisc_keys("ADRS"))

tc <-cdisc_data(adsl, adlb, adrs)

\dontrun{
load_datasets(tc)
}
}
