% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_datasets}
\alias{load_datasets}
\alias{load_datasets.RelationalData}
\alias{load_datasets.RelationalDataConnector}
\alias{load_datasets.RelationalDataList}
\title{Load datasets}
\usage{
load_datasets(x)

\method{load_datasets}{RelationalData}(x)

\method{load_datasets}{RelationalDataConnector}(x)

\method{load_datasets}{RelationalDataList}(x)
}
\arguments{
\item{x}{(\code{object}) of class\link{RawDatasetConnector} or \link{NamedDataset}. If of
class \code{character} will be treated as file to read.}
}
\value{
object of the same class as \code{x}
}
\description{
Load datasets
}
\examples{

# RelationalData ------
library(random.cdisc.data)
adsl <- cdisc_dataset(dataname = "ADSL", # RelationalDataset
                      data = radsl(cached = TRUE),
                      code = "library(random.cdisc.data)\nADSL <- radsl(cached = TRUE)")

adtte <- cdisc_dataset(dataname = "ADTTE", # RelationalDataset
                       data = radtte(cached = TRUE),
                       code = "library(random.cdisc.data)\nADTTE <- radtte(cached = TRUE)")

rd <- teal_data(adsl, adtte)

\dontrun{
# return a warning
load_datasets(rd)
}

# RelationalDataConnector --------
library(random.cdisc.data)
adsl <- rcd_cdisc_dataset_connector(dataname = "ADSL", fun = radsl, cached = TRUE)
adrs <- rcd_cdisc_dataset_connector(dataname = "ADRS", fun = radrs, ADSL = adsl)

rdc <- rcd_cdisc_data(adsl, adrs)

\dontrun{
load_datasets(rdc)
}

# RelationalDataList --------
library(random.cdisc.data)
adsl <- rcd_cdisc_dataset_connector(dataname = "ADSL", fun = radsl, cached = TRUE)
adlb <- rcd_cdisc_dataset_connector(dataname = "ADLB", fun = radlb, cached = TRUE)
adrs <- rcd_cdisc_dataset_connector("ADRS", radrs, ADSL = adsl)

tc <- teal_data(
  rcd_cdisc_data(adsl, adlb),
  rcd_cdisc_data(adrs)
)

\dontrun{
load_datasets(tc)
}
}
