% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_datasets}
\alias{load_datasets}
\alias{load_datasets.RawDataset}
\alias{load_datasets.RawDatasetConnector}
\alias{load_datasets.RelationalDataConnector}
\alias{load_datasets.RelationalData}
\title{Load datasets}
\usage{
load_datasets(x, ...)

\method{load_datasets}{RawDataset}(x, ...)

\method{load_datasets}{RawDatasetConnector}(x, args = NULL, try = FALSE, ...)

\method{load_datasets}{RelationalDataConnector}(x, ...)

\method{load_datasets}{RelationalData}(x, ...)
}
\arguments{
\item{x}{(\code{object} of class \code{\link{RelationalData}}, \code{\link{RawDataset}} or
\code{\link{RawDatasetConnector}})}

\item{...}{(not used)\cr
only for support of S3}

\item{args}{(\code{NULL} or named \code{list})\cr
additional dynamic arguments passed to function which loads the data. Applicable only on
\code{\link{RawDatasetConnector}})}

\item{try}{(\code{logical})\cr
whether perform function evaluation inside \code{try} clause. Applicable only on
\code{\link{RawDatasetConnector}})}
}
\value{
If executed in the interactive session shiny app is opened to load the data. If executed in
shiny application - it returns shiny server module.
}
\description{
\ifelse{html}{\out{<a href='https://www.tidyverse.org/lifecycle/#experimental'><img src='figures/lifecycle-experimental.svg' alt='Experimental lifecycle'></a>}}{\strong{Experimental}}
}
\examples{

# RawDataset ------
library(random.cdisc.data)
ADSL <- radsl(cached = TRUE)
ADSL_dataset <- raw_dataset(x = ADSL)

load_datasets(ADSL_dataset)

# RawDatasetConnector ------
library(random.cdisc.data)
adsl <- rcd_cdisc_dataset_connector("ADSL", fun = radsl, cached = FALSE)
load_datasets(adsl)
get_dataset(adsl)

adae <- rcd_cdisc_dataset_connector("ADAE", fun = radae, ADSL = adsl, max_n_aes = 2L)
load_datasets(adae)

# RelationalDataConnector --------
library(random.cdisc.data)
adsl <- rcd_cdisc_dataset_connector(dataname = "ADSL", fun = radsl, cached = FALSE)
adrs <- rcd_cdisc_dataset_connector(dataname = "ADRS", fun = radrs, ADSL = adsl)

rdc <- rcd_data(adsl, adrs)

\dontrun{
load_datasets(rdc)
}

# RelationalData --------
library(random.cdisc.data)
adsl <- rcd_cdisc_dataset_connector(dataname = "ADSL", fun = radsl, cached = FALSE)
adlb <- rcd_cdisc_dataset_connector(dataname = "ADLB", fun = radlb, cached = FALSE)
adrs <- rcd_cdisc_dataset_connector("ADRS", radrs, ADSL = adsl)

tc <- cdisc_data(
  rcd_data(adsl, adlb),
  rcd_data(adrs)
)

\dontrun{
load_datasets(tc)
}
}
