% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_dataset.R
\name{load_datasets}
\alias{load_datasets}
\alias{load_datasets.TealDataset}
\alias{load_datasets.TealDatasetConnector}
\alias{load_datasets.TealDataConnector}
\alias{load_datasets.TealData}
\title{Load datasets}
\usage{
load_datasets(x, ...)

\method{load_datasets}{TealDataset}(x, ...)

\method{load_datasets}{TealDatasetConnector}(x, args = NULL, try = FALSE, ...)

\method{load_datasets}{TealDataConnector}(x, ...)

\method{load_datasets}{TealData}(x, ...)
}
\arguments{
\item{x}{(\code{\link{TealData}}, \code{\link{TealDataset}} or \code{\link{TealDatasetConnector}})}

\item{...}{(not used)\cr
only for support of S3}

\item{args}{(\code{NULL} or named \code{list})\cr
additional dynamic arguments passed to function which loads the data. Applicable only on \code{\link{TealDatasetConnector}})}

\item{try}{(\code{logical})\cr
whether perform function evaluation inside \code{try} clause. Applicable only on \code{\link{TealDatasetConnector}})}
}
\value{
If executed in the interactive session shiny app is opened to load the data. If executed in
shiny application - it returns shiny server module.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
}
\examples{

# TealDataset ------
library(scda)
ADSL <- synthetic_cdisc_data("latest")$adsl
x <- dataset("ADSL", x = ADSL)

load_datasets(x)

# TealDatasetConnector ------
library(scda)
pull_fun_adsl <- callable_function(
  function() {
    synthetic_cdisc_data("latest")$adsl
  }
)
adsl <- dataset_connector("ADSL", pull_fun_adsl)
load_datasets(adsl)
get_dataset(adsl)

pull_fun_adae <- callable_function(
  function() {
    synthetic_cdisc_data("latest")$adae
  }
)
adae <- dataset_connector("ADAE", pull_fun_adae)
load_datasets(adae)

# TealDataConnector --------
library(scda)
adsl_cf <- callable_function(
  function() {
    synthetic_cdisc_data("latest")$adsl
  }
)
adsl <- cdisc_dataset_connector(
  dataname = "ADSL",
  pull_callable = adsl_cf,
  keys = get_cdisc_keys("ADSL")
)
adrs_cf <- callable_function(
  function() {
    synthetic_cdisc_data("latest")$adrs
  }
)
adrs <- cdisc_dataset_connector(
  dataname = "ADRS",
  pull_callable = adrs_cf,
  keys = get_cdisc_keys("ADRS")
)

rdc <- cdisc_data(adsl, adrs)
\dontrun{
load_datasets(rdc)
}

# TealData --------
library(scda)
adsl_cf <- callable_function(
  function() {
    synthetic_cdisc_data("latest")$adsl
  }
)
adsl <- cdisc_dataset_connector(
  dataname = "ADSL",
  pull_callable = adsl_cf,
  keys = get_cdisc_keys("ADSL")
)
adlb_cf <- callable_function(
  function() {
    synthetic_cdisc_data("latest")$adlb
  }
)
adlb <- cdisc_dataset_connector(
  dataname = "ADLB",
  pull_callable = adlb_cf,
  keys = get_cdisc_keys("ADLB")
)
adrs_cf <- callable_function(
  function() {
    synthetic_cdisc_data("latest")$adrs
  }
)
adrs <- cdisc_dataset_connector(
  dataname = "ADRS",
  pull_callable = adrs_cf,
  keys = get_cdisc_keys("ADRS")
)

tc <- cdisc_data(adsl, adlb, adrs)
\dontrun{
load_datasets(tc)
}
}
