% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAETealDataset.R
\name{mae_dataset}
\alias{mae_dataset}
\title{The constructor of \code{MAETealDataset}}
\usage{
mae_dataset(
  dataname,
  x,
  label = data_label(x),
  code = character(0),
  vars = list()
)
}
\arguments{
\item{dataname}{(\code{character})\cr
A given name for the dataset it may not contain spaces}

\item{x}{(\code{MultiAssayExperiment})}

\item{label}{(\code{character})\cr
Label to describe the dataset}

\item{code}{(\code{character})\cr
A character string defining the code needed to produce the data set in \code{x}}

\item{vars}{(named \code{list})) \cr
In case when this object code depends on other \code{TealDataset} object(s) or
other constant value, this/these object(s) should be included as named
element(s) of the list. For example if this object code needs \code{ADSL}
object we should specify \verb{vars = list(ADSL = <adsl object>)}.
It's recommended to include \code{TealDataset} or \code{TealDatasetConnector} objects to
the \code{vars} list to preserve reproducibility. Please note that \code{vars}
are included to this object as local \code{vars} and they cannot be modified
within another dataset.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
\examples{
# Simple example
mae_d <- dataset("MAE", MultiAssayExperiment::miniACC)
mae_d$get_dataname()
mae_d$get_dataset_label()
mae_d$get_code()
mae_d$get_raw_data()
}
