% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{modify_title}
\alias{modify_title}
\title{Add a custom Title to \code{teal} application}
\usage{
modify_title(
  app,
  title = "teal app",
  favicon =
    "https://raw.githubusercontent.com/insightsengineering/hex-stickers/main/PNG/nest.png"
)
}
\arguments{
\item{app}{(\code{list}) The \code{teal} app environment.}

\item{title}{(\code{shiny.tag} or \code{character(1)}) The title to be used}

\item{favicon}{(\code{character}) The path for the icon for the title.
The image/icon path can be remote or the static path accessible by \code{shiny}, like the \verb{www/}}
}
\value{
The modified app object
}
\description{
Add a custom Title to \code{teal} application
}
\examples{
app <- init(
  data = teal_data(IRIS = iris, MTCARS = mtcars),
  modules = modules(example_module())
) |>
  modify_title("Custom title")

if (interactive()) {
  shinyApp(app$ui, app$server)
}
}
