% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{module}
\alias{module}
\title{Create a module with a new shiny page}
\usage{
module(label, server, ui, filters, server_args = NULL, ui_args = NULL)
}
\arguments{
\item{label}{(\code{character}) Label shown in the navigation item for the module.}

\item{server}{(\code{function}) Shiny server module function, see
\code{\link[shiny]{callModule}}.}

\item{ui}{(\code{function}) Shiny ui module function (see \code{\link[shiny]{callModule}})
with additional teal-specific \code{datasets} argument.}

\item{filters}{(\code{character}) A vector with datanames that are relevant for the item. The
filter panel will automatically update the shown filters to include only
filters in the listed datasets. \code{NULL} will hide the filter panel,
and the keyword \code{'all'} will show the filters of all datasets. The
argument can be thought of as \code{'active_datanames'} and may be renamed
in future versions of teal.}

\item{server_args}{(\code{list}) Named list with additional arguments passed on to the
server function. Note that the \code{FilteredDatasets} object gets
automatically passed to the server function as arguments \code{datasets}.}

\item{ui_args}{(\code{list}) Named list with additional arguments passed on to the
ui function. The argument \code{'teal_datasets'} will always be
replaced by the \code{FilteredData} object.}
}
\description{
Tab items allows you to add a shiny module to the teal app
}
