% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_bookmark_manager.R
\name{module_bookmark_manager}
\alias{module_bookmark_manager}
\alias{bookmark_module_ui}
\alias{bookmark}
\alias{bookmark_manager}
\alias{bookmark_manager_module}
\alias{bookmark_manager_srv}
\title{App state management.}
\usage{
bookmark_module_ui(id, modules)

bookmark_manager_srv(id, modules)
}
\arguments{
\item{id}{(\code{character(1)}) \code{shiny} module instance id.}
}
\value{
\code{reactiveVal} containing a named list of bookmark URLs.
}
\description{
Capture and restore the global (app) input state.
}
\details{
This module introduces bookmarks into \code{teal} apps: the \code{shiny} bookmarking mechanism becomes enabled
and server-side bookmarks can be created.

The bookmark manager is accessed with the bookmark icon in the \code{\link{wunder_bar}}.
The manager's modal displays the bookmark URL, which can be copied to the clipboard.
}
\section{Server logic}{

A bookmark is a URL that contains the app address with a \verb{/?_state_id_=<bookmark_dir>} suffix.
\verb{<bookmark_dir>} is a directory created on the server, where the state of the application is saved.
Accessing the bookmark URL opens a new session of the app that starts in the previously saved state.
}

\section{Note}{

\code{shinyOptions("bookmarkStore" = "server")} is set in \code{teal} by default on package load. Using the
\code{url} option is not supported.

\code{teal} does not guarantee that all \code{teal_module} elements are bookmarkable. Those that are, have the
\code{teal_bookmarkable} attribute set to \code{TRUE}. The bookmark manager modal displays a warning if any
modules are not bookmarkable. In order to set external \code{teal_module} elements as bookmarkable, the
\code{teal_bookmarkable} attribute must be set to \code{TRUE} by the module developer.
}

\keyword{internal}
