% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_filter_manager.R
\name{module_filter_manager_modal}
\alias{module_filter_manager_modal}
\alias{filter_manager_modal_ui}
\alias{filter_manager_modal_srv}
\title{Filter manager modal}
\usage{
filter_manager_modal_ui(id)

filter_manager_modal_srv(id, filtered_data_list, filter)
}
\arguments{
\item{id}{(\code{character(1)})\cr
\code{shiny} module id.}

\item{filtered_data_list}{(\verb{named list})\cr
A list, possibly nested, of \code{FilteredData} objects.
Each \code{FilteredData} will be served to one module in the \code{teal} application.
The structure of the list must reflect the nesting of modules in tabs
and names of the list must be the same as labels of their respective modules.}

\item{filter}{(\code{teal_slices})\cr
Specification of initial filter. Filters can be specified using \code{\link[=teal_slices]{teal_slices()}}.
Old way of specifying filters through a list is deprecated and will be removed in the
next release. Please fix your applications to use \code{\link[=teal_slices]{teal_slices()}}.}
}
\description{
Opens modal containing the filter manager \code{ui}.
}
\seealso{
examples found here: \code{vignette("internal_function_examples", package = "teal")}.
}
\keyword{internal}
