% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_filter_manager.R
\name{module_filter_manager_modal}
\alias{module_filter_manager_modal}
\alias{filter_manager_modal_ui}
\alias{filter_manager_modal_srv}
\title{Filter manager modal}
\usage{
filter_manager_modal_ui(id)

filter_manager_modal_srv(id, filtered_data_list, filter)
}
\arguments{
\item{id}{(\code{character(1)})
\code{shiny} module id.}

\item{filtered_data_list}{(named \code{list})
A list, possibly nested, of \code{FilteredData} objects.
Each \code{FilteredData} will be served to one module in the \code{teal} application.
The structure of the list must reflect the nesting of modules in tabs
and the names of the list must match the labels of their respective modules.}

\item{filter}{(\code{teal_slices})
Specifies the initial filter using \code{\link[=teal_slices]{teal_slices()}}.}
}
\description{
Opens a modal containing the filter manager UI.
}
\keyword{internal}
