% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_nested_tabs.R
\name{module_nested_tabs}
\alias{module_nested_tabs}
\alias{ui_nested_tabs}
\alias{ui_nested_tabs.default}
\alias{ui_nested_tabs.teal_modules}
\alias{ui_nested_tabs.teal_module}
\alias{srv_nested_tabs}
\alias{srv_nested_tabs.default}
\alias{srv_nested_tabs.teal_modules}
\alias{srv_nested_tabs.teal_module}
\title{Create a UI of nested tabs of \code{teal_modules}}
\usage{
ui_nested_tabs(id, modules, datasets, depth = 0L, is_module_specific = FALSE)

\method{ui_nested_tabs}{default}(id, modules, datasets, depth = 0L, is_module_specific = FALSE)

\method{ui_nested_tabs}{teal_modules}(id, modules, datasets, depth = 0L, is_module_specific = FALSE)

\method{ui_nested_tabs}{teal_module}(id, modules, datasets, depth = 0L, is_module_specific = FALSE)

srv_nested_tabs(
  id,
  datasets,
  modules,
  is_module_specific = FALSE,
  reporter = teal.reporter::Reporter$new()
)

\method{srv_nested_tabs}{default}(
  id,
  datasets,
  modules,
  is_module_specific = FALSE,
  reporter = teal.reporter::Reporter$new()
)

\method{srv_nested_tabs}{teal_modules}(
  id,
  datasets,
  modules,
  is_module_specific = FALSE,
  reporter = teal.reporter::Reporter$new()
)

\method{srv_nested_tabs}{teal_module}(
  id,
  datasets,
  modules,
  is_module_specific = TRUE,
  reporter = teal.reporter::Reporter$new()
)
}
\arguments{
\item{id}{(\code{character(1)})
module id}

\item{modules}{(\code{teal_modules}) object containing the output modules which
will be displayed in the \code{teal} application. See \code{\link[=modules]{modules()}} and \code{\link[=module]{module()}} for
more details.}

\item{datasets}{(named \code{list} of \code{FilteredData})
object to store filter state and filtered datasets, shared across modules. For more
details see \code{\link[teal.slice:FilteredData]{teal.slice::FilteredData}}. Structure of the list must be the same as structure
of the \code{modules} argument and list names must correspond to the labels in \code{modules}.
When filter is not module-specific then list contains the same object in all elements.}

\item{depth}{(\code{integer(1)})
number which helps to determine depth of the modules nesting.}

\item{is_module_specific}{(\code{logical(1)})
flag determining if the filter panel is global or module-specific.
When set to \code{TRUE}, a filter panel is called inside of each module tab.}

\item{reporter}{(\code{Reporter}) object from \code{teal.reporter}}
}
\value{
depending on class of \code{modules}, \code{ui_nested_tabs} returns:
\itemize{
\item \code{teal_module}: instantiated UI of the module.
\item \code{teal_modules}: \code{tabsetPanel} with each tab corresponding to recursively
calling this function on it.
\code{srv_nested_tabs} returns a reactive which returns the active module that corresponds to the selected tab.
}
}
\description{
Create a UI of nested tabs of \code{teal_modules}
}
\section{\code{ui_nested_tabs}}{

Each \code{teal_modules} is translated to a \code{tabsetPanel} and each
of its children is another tab-module called recursively. The UI of a
\code{teal_module} is obtained by calling the UI function on it.

The \code{datasets} argument is required to resolve the \code{teal} arguments in an
isolated context (with respect to reactivity)
}

\section{\code{srv_nested_tabs}}{

This module calls recursively all elements of the \code{modules} returns one which
is currently active.
\itemize{
\item \code{teal_module} returns self as a active module.
\item \code{teal_modules} also returns module active within self which is determined by the \code{input$active_tab}.
}
}

\examples{
# use non-exported function from teal
include_teal_css_js <- getFromNamespace("include_teal_css_js", "teal")
teal_data_to_filtered_data <- getFromNamespace("teal_data_to_filtered_data", "teal")
ui_nested_tabs <- getFromNamespace("ui_nested_tabs", "teal")
srv_nested_tabs <- getFromNamespace("srv_nested_tabs", "teal")

# creates `teal_data`
data <- teal_data(iris = iris, mtcars = mtcars)
datanames <- datanames(data)

# creates a hierarchy of `teal_modules` from which a `teal` app can be created.
mods <- modules(
  label = "d1",
  modules(
    label = "d2",
    modules(
      label = "d3",
      example_module(label = "aaa1", datanames = datanames),
      example_module(label = "aaa2", datanames = datanames)
    ),
    example_module(label = "bbb", datanames = datanames)
  ),
  example_module(label = "ccc", datanames = datanames)
)

# creates nested list aligned with the module hierarchy created above,
# each leaf holding the same `FilteredData` object.
datasets <- teal_data_to_filtered_data(data)
datasets <- list(
  "d2" = list(
    "d3" = list(
      "aaa1" = datasets,
      "aaa2" = datasets
    ),
    "bbb" = datasets
  ),
  "ccc" = datasets
)

ui <- function() {
  tagList(
    include_teal_css_js(),
    textOutput("info"),
    fluidPage( # needed for nice tabs
      ui_nested_tabs("dummy", modules = mods, datasets = datasets)
    )
  )
}
server <- function(input, output, session) {
  active_module <- srv_nested_tabs(
    "dummy",
    datasets = datasets,
    modules = mods
  )
  output$info <- renderText({
    paste0("The currently active tab name is ", active_module()$label)
  })
}
if (interactive()) {
  shinyApp(ui, server)
}

}
\keyword{internal}
