% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.0_filter_panel.R, R/module_teal.R
\name{module_teal}
\alias{module_teal}
\alias{ui_filter_panel}
\alias{srv_filter_panel}
\alias{ui_teal}
\alias{srv_teal}
\title{Filter panel module in teal}
\usage{
ui_filter_panel(id)

srv_filter_panel(id, datasets, active_datanames)

ui_teal(
  id,
  modules,
  data = NULL,
  title = build_app_title(),
  header = tags$p(),
  footer = tags$p()
)

srv_teal(id, data, modules, filter = teal_slices())
}
\arguments{
\item{id}{(\code{character(1)}) module id}

\item{datasets}{(\code{reactive}) reactive \code{FilteredData} object}

\item{modules}{(\code{list} or \code{teal_modules} or \code{teal_module})
nested list of \code{teal_modules} or \code{teal_module} objects or a single
\code{teal_modules} or \code{teal_module} object. These are the specific output modules which
will be displayed in the \code{teal} application. See \code{\link[=modules]{modules()}} and \code{\link[=module]{module()}} for
more details.}

\item{data}{(\code{teal_data}, \code{teal_data_module} or \code{reactive} returning \code{teal_data})}

\item{title}{(\code{shiny.tag} or \code{character(1)})
The browser window title. Defaults to a title "teal app" with the icon of NEST.
Can be created using the \code{build_app_title()} or
by passing a valid \code{shiny.tag} which is a head tag with title and link tag.}

\item{header}{(\code{shiny.tag} or \code{character(1)})
The header of the app.}

\item{footer}{(\code{shiny.tag} or \code{character(1)})
The footer of the app.}

\item{filter}{(\code{teal_slices}) filter object}
}
\value{
Returns a \code{reactive} expression which returns the currently active module.
}
\description{
Reactive filter panel module in teal

This module is a central point of the \code{teal} app. It is called by \code{\link[=init]{init()}} but can be also
used as a standalone module in your custom application. It is responsible for creating the main
\code{shiny} app layout and initializing all the necessary components:
\itemize{
\item \code{\link{module_data}} - for handling the \code{data}.
\item \code{\link{module_teal_module}} - for handling the \code{modules}.
\item \code{\link{module_filter_manager}} - for handling the \code{filter}.
\item \code{\link{module_snapshot_manager}} - for handling the \code{snapshots}.
\item \code{\link{module_bookmark_manager}} - for handling the \code{bookmarks}.
}
}
\details{
This module establishes reactive connection between the \code{data} and every other component in the app.
Reactive change of the \code{data} triggers reload of the app and possibly keeping all inputs settings
the same so the user can continue where one left off.
Similar applies to \code{\link{module_bookmark_manager}} which allows to start a new session with restored
inputs.
}
\keyword{internal}
