% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_teal.R
\name{module_teal}
\alias{module_teal}
\alias{ui_teal}
\alias{srv_teal}
\title{\code{teal} main app module}
\usage{
ui_teal(
  id,
  modules,
  data = NULL,
  title = build_app_title(),
  header = tags$p(),
  footer = tags$p()
)

srv_teal(id, data, modules, filter = teal_slices())
}
\arguments{
\item{id}{(\code{character}) optional
string specifying the \code{shiny} module id in cases it is used as a \code{shiny} module
rather than a standalone \code{shiny} app. This is a legacy feature.}

\item{modules}{(\code{list} or \code{teal_modules} or \code{teal_module})
nested list of \code{teal_modules} or \code{teal_module} objects or a single
\code{teal_modules} or \code{teal_module} object. These are the specific output modules which
will be displayed in the \code{teal} application. See \code{\link[=modules]{modules()}} and \code{\link[=module]{module()}} for
more details.}

\item{data}{(\code{teal_data}, \code{teal_data_module} or \code{reactive} returning \code{teal_data})}

\item{title}{(\code{shiny.tag} or \code{character(1)})
The browser window title. Defaults to a title "teal app" with the icon of NEST.
Can be created using the \code{build_app_title()} or
by passing a valid \code{shiny.tag} which is a head tag with title and link tag.}

\item{header}{(\code{shiny.tag} or \code{character(1)})
The header of the app.}

\item{footer}{(\code{shiny.tag} or \code{character(1)})
The footer of the app.}

\item{filter}{(\code{teal_slices})
Specifies the initial filter using \code{\link[=teal_slices]{teal_slices()}}.}
}
\value{
Returns a \code{reactive} expression which returns the currently active module.
}
\description{
Module to create a \code{teal} app. This module (\code{ui} and \code{server}) is called directly by \code{\link[=init]{init()}} after
initial argument checking and setting default values. This module can be called directly and
included in your custom application.
}
\details{
Please note that \code{\link[=init]{init()}} adds \code{reporter_previewer_module} automatically, which is not a case
when calling \code{ui/srv_teal} directly.

Module is responsible for creating the main \code{shiny} app layout and initializing all the necessary
components. This module establishes reactive connection between the input \code{data} and every other
component in the app. Reactive change of the \code{data} triggers reload of the app and possibly
keeping all inputs settings the same so the user can continue where one left off.
}
\section{data flow in \code{teal} application}{

\code{teal} supports multiple data inputs (see \code{data} in \code{\link{module_init_data}}) but eventually, they are
all converted to \code{reactive} returning \code{teal_data} in \code{module_teal}. There are several operations
on this \verb{reactive teal_data} object:
\itemize{
\item data loading in \code{\link{module_init_data}}
\item data filtering in \code{\link{module_filter_data}}
\item data transformation in \code{\link{module_transform_data}}
}
}

