% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_teal.R
\name{module_teal}
\alias{module_teal}
\alias{ui_teal}
\alias{srv_teal}
\title{\code{teal} main module}
\usage{
ui_teal(
  id,
  modules,
  data = NULL,
  title = build_app_title(),
  header = tags$p(),
  footer = tags$p()
)

srv_teal(id, data, modules, filter = teal_slices())
}
\arguments{
\item{id}{(\code{character}) Optionally,
a string specifying the \code{shiny} module id in cases it is used as a \code{shiny} module
rather than a standalone \code{shiny} app. This is a legacy feature.}

\item{modules}{(\code{list} or \code{teal_modules} or \code{teal_module})
Nested list of \code{teal_modules} or \code{teal_module} objects or a single
\code{teal_modules} or \code{teal_module} object. These are the specific output modules which
will be displayed in the \code{teal} application. See \code{\link[=modules]{modules()}} and \code{\link[=module]{module()}} for
more details.}

\item{data}{(\code{teal_data}, \code{teal_data_module}, or \code{reactive} returning \code{teal_data})
The \code{ui} component of this module does not require \code{data} if \code{teal_data_module} is not provided.
The \code{data} argument in the \code{ui} is included solely for the \verb{$ui} function of the
\code{teal_data_module}. Otherwise, it can be disregarded, ensuring that \code{ui_teal} does not depend on
the reactive data of the enclosing application.}

\item{title}{(\code{shiny.tag} or \code{character(1)}) Optionally,
the browser window title. Defaults to a title "teal app" with the icon of NEST.
Can be created using the \code{build_app_title()} or
by passing a valid \code{shiny.tag} which is a head tag with title and link tag.}

\item{header}{(\code{shiny.tag} or \code{character(1)}) Optionally,
the header of the app.}

\item{footer}{(\code{shiny.tag} or \code{character(1)}) Optionally,
the footer of the app.}

\item{filter}{(\code{teal_slices}) Optionally,
specifies the initial filter using \code{\link[=teal_slices]{teal_slices()}}.}
}
\value{
\code{NULL} invisibly
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
Module to create a \code{teal} app. This module can be called directly instead of \code{\link[=init]{init()}} and
included in your custom application. Please note that \code{\link[=init]{init()}} adds \code{reporter_previewer_module}
automatically, which is not a case when calling \code{ui/srv_teal} directly.
}
\details{
Module is responsible for creating the main \code{shiny} app layout and initializing all the necessary
components. This module establishes reactive connection between the input \code{data} and every other
component in the app. Reactive change of the \code{data} passed as an argument, reloads the app and
possibly keeps all input settings the same so the user can continue where one left off.
\subsection{data flow in \code{teal} application}{

This module supports multiple data inputs but eventually, they are all converted to \code{reactive}
returning \code{teal_data} in this module. On this \verb{reactive teal_data} object several actions are
performed:
\itemize{
\item data loading in \code{\link{module_init_data}}
\item data filtering in \code{\link{module_filter_data}}
\item data transformation in \code{\link{module_transform_data}}
}
}
}
