% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.0-module_teal_module.R,
%   R/1.0-module_transform_data.R
\name{module_teal_module}
\alias{module_teal_module}
\alias{ui_teal_module}
\alias{ui_teal_module.default}
\alias{ui_teal_module.teal_modules}
\alias{ui_teal_module.teal_module}
\alias{srv_teal_module}
\alias{srv_teal_module.default}
\alias{srv_teal_module.teal_modules}
\alias{srv_teal_module.teal_module}
\alias{ui_teal_transform_data}
\alias{srv_teal_transform_data}
\title{Create a UI of nested tabs of \code{teal_modules}}
\usage{
ui_teal_module(id, modules, depth = 0L)

\method{ui_teal_module}{default}(id, modules, depth = 0L)

\method{ui_teal_module}{teal_modules}(id, modules, depth = 0L)

\method{ui_teal_module}{teal_module}(id, modules, depth = 0L)

srv_teal_module(
  id,
  data_rv,
  datasets,
  modules,
  reporter = teal.reporter::Reporter$new()
)

\method{srv_teal_module}{default}(
  id,
  data_rv,
  datasets,
  modules,
  reporter = teal.reporter::Reporter$new()
)

\method{srv_teal_module}{teal_modules}(
  id,
  data_rv,
  datasets,
  modules,
  reporter = teal.reporter::Reporter$new()
)

\method{srv_teal_module}{teal_module}(
  id,
  data_rv,
  datasets,
  modules,
  reporter = teal.reporter::Reporter$new()
)

ui_teal_transform_data(id, modules)

srv_teal_transform_data(id, data, modules)
}
\arguments{
\item{id}{(\code{character(1)}) Module id}

\item{modules}{(\code{teal_module})}

\item{depth}{(\code{integer(1)})
number which helps to determine depth of the modules nesting.}

\item{datasets}{(named \code{list} of \code{FilteredData})
object to store filter state and filtered datasets, shared across modules. For more
details see \code{\link[teal.slice:FilteredData]{teal.slice::FilteredData}}. Structure of the list must be the same as structure
of the \code{modules} argument and list names must correspond to the labels in \code{modules}.
When filter is not module-specific then list contains the same object in all elements.}

\item{reporter}{(\code{Reporter}) object from \code{teal.reporter}}

\item{data}{(\code{reactive}) \code{teal_data}}

\item{is_module_specific}{(\code{logical(1)})
flag determining if the filter panel is global or module-specific.
When set to \code{TRUE}, a filter panel is called inside of each module tab.}
}
\value{
Depending on the class of \code{modules}, \code{ui_teal_module} returns:
\itemize{
\item \code{teal_module}: instantiated UI of the module.
\item \code{teal_modules}: \code{bslib::navset_tab} with each tab corresponding to recursively
calling this function on it.
}

\code{srv_teal_module} returns a reactive which returns the active module that corresponds to the selected tab.

\code{reactive} \code{teal_data}
}
\description{
Function executes the transform module and returns the transformed data.
Transformations' order is determined by the order of the modules in the \code{modules} argument.
Reactive data output of the previous module is used as input for the next module, so the final
data is the product of all consecutive transformations.
}
\section{\code{ui_teal_module}}{

Each \code{teal_modules} is translated to a \code{bslib::navset_tab} and each
of its children is another tab-module called recursively. The UI of a
\code{teal_module} is obtained by calling its UI function.

The \code{datasets} argument is required to resolve the \code{teal} arguments in an
isolated context (with respect to reactivity).
}

\section{\code{srv_teal_module}}{

This module recursively calls all elements of \code{modules} and returns currently active one.
\itemize{
\item \code{teal_module} returns self as a active module.
\item \code{teal_modules} also returns module active within self which is determined by the \code{input$active_tab}.
}
}

\keyword{internal}
