% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_transform_data.R
\name{module_transform_data}
\alias{module_transform_data}
\alias{srv_transform_teal_data}
\alias{ui_transform_teal_data}
\title{Module to decorate \code{reactive} \code{teal_data}}
\usage{
srv_transform_teal_data(
  id,
  data,
  transformators,
  modules = lifecycle::deprecated(),
  is_transform_failed = lifecycle::deprecated(),
  expr
)

ui_transform_teal_data(id, transformators, ...)
}
\arguments{
\item{id}{(\code{character(1)}) \code{shiny} module instance id.}

\item{data}{(\code{teal_data}, \code{teal_data_module}, or \code{reactive} returning \code{teal_data})
The data which application will depend on.}

\item{transformators}{(\code{list} of \code{teal_transform_module}) decorator modules to apply sequentially
to \code{data}. Each transformator receives the output of the previous one as input.}

\item{modules}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} No longer used.}

\item{is_transform_failed}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} No longer used.}

\item{expr}{(\code{expression} or \code{reactive}) optional expression evaluated on top of the decorated
output. Useful for post-processing after all transformators have run.}

\item{...}{additional arguments passed to \code{.ui_transform_teal_data} (e.g. \code{class}).}

\item{class}{(\code{character(1)}) CSS class added to the \code{div} wrapper of each transformator panel.}
}
\value{
\code{reactive} \code{teal_data}

A \code{list} of \code{bslib::accordion} UI elements, one per transformator, or \code{NULL} if
\code{transformators} is empty.
}
\description{
Module to decorate \code{reactive} \code{teal_data}
}
\details{
Module allows to modify elements in \code{teal_data} object. This module is commonly used in \code{teal_module}
consuming \code{\link[=teal_transform_module]{teal_transform_module()}} in order to "decorate" its output.

\code{ui_transform_teal_data} is a thin wrapper around the internal \code{.ui_transform_teal_data}.
}
