% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_transform_data.R
\name{module_transform_data}
\alias{module_transform_data}
\alias{ui_transform_data}
\alias{srv_transform_data}
\title{Module to transform \code{reactive} \code{teal_data}}
\usage{
ui_transform_data(id, transformers = list(), class = "well")

srv_transform_data(
  id,
  data,
  transformers = list(),
  modules,
  is_transformer_failed = reactiveValues()
)
}
\arguments{
\item{id}{(\code{character(1)}) Module id}

\item{transformers}{(\code{list} of \code{teal_data_module}) that will be applied to transform the data.
Each transform module UI will appear in the \code{teal}'s sidebar panel.
Transformers' \code{datanames} are added to the \code{datanames}. See \code{\link[=teal_transform_module]{teal_transform_module()}}.}

\item{data}{(\verb{reactive teal_data})}

\item{modules}{(\code{teal_modules} or \code{teal_module}) For \code{datanames} validation purpose}

\item{is_transformer_failed}{(\code{reactiveValues}) contains \code{logical} flags named after each transformer.
Help to determine if any previous transformer failed, so that following transformers can be disabled
and display a generic failure message.}
}
\value{
\code{reactive} \code{teal_data}
}
\description{
Module calls multiple \code{\link{module_teal_data}} in sequence so that \verb{reactive teal_data} output
from one module is handed over to the following module's input.
}
\keyword{internal}
