% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_wunder_bar.R
\name{module_wunder_bar}
\alias{module_wunder_bar}
\alias{wunder_bar_ui}
\alias{wunder_bar}
\alias{wunder_bar_module}
\alias{wunder_bar_srv}
\title{Manager bar module}
\usage{
wunder_bar_ui(id)

wunder_bar_srv(id, datasets, filter, modules)
}
\arguments{
\item{id}{(\code{character(1)}) \code{shiny} module instance id.}

\item{datasets}{(named \code{list})
A list, possibly nested, of \code{FilteredData} objects.
Each \code{FilteredData} will be served to one module in the \code{teal} application.
The structure of the list must reflect the nesting of modules in tabs
and the names of the list must match the labels of their respective modules.}

\item{filter}{(\code{teal_slices})
Specifies the initial filter using \code{\link[=teal_slices]{teal_slices()}}.}
}
\value{
Nothing is returned.
}
\description{
Bar of buttons that open modal dialogs.
}
\details{
Creates a bar of buttons that open modal dialogs where manager modules reside.
Currently contains three modules:
\itemize{
\item \code{\link{module_filter_manager}}
\item \code{\link{module_snapshot_manager}}
\item \code{\link{module_bookmark_manager}}
}

The bar is placed in the \code{teal} app UI, next to the filter panel hamburger.
}
\keyword{internal}
