% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabs.R
\name{modules}
\alias{modules}
\title{Create a collection of \code{module} and \code{modules} object}
\usage{
modules(label, ...)
}
\arguments{
\item{label}{label of modules collection}

\item{...}{\code{\link{module}} and \code{\link{modules}} object}
}
\value{
object of class \code{teal_modules}
}
\description{
Modules collects a tree of \code{\link{module}} and \code{\link{modules}}
objects. This is useful to define the navigation structure of a teal app.
}
