% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{modules}
\alias{modules}
\title{Create a collection of \code{module} and \code{modules} object}
\usage{
modules(label, ...)
}
\arguments{
\item{label}{(\code{character(1)})\cr
label of modules collection}

\item{...}{(\code{teal_module} or \code{teal_modules})\cr
see \code{\link[=module]{module()}} and \code{\link[=modules]{modules()}} for more details}
}
\value{
object of class \code{teal_modules}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#maturing}{\figure{lifecycle-maturing.svg}{options: alt='[Maturing]'}}}{\strong{[Maturing]}}
Modules collects a tree of \code{\link[=module]{module()}} and \code{\link[=modules]{modules()}}
objects. This is useful to define the navigation structure of a teal app.
}
