% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{format.teal_modules}
\alias{format.teal_modules}
\title{Converts \code{teal_modules} to a string}
\usage{
\method{format}{teal_modules}(x, indent = 0, ...)
}
\arguments{
\item{x}{(\code{teal_modules}) to print}

\item{indent}{(\code{integer}) indent level;
each nested \code{teal_modules} or \code{teal_module} is indented one level more}
}
\value{
(\code{character})
}
\description{
Converts \code{teal_modules} to a string
}
