% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{modules}
\alias{modules}
\alias{format.teal_modules}
\alias{print.teal_modules}
\title{Create a \code{teal_modules} object.}
\usage{
modules(..., label = "root")

\method{format}{teal_modules}(x, indent = 0, ...)

\method{print}{teal_modules}(x, ...)
}
\arguments{
\item{...}{\itemize{
\item For \code{modules()}: (\code{teal_module} or \code{teal_modules}) objects to wrap into a tab.
\item For \code{format()} and \code{print()}: arguments passed to other methods.
}}

\item{label}{(\code{character(1)}) label of modules collection (default \code{"root"}).
If using the \code{label} argument then it must be explicitly named.
For example \code{modules("lab", ...)} should be converted to \code{modules(label = "lab", ...)}}

\item{x}{(\code{teal_modules})}

\item{indent}{(\code{integer}) indent level;
each nested \code{teal_modules} or \code{teal_module} is indented one level more}
}
\value{
object of class \code{teal_modules}. Object contains following fields
\itemize{
\item \code{label}: taken from the \code{label} argument
\item \code{children}: a list containing objects passed in \code{...}. List elements are named after
their \code{label} attribute converted to a valid \code{shiny} id.
}

(\code{character})
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Collects \code{teal_module} or \code{teal_modules} objects to build the nested tab structure of the \code{teal} application.
}
\details{
\code{modules()} shapes the internal structure of a \code{teal} application by organizing \code{teal} modules in the navigation
panel. It wraps \code{teal_module} and \code{teal_modules} objects in a \code{teal_modules} object,
which results in a nested structure corresponding to the nested tabs in the final application.
}
\examples{
app <- init(
  data = teal_data(iris = iris),
  modules = modules(
    label = "Modules",
    modules(
      label = "Module",
      module(
        label = "Inner module",
        server = function(id, data) {
          moduleServer(
            id,
            module = function(input, output, session) {
              output$data <- renderDataTable(data()[["iris"]])
            }
          )
        },
        ui = function(id) {
          ns <- NS(id)
          tagList(dataTableOutput(ns("data")))
        },
        datanames = "all"
      )
    ),
    module(
      label = "Another module",
      server = function(id) {
        moduleServer(
          id,
          module = function(input, output, session) {
            output$text <- renderText("Another module")
          }
        )
      },
      ui = function(id) {
        ns <- NS(id)
        tagList(textOutput(ns("text")))
      },
      datanames = NULL
    )
  )
)
if (interactive()) {
  shinyApp(app$ui, app$server)
}
}
