% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{modules_depth}
\alias{modules_depth}
\title{check that modules has not more than depth 2}
\usage{
modules_depth(x, depth = 0)
}
\arguments{
\item{x}{\code{teal.modules} object}

\item{depth}{optional, integer determinint current depth level}
}
\value{
depth level for given module
}
\description{
check that modules has not more than depth 2
}
\examples{
m <- module("aaa", server = function(input, output, session, datasets){}, ui = NULL, filters = NULL)
x <- modules(
  "d1",
  modules(
    "d2",
    modules(
      "d3",
      m, m, m
    ),
    m
  ),
  m
)

x <- modules(
  "a",
  modules(
    "b", m
  ),
  m
)

teal:::modules_depth(x)
}
