% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{modules_depth}
\alias{modules_depth}
\title{Get module depth}
\usage{
modules_depth(modules, depth = 0)
}
\arguments{
\item{modules}{\code{teal_module} or \code{teal_modules} object, the latter
can be used for nested tabs, see \verb{\link{ui_nested_tabs}}}

\item{depth}{optional, integer determining current depth level}
}
\value{
depth level for given module
}
\description{
Depth starts at 0, so a single \code{teal.module} has depth 0.
Nesting it increases overall depth by 1.
}
\examples{
create_mod <- function(module_name) module(
  module_name,
  server = function(input, output, session, datasets) {},
  ui = function(id, ...) { tags$p(id) },
  filters = 'all'
)
mods <- modules(
  "d1",
  modules(
    "d2",
    modules(
      "d3",
      create_mod("aaa1"), create_mod("aaa2"), create_mod("aaa3")
    ),
    create_mod("bbb")
  ),
  create_mod("ccc")
)
stopifnot(teal:::modules_depth(mods) == 3)

mods <- modules(
  "a",
  modules(
    "b1", create_mod("c")
  ),
  create_mod("b2")
)
stopifnot(teal:::modules_depth(mods) == 2)
}
