% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{modules_depth}
\alias{modules_depth}
\title{Get module depth}
\usage{
modules_depth(modules, depth = 0L)
}
\arguments{
\item{modules}{(\code{list} or \code{teal_modules} or \code{teal_module})
Nested list of \code{teal_modules} or \code{teal_module} objects or a single
\code{teal_modules} or \code{teal_module} object. These are the specific output modules which
will be displayed in the \code{teal} application. See \code{\link[=modules]{modules()}} and \code{\link[=module]{module()}} for
more details.}

\item{depth}{optional integer determining current depth level}
}
\value{
Depth level for given module.
}
\description{
Depth starts at 0, so a single \code{teal.module} has depth 0.
Nesting it increases overall depth by 1.
}
\keyword{internal}
