% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/named_choices.R
\name{named_choices}
\alias{named_choices}
\title{Set "<choice>: <label>" type of Names}
\usage{
named_choices(choices, labels, select = NULL)
}
\arguments{
\item{choices}{a character vector}

\item{labels}{vector containing labels to be applied to \code{choices}}

\item{select}{a character vector that is a subset of \code{choices}. This is useful if 
only a few variables need to be named. If this argument is used, the returned vector will 
match it's order.}
}
\value{
a named character vector
}
\description{
This is often useful for \code{\link[teal]{choices_selected}} as it marks up the dropdown boxes
for \code{\link[shiny]{selectInput}}.
}
\details{
If either \code{choices} or \code{labels} are factors, they are coerced 
  to character.
  
  Duplicated elements from \code{choices} get removed.
}
\examples{

library(random.cdisc.data)
library(tern)
ADSL <- radsl(N=10, seed = 1)
ADTTE <- radtte(ADSL, seed = 1)

choices1 <- named_choices(names(ADSL), tern::var_labels(ADSL))
choices2 <- named_choices(ADTTE$PARAMCD, ADTTE$PARAM)

# if only a subset of variables are needed, use select argument
choices3 <- named_choices(names(ADSL), tern::var_labels(ADSL), 
                         select = c("ARMCD", "ARM"))


\dontrun{

library(shiny)

shinyApp(
  ui = fluidPage(selectInput("c1", label = "Choices from ADSL",
                             choices = choices1,
                             selected = choices1[1]),
                 
                 selectInput("c2", label = "Choices from ADTTE",
                             choices = choices2,
                             selected = choices2[1]),
                             
                 selectInput("c3", label = "Arm choices from ADSL",
                             choices = choices3,
                             selected = choices3[1])),
  
  server = function(input, output) {}
)

}
}
