% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NamedDataset.R
\name{named_dataset}
\alias{named_dataset}
\title{Constructor for \link{NamedDataset} class}
\usage{
named_dataset(dataname, x, code = character(0), label = character(0))
}
\arguments{
\item{dataname}{(\code{character}) A given name for the dataset
it may not contain spaces}

\item{x}{(\code{data.frame})}

\item{code}{(\code{character}) A character string defining the code
needed to produce the data set in \code{x}}

\item{label}{(\code{character}) Label to describe the dataset}
}
\value{
\link{NamedDataset} object
}
\description{
Constructor for \link{NamedDataset} class
}
\examples{
library(random.cdisc.data)
ADSL <- radsl(cached = TRUE)
ADSL_dataset <- named_dataset(dataname = "ADSL", x = ADSL)

ADSL_dataset$get_dataname()

ADSL_dataset <- named_dataset(dataname = "ADSL",
  x = ADSL,
  label = "AdAM subject-level dataset",
  code = "ADSL <- radsl(cached = TRUE)"
)

ADSL_dataset$get_dataset_label()
ADSL_dataset$get_code()

}
