% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NamedDatasetConnector.R
\name{named_dataset_connector}
\alias{named_dataset_connector}
\title{Create a new \code{NamedDatasetConnector} object}
\usage{
named_dataset_connector(
  pull_fun,
  dataname,
  code = character(0),
  label = character(0),
  vars = list()
)
}
\arguments{
\item{pull_fun}{(\code{CallableFunction})\cr
Set the pulling function \link{CallableFunction} to load \code{data.frame}.
\code{dataname} will be used as name of object to be assigned.}

\item{dataname}{(\code{character})\cr
A given name for the dataset it may not contain spaces}

\item{code}{(\code{character})\cr
A character string defining the code needed to produce the data set in \code{x}}

\item{label}{(\code{character})\cr
Label to describe the dataset}

\item{vars}{(list)\cr
In case when this object code depends on the \code{raw_data} from the other
\code{NamedDataset}, \code{NamedDatasetConnector} object(s) or other constant value,
this/these object(s) should be included}
}
\value{
new \code{NamedDatasetConnector} object
}
\description{
Create \code{NamedDatasetConnector} from \link{callable_function}.
}
