% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choices_selected.R
\name{no_selected_as_NULL}
\alias{no_selected_as_NULL}
\title{Check select choices for no choice made}
\usage{
no_selected_as_NULL(x)
}
\arguments{
\item{x}{(\code{character}) Word that shall be checked for
NULL, empty, "--no-selection"}
}
\value{
the word or NULL
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
}
