% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optionalInput.R
\name{optionalSliderInput}
\alias{optionalSliderInput}
\title{if min or max are \code{NA} then the slider widget will be hidden}
\usage{
optionalSliderInput(inputId, label, min, max, value, ...)
}
\arguments{
\item{inputId}{The \code{input} slot that will be used to access the value.}

\item{label}{Display label for the control, or \code{NULL} for no label.}

\item{min}{The minimum value (inclusive) that can be selected.}

\item{max}{The maximum value (inclusive) that can be selected.}

\item{value}{The initial value of the slider. A numeric vector of length one
will create a regular slider; a numeric vector of length two will create a
double-ended range slider. A warning will be issued if the value doesn't
fit between \code{min} and \code{max}.}

\item{...}{optional arguments to \code{sliderInput}}
}
\description{
Hidden input widgets are useful to have the \code{input[[inputId]]} variable
on available in the server funciont but no corresponding visual clutter from
input widgets that provide only a single choice.
}
\examples{
optionalSliderInput("a", "b", 0, 1, 0.2)
}
