% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/header_filters.R
\name{output_header}
\alias{output_header}
\title{Generates header text for analysis items}
\usage{
output_header(title, description = NULL, libraries = NULL,
  git_repo = NULL, data = NULL)
}
\arguments{
\item{title}{A character title of the module}

\item{description}{A character description of the module with additional
information not reflected in the title}

\item{libraries}{A character vector of the names of libraries required for
the module.}

\item{git_repo}{A character vector of the link to the GitHub repository
containing the R-code for the teal module. Optional argument}

\item{data}{optionally either an FilteredData object, or a named list of
datasets. The data sets need the \code{import} attribute and optionally the
\code{md5sum} attribute.}
}
\value{
A character string for the header text
}
\examples{


}
