% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_code_chunks.R
\name{parse_code_chunks}
\alias{parse_code_chunks}
\title{Parse Code Chunks}
\usage{
parse_code_chunks(file)
}
\arguments{
\item{file}{path to the file containing the code chunks}
}
\value{
A list of individual code chunks, with each element consisting
  of a character vector of length the number of lines from `file`
  that corresponds to the specific code chunk
}
\description{
Save code chunks between specified start and end tags
  from a code script as a list of individual code chunks, each named
  according to the code chunk name included in the tags
}
\details{
Code chunks should be non-overlapping, start with a valid start tag
  beginning after '#' (e.g. "# @start_chunk1") and end with a valid end tag,
  with non-duplicated code chunk names that adhere to variable naming rules in R.
  Otherwise, the function will throw an error
}
\examples{

file <- tempfile()

cat("
# @start_aaa
x <- 10
y <- -10
x*y
# @end_aaa

#   @start_aaa_part2
x <- data.frame(y = 1:3, z = 1:3)
x$y == x$z
#@end_aaa_part2

x <- c(\\"abc\\", \\"abcd\\")     # @start_bbb some text
y <- pi * 2
round(y, 2)
   #@end_bbb more text

#@start_ not part of a function
plot(1:10) #@end_
   ", file = file, append = FALSE)

parse_code_chunks(file)

}

