% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pdeparse}
\alias{pdeparse}
\title{Deparses and concatenates an expression}
\usage{
pdeparse(x, width.cutoff = 500L)
}
\arguments{
\item{x}{(\code{language})\cr}

\item{width.cutoff}{integer in \eqn{[20, 500]} determining the cutoff
    (in bytes) at which line-breaking is tried.}
}
\value{
\code{character}
}
\description{
Deparses and concatenates an expression
}
