% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{print.teal_module}
\alias{print.teal_module}
\title{Print \code{teal_module}}
\usage{
\method{print}{teal_module}(x, ...)
}
\arguments{
\item{x}{\code{teal_module}}

\item{...}{parameters passed to \code{toString}}
}
\description{
Print \code{teal_module}
}
