% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{print.teal_modules}
\alias{print.teal_modules}
\title{Print \code{teal_modules}}
\usage{
\method{print}{teal_modules}(x, ...)
}
\arguments{
\item{x}{\code{teal_modules}}

\item{...}{parameters passed to \code{toString}}
}
\description{
Print \code{teal_modules}
}
