% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dataset.R
\name{raw_dataset}
\alias{raw_dataset}
\title{Constructor for \code{\link{Dataset}} class}
\usage{
raw_dataset(x)
}
\arguments{
\item{x}{(\code{data.frame} or \code{MultiAssayExperiment})\cr}
}
\value{
\code{\link{Dataset}} object
}
\description{
\ifelse{html}{\figure{lifecycle-defunct.svg}{options: alt='Defunct lifecycle'}}{\strong{Defunct}}
}
\examples{
# Simple example
dataset("iris", iris)

# Example with more arguments
library(random.cdisc.data)
ADSL <- radsl(cached = TRUE)
ADSL_dataset <- dataset(dataname = "ADSL", x = ADSL)

ADSL_dataset$get_dataname()

ADSL_dataset <- dataset(
  dataname = "ADSL",
  x = ADSL,
  label = "AdAM subject-level dataset",
  code = "ADSL <- radsl(cached = TRUE)"
)

ADSL_dataset$get_dataset_label()
ADSL_dataset$get_code()
# Simple example
\dontrun{
library(MultiAssayExperiment)
MAE_dataset <- dataset("MAE", miniACC, keys = c("STUDYID", "USUBJID"))
MAE_dataset$get_dataname()
MAE_dataset$get_dataset_label()
MAE_dataset$get_code()
MAE_dataset$get_raw_data()
}
}
