% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dataset.R
\name{raw_dataset}
\alias{raw_dataset}
\title{Constructor for \code{\link{Dataset}} class}
\usage{
raw_dataset(x)
}
\arguments{
\item{x}{(\code{data.frame} or \code{MultiAssayExperiment})\cr}
}
\value{
\code{\link{Dataset}} object
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#defunct}{\figure{lifecycle-defunct.svg}{options: alt='[Defunct]'}}}{\strong{[Defunct]}}
}
\examples{
# Simple example
dataset("iris", iris)

# Example with more arguments
library(scda)
ADSL <- synthetic_cdisc_data("latest")$adsl
ADSL_dataset <- dataset(dataname = "ADSL", x = ADSL)

ADSL_dataset$get_dataname()

ADSL_dataset <- dataset(
  dataname = "ADSL",
  x = ADSL,
  label = "AdAM subject-level dataset",
  code = "ADSL <- synthetic_cdisc_data(\"latest\")$adsl"
)

ADSL_dataset$get_dataset_label()
ADSL_dataset$get_code()
}
