% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DatasetConnector_constructors.R
\name{raw_dataset_connector}
\alias{raw_dataset_connector}
\title{Create a new \code{DatasetConnector} object}
\usage{
raw_dataset_connector(pull_callable)
}
\arguments{
\item{pull_callable}{(\code{CallableFunction})\cr
function with necessary arguments set to fetch data from connection.}
}
\value{
new \code{DatasetConnector} object
}
\description{
\ifelse{html}{\figure{lifecycle-defunct.svg}{options: alt='Defunct lifecycle'}}{\strong{Defunct}}
}
\details{
Create \code{DatasetConnector} from \code{\link{callable_function}}.
}
