% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RawDatasetConnector.R
\name{raw_dataset_connector}
\alias{raw_dataset_connector}
\title{Create \code{RawDatasetConnector} object}
\usage{
raw_dataset_connector(pull_fun)
}
\arguments{
\item{pull_fun}{(\code{CallableFunction})\cr
function with necessary arguments set to fetch data from connection.}
}
\value{
\code{RawDatasetConnector} object
}
\description{
Create \link{RawDatasetConnector} object to execute specific call to fetch data
}
\examples{
ds <- raw_dataset_connector(pull_fun = callable_function(data.frame))
set_args(ds, list(x = 1:5, y = letters[1:5]))
ds$pull_dataset()
ds$get_raw_data()
ds$get_code()
}
