% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DatasetConnector_constructors.R
\name{raw_dataset_connector}
\alias{raw_dataset_connector}
\title{Create a new \code{DatasetConnector} object}
\usage{
raw_dataset_connector(pull_callable)
}
\arguments{
\item{pull_callable}{(\code{CallableFunction})\cr
function with necessary arguments set to fetch data from connection.}
}
\value{
new \code{DatasetConnector} object
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#defunct}{\figure{lifecycle-defunct.svg}{options: alt='[Defunct]'}}}{\strong{[Defunct]}}
}
\details{
Create \code{DatasetConnector} from \code{\link{callable_function}}.
}
\examples{
library(MultiAssayExperiment)
# data.frame example
pull_fun2 <- callable_function(data.frame)
pull_fun2$set_args(args = list(a = c(1,2,3)))
dataset_connector("test", pull_fun2)

# MultiAssayExperiment example
pull_fun <- callable_function(
  function() { library("MultiAssayExperiment"); data("miniACC"); return(miniACC) })
dataset_connector(
  "miniacc",
  pull_fun,
  code = 'library("MultiAssayExperiment"); data("miniACC"); return(miniACC)'
)

}
