% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RawDatasetConnector.R
\name{RawDatasetConnector}
\alias{RawDatasetConnector}
\title{A \code{RawDatasetConnector} class of objects}
\description{
A \code{RawDatasetConnector} class of objects

A \code{RawDatasetConnector} class of objects
}
\details{
Objects of this class store the connection function to fetch a single dataset.
Note that for some specific connection types (e.g. \code{RICE} or \code{SAICE}),
an object of class \code{DataConnection} must be provided.
Data can be pulled via the \code{pull} method and accessed directly
through the \code{dataset} active binding.
Pulled data inherits from the class \link{RelationalDataset}.
}
\examples{

## ------------------------------------------------
## Method `RawDatasetConnector$set_ui_input`
## ------------------------------------------------

ds <- raw_dataset_connector(pull_fun = callable_function(data.frame))
ds$set_ui_input(list(z = "character", w = "character", xx = "numeric"))
\dontrun{
ds$launch()
}

## ------------------------------------------------
## Method `RawDatasetConnector$launch`
## ------------------------------------------------

ds <- raw_dataset_connector(pull_fun = callable_function(data.frame))
ds$set_ui_input(list(z = "character", w = "character", xx = "numeric"))
\dontrun{
ds$launch()
}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{RawDatasetConnector$new()}}
\item \href{#method-set_args}{\code{RawDatasetConnector$set_args()}}
\item \href{#method-get_code}{\code{RawDatasetConnector$get_code()}}
\item \href{#method-get_dataset}{\code{RawDatasetConnector$get_dataset()}}
\item \href{#method-get_pull_fun}{\code{RawDatasetConnector$get_pull_fun()}}
\item \href{#method-get_raw_data}{\code{RawDatasetConnector$get_raw_data()}}
\item \href{#method-pull}{\code{RawDatasetConnector$pull()}}
\item \href{#method-is_pulled}{\code{RawDatasetConnector$is_pulled()}}
\item \href{#method-set_ui_input}{\code{RawDatasetConnector$set_ui_input()}}
\item \href{#method-get_ui}{\code{RawDatasetConnector$get_ui()}}
\item \href{#method-get_server}{\code{RawDatasetConnector$get_server()}}
\item \href{#method-launch}{\code{RawDatasetConnector$launch()}}
\item \href{#method-clone}{\code{RawDatasetConnector$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
Create a new \code{RawDatasetConnector} object. Set the pulling function
\link{CallableFunction} which returns a \code{data.frame}, e.g. by reading
from a function or creating it on the fly.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RawDatasetConnector$new(pull_fun, vars = list())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pull_fun}}{(\code{CallableFunction})\cr
function to pull the data.}

\item{\code{vars}}{(list)\cr
In case when this object code depends on the \code{raw_data} from the other
\code{RelationalDataset}, \code{RelationalDatasetConnector} object(s) or other constant value,
this/these object(s) should be included}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
new \code{RawDatasetConnector} object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_args"></a>}}
\subsection{Method \code{set_args()}}{
Set arguments to the pulling function
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RawDatasetConnector$set_args(args)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{args}}{(\code{NULL} or named \code{list}) dynamic arguments to function}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
nothing
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_code"></a>}}
\subsection{Method \code{get_code()}}{
Get code to get data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RawDatasetConnector$get_code(deparse = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deparse}}{(\code{logical})\cr
whether return deparsed form of a call}

\item{\code{args}}{(\code{NULL} or named \code{list})\cr
dynamic arguments to function which loads data}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
optionally deparsed \code{call} object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_dataset"></a>}}
\subsection{Method \code{get_dataset()}}{
Get dataset
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RawDatasetConnector$get_dataset()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
dataset (\code{<...>Dataset})
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_pull_fun"></a>}}
\subsection{Method \code{get_pull_fun()}}{
Get pull function
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RawDatasetConnector$get_pull_fun()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
dataset (\code{CallableFunction})
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_raw_data"></a>}}
\subsection{Method \code{get_raw_data()}}{
Get raw data from dataset
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RawDatasetConnector$get_raw_data()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
data (data.frame)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-pull"></a>}}
\subsection{Method \code{pull()}}{
Pull the data

Read or create data using \code{pull_fun} specified in the constructor.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RawDatasetConnector$pull(args = NULL, try = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{args}}{(\code{NULL} or named \code{list})\cr
additional dynamic arguments for pull function. \code{args} can be omitted if \code{pull_fun}
from constructor already contains all necessary arguments to pull data. One can try
to execute \code{pull_fun} directly by \code{x$pull_fun$run()} or to get code using
\code{x$pull_fun$get_code()}. \code{args} specified in pull are used temporary to get data but
not saved in code.}

\item{\code{try}}{(\code{logical} value)\cr
whether perform function evaluation inside \code{try} clause}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{NULL} if successful or \code{try-error} if not.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-is_pulled"></a>}}
\subsection{Method \code{is_pulled()}}{
Check if dataset has already been pulled.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RawDatasetConnector$is_pulled()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{TRUE} if connector has been already pulled, else \code{FALSE}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_ui_input"></a>}}
\subsection{Method \code{set_ui_input()}}{
Sets the shiny UI according to the given inputs.
It creates a checkboxInput for logical type, numericInput for numeric type,
and textInput for character type. No other types are allowed.
Inputs provide only scalar (length of 1) variables.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RawDatasetConnector$set_ui_input(inputs = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{inputs}}{(\code{list}) Named list with specification of what inputs
should be provided in the UI.
Elements of the list should be one of \code{c("logical", "numeric", "character")}
with length 1.
Names of the list elements will become the labels of the respective UI widgets.
Nested lists are not allowed.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Invisible self for chaining.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{ds <- raw_dataset_connector(pull_fun = callable_function(data.frame))
ds$set_ui_input(list(z = "character", w = "character", xx = "numeric"))
\dontrun{
ds$launch()
}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_ui"></a>}}
\subsection{Method \code{get_ui()}}{
Get shiny ui function
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RawDatasetConnector$get_ui(id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{(\code{character}) namespace id}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
shiny UI in given namespace id
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_server"></a>}}
\subsection{Method \code{get_server()}}{
Get shiny server function
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RawDatasetConnector$get_server()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
shiny server function
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-launch"></a>}}
\subsection{Method \code{launch()}}{
Launches a shiny app.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RawDatasetConnector$launch()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Shiny app
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{ds <- raw_dataset_connector(pull_fun = callable_function(data.frame))
ds$set_ui_input(list(z = "character", w = "character", xx = "numeric"))
\dontrun{
ds$launch()
}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RawDatasetConnector$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
