% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_code.R
\name{read_lib_names}
\alias{read_lib_names}
\title{Libraries names from preprocessing code}
\usage{
read_lib_names(lines)
}
\arguments{
\item{lines}{(\code{character}) of preprocessing code.}
}
\value{
libraries names loaded in preprocessing code
}
\description{
Reads library names from preprocessing code
}
