% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_script}
\alias{read_script}
\title{Read .R file into character}
\usage{
read_script(file, dataname = NULL)
}
\arguments{
\item{file}{(\code{character}) File to be parsed into code}

\item{dataname}{(\code{character}) dataset name to subset code from chunks}
}
\value{
(\code{character}) vector with the code
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
Comments will be excluded
}
\examples{
file_example <- tempfile()
writeLines(c("x <- 2", "#second line comment", "x <- x + 2"), file_example)

read_script(file_example)
}
