% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RelationalDatasetConnector_constructors.R
\name{relational_dataset_connector}
\alias{relational_dataset_connector}
\alias{relational_dataset_connector_file}
\alias{rcd_dataset_connector}
\alias{rds_dataset_connector}
\alias{script_dataset_connector}
\alias{code_dataset_connector}
\alias{rice_dataset_connector}
\alias{teradata_dataset_connector}
\alias{rds_cdisc_dataset_connector}
\alias{rcd_cdisc_dataset_connector}
\alias{rice_cdisc_dataset_connector}
\alias{script_cdisc_dataset_connector}
\alias{code_cdisc_dataset_connector}
\alias{teradata_cdisc_dataset_connector}
\title{Create a new \code{RelationalDatasetConnector} object}
\usage{
relational_dataset_connector(
  pull_fun,
  dataname,
  keys,
  code = character(0),
  label = character(0),
  vars = list()
)

relational_dataset_connector_file(x)

rcd_dataset_connector(
  dataname,
  fun,
  keys,
  code = character(0),
  script = character(0),
  label = character(0),
  ...
)

rds_dataset_connector(
  dataname,
  file,
  keys,
  code = character(0),
  script = character(0),
  label = character(0),
  ...
)

script_dataset_connector(
  dataname,
  file,
  keys,
  code = character(0),
  script = character(0),
  label = character(0),
  ...
)

code_dataset_connector(
  dataname,
  code,
  keys,
  mutate_code = character(0),
  label = character(0),
  ...
)

rice_dataset_connector(
  dataname,
  path,
  keys,
  code = character(0),
  script = character(0),
  label = character(0),
  ...
)

teradata_dataset_connector(
  dataname,
  table,
  keys,
  code = character(0),
  script = character(0),
  label = character(0),
  ...
)

rds_cdisc_dataset_connector(
  dataname,
  file,
  code = character(0),
  script = character(0),
  label = character(0),
  ...
)

rcd_cdisc_dataset_connector(
  dataname,
  fun,
  code = character(0),
  script = character(0),
  label = character(0),
  ...
)

rice_cdisc_dataset_connector(
  dataname,
  path,
  code = character(0),
  script = character(0),
  label = character(0),
  ...
)

script_cdisc_dataset_connector(
  dataname,
  file,
  code = character(0),
  script = character(0),
  label = character(0),
  ...
)

code_cdisc_dataset_connector(
  dataname,
  code = character(0),
  mutate_code = character(0),
  label = character(0),
  ...
)

teradata_cdisc_dataset_connector(
  dataname,
  table,
  code = character(0),
  script = character(0),
  label = character(0),
  ...
)
}
\arguments{
\item{pull_fun}{(\code{CallableFunction})\cr
function with necessary arguments set to fetch data from connection.}

\item{dataname}{(\code{character})\cr
A given name for the dataset it may not contain spaces}

\item{keys}{(\code{keys})\cr
object of S3 class keys containing foreign, primary keys and parent information}

\item{code}{(\code{character})\cr
A character string defining the code needed to produce the data set in \code{x}}

\item{label}{(\code{character})\cr
Label to describe the dataset}

\item{vars}{(list)\cr
In case when this object code depends on the \code{raw_data} from the other
\code{NamedDataset}, \code{NamedDatasetConnector} object(s) or other constant value,
this/these object(s) should be included}

\item{x}{(\code{character}) string giving the pathname of the file to read from.}

\item{fun}{(\code{function})\cr
any R function which generates \code{data.frame}, especially functions from
\code{random.cdisc.data} like \code{\link[random.cdisc.data]{radsl}}}

\item{script}{(\code{character})\cr
file that contains R Code that can be read using \link{read_script}.
Preferred before \code{code} argument}

\item{...}{(\code{optional})\cr
additional arguments applied to pull function}

\item{file}{(\code{character})\cr
path to code for \code{source}}

\item{mutate_code}{(\code{character})\cr
Vector with additional code that can be supplied to mutate the dataset.}

\item{path}{(\code{character})\cr
path to the file}

\item{table}{(\code{character}) table name}
}
\value{
new \code{RelationalDatasetConnector} object

\code{RelationalDatasetConnector} object
}
\description{
Create \code{RelationalDatasetConnector} from \link{callable_function}.

Please note that the script has to end with a call creating desired object. The error will be raised otherwise.

\code{rcd_dataset_connector} - Create a \code{RelationalDatasetConnector} from any R function.

\code{rds_dataset_connector} - Create a \code{RelationalDatasetConnector} from \code{RDS} file.

\code{script_dataset_connector} - Create a \code{RelationalDatasetConnector} from \code{.R} file.

\code{code_dataset_connector} - Create a \code{RelationalDatasetConnector}
  from a string.

\code{rice_dataset_connector} -
Create a \code{RelationalDatasetConnector} from \code{RICE} dataset.

\code{teradata_dataset_connector} -
Create a \code{RelationalDatasetConnector} from \code{TERADATA} dataset

\code{rds_cdisc_dataset_connector} -
Create a \code{RelationalDatasetConnector} from \code{RDS} file with keys automatically
assigned by \code{dataname}

\code{rcd_cdisc_dataset_connector} -
Create a \code{DatasetConnector} from any R function with keys assigned automatically
by \code{dataname}.

\code{rice_cdisc_dataset_connector} -
Create a \code{RelationalDatasetConnector} from \code{RICE} dataset with keys assigned
automatically by \code{dataname}.

\code{script_cdisc_dataset_connector} -
Create a \code{RelationalDatasetConnector} from \code{script} file with keys assigned
automatically by \code{dataname}.

\code{code_cdisc_dataset_connector} -
Create a \code{RelationalDatasetConnector} from \code{code} string with keys assigned
automatically by \code{dataname}.

\code{teradata_cdisc_dataset_connector} -
Create a \code{RelationalDatasetConnector} from \code{TERADATA} dataset with keys assigned
automatically by \code{dataname}.
}
\note{
Do not include assignment in \code{code} argument. \code{code} string
  should be a single function call which returns an object. Use \code{mutate_code}
  to make additional transformations.
}
\examples{
# simple example
file_example <- tempfile(fileext = ".R")
writeLines(
  text = c(
    "library(teal)
     library(random.cdisc.data)

     pull_fun <- callable_function(radsl)
     pull_fun$set_args(list(cached = TRUE))
     relational_dataset_connector(pull_fun, \"ADSL\", get_cdisc_keys(\"ADSL\"))"
  ),
  con = file_example
)
x <- relational_dataset_connector_file(file_example)
get_code(x)
library(random.cdisc.data)
x <- rcd_dataset_connector(
  dataname = "ADSL",
  fun = radsl,
  keys = get_cdisc_keys("ADSL"),
  cached = TRUE
)
x$get_code()
load_dataset(x)
get_dataset(x)
x$get_raw_data()
\dontrun{
x <- rds_dataset_connector(
  dataname = "ADSL",
  file = "path/to/file.RDS",
  keys = get_cdisc_keys("ADSL")
)
x$get_code()
}
\dontrun{
x <- script_dataset_connector(
  dataname = "ADSL",
  file = "path/to/script.R",
  keys = get_cdisc_keys("ADSL")
)
x$get_code()
}
x <- code_dataset_connector(
  dataname = "ADSL",
  keys = get_cdisc_keys("ADSL"),
  code = "radsl(cached = TRUE)"
)

x$get_code()

mutate_dataset(x, code = "ADSL$new_variable <- 1")
x$get_code()
x <- rice_dataset_connector(
  dataname = "ADSL",
  path = "/path/to/file.sas7bdat",
  keys = get_cdisc_keys("ADSL")
)
x$get_code()
\dontrun{
load_dataset(x)
get_dataset(x)
x$get_raw_data()
}
}
