% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DatasetConnector_constructors.R
\name{relational_dataset_connector_file}
\alias{relational_dataset_connector_file}
\title{Load \code{DatasetConnector} object from a file}
\usage{
relational_dataset_connector_file(path)
}
\arguments{
\item{path}{(\code{character}) string giving the pathname of the file to read from.}
}
\value{
\code{DatasetConnector} object
}
\description{
\ifelse{html}{\figure{lifecycle-soft-deprecated.svg}{options: alt='Soft-deprecated lifecycle'}}{\strong{Soft-deprecated}}
}
\details{
Please note that the script has to end with a call creating desired object. The error will
be raised otherwise.
}
\examples{
# simple example
file_example <- tempfile(fileext = ".R")
writeLines(
  text = c(
    "library(teal)
     library(scda)

     pull_callable <- callable_function(function() {synthetic_cdisc_data('latest')$adsl})
     dataset_connector(\"ADSL\", pull_callable, get_cdisc_keys(\"ADSL\"))"
  ),
  con = file_example
)
x <- dataset_connector_file(file_example)
get_code(x)
}
