% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RelationalData.R
\name{RelationalData}
\alias{RelationalData}
\title{\code{RelationalData} class}
\description{
Class combines multiple \code{RelationalDataset} objects.
}
\examples{
x <- teal:::RelationalDataset$new(
  x = data.frame(x = c(1, 2), y = c("a", "b"), stringsAsFactors = FALSE),
  keys = keys(primary = "y", foreign = NULL, parent = NULL),
  dataname = "XY",
  code = "XY <- data.frame(x = c(1, 2), y = c('aa', 'bb'),
                           stringsAsFactors = FALSE)"
)

x2 <- teal:::RelationalDataset$new(
  x = data.frame(x = c(1, 2), y = c("a", "b"), stringsAsFactors = FALSE),
  keys = keys(primary = "y", foreign = NULL, parent = NULL),
  dataname = "XYZ",
  code = "XYZ <- data.frame(x = c(1, 2), y = c('aa', 'bb'),
                           stringsAsFactors = FALSE)"
)

rd <- teal:::RelationalData$new(x, x2)
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{RelationalData$new()}}
\item \href{#method-get_cdisc_data}{\code{RelationalData$get_cdisc_data()}}
\item \href{#method-get_datanames}{\code{RelationalData$get_datanames()}}
\item \href{#method-get_code}{\code{RelationalData$get_code()}}
\item \href{#method-get_dataset}{\code{RelationalData$get_dataset()}}
\item \href{#method-get_datasets}{\code{RelationalData$get_datasets()}}
\item \href{#method-is_pulled}{\code{RelationalData$is_pulled()}}
\item \href{#method-check}{\code{RelationalData$check()}}
\item \href{#method-clone}{\code{RelationalData$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
Create a new \code{RelationalData} object from multiple
\code{RelationalDataset} objects.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RelationalData$new(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{(\code{RelationalDataset})\cr
at least one object.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
new \code{RelationalData} object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_cdisc_data"></a>}}
\subsection{Method \code{get_cdisc_data()}}{
Get \code{cdisc_data} object from multiple \code{RelationalDataset} objects.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RelationalData$get_cdisc_data()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{cdisc_data} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_datanames"></a>}}
\subsection{Method \code{get_datanames()}}{
Get names of the datasets.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RelationalData$get_datanames()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{character} vector with names of all datasets.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_code"></a>}}
\subsection{Method \code{get_code()}}{
Get code for all datasets.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RelationalData$get_code()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
(\code{character}) vector of code to generate datasets.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_dataset"></a>}}
\subsection{Method \code{get_dataset()}}{
Get \code{RelationalDataset} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RelationalData$get_dataset(dataname = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dataname}}{(\code{character} value)\cr
name of dataset to be returned. If \code{NULL}, all datasets are returned.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{RelationalDataset}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_datasets"></a>}}
\subsection{Method \code{get_datasets()}}{
Get \code{list} of \code{RelationalDataset} objects.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RelationalData$get_datasets()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{list} of \code{RelationalDataset}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-is_pulled"></a>}}
\subsection{Method \code{is_pulled()}}{
Check if dataset has already been pulled.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RelationalData$is_pulled()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{TRUE} if dataset has been already pulled, else \code{FALSE}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-check"></a>}}
\subsection{Method \code{check()}}{
Check if the  object raw data is reproducible from the
  \code{get_code()} code.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RelationalData$check()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{TRUE} if the dataset generated from evaluating the
  \code{get_code()} code is identical to the raw data, else \code{FALSE}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RelationalData$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
