% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RelationalData.R
\name{RelationalData}
\alias{RelationalData}
\title{\code{RelationalData} class}
\description{
Class combines multiple \code{RelationalDataset} objects.
}
\examples{
x <- relational_dataset(
  x = data.frame(x = c(1, 2), y = c("a", "b"), stringsAsFactors = FALSE),
  keys = keys(primary = "y", foreign = NULL, parent = NULL),
  dataname = "XY",
  code = "XY <- data.frame(x = c(1, 2), y = c('a', 'b'),
                           stringsAsFactors = FALSE)",
  label = character(0)
)

x2 <- relational_dataset(
  x = data.frame(x = c(1, 2), y = c("a", "b"), stringsAsFactors = FALSE),
  keys = keys(primary = "y", foreign = NULL, parent = NULL),
  dataname = "XYZ",
  code = "XYZ <- data.frame(x = c(1, 2), y = c('a', 'b'),
                           stringsAsFactors = FALSE)",
  label = character(0)
)

rd <- teal:::RelationalData$new(x, x2)
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{RelationalData$new()}}
\item \href{#method-check}{\code{RelationalData$check()}}
\item \href{#method-get_check_result}{\code{RelationalData$get_check_result()}}
\item \href{#method-get_items}{\code{RelationalData$get_items()}}
\item \href{#method-get_datanames}{\code{RelationalData$get_datanames()}}
\item \href{#method-get_code}{\code{RelationalData$get_code()}}
\item \href{#method-get_code_class}{\code{RelationalData$get_code_class()}}
\item \href{#method-mutate}{\code{RelationalData$mutate()}}
\item \href{#method-mutate_dataset}{\code{RelationalData$mutate_dataset()}}
\item \href{#method-get_dataset}{\code{RelationalData$get_dataset()}}
\item \href{#method-get_datasets}{\code{RelationalData$get_datasets()}}
\item \href{#method-is_pulled}{\code{RelationalData$is_pulled()}}
\item \href{#method-set_check}{\code{RelationalData$set_check()}}
\item \href{#method-clone}{\code{RelationalData$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
Create a new \code{RelationalData} object from multiple
\code{RelationalDataset} objects.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RelationalData$new(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{(\code{RelationalDataset})\cr
at least one object.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
new \code{RelationalData} object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-check"></a>}}
\subsection{Method \code{check()}}{
Check if the object raw data is reproducible from the \code{get_code()} code.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RelationalData$check()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{TRUE} if all the datasets generated from evaluating the
  \code{get_code()} code are identical to the raw data, else \code{FALSE}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_check_result"></a>}}
\subsection{Method \code{get_check_result()}}{
Get result of reproducibility check
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RelationalData$get_check_result()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{NULL} if check has not been called yet, \code{TRUE} / \code{FALSE} otherwise
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_items"></a>}}
\subsection{Method \code{get_items()}}{
Get all datasets and all dataset connectors

  name of dataset connector to be returned. If \code{NULL}, all connectors are returned.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RelationalData$get_items(dataname = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dataname}}{(\code{character} value)\cr}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{list} with all datasets and all connectors
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_datanames"></a>}}
\subsection{Method \code{get_datanames()}}{
Get names of the datasets.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RelationalData$get_datanames()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{character} vector with names of all datasets.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_code"></a>}}
\subsection{Method \code{get_code()}}{
Get code for all datasets.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RelationalData$get_code(dataname = NULL, deparse = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dataname}}{(\code{character}) dataname or \code{NULL} for all datasets}

\item{\code{deparse}}{(\code{logical}) whether to return the deparsed form of a call}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
(\code{character}) vector of code to generate datasets.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_code_class"></a>}}
\subsection{Method \code{get_code_class()}}{
Get internal \code{CodeClass} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RelationalData$get_code_class()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{CodeClass}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-mutate"></a>}}
\subsection{Method \code{mutate()}}{
Mutate data by code
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RelationalData$mutate(code, vars = list())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{code}}{(\code{character}) Code to mutate the dataset. Must contain the
\code{dataset$dataname}}

\item{\code{vars}}{(list)\cr
In case when this object code depends on the \code{raw_data} from the other
\code{RelationalDataset}, \code{RelationalDatasetConnector} object(s) or other constant value,
this/these object(s) should be included}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
self invisibly for chaining
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-mutate_dataset"></a>}}
\subsection{Method \code{mutate_dataset()}}{
Mutate dataset by code
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RelationalData$mutate_dataset(dataname, code, vars = list())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dataname}}{(\code{character}) Dataname to be mutated}

\item{\code{code}}{(\code{character}) Code to mutate the dataset. Must contain the
\code{dataset$dataname}}

\item{\code{vars}}{(list)\cr
In case when this object code depends on the \code{raw_data} from the other
\code{RelationalDataset}, \code{RelationalDatasetConnector} object(s) or other constant value,
this/these object(s) should be included}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
self invisibly for chaining
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_dataset"></a>}}
\subsection{Method \code{get_dataset()}}{
Get \code{RelationalDataset} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RelationalData$get_dataset(dataname = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dataname}}{(\code{character} value)\cr
name of dataset to be returned. If \code{NULL}, all datasets are returned.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{RelationalDataset}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_datasets"></a>}}
\subsection{Method \code{get_datasets()}}{
Get \code{list} of \code{RelationalDataset} objects.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RelationalData$get_datasets()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{list} of \code{RelationalDataset}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-is_pulled"></a>}}
\subsection{Method \code{is_pulled()}}{
Check if dataset has already been pulled.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RelationalData$is_pulled()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{TRUE} if dataset has been already pulled, else \code{FALSE}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_check"></a>}}
\subsection{Method \code{set_check()}}{
Set reproducibility check
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RelationalData$set_check(check = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{check}}{(\code{logical}) whether to perform reproducibility check.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{self} invisibly for chaining.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RelationalData$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
