% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{resolve_modules_datanames}
\alias{resolve_modules_datanames}
\title{Resolve \code{datanames} for the modules}
\usage{
resolve_modules_datanames(modules, datanames, join_keys)
}
\arguments{
\item{modules}{(\code{teal_modules}) object}

\item{datanames}{(\code{character}) names of datasets available in the \code{data} object}

\item{join_keys}{(\code{join_keys}) object}
}
\value{
\code{teal_modules} with resolved \code{datanames}.
}
\description{
Modifies \code{module$datanames} to include names of the parent dataset (taken from \code{join_keys}).
When \code{datanames} is set to \code{"all"} it is replaced with all available datasets names.
}
\keyword{internal}
