% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RelationalDataConnector_wrappers.R
\name{rice_cdisc_data}
\alias{rice_cdisc_data}
\title{Data connector for \code{RICE}}
\usage{
rice_cdisc_data(..., additional_ui = NULL)
}
\arguments{
\item{...}{(\code{DatasetConnector}) dataset connectors created using \link{rice_dataset_connector}}

\item{additional_ui}{\code{shiny.tag} additional user interface to be visible over login panel}
}
\value{
An object of class \code{DataConnector}
}
\description{
Build data connector for \code{RICE} datasets
}
\examples{
\dontrun{
x <- rice_cdisc_data(
  rice_dataset_connector("ADSL", "/path/to/ADSL"),
  rice_dataset_connector("ADLB", "/path/to/ADLB")
)
app <- init(
  data = x,
  modules = root_modules(
    module(
      "ADSL AGE histogram",
      server = function(input, output, session, datasets) {
        output$hist <- renderPlot({
          hist(datasets$get_data("ADSL", filtered = TRUE)$AGE)
        })
      },
      ui = function(id, ...) {ns <- NS(id); plotOutput(ns('hist'))},
      filters = "ADSL"
    )
  ),
  header = tags$h1("Sample App")
)
shinyApp(app$ui, app$server)
}
}
