% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataConnection.R
\name{rice_connection}
\alias{rice_connection}
\title{Open connection to \code{rice}}
\usage{
rice_connection(open_args = list(), close_args = list(), ping_args = list())
}
\arguments{
\item{open_args}{optional, named (\code{list}) of additional parameters for the connection's
\code{rice::rice_session_open}. Please note that the \code{password} argument will be
overwritten with \code{random.cdisc.data}.}

\item{close_args}{optional, named (\code{list}) of additional parameters for the connection's
\code{rice::rice_session_close} close function. Please note that the \code{message} argument
will be overwritten with \code{FALSE}.}

\item{ping_args}{optional, named (\code{list}) of additional parameters for the connection's
\code{rice::rice_session_active} ping function.}
}
\value{
\code{DataConnection} type of object
}
\description{
Open connection to \code{rice}
}
