% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RelationalDataConnector_constructors.R
\name{rice_data}
\alias{rice_data}
\title{Data connector for \code{RICE}}
\usage{
rice_data(..., connection = rice_connection(), additional_ui = NULL)
}
\arguments{
\item{...}{(\code{RelationalDatasetConnector} objects)\cr
dataset connectors created using \link{rice_dataset_connector}}

\item{connection}{(\code{DataConnection}) object returned from \code{rice_connection}.}

\item{additional_ui}{(\code{shiny.tag})\cr
additional user interface to be visible over login panel}
}
\value{
An object of class \code{RelatonalDataConnector}
}
\description{
\ifelse{html}{\out{<a href='https://www.tidyverse.org/lifecycle/#experimental'><img src='figures/lifecycle-experimental.svg' alt='Experimental lifecycle'></a>}}{\strong{Experimental}}
Build data connector for \code{RICE} datasets
}
\examples{

x <- rice_data(
  rice_cdisc_dataset_connector("ADSL", "/path/to/ADSL"),
  rice_cdisc_dataset_connector("ADLB", "/path/to/ADLB")
)
app <- init(
  data = cdisc_data(x),
  modules = root_modules(
    module(
      "ADSL AGE histogram",
      server = function(input, output, session, datasets) {
        output$hist <- renderPlot({
          hist(datasets$get_data("ADSL", filtered = TRUE)$AGE)
        })
      },
      ui = function(id, ...) {ns <- NS(id); plotOutput(ns('hist'))},
      filters = "ADSL"
    )
  ),
  header = tags$h1("Sample App")
)
\dontrun{
shinyApp(app$ui, app$server)
}
}
