% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TealDataConnector_constructors.R
\name{rice_data}
\alias{rice_data}
\title{Superseded (moved into separate package): \code{TealDataConnector} connector for \code{RICE}}
\usage{
rice_data(..., connection = rice_connection(), additional_ui = NULL)
}
\arguments{
\item{...}{(\code{TealDatasetConnector} objects)\cr
dataset connectors created using \code{rice_dataset_connector}}

\item{connection}{(\code{TealDataConnection}) object returned from \code{rice_connection}.}

\item{additional_ui}{(\code{shiny.tag})\cr
additional user interface to be visible over login panel}
}
\value{
An object of class \code{TealDataConnector}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#superseded}{\figure{lifecycle-superseded.svg}{options: alt='[Superseded]'}}}{\strong{[Superseded]}}
Build data connector for \code{RICE} datasets
}
