% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DatasetConnector_constructors.R
\name{rice_dataset_connector}
\alias{rice_dataset_connector}
\title{Rice \code{DatasetConnector}}
\usage{
rice_dataset_connector(
  dataname,
  path,
  keys = character(0),
  label = character(0),
  code = character(0),
  script = character(0),
  ...
)
}
\arguments{
\item{dataname}{(\code{character})\cr
A given name for the dataset it may not contain spaces}

\item{path}{(\code{character})\cr
path to the file}

\item{keys}{optional, (\code{character})\cr
vector of dataset primary keys column names}

\item{label}{(\code{character})\cr
Label to describe the dataset.}

\item{code}{(\code{character})\cr
A character string defining code to modify \code{raw_data} from this dataset. To modify
current dataset code should contain at least one assignment to object defined in \code{dataname}
argument. For example if \code{dataname = ADSL} example code should contain
\verb{ADSL <- <some R code>}. Can't be used simultaneously with \code{script}}

\item{script}{(\code{character})\cr
Alternatively to \code{code} - location of the file containing modification code.
Can't be used simultaneously with \code{script}.}

\item{...}{(\code{optional})\cr
additional arguments applied to pull function}
}
\description{
\ifelse{html}{\out{<a href='https://www.tidyverse.org/lifecycle/#experimental'><img src='figures/lifecycle-experimental.svg' alt='Experimental lifecycle'></a>}}{\strong{Experimental}}
}
\details{
Create a \code{DatasetConnector} from \code{RICE}.
}
\examples{
x <- rice_dataset_connector(
  dataname = "ADSL",
  path = "/path/to/file.sas7bdat"
)
x$get_code()
\dontrun{
load_dataset(x)
get_dataset(x)
x$get_raw_data()
}
}
