% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{root_modules}
\alias{root_modules}
\title{Create the root modules container}
\usage{
root_modules(...)
}
\arguments{
\item{...}{\code{\link{module}} and \code{\link{modules}} object}
}
\description{
sets the label to root in \code{\link{modules}}
}
