% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{root_modules}
\alias{root_modules}
\title{Deprecated: Creates the root modules container}
\usage{
root_modules(...)
}
\arguments{
\item{...}{parameters passed to \code{toString}}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
This function is deprecated, you should call \code{teal::modules} directly instead.
}
\details{
The function \code{\link[=modules]{modules()}} can also be used. The purpose of this
function is to not confuse the end-user as the label of the top-module
will not be displayed as a tab name (because the root is only one element
which has multiple children).
}
\examples{
library(shiny)

app <- init(
  data = teal_data(dataset("iris", iris)),
  modules = modules(
    module(
      label = "Module",
      server = function(id, data) {
        moduleServer(
          id,
          module = function(input, output, session) {
            output$data <- renderDataTable(data[["iris"]]())
          }
        )
      },
      ui = function(id) {
        ns <- NS(id)
        tagList(dataTableOutput(ns("data")))
      },
      filters = "all"
    ),
    module(
      label = "Another module",
      server = function(id) {
        moduleServer(
          id,
          module = function(input, output, session) {
            output$text <- renderText("Another module")
          }
        )
      },
      ui = function(id) {
        ns <- NS(id)
        tagList(textOutput(ns("text")))
      },
      filters = "all"
    )
  )
)
\dontrun{
runApp(app)
}
}
