% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DatasetConnector_constructors.R
\name{scda_dataset_connector}
\alias{scda_dataset_connector}
\alias{scda_cdisc_dataset_connector}
\title{\code{scda} \code{DatasetConnector}}
\usage{
scda_dataset_connector(
  dataname,
  scda_dataname,
  scda_name = "latest",
  keys = character(0),
  label = character(0),
  code = character(0),
  script = character(0)
)

scda_cdisc_dataset_connector(
  dataname,
  scda_dataname,
  scda_name = "latest",
  keys = get_cdisc_keys(dataname),
  parent = if (identical(dataname, "ADSL")) character(0L) else "ADSL",
  label = character(0),
  code = character(0),
  script = character(0)
)
}
\arguments{
\item{dataname}{(\code{character})\cr
A given name for the dataset it may not contain spaces}

\item{scda_dataname}{(\code{character}) which scda dataset to use (e.g. "adsl").}

\item{scda_name}{(\code{character}) which version of scda data to take, default "latest".}

\item{keys}{optional, (\code{character})\cr
vector of dataset primary keys column names}

\item{label}{(\code{character})\cr
Label to describe the dataset.}

\item{code}{(\code{character})\cr
A character string defining code to modify \code{raw_data} from this dataset. To modify
current dataset code should contain at least one assignment to object defined in \code{dataname}
argument. For example if \code{dataname = ADSL} example code should contain
\verb{ADSL <- <some R code>}. Can't be used simultaneously with \code{script}}

\item{script}{(\code{character})\cr
Alternatively to \code{code} - location of the file containing modification code.
Can't be used simultaneously with \code{script}.}

\item{parent}{optional, (\code{character}) \cr
parent dataset name}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\details{
Create a \code{DatasetConnector} for dataset in \code{scda}

Create a \code{CDISCDatasetConnector} from \code{scda} data
}
\examples{
library(scda)
x <- scda_dataset_connector(
  dataname = "ADSL", scda_dataname = "adsl",
)
x$get_code()
load_dataset(x)
get_dataset(x)
x$get_raw_data()
}
