% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decorators-helpers.R
\name{select_decorators}
\alias{select_decorators}
\title{Subset decorators based on the scope}
\usage{
select_decorators(decorators, scope)
}
\arguments{
\item{decorators}{(named \code{list}) a named list of decorators to subset.}

\item{scope}{(\code{character(1)}) a decorator name to include.}
}
\value{
A \code{list} of \code{teal_transform_module} objects matching the given \code{scope} and \code{all}.
Returns an empty list if \code{scope} and \code{all} is not found in \code{decorators}.
}
\description{
Subset decorators based on the scope
}
\keyword{internal}
