% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_args.R
\name{set_args}
\alias{set_args}
\alias{set_args.CallableFunction}
\alias{set_args.CallableCode}
\alias{set_args.TealDatasetConnector}
\title{Set arguments of a \code{CallableFunction}}
\usage{
set_args(x, args)

\method{set_args}{CallableFunction}(x, args)

\method{set_args}{CallableCode}(x, args)

\method{set_args}{TealDatasetConnector}(x, args)
}
\arguments{
\item{x}{\code{CallableFunction} or \code{TealDatasetConnector})}

\item{args}{(\code{NULL} or named \code{list}) dynamic arguments to function}
}
\value{
nothing
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
Set arguments of a \code{CallableFunction}
}
\examples{
library(scda)
f <- function(df) {
  synthetic_cdisc_data("latest")[[df]]
}
fun <- callable_function(f)
set_args(fun, list(df = "adsl"))
library(scda)
f <- function(df) {
  synthetic_cdisc_data("latest")[[df]]
}
code <- callable_code("f()")
set_args(code, list(df = "adsl"))
ds <- dataset_connector("x", pull_callable = callable_function(data.frame))
set_args(ds, list(x = 1:5, y = letters[1:5]))
}
