% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilteredData.R
\name{set_single_filter_state}
\alias{set_single_filter_state}
\title{Set the filter state for a single variable}
\usage{
set_single_filter_state(datasets, dataname, varname, state)
}
\arguments{
\item{datasets}{`FilteredData`}

\item{dataname}{`character` dataname}

\item{varname}{`character` variable name}

\item{state}{`list` new state for the variable, can be
`default_filter()`}
}
\description{
See `FilteredData$set_filter_state`.
}
