% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_rcode_modal.R
\name{show_rcode_modal}
\alias{show_rcode_modal}
\title{Show \code{R} code modal}
\usage{
show_rcode_modal(title = NULL, rcode, session = getDefaultReactiveDomain())
}
\arguments{
\item{title}{(\code{character(1)})
Title of the modal, displayed in the first comment of the \code{R} code.}

\item{rcode}{(\code{character})
vector with \code{R} code to show inside the modal.}

\item{session}{(\code{ShinySession} optional)
\code{shiny} session object, if missing then \code{\link[shiny:domains]{shiny::getDefaultReactiveDomain()}} is used.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Use the \code{\link[shiny:showModal]{shiny::showModal()}} function to show the \code{R} code inside.
}
\references{
\code{\link[shiny:showModal]{shiny::showModal()}}
}
