% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teal_slices-store.R
\name{slices_store}
\alias{slices_store}
\title{Store teal_slices object to a file}
\usage{
slices_store(tss, file)
}
\arguments{
\item{tss}{(\code{teal_slices}) object to be stored.}

\item{file}{(\code{character(1)}) The file path where \code{teal_slices} object will be saved.
The file extension should be \code{".json"}.}
}
\value{
\code{NULL}, invisibly.
}
\description{
This function takes a \code{teal_slices} object and saves it to a file in \code{JSON} format.
The \code{teal_slices} object contains information about filter states and can be used to
create, modify, and delete filter states. The saved file can be later loaded using
the \code{slices_restore} function.
}
\details{
\code{Date} class is stored in \code{"ISO8601"} format (\code{YYYY-MM-DD}). \code{POSIX*t} classes are converted to a
character by using \code{format.POSIX*t(usetz = TRUE, tz = "UTC")} (\verb{YYYY-MM-DD \{N\}\{N\}:\{N\}\{N\}:\{N\}\{N\} UTC}, where
\verb{\{N\} = [0-9]} is a number and \code{UTC} is \verb{Coordinated Universal Time} timezone short-code).
This format is assumed during \code{slices_restore}. All \code{POSIX*t} objects in \code{selected} or \code{choices} fields of
\code{teal_slice} objects are always printed in \code{UTC} timezone as well.
}
\keyword{internal}
