% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TealDataConnection_constructors.R
\name{snowflake_connection}
\alias{snowflake_connection}
\title{Superseded (moved into separate package): Open connection to \code{Snowflake}}
\usage{
snowflake_connection(
  open_args = list(),
  close_args = list(),
  ping_args = list()
)
}
\arguments{
\item{open_args}{optional, named (\code{list}) of additional parameters for the connection's
\code{snowflake_connection_function} open function.}

\item{close_args}{optional, named (\code{list}) of additional parameters for the connection's
\code{\link[DBI:dbDisconnect]{DBI::dbDisconnect()}} close function.}

\item{ping_args}{optional, named (\code{list}) of additional parameters for the connection's
\code{\link[DBI:dbIsValid]{DBI::dbIsValid()}} ping function.}
}
\value{
(\code{TealDataConnection}) type of object
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#superseded}{\figure{lifecycle-superseded.svg}{options: alt='[Superseded]'}}}{\strong{[Superseded]}}
}
