% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataConnection.R
\name{snowflake_connection}
\alias{snowflake_connection}
\title{Open connection to \code{Snowflake}}
\usage{
snowflake_connection(
  open_args = list(),
  close_args = list(),
  ping_args = list()
)
}
\arguments{
\item{open_args}{optional, named (\code{list}) of additional parameters for the connection's
\code{teal::snowflake_connection_function} open function.}

\item{close_args}{optional, named (\code{list}) of additional parameters for the connection's
\code{\link[DBI]{dbDisconnect}} close function.}

\item{ping_args}{optional, named (\code{list}) of additional parameters for the connection's
\code{\link[DBI]{dbIsValid}} ping function.}
}
\value{
(\code{DataConnection}) type of object
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#defunct}{\figure{lifecycle-defunct.svg}{options: alt='[Defunct]'}}}{\strong{[Defunct]}}
}
