% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TealDataConnection.R
\name{snowflake_connection_function}
\alias{snowflake_connection_function}
\title{Superseded (moved into separate package): Helper function to connect to \code{Snowflake}}
\usage{
snowflake_connection_function(
  username = askpass::askpass("Please enter your username"),
  password = askpass::askpass("Please enter your password"),
  role,
  database,
  schema,
  warehouse,
  server,
  port = 443,
  driver = "SnowflakeDSIIDriver",
  token_provider
)
}
\arguments{
\item{username}{the username used to collect the auth token to connect to snowflake.}

\item{password}{the password used to collect the auth token to connect to snowflake}

\item{role}{the user role used to connect to \code{Snowflake}.}

\item{database}{the \code{Snowflake} database to connect to.}

\item{schema}{the \code{Snowflake} schema to connect to.}

\item{warehouse}{the \code{Snowflake} warehouse to connect to.}

\item{server}{the \code{Snowflake}server to connect to.}

\item{port}{the port to connect to the \code{Snowflake} instance.}

\item{driver}{the driver to use to connect to the \code{Snowflake} instance.}

\item{token_provider}{location of the auth token provider needed to access \code{Snowflake}.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#superseded}{\figure{lifecycle-superseded.svg}{options: alt='[Superseded]'}}}{\strong{[Superseded]}}

This is used by \code{snowflake_connection} and does not need to be called directly
}
