% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TealDataConnector_constructors.R
\name{snowflake_data}
\alias{snowflake_data}
\title{Superseded (moved into separate package): \code{TealDataConnector} connector for \code{SNOWFLAKE}}
\usage{
snowflake_data(..., connection)
}
\arguments{
\item{...}{(\code{TealDatasetConnector}) dataset connectors created using \link{snowflake_dataset_connector}}

\item{connection}{(\code{TealDataConnection}) object returned from \code{snowflake_connection}.}
}
\value{
An object of class \code{TealDataConnector}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#superseded}{\figure{lifecycle-superseded.svg}{options: alt='[Superseded]'}}}{\strong{[Superseded]}}
Build data connector for \code{SNOWFLAKE} functions or datasets
}
\details{
Note the server location and token_provider must be provided as arguments to
the snowflake_connection function, see example below.
}
