% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{split_by_sep}
\alias{split_by_sep}
\title{Split by sepearator}
\usage{
split_by_sep(x, sep)
}
\arguments{
\item{x}{(\code{character}) Character (single)}

\item{sep}{(\code{character}) Separator}
}
\description{
Split by sepearator
}
