% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_filter_panel.R
\name{srv_filter_panel}
\alias{srv_filter_panel}
\title{Server function for filter panel}
\usage{
srv_filter_panel(
  input,
  output,
  session,
  datasets,
  active_datanames = function() "all"
)
}
\arguments{
\item{input}{(\verb{Shiny input object})}

\item{output}{(\verb{Shiny output object})}

\item{session}{(\verb{Shiny session object})}

\item{datasets}{(\code{FilteredData})\cr
object to store filter state and filtered datasets, shared across modules}

\item{active_datanames}{\verb{function / reactive} returning datanames that
should be shown on the filter panel,
must be a subset of the \code{datanames} argument provided to \code{ui_filter_panel};
if the function returns \code{NULL} (as opposed to \code{character(0)}), the filter
panel will be hidden}
}
\description{
Server function for filter panel
}
