% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_set_bookmark_state.R
\name{srv_init_filter_state}
\alias{srv_init_filter_state}
\title{Server module to set initial filter state}
\usage{
srv_init_filter_state(id, datasets, filter)
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the module's
UI function.}

\item{datasets}{(\code{FilteredData})\cr
object to store filter state and filtered datasets, shared across modules. For more
details see \code{\link{FilteredData}}.}

\item{filter}{(\code{list})\cr
You can define filters that show when the app starts.
The general pattern is:
\code{list(iris = list(Species = ..., Sepal.Length = ...), mtcars = ...)}.
An example is:
\code{list(iris = list(Species = c("setosa", "versicolor")))}.
More generally, the filters for the variable, e.g. \code{Species} can be
specified as follows:
\code{list(Species = c("setosa", "versicolor"))},
\code{list(Species = list(c("setosa", "versicolor"), keep_na = TRUE))},
or equivalently with:
\code{list(Species = c("setosa", "versicolor", NA))},
or for the default filter (not very restrictive):
\code{list(Species = default_filter())}

Instead of \code{choices} above, use the following names:
\itemize{
\item \code{numerical}: \code{range}
\item \code{factor}: \code{choices}
\item \code{logical}: \code{logical}
A general example is:
\code{list( iris = list(Sepal.Length = default_filter(), Species = c("M", NA)), mtcars = list(mpg = default_filter())   )}
Ignored if the app is restored from a bookmarked state.
}}
}
\value{
\code{shiny} module which returns \code{NULL}
}
\description{
Initial filter state is set to:
\enumerate{
\item state defined by developer in the \code{filter} argument of \code{init}
\item When app is restored from with bookmark state, then filter-state from
the URL attributes is set. Bookmark state has a priority over \code{filter}
argument.
}
}
\details{
The Shiny bookmarking functionality by default only stores inputs.
We need to add \code{FilteredData} object to the state so we restore it as well.
To test bookmarking, include the \code{bookmark_module}, click on the bookmark
button and then get the link. Keep the Shiny app running and open the
obtained link in another browser tab.
}
\keyword{internal}
