% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_nested_tabs.R
\name{srv_nested_tabs}
\alias{srv_nested_tabs}
\title{Server function that returns currently active module}
\usage{
srv_nested_tabs(id, datasets, modules)
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the module's
UI function.}

\item{datasets}{(\code{FilteredData})\cr
object to store filter state and filtered datasets, shared across modules}

\item{modules}{(\code{teal_module} or \code{teal_modules})
single or multiple modules. \code{teal_modules}  can be used for nested
tabs, see \code{ui_nested_tabs}}
}
\value{
\code{reactive} which returns the active module that corresponds to the selected tab
}
\description{
Server function that returns currently active module
}
