% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{srv_shiny_module_arguments}
\alias{srv_shiny_module_arguments}
\title{Function to inherit Shiny module arguments that must always be present}
\usage{
srv_shiny_module_arguments(input, output, session, datasets, modules)
}
\arguments{
\item{input}{(\verb{Shiny input object})}

\item{output}{(\verb{Shiny output object})}

\item{session}{(\verb{Shiny session object})}

\item{datasets}{(\code{FilteredData})\cr
object to store filter state and filtered datasets, shared across modules}

\item{modules}{(\code{teal_module} or \code{teal_modules})
single or multiple modules. \code{teal_modules}  can be used for nested
tabs, see \link{ui_nested_tabs}}
}
\description{
Function to inherit Shiny module arguments that must always be present
}
