% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_tabs_with_filters.R
\name{srv_tabs_with_filters}
\alias{srv_tabs_with_filters}
\title{Server function}
\usage{
srv_tabs_with_filters(input, output, session, datasets, modules)
}
\arguments{
\item{input}{\verb{Shiny input object}}

\item{output}{\verb{Shiny output object}}

\item{session}{\verb{Shiny session object}}

\item{datasets}{\code{FilteredData} object to store filter state and filtered
datasets, shared across modules}

\item{modules}{\code{teal_module} or \code{teal_modules} object, the latter
can be used for nested tabs, see \verb{\link{ui_nested_tabs}}}
}
\value{
\code{reactive} currently selected active_module
}
\description{
Server function
}
