% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_tabs_with_filters.R
\name{srv_tabs_with_filters}
\alias{srv_tabs_with_filters}
\title{Server function}
\usage{
srv_tabs_with_filters(id, datasets, modules)
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the module's
UI function.}

\item{datasets}{(\code{FilteredData})\cr
object to store filter state and filtered datasets, shared across modules}

\item{modules}{(\code{teal_module} or \code{teal_modules})
single or multiple modules. \code{teal_modules}  can be used for nested
tabs, see \code{ui_nested_tabs}}
}
\value{
\code{reactive} currently selected active_module
}
\description{
Server function
}
