% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_teal_with_splash.R
\name{srv_teal_with_splash}
\alias{srv_teal_with_splash}
\title{Server function that loads the data through reactive loading and then delegates
to \code{\link{srv_teal}}.}
\usage{
srv_teal_with_splash(id, data, modules, filter = list())
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the module's
UI function.}

\item{data}{\code{DataAbstract} R6 object and container for data}

\item{modules}{(\code{teal_module} or \code{teal_modules})
single or multiple modules. \code{teal_modules}  can be used for nested
tabs, see \code{ui_nested_tabs}}

\item{filter}{(\code{list}) You can define filters that show when
the app starts.
The general pattern is:
\code{list(iris = list(Species = ..., Sepal.Length = ...), mtcars = ...)}.
An example is:
\code{list(iris = list(Species = c("setosa", "versicolor")))}.
More generally, the filters for the variable, e.g. \code{Species} can be
specified as follows:
\code{list(Species = c("setosa", "versicolor"))},
\code{list(Species = list(c("setosa", "versicolor"), keep_na = TRUE))},
or equivalently with:
\code{list(Species = c("setosa", "versicolor", NA))},
or for the default filter (not very restrictive):
\code{list(Species = default_filter())}

Instead of \code{choices} above, use the following names:
\itemize{
\item \code{numerical}: \code{range}
\item \code{factor}: \code{choices}
\item \code{logical}: \code{logical}
A general example is:
\code{list( iris = list(Sepal.Length = default_filter(), Species = c("M", NA)), mtcars = list(mpg = default_filter())   )}
Ignored if the app is restored from a bookmarked state.
}}
}
\value{
\code{reactive}, return value of \code{\link{srv_teal}}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#maturing}{\figure{lifecycle-maturing.svg}{options: alt='[Maturing]'}}}{\strong{[Maturing]}}
Please also refer to the doc of \code{\link{init}}.
}
