% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_teal_with_splash.R
\name{srv_teal_with_splash}
\alias{srv_teal_with_splash}
\title{\code{server} function that loads the data through reactive loading and then delegates
to \code{\link[=srv_teal]{srv_teal()}}.}
\usage{
srv_teal_with_splash(id, data, modules, filter = teal_slices())
}
\arguments{
\item{id}{(\code{character})\cr
module id to embed it, if provided,
the \code{server} function must be called with \code{\link[shiny:moduleServer]{shiny::moduleServer()}};
See the vignette for an example. However, \code{\link[=ui_teal_with_splash]{ui_teal_with_splash()}}
is then preferred to this function.}

\item{data}{(\code{teal_data}, \code{teal_data_module}, \verb{named list})\cr
\code{teal_data} object as returned by \code{\link[teal.data:teal_data]{teal.data::teal_data()}} or
\code{teal_data_module} or simply a list of a named list of objects
(\code{data.frame} or \code{MultiAssayExperiment}).}

\item{modules}{\code{teal_modules} object containing the output modules which
will be displayed in the \code{teal} application. See \code{\link[=modules]{modules()}} and \code{\link[=module]{module()}} for
more details.}

\item{filter}{(\code{teal_slices})\cr
Specification of initial filter. Filters can be specified using \code{\link[=teal_slices]{teal_slices()}}.
Old way of specifying filters through a list is deprecated and will be removed in the
next release. Please fix your applications to use \code{\link[=teal_slices]{teal_slices()}}.}
}
\value{
\code{reactive} containing \code{teal_data} object when data is loaded.
If data is not loaded yet, \code{reactive} returns \code{NULL}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
Please also refer to the doc of \code{\link[=init]{init()}}.
}
