% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_layout.R
\name{standard_layout}
\alias{standard_layout}
\title{Create a standard ui layout with output on the right and an encoding panel on
the left}
\usage{
standard_layout(output, encoding = NULL, forms = NULL, pre_output = NULL,
  post_output = NULL)
}
\description{
This is the layout used for the teal modules in the \code{beam} package
}
