% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-module.R
\name{submit_button_module}
\alias{submit_button_module}
\alias{submit_button_ui}
\alias{submit_button_server}
\title{Run code and mask inputs}
\usage{
submit_button_ui(id)

submit_button_server(id, ...)
}
\arguments{
\item{id}{(\code{character}) \code{shiny} module id.}

\item{...}{(\code{list}) arguments passed to \code{\link[=eval_and_mask]{eval_and_mask()}}.}
}
\value{
\code{shiny} module
}
\description{
Delayed Data Loading module with login and password input.
}
