% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tdata.R
\name{tdata2env}
\alias{tdata2env}
\title{Function to convert a \code{tdata} object to an \code{environment}
Any \code{reactives} inside \code{tdata} are first evaluated}
\usage{
tdata2env(data)
}
\arguments{
\item{data}{a \code{tdata} object}
}
\value{
an \code{environment}
}
\description{
Function to convert a \code{tdata} object to an \code{environment}
Any \code{reactives} inside \code{tdata} are first evaluated
}
\examples{

data <- new_tdata(
  data = list(iris = iris, mtcars = reactive(mtcars)),
  code = "iris <- iris
    mtcars = mtcars"
)

my_env <- isolate(tdata2env(data))

}
