% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teal.R
\docType{package}
\name{teal-package}
\alias{teal}
\alias{teal-package}
\title{teal: Interactive Exploration of Clinical Trials Data}
\description{
The teal package provides a shiny based framework for creating an
interactive data analysis environment.
}
\details{
To learn mode about the package either read the project website at
\url{Project Website} or read the \code{\link{init}} manual pages.
}
\author{
\strong{Maintainer}: NEST \email{basel.nestcicd@roche.com}

Authors:
\itemize{
  \item Adrian Waddell \email{adrian.waddell@roche.com}
  \item Pawel Rucki \email{pawel.rucki@contractors.roche.com}
  \item Dawid Kaledkowski \email{dawid.kaledkowski@contractors.roche.com}
  \item Sebastian Wolf \email{sebastian.wolf.sw1@roche.com}
  \item Maximilian Mordig \email{maximilian_oliver.mordig@roche.com}
  \item Bartosz Baranowski \email{bartosz.baranowski@roche.com}
  \item Ondrej Slama \email{ondrej.slama@roche.com}
  \item Nikolas Burkoff \email{nikolas.burkoff@roche.com}
  \item Maciej Nasinski \email{maciej.nasinski@contractors.roche.com}
}

Other contributors:
\itemize{
  \item Tadeusz Lewandowski \email{tadeusz.lewandowski@roche.com} [contributor]
  \item Roche [copyright holder, funder]
}

}
\keyword{internal}
