% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teal.R
\docType{package}
\name{teal-package}
\alias{teal}
\alias{teal-package}
\title{teal: Interactive Exploration of Clinical Trials Data}
\description{
The teal package provides a shiny based framework for creating an
interactive data analysis environment.
}
\details{
To learn mode about the package either read the project website at
\url{Project Website} or read the \code{\link{init}} manual pages.
}
\author{
\strong{Maintainer}: Dawid Kaledkowski \email{dawid.kaledkowski@roche.com}

Authors:
\itemize{
  \item Pawel Rucki \email{pawel.rucki@roche.com}
  \item Nikolas Burkoff \email{nikolas.burkoff@roche.com}
  \item Mahmoud Hallal \email{mahmoud.hallal@roche.com}
  \item Maciej Nasinski \email{maciej.nasinski@contractors.roche.com}
  \item Konrad Pagacz \email{konrad.pagacz@contractors.roche.com}
  \item Junlue Zhao \email{zhaoj88@gene.com}
  \item Adrian Waddell \email{adrian.waddell@gene.com}
}

Other contributors:
\itemize{
  \item F. Hoffmann-La Roche AG [copyright holder, funder]
  \item Maximilian Mordig [contributor]
}

}
\keyword{internal}
