% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teal-package.R
\docType{package}
\name{teal}
\alias{teal}
\alias{teal-package}
\title{teal: Interactive Exploration of Clinical Trials Data}
\description{
The teal package provides a shiny based framework for creating an
interactinve data analysis environment.
}
