% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teal_data.R
\name{teal_data}
\alias{teal_data}
\title{Teal data}
\usage{
teal_data(..., join_keys, code = "", check = FALSE)
}
\arguments{
\item{...}{(\code{TealDataConnector}, \code{TealDataset}, \code{TealDatasetConnector})\cr
objects}

\item{join_keys}{(\code{JoinKeys}) or a single (\code{JoinKeySet})\cr
(optional) object with dataset column relationships used for joining.
If empty then no joins between pairs of objects}

\item{code}{(\code{character}) code to reproduce the datasets.}

\item{check}{(\code{logical}) reproducibility check - whether evaluated preprocessing code gives the same objects
as provided in arguments. Check is run only if flag is true and preprocessing code is not empty.}
}
\value{
(\code{TealData})
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
Universal function to pass data to teal application
}
\examples{
x1 <- dataset(
  "x1",
  iris,
  code = "x1 <- iris"
)

x2 <- dataset(
  "x2",
  mtcars,
  code = "x2 <- mtcars"
)

teal_data(x1, x2)
}
