% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teal_data.R
\name{teal_data_file}
\alias{teal_data_file}
\title{Load \code{RelationalData} object from a file}
\usage{
teal_data_file(x)
}
\arguments{
\item{x}{A (\code{connection}) or a (\code{character}) string giving the pathname
of the file or URL to read from. "" indicates the connection \code{stdin}.}
}
\value{
\code{RelationalData} object if file returns a \code{RelationalData}
  object.
}
\description{
Load \code{RelationalData} object from a file
}
\examples{
file_example <- tempfile(fileext = ".R")
writeLines(
  text = c(
    "library(teal)
     library(random.cdisc.data)
     library(dplyr)

     adsl <- cdisc_dataset(dataname = \"ADSL\", # RelationalDataset
                           data = radsl(cached = TRUE),
                           code = \"library(random.cdisc.data)\nADSL <- radsl(cached = TRUE)\")

     adtte <- cdisc_dataset(dataname = \"ADTTE\", # RelationalDataset
                            data = radtte(cached = TRUE),
                            code = \"library(random.cdisc.data)\nADTTE <- radtte(cached = TRUE)\")

     teal_data(adsl, adtte)"
  ),
  con = file_example
)

teal_data_file(file_example)

}
