% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teal_data.R
\name{teal_data_file}
\alias{teal_data_file}
\title{Load \code{RelationalData} object from a file}
\usage{
teal_data_file(x, code = get_code(x))
}
\arguments{
\item{x}{(\code{character}) string giving the pathname of the file to read from.}

\item{code}{(\code{character}) reproducible code to re-create object}
}
\value{
\code{RelationalData} object
}
\description{
Please note that the script has to end with a call creating desired object. The error will be raised otherwise.
}
\examples{
# simple example
file_example <- tempfile(fileext = ".R")
writeLines(
  text = c(
    "library(teal)
     library(random.cdisc.data)

     adsl <- cdisc_dataset(dataname = \"ADSL\",
                           data = radsl(cached = TRUE),
                           code = \"library(random.cdisc.data)\nADSL <- radsl(cached = TRUE)\")

     adtte <- cdisc_dataset(dataname = \"ADTTE\",
                            data = radtte(cached = TRUE),
                            code = \"library(random.cdisc.data)\nADTTE <- radtte(cached = TRUE)\")

     cdisc_data(adsl, adtte)"
  ),
  con = file_example
)
x <- teal_data_file(file_example, code = character(0))
get_code(x)

# exaample with custom datasets
file_example <- tempfile(fileext = ".R")
writeLines(
  text = c(
    "library(teal)
    library(random.cdisc.data)

    # code>
    ADSL <- radsl(cached = TRUE)
    ADTTE <- radtte(cached = TRUE)

    ADSL$n <- nrow(ADTTE)
    ADTTE$n <- nrow(ADSL)
    # <code

    ADSL <- cdisc_dataset(dataname = \"ADSL\", data = ADSL)
    ADTTE <- cdisc_dataset(dataname = \"ADTTE\", data = ADTTE)

    cdisc_data(ADSL, ADTTE)"
  ),
  con = file_example
)
x <- teal_data_file(file_example)
get_code(x)
}
