% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{teal_data_to_filtered_data}
\alias{teal_data_to_filtered_data}
\title{Create a \code{FilteredData}}
\usage{
teal_data_to_filtered_data(x, datanames = names(x))
}
\arguments{
\item{x}{(\code{teal_data}) object}

\item{datanames}{(\code{character}) vector of data set names to include; must be subset of \code{names(x)}}
}
\value{
A \code{FilteredData} object.
}
\description{
Create a \code{FilteredData} object from a \code{teal_data} object.
}
\keyword{internal}
