% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teal_data_utils.R
\name{teal_data_utilities}
\alias{teal_data_utilities}
\alias{.append_evaluated_code}
\alias{.append_modified_data}
\title{\code{teal_data} utils}
\usage{
.append_evaluated_code(data, code)

.append_modified_data(data, objects)
}
\arguments{
\item{data}{(\code{teal_data})}

\item{code}{(\code{character}) code to append to the object's code slot.}

\item{objects}{(\code{list}) objects to append to object's environment.}
}
\value{
modified \code{teal_data}
}
\description{
In \code{teal} we need to recreate the \code{teal_data} object due to two operations:
\itemize{
\item we need to append filter-data code and objects which have been evaluated in \code{FilteredData} and
we want to avoid double-evaluation.
\item we need to subset \code{teal_data} to \code{datanames} used by the module, to shorten obtainable R-code
}
}
\details{
Due to above recreation of \code{teal_data} object can't be done simply by using public
\code{teal.code} and \code{teal.data} methods.
}
\keyword{internal}
