% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teal_modifiers.R
\name{teal_extend_server}
\alias{teal_extend_server}
\title{Add a Custom Server Logic to \code{teal} Application}
\usage{
teal_extend_server(x, custom_server, module_id = character(0))
}
\arguments{
\item{x}{(\code{teal_app}) A \code{teal_app} object created using the \code{init} function.}

\item{custom_server}{(\verb{function(input, output, session)} or \verb{function(id, ...)})
The custom server function or server module to set.}

\item{module_id}{(\code{character(1)}) The ID of the module when a module server function is passed.}
}
\description{
Adds a custom server function to the \code{teal} app. This function can define additional server logic.
}
\keyword{internal}
