% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teal_modifiers.R
\name{teal_modifiers}
\alias{teal_modifiers}
\alias{modify_title}
\alias{modify_header}
\alias{modify_footer}
\title{Replace UI Elements in \code{teal} UI objects}
\usage{
modify_title(
  x,
  title = "teal app",
  favicon =
    "https://raw.githubusercontent.com/insightsengineering/hex-stickers/main/PNG/nest.png"
)

modify_header(x, element = tags$p())

modify_footer(x, element = tags$p())
}
\arguments{
\item{x}{One of:
\itemize{
\item List containing \code{ui} and \code{server} components (teal app/module)
\item Function with \code{request} parameter (teal UI function)
\item Function with \code{id} parameter (module UI function)
}}

\item{title}{(\code{shiny.tag} or \code{character(1)}) The new title to be used.}

\item{favicon}{(\code{character}) The path for the icon for the title.
The image/icon path can be remote or the static path accessible by \code{shiny}, like the \verb{www/}}

\item{element}{Replacement UI element (shiny tag or HTML)}

\item{selector}{(\code{character(1)}) CSS selector to find elements to replace}
}
\description{
Replace UI Elements in \code{teal} UI objects

Add a custom Title to \code{teal} application

Add a Header to \code{teal} Application

Add a Footer to \code{teal} Application
}
\examples{
app <- init(
  data = teal_data(IRIS = iris, MTCARS = mtcars),
  modules = modules(example_module())
) |>
  modify_title(title = "Custom title")

if (interactive()) {
  shinyApp(app$ui, app$server)
}
app <- init(
  data = teal_data(IRIS = iris),
  modules = modules(example_module())
) |>
  modify_header(element = tags$div(h3("Custom header")))

if (interactive()) {
  shinyApp(app$ui, app$server)
}
app <- init(
  data = teal_data(IRIS = iris),
  modules = modules(example_module())
) |>
  modify_footer(element = "Custom footer")

if (interactive()) {
  shinyApp(app$ui, app$server)
}
}
\keyword{internal}
