% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataConnection.R
\name{teradata_connection}
\alias{teradata_connection}
\title{Open connection to \code{Teradata}}
\usage{
teradata_connection(
  open_args = list(),
  close_args = list(),
  ping_args = list()
)
}
\arguments{
\item{open_args}{optional, named (\code{list}) of additional parameters for the connection's
\code{RocheTeradata::connect_teradata} open function. Please note that the \code{type}
argument will be overwritten with \code{ODBC}.}

\item{close_args}{optional, named (\code{list}) of additional parameters for the connection's
\code{\link[DBI]{dbDisconnect}} close function.}

\item{ping_args}{optional, named (\code{list}) of additional parameters for the connection's
\code{\link[DBI]{dbIsValid}} ping function.}
}
\value{
\code{DataConnection} type of object
}
\description{
\ifelse{html}{\out{<a href='https://www.tidyverse.org/lifecycle/#experimental'><img src='figures/lifecycle-experimental.svg' alt='Experimental lifecycle'></a>}}{\strong{Experimental}}
}
