% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TealDataConnector_constructors.R
\name{teradata_data}
\alias{teradata_data}
\title{Superseded (moved into separate package): \code{TealDataConnector} connector for \code{TERADATA}}
\usage{
teradata_data(..., connection = teradata_connection())
}
\arguments{
\item{...}{(\code{TealDatasetConnector}) dataset connectors created using \link{teradata_dataset_connector}}

\item{connection}{(\code{TealDataConnection}) object returned from \code{teradata_connection}.}
}
\value{
An object of class \code{TealDataConnector}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#superseded}{\figure{lifecycle-superseded.svg}{options: alt='[Superseded]'}}}{\strong{[Superseded]}}
Build data connector for \code{TERADATA} functions or datasets
}
