% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RelationalDataConnector_constructors.R
\name{teradata_data}
\alias{teradata_data}
\title{\code{RelationalDataConnector} connector for \code{TERADATA}}
\usage{
teradata_data(..., connection = teradata_connection())
}
\arguments{
\item{...}{(\code{DatasetConnector}) dataset connectors created using \code{\link{teradata_dataset_connector}}}

\item{connection}{(\code{DataConnection}) object returned from \code{teradata_connection}.}
}
\value{
An object of class \code{RelationalDataConnector}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#defunct}{\figure{lifecycle-defunct.svg}{options: alt='[Defunct]'}}}{\strong{[Defunct]}}
Build data connector for \code{TERADATA} functions or datasets
}
