% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_data_table.R
\name{tm_data_table}
\alias{tm_data_table}
\title{Data Table Viewer Teal Module}
\usage{
tm_data_table(label = "data table", variables_selected = NULL)
}
\arguments{
\item{label}{label shown in the navigation for the item}

\item{variables_selected}{a named list that says which variables should be
initially  shown for which dataset}
}
\description{
A data table viewer shows the data using a paginated table.
}
