% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{to_relational_data}
\alias{to_relational_data}
\title{S3 generic for \code{to_relational_data} function.}
\usage{
to_relational_data(data, data_call = NULL)
}
\arguments{
\item{data}{\code{RelationalData}, \code{Dataset}, \code{data.frame}, \code{list} or \code{function} returning a named list.}
}
\value{
list of \code{RelationalData} objects
}
\description{
Takes the input of data argument and translates them to relational data objects.
}
