% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{toString.teal_module}
\alias{toString.teal_module}
\title{Convert `teal_module` to a string}
\usage{
\method{toString}{teal_module}(x, indent = 0, ...)
}
\arguments{
\item{x}{`teal_module`}

\item{indent}{\code{integer} indent level;
each submodule is indented one level more}

\item{...}{ignored}
}
\description{
Convert `teal_module` to a string
}
