% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{toString.teal_modules}
\alias{toString.teal_modules}
\title{Convert \code{teal_modules} to a string}
\usage{
\method{toString}{teal_modules}(x, indent = 0, ...)
}
\arguments{
\item{x}{(\code{teal_modules}) to print}

\item{indent}{(\code{integer}) indent level;
each submodule is indented one level more}

\item{...}{(optional) additional parameters to pass to recursive calls of \code{toString}}
}
\value{
(\code{character} value)
}
\description{
The first line prints the \code{modules} label.
The consecutive lines recursively list each submodule.
}
