% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_teal_with_splash.R
\name{ui_teal_with_splash}
\alias{ui_teal_with_splash}
\title{UI to show a splash screen in the beginning, then delegate to \verb{\link{srv_teal}}}
\usage{
ui_teal_with_splash(
  id,
  data,
  title,
  header = tags$p("Add Title Here"),
  footer = tags$p("Add Footer Here")
)
}
\arguments{
\item{id}{module id}

\item{data}{\code{RealtionalDataList} R6 object and container for data}

\item{title}{(\code{character}) The browser window title (defaults to the host URL of the page)}

\item{header}{\verb{shiny.tag or character} header to display above the app}

\item{footer}{\verb{shiny.tag or character} footer to display below the app}
}
\description{
The splash screen could be used to query for a password to fetch the data.
\verb{\link{init}} is a very thin wrapper around this module useful for end-users which
assumes that it is a top-level module and cannot be embedded.
This function instead adheres to the Shiny module conventions.
}
\details{
If data is obtained through delayed loading, its splash screen is used. Otherwise,
a default splash screen is shown.

Please also refer to the doc of \verb{\link{init}}.
}
