% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_teal_with_splash.R
\name{ui_teal_with_splash}
\alias{ui_teal_with_splash}
\title{UI to show a splash screen in the beginning, then delegate to \code{\link{srv_teal}}}
\usage{
ui_teal_with_splash(
  id,
  data,
  title,
  header = tags$p("Add Title Here"),
  footer = tags$p("Add Footer Here")
)
}
\arguments{
\item{id}{(\code{character(1)})\cr
module id}

\item{data}{(\code{TealData} or \code{TealDataset} or \code{TealDatasetConnector} or \code{list} or \code{data.frame})\cr
\code{R6} object as returned by \code{\link[=cdisc_data]{cdisc_data()}}, \code{\link[=teal_data]{teal_data()}}, \code{\link[=cdisc_dataset]{cdisc_dataset()}}, \code{\link[=dataset]{dataset()}},
\code{\link[=dataset_connector]{dataset_connector()}} or \code{\link[=cdisc_dataset_connector]{cdisc_dataset_connector()}} or a single \code{data.frame} or a list of
the previous objects or function returning a named list.
NOTE: teal does not guarantee reproducibility of the code when names of the list elements
do not match the original object names. To ensure reproducibility please use \code{\link[=teal_data]{teal_data()}}
or \code{\link[=cdisc_data]{cdisc_data()}} with \code{check = TRUE} enabled.}

\item{title}{(\code{NULL} or \code{character})\cr
The browser window title (defaults to the host URL of the page).}

\item{header}{(\code{character} or \code{shiny.tag}) \cr
the header of the app. Note shiny code placed here (and in the footer
argument) will be placed in the app's \code{ui} function so code which needs to be placed in the \code{ui} function
(such as loading css via \code{\link[htmltools:htmlDependency]{htmltools::htmlDependency()}}) should be included here.}

\item{footer}{(\code{character} or \code{shiny.tag})\cr
the footer of the app}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#maturing}{\figure{lifecycle-maturing.svg}{options: alt='[Maturing]'}}}{\strong{[Maturing]}}
The splash screen could be used to query for a password to fetch the data.
\code{\link{init}} is a very thin wrapper around this module useful for end-users which
assumes that it is a top-level module and cannot be embedded.
This function instead adheres to the Shiny module conventions.

If data is obtained through delayed loading, its splash screen is used. Otherwise,
a default splash screen is shown.

Please also refer to the doc of \code{\link[=init]{init()}}.
}
