% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_teal_with_splash.R
\name{ui_teal_with_splash}
\alias{ui_teal_with_splash}
\title{UI to show a splash screen in the beginning, then delegate to \code{\link[=srv_teal]{srv_teal()}}}
\usage{
ui_teal_with_splash(
  id,
  data,
  title,
  header = tags$p("Add Title Here"),
  footer = tags$p("Add Footer Here")
)
}
\arguments{
\item{id}{(\code{character(1)})\cr
module id}

\item{data}{(\code{teal_data}, \code{teal_data_module}, \verb{named list})\cr
\code{teal_data} object as returned by \code{\link[teal.data:teal_data]{teal.data::teal_data()}} or
\code{teal_data_module} or simply a list of a named list of objects
(\code{data.frame} or \code{MultiAssayExperiment}).}

\item{title}{(\code{shiny.tag} or \code{character})\cr
The browser window title. Defaults to a title "Teal app" with the icon of NEST.
Can be created using the \code{build_app_title()} or
by passing a valid \code{shiny.tag} which is a head tag with title and link tag.}

\item{header}{(\code{shiny.tag} or \code{character}) \cr
the header of the app. Note shiny code placed here (and in the footer
argument) will be placed in the app's \code{ui} function so code which needs to be placed in the \code{ui} function
(such as loading \code{CSS} via \code{\link[htmltools:htmlDependency]{htmltools::htmlDependency()}}) should be included here.}

\item{footer}{(\code{shiny.tag} or \code{character})\cr
the footer of the app}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
The splash screen could be used to query for a password to fetch the data.
\code{\link[=init]{init()}} is a very thin wrapper around this module useful for end-users which
assumes that it is a top-level module and cannot be embedded.
This function instead adheres to the Shiny module conventions.

If data is obtained through delayed loading, its splash screen is used. Otherwise,
a default splash screen is shown.

Please also refer to the doc of \code{\link[=init]{init()}}.
}
