% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validations.R
\name{validate_has_elements}
\alias{validate_has_elements}
\title{Validates that vector has length greater than 0}
\usage{
validate_has_elements(x, msg)
}
\arguments{
\item{x}{vector}

\item{msg}{message to display}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
}
\details{
This function is a wrapper for \code{shiny::validate}.
}
\examples{
library(scda)
ADSL <- synthetic_cdisc_data("latest")$adsl

ui <- fluidPage(
  selectInput("ref_arm", "Select reference treatment",
    choices = c("ARM A", "ARM B", "ARM X"), selected = "ARM A"
  ),
  selectInput("comp_arm", "Select comparison treatment",
    choices = c("ARM C", "ARM Y", "ARM Z"), selected = "ARM C"
  ),
  verbatimTextOutput("arm_summary")
)

server <- function(input, output) {
  output$arm_summary <- renderText({
    ref_arm <- ADSL$ARMCD[input$ref_arm == ADSL$ARMCD]
    comp_arm <- ADSL$ARMCD[input$comp_arm == ADSL$ARMCD]

    validate_has_elements(ref_arm, "Need reference treatment.")
    validate_has_elements(comp_arm, "Need comparison treatment.")

    paste0(
      "Number of patients in: reference treatment=",
      length(ref_arm), " comparions treatment=", length(comp_arm)
    )
  })
}
\dontrun{
shinyApp(ui, server)
}
}
