% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validations.R
\name{validate_has_variable}
\alias{validate_has_variable}
\title{Validates that dataset contains specific variable}
\usage{
validate_has_variable(data, varname, msg)
}
\arguments{
\item{data}{a data.frame}

\item{varname}{name of variable in \code{data}}

\item{msg}{message to display if \code{data} does not include \code{varname}}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
}
\details{
This function is a wrapper for \code{shiny::validate}.
}
\examples{
library(scda)
ADSL <- synthetic_cdisc_data("latest")$adsl

ui <- fluidPage(
  selectInput("arm", "Select treatment",
    choices = c("ARM", "ARMCD", "ACTARM", "ACTARMCD", "TRT01P", "TRT01A"),
    selected = "ARM", multiple = TRUE
  ),
  verbatimTextOutput("arm_summary")
)

server <- function(input, output) {
  output$arm_summary <- renderText({
    validate_has_variable(ADSL, input$arm)

    paste0("Selected treatment variables: ", paste(input$arm, collapse = ", "))
  })
}
\dontrun{
shinyApp(ui, server)
}
}
