% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validations.R
\name{validate_in}
\alias{validate_in}
\title{Validates that vector includes all expected values}
\usage{
validate_in(x, choices, msg)
}
\arguments{
\item{x}{values to test. All must be in \code{choices}}

\item{choices}{a vector to test for values of \code{x}}

\item{msg}{warning message to display}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
}
\details{
This function is a wrapper for \code{shiny::validate}.
}
\examples{
library(scda)

ADSL <- synthetic_cdisc_data("latest")$adsl
ADRS <- synthetic_cdisc_data("latest")$adrs

ui <- fluidPage(
  selectInput(
    "rsp",
    "Select response parameter",
    choices = c("BESRSPI", "INVET", "CBRSPI"),
    selected = "BESRSPI",
    multiple = FALSE
  ),
  verbatimTextOutput("rsp_summary")
)

server <- function(input, output) {
  output$rsp_summary <- renderPrint({
    validate_in(input$rsp, ADRS$PARAMCD, "Parameter does not exist.")
    nrow(ADRS[ADRS$PARAMCD == input$rsp, ])
  })
}
if (interactive()) {
  shinyApp(ui, server)
}

}
