% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_inputs.R
\name{validate_input}
\alias{validate_input}
\title{Validate input}
\usage{
validate_input(
  inputId,
  condition = function(x) TRUE,
  message = "",
  session = shiny::getDefaultReactiveDomain()
)
}
\arguments{
\item{inputId}{(\code{character}) Character of input ID(s) to validate}

\item{condition}{(\code{logical(1)}, \verb{function(x)}) Logical value or function returning logical value.
Condition should determine expected state, \code{FALSE} throws.}

\item{message}{(\code{character(1)}) Character string of validation message to display}

\item{session}{Shiny session object}
}
\value{
\code{NULL} or \code{shiny.silent.error} when condition is not met
}
\description{
Validate input
}
\keyword{internal}
