% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validations.R
\name{validate_no_intersection}
\alias{validate_no_intersection}
\title{Validates no intersection between two vectors}
\usage{
validate_no_intersection(x, y, msg)
}
\arguments{
\item{x}{vector}

\item{y}{vector}

\item{msg}{message to display if \code{x} and \code{y} intersect}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
}
\details{
This function is a wrapper for \code{shiny::validate}.
}
\examples{
library(scda)
ADSL <- synthetic_cdisc_data("latest")$adsl

ui <- fluidPage(
  selectInput("ref_arm", "Select reference treatment",
    choices = c("ARM A", "ARM B", "ARM C"),
    selected = "ARM A"
  ),
  selectInput("comp_arm", "Select comparison treatment",
    choices = c("ARM A", "ARM B", "ARM C"),
    selected = "ARM C"
  ),
  verbatimTextOutput("arm_summary")
)

server <- function(input, output) {
  output$arm_summary <- renderText({
    ref_arm <- ADSL$ARMCD[input$ref_arm == ADSL$ARMCD]
    comp_arm <- ADSL$ARMCD[input$comp_arm == ADSL$ARMCD]

    validate_no_intersection(
      comp_arm, ref_arm,
      "reference and comparison treatments cannot overlap"
    )
    paste0(
      "Number of patients in: reference treatment=", length(ref_arm),
      " comparions treatment=", length(comp_arm)
    )
  })
}
if (interactive()) {
  shinyApp(ui, server)
}

}
