% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validations.R
\name{validate_one_row_per_id}
\alias{validate_one_row_per_id}
\title{Validate that dataset has unique rows for key variables}
\usage{
validate_one_row_per_id(x, key = c("USUBJID", "STUDYID"))
}
\arguments{
\item{x}{a data.frame}

\item{key}{a vector of ID variables from \code{x} that identify unique records}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
}
\details{
This function is a wrapper for \code{shiny::validate}.
}
\examples{
library(scda)
ADSL <- synthetic_cdisc_data("latest")$adsl

ui <- fluidPage(
  sliderInput("obs", "Max Age",
    min = 0, max = 100, value = 50
  ),
  verbatimTextOutput("age_summary")
)

server <- function(input, output) {
  output$age_summary <- renderText({
    ADSL_f <- ADSL[ADSL$AGE <= input$obs, ]
    validate_one_row_per_id(ADSL_f, key = c("STUDYID"))

    paste0("Mean age :", mean(ADSL_f$AGE))
  })
}
\dontrun{
shinyApp(ui, server)
}

}
