% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_variable_browser.R
\name{variable_browser_item}
\alias{variable_browser_item}
\title{Deprecated: Variable Browser Teal Module}
\usage{
variable_browser_item(...)
}
\arguments{
\item{...}{arguments passed on to \code{\link{tm_variable_browser}}}
}
\description{
Please use the \code{\link{tm_variable_browser}} function instead.
}
