% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choices_labeled.R
\name{variable_choices}
\alias{variable_choices}
\alias{variable_choices.character}
\alias{variable_choices.data.frame}
\alias{variable_choices.TealDataset}
\alias{variable_choices.TealDatasetConnector}
\title{Wrapper on \link{choices_labeled} to label variables basing on existing labels in data}
\usage{
variable_choices(data, subset = NULL, fill = FALSE, key = NULL)

\method{variable_choices}{character}(data, subset = NULL, fill = FALSE, key = NULL)

\method{variable_choices}{data.frame}(data, subset = NULL, fill = FALSE, key = NULL)

\method{variable_choices}{TealDataset}(data, subset = NULL, fill = FALSE, key = get_keys(data))

\method{variable_choices}{TealDatasetConnector}(data, subset = NULL, fill = FALSE, key = get_keys(data))
}
\arguments{
\item{data}{(\code{data.frame}, \code{character}, \code{TealDataset}, \code{TealDatasetConnector})
If \code{data.frame}, then data to extract labels from
If \code{character}, then name of the dataset to extract data from once available
If \code{TealDataset} or \code{TealDatasetConnector}, then raw data to extract labels from.}

\item{subset}{(\code{character} or \code{function})
If \code{character}, then a vector of column names.
If \code{function}, then this function is used to determine the possible columns (e.g. all factor columns).
In this case, the function must take only single argument "data" and return a character vector.
See examples for more details.}

\item{fill}{boolean in case the \code{label} attribute does not exist if
\code{TRUE} the variable names is returned, otherwise \code{NA}}

\item{key}{(\code{character}) vector with names of the variables, which are part of the primary key
of the \code{data} argument. This is an optional argument, which allows to identify variables
associated with the primary key and display the appropriate icon for them in the
\link{optionalSelectInput} widget.}
}
\value{
named character vector with additional attributes or \code{delayed_data} object
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#maturing}{\figure{lifecycle-maturing.svg}{options: alt='[Maturing]'}}}{\strong{[Maturing]}}
}
\examples{
library(scda)
ADRS <- synthetic_cdisc_data("latest")$adrs

variable_choices(ADRS)
variable_choices(ADRS, subset = c("PARAM", "PARAMCD"))
variable_choices(ADRS, subset = c("", "PARAM", "PARAMCD"))
variable_choices(ADRS, subset = c("", "PARAM", "PARAMCD"), key = get_cdisc_keys("ADRS"))

# delayed version
variable_choices("ADRS", subset = c("USUBJID", "STUDYID"))

# also works with TealDataset and TealDatasetConnector
ADRS_dataset <- dataset("ADRS", ADRS, key = get_cdisc_keys("ADRS"))
variable_choices(ADRS_dataset)

ADRS_conn <- dataset_connector(
  "ADRS",
  pull_callable = callable_code("radrs(cached = TRUE)"),
  key = get_cdisc_keys("ADRS"))
variable_choices(ADRS_conn)

# functional subset (with delayed data) - return only factor variables
variable_choices("ADRS", subset = function(data) {
  idx <- vapply(data, is.factor, logical(1))
  return(names(data)[idx])
})
}
