% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choices_labeled.R
\name{variable_choices}
\alias{variable_choices}
\title{Wrapper on \code{\link{choices_labeled}} to label variables basing on existing labels in data}
\usage{
variable_choices(data, subset = NULL, fill = FALSE)
}
\arguments{
\item{data}{(\code{data.frame}) or (\code{character})
If \code{data.frame}, then data to extract labels from.
If \code{character}, then name of the dataset to extract data from once available.}

\item{subset}{(\code{character} or \code{function})
If \code{character}, then a vector of column names.
If \code{function}, then this function is used to determine the possible columns (e.g. all factor columns).
In this case, the function must take only single argument "data" and return a character vector.
See examples for more details.}

\item{fill}{boolean in case the \code{label} attribute does not exist if
\code{TRUE} the variable names is returned, otherwise \code{NA}}
}
\value{
named character vector with additional attributes or \code{delayed_data} object
}
\description{
Wrapper on \code{\link{choices_labeled}} to label variables basing on existing labels in data
}
\examples{
library(random.cdisc.data)
ADRS <- radrs(cached = TRUE)

variable_choices(ADRS)
variable_choices(ADRS, subset = c("PARAM", "PARAMCD"))
variable_choices(ADRS, subset = c("", "PARAM", "PARAMCD"))

# delayed version
variable_choices("ADRS", subset = c("USUBJID", "STUDYID"))

# functional subset (with delayed data) - return only factor variables
variable_choices("ADRS", subset = function(data) {
  idx <- vapply(data, is.factor, logical(1))
  return(names(data)[idx])
})
}
