% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_label.R
\name{variable_labels}
\alias{variable_labels}
\title{Extracts column labels from a CDISC dataset.}
\usage{
variable_labels(data, columns = NULL, fill = TRUE)
}
\arguments{
\item{data}{(\code{data.frame}) any CDISC data set}

\item{columns}{optional, (\code{character}) column names to extract the labels from. If (\code{NULL}) then all
columns are being used.}

\item{fill}{(\code{logical(1)}) if \code{TRUE}, the function will return variable names for columns with non-existent labels;
otherwise will return \code{NA} for them}
}
\value{
labels of the columns
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
}
\examples{
library(scda)
ADSL <- synthetic_cdisc_data("latest")$adsl

variable_labels(ADSL)
variable_labels(ADSL, c("AGE", "RACE", "BMRKR1"))
variable_labels(ADSL, c("AGE", "RACE", "BMRKR1", "xyz"))

ADSL$NEW_COL <- 1
variable_labels(ADSL, c("AGE", "RACE", "BMRKR1", "NEW_COL"))
variable_labels(ADSL, c("AGE", "RACE", "BMRKR1", "NEW_COL"), fill = FALSE)

}
