% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optionalInput.R
\name{variable_type_icons}
\alias{variable_type_icons}
\title{Get icons to represent variable types in dataset}
\usage{
variable_type_icons(var_type)
}
\arguments{
\item{var_type}{(\code{character} vector) of R internal types (classes).}
}
\value{
(atomic vector of \code{character}) vector of HTML icons corresponding to
data type in each column.
}
\description{
Get icons to represent variable types in dataset
}
\examples{
teal:::variable_type_icons(c(
  "integer", "numeric", "logical", "Date", "POSIXct", "POSIXlt",
  "factor", "character", "unknown", ""
))
}
\keyword{internal}
