% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teal_data_module.R
\name{within.teal_data_module}
\alias{within.teal_data_module}
\title{Evaluate expression in \code{qenv} object.}
\usage{
\method{within}{teal_data_module}(data, expr, ...)
}
\arguments{
\item{data}{\code{qenv} object}

\item{expr}{\code{expression} to evaluate}

\item{...}{\code{name:value} pairs to inject values into \code{expr}}
}
\value{
Returns a \code{qenv} object with \code{expr} evaluated. If evaluation raises an error, a \code{qenv.error} is returned.
}
\description{
Convenience function for evaluating inline code inside the environment of a \code{qenv}.
}
\details{
This is a wrapper for \code{eval_code} that provides a simplified way of passing code for evaluation.
It accepts only inline expressions (both simple and compound) and allows for injecting values into \code{expr}
through the \code{...} argument: as \code{name:value} pairs are passed to \code{...},
\code{name} in \code{expr} will be replaced with \code{value}.
}
\section{Using language objects}{


Passing language objects to \code{expr} is generally not intended but can be achieved with \code{do.call}.
Only single \code{expression}s will work and substitution is not available. See examples.

}

\examples{
q <- new_qenv()

# execute code
q <- within(q, {
  i <- iris
})
q <- within(q, {
  m <- mtcars
  f <- faithful
})
q
get_code(q)

# inject values into code
q <- new_qenv()
q <- within(q, i <- iris)
within(q, print(dim(subset(i, Species == "virginica"))))
within(q, print(dim(subset(i, Species == species)))) # fails
within(q, print(dim(subset(i, Species == species))), species = "versicolor")
species_external <- "versicolor"
within(q, print(dim(subset(i, Species == species))), species = species_external)

# pass language objects
expr <- expression(i <- iris, m <- mtcars)
within(q, expr) # fails
do.call(within, list(q, expr))

exprlist <- list(expression(i <- iris), expression(m <- mtcars))
within(q, exprlist) # fails
do.call(within, list(q, do.call(c, exprlist)))
}
\seealso{
\code{\link[teal.code]{eval_code}}, \code{\link[base:with]{base::within}}
}
